configs: New config in drivers/bus for v5.4-rc1
Message-id: <20190918194030.6474-1-jcline@redhat.com>
Patchwork-id: 272480
O-Subject: [ARK INTERNAL PATCH] [redhat] configs: New config in drivers/bus for v5.4-rc1
Bugzilla:
RH-Acked-by: John Linville <linville@redhat.com>

Upstream Status: RHEL only.

Hi,

As part of the rebase effort to v5.4-rc1, New config option(s) were
exposed. Please review the options below.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

* drivers/bus/Kconfig (help): MOXTET (tristate)
    Say yes here to add support for the module configuration bus found
    on CZ.NIC's Turris Mox. This is needed for the ability to discover
    the order in which the modules are connected and to get/set some of
    their settings. For example the GPIOs on Mox SFP module are
    configured through this bus.

Cc: John Linville <linville@redhat.com>
Signed-off-by: Jeremy Cline <jcline@redhat.com>