configs: New config in drivers/iio for v5.4-rc1
Message-id: <20190918194220.7803-1-jcline@redhat.com>
Patchwork-id: 272488
O-Subject: [ARK INTERNAL PATCH] [redhat] configs: New config in drivers/iio for v5.4-rc1
Bugzilla:
RH-Acked-by: Laura Abbott <labbott@redhat.com>

Upstream Status: RHEL only.

Hi,

As part of the rebase effort to v5.4-rc1, New config option(s) were
exposed. Please review the options below.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

* drivers/iio/imu/Kconfig (help): ADIS16460 (tristate)
    Say yes here to build support for Analog Devices ADIS16460 inertial
    sensor.

    To compile this driver as a module, choose M here: the module will be
    called adis16460.
* drivers/iio/potentiometer/Kconfig (help): MAX5432 (tristate)
    Say yes here to build support for the Maxim
    MAX5432, MAX5433, MAX5434 and MAX5435 digital
    potentiometer chips.

    To compile this driver as a module, choose M here: the
    module will be called max5432.
* drivers/iio/light/Kconfig (help): NOA1305 (tristate)
    Say Y here if you want to build support for the ON Semiconductor
    NOA1305 ambient light sensor.

    To compile this driver as a module, choose M here:
    The module will be called noa1305.

Signed-off-by: Jeremy Cline <jcline@redhat.com>