configs: New config in drivers/usb for v5.4-rc1
Message-id: <20190918194258.8325-1-jcline@redhat.com>
Patchwork-id: 272492
O-Subject: [ARK INTERNAL PATCH] [redhat] configs: New config in drivers/usb for v5.4-rc1
Bugzilla:
RH-Acked-by: John Linville <linville@redhat.com>

Upstream Status: RHEL only.

Hi,

As part of the rebase effort to v5.4-rc1, New config option(s) were
exposed. Please review the options below.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

* drivers/usb/cdns3/Kconfig (help): USB_CDNS3 (tristate)
    Say Y here if your system has a Cadence USB3 dual-role controller.
    It supports: dual-role switch, Host-only, and Peripheral-only.

    If you choose to build this driver is a dynamically linked
    as module, the module will be called cdns3.ko.
* drivers/usb/common/Kconfig (help): USB_CONN_GPIO (tristate)
    The driver supports USB role switch between host and device via GPIO
    based USB cable detection, used typically if an input GPIO is used
    to detect USB ID pin, and another input GPIO may be also used to detect
    Vbus pin at the same time, it also can be used to enable/disable
    device if an input GPIO is only used to detect Vbus pin.

    To compile the driver as a module, choose M here: the module will
    be called usb-conn-gpio.ko

Cc: Torez Smith <torez@redhat.com>
Cc: Don Zickus <dzickus@redhat.com>
Signed-off-by: Jeremy Cline <jcline@redhat.com>