configs: New config in drivers/pci for v5.4-rc1
Message-id: <20191014162300.16308-14-jcline@redhat.com>
Patchwork-id: 278875
O-Subject: [ARK INTERNAL PATCH] configs: New config in drivers/pci for v5.4-rc1
Bugzilla:
RH-Acked-by: Myron Stowe <mstowe@redhat.com>

Upstream Status: RHEL only.

Hi,

As part of the rebase effort to v5.4-rc1, New config option(s) were
exposed. Please review the options below.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

* drivers/pci/controller/dwc/Kconfig (help): PCIE_AL (bool)
    Say Y here to enable support of the Amazon's Annapurna Labs PCIe
    controller IP on Amazon SoCs. The PCIe controller uses the DesignWare
    core plus Annapurna Labs proprietary hardware wrappers. This is
    required only for DT-based platforms. ACPI platforms with the
    Annapurna Labs PCIe controller don't need to enable this.

Cc: Myron Stowe <myron.stowe@redhat.com>
Cc: Prarit Bhargava <prarit@redhat.com>
Signed-off-by: Jeremy Cline <jcline@redhat.com>