configs: New config in drivers/platform for v5.4-rc1
Message-id: <20191014162300.16308-16-jcline@redhat.com>
Patchwork-id: 278883
O-Subject: [ARK INTERNAL PATCH] configs: New config in drivers/platform for v5.4-rc1
Bugzilla:
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: Laura Abbott <labbott@redhat.com>

Upstream Status: RHEL only.

Hi,

As part of the rebase effort to v5.4-rc1, New config option(s) were
exposed. Please review the options below.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

* drivers/platform/chrome/Kconfig (help): CROS_EC (tristate)
    If you say Y here you get support for the ChromeOS Embedded
    Controller (EC) providing keyboard, battery and power services.
    You also need to enable the driver for the bus you are using. The
    protocol for talking to the EC is defined by the bus driver.

    To compile this driver as a module, choose M here: the
    module will be called cros_ec.

Signed-off-by: Jeremy Cline <jcline@redhat.com>