[redhat] Align some configs for Fedora
Because of the nature of keeping two sets of configs, the
Fedora configs in this tree are out of sync with work
that's happening in the official Fedora dist-git repo.
Bring this back in sync.

As part of this work, prune a bunch of dead configuration
options. None of the changes meaningfully affect the generated
config output for ARK as compared against the full configs in the
srpms.

Signed-off-by: Laura Abbott <labbott@redhat.com>