configs: Turn off ISDN
Message-id: <20191121191829.30226-1-labbott@redhat.com>
Patchwork-id: 287874
O-Subject: [ARK INTERNAL PATCH] [redhat] configs: Turn off ISDN
Bugzilla:
RH-Acked-by: Jiri Benc <jbenc@redhat.com>
RH-Acked-by: John Linville <linville@redhat.com>

This is slowly going away upstream. Nobody seems to want it so just
turn it off.

Cc: Jiri Benc <jbenc@redhat.com>
Cc: Neil Horman <nhorman@redhat.com>
Cc: John W. Linville <linville@redhat.com>
Signed-off-by: Laura Abbott <labbott@redhat.com>
---
I know we're still waiting to hear back from the BU but if there's
general agreement this is what we want let's just turn it off. It's
really cheap to turn it back on later.

Signed-off-by: Jeremy Cline <jcline@redhat.com>