[redhat] New configs in arch/powerpc
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 CONFIG_PPC_UV:

 This option paravirtualizes the kernel to run in POWER platforms that
 supports the Protected Execution Facility (PEF). On such platforms,
 the ultravisor firmware runs at a privilege level above the
 hypervisor.

 If unsure, say "N".

 Symbol: PPC_UV [=n]
 Type  : bool
 Prompt: Ultravisor support
   Location:
     -> Kernel options
   Defined at arch/powerpc/Kconfig:455
   Depends on: KVM_BOOK3S_HV_POSSIBLE [=y]
   Selects: ZONE_DEVICE [=y] && DEV_PAGEMAP_OPS [=y] && DEVICE_PRIVATE [=y] && MEMORY_HOTPLUG [=y] && MEMORY_HOTREMOVE [=y]

---

Cc: Thomas Huth <thuth@redhat.com>
Cc: David Gibson <dgibson@redhat.com>
Cc: Laurent Vivier <lvivier@redhat.com>