[redhat] New configs in drivers/base
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 CONFIG_FW_CACHE:

 Because firmware caching generates uevent messages that are sent
 over a netlink socket, it can prevent suspend on many platforms.
 It is also not always useful, so on such platforms we have the
 option.

 If unsure, say Y.

 Symbol: FW_CACHE [=y]
 Type  : bool
 Prompt: Enable firmware caching during suspend
   Location:
     -> Device Drivers
       -> Generic Driver Options
         -> Firmware loader
           -> Firmware loading facility (FW_LOADER [=y])
   Defined at drivers/base/firmware_loader/Kconfig:172
   Depends on: FW_LOADER [=y] && PM_SLEEP [=y]

---

Cc: Prarit Bhargava <prarit@redhat.com>
Cc: Neil Horman <nhorman@redhat.com>