[redhat] New configs in drivers/gpu
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 CONFIG_DRM_AMD_DC_HDCP:

 Choose this option
 if you want to support
 HDCP authentication

 Symbol: DRM_AMD_DC_HDCP [=n]
 Type  : bool
 Prompt: Enable HDCP support in DC
   Location:
     -> Device Drivers
       -> Graphics support
         -> Display Engine Configuration
   Defined at drivers/gpu/drm/amd/display/Kconfig:47
   Depends on: HAS_IOMEM [=y] && DRM [=m] && DRM_AMDGPU [=m] && DRM_AMD_DC [=y]

---

Cc: David Airlie <airlied@redhat.com>
Cc: Adam Jackson <ajax@redhat.com>
Cc: Rob Clark <rclark@redhat.com>