[redhat] New configs in drivers/soc
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 CONFIG_FSL_RCPM:

 The NXP QorIQ Processors based on ARM Core have RCPM module
 (Run Control and Power Management), which performs all device-level
 tasks associated with power management, such as wakeup source control.
 Note that currently this driver will not support PowerPC based
 QorIQ processor.
 Symbol: FSL_RCPM [=n]
 Type  : bool
 Prompt: Freescale RCPM support
   Location:
     -> Device Drivers
       -> SOC (System On Chip) specific Drivers
         -> NXP/Freescale QorIQ SoC drivers
   Defined at drivers/soc/fsl/Kconfig:44
   Depends on: PM_SLEEP [=y] && (ARM || ARM64 [=y])

---