[redhat] Explicitly use python3 for building the htmldocs
In the never-ending slow-motion train wreck that is the Python
ecosystem, we need to explicitly specify Python 3 because the build root
does not have /usr/bin/python which causes the documentation build to
fail. Even if it did, the Python 3 version of Sphinx is being used to
build the documentation so unless /usr/bin/python was Python 3 it'd
still break.

Cc: Prarit Bhargava <prarit@redhat.com>
Cc: Herton R. Krzesinski <herton@redhat.com>
Cc: Bruno Meneguele <bmeneg@redhat.com>
Cc: Juri Lelli <jlelli@redhat.com>
Cc: Don Zickus <dzickus@redhat.com>
Signed-off-by: Jeremy Cline <jcline@redhat.com>