[redhat] Explicitly disable GGC_PLUGINS for ARK
It was reported that ARK started to fail to build in some build
environments starting with v5.5-rc5 because CONFIG_GCC_PLUGINS was being
changed from unset to y. This is likely due to upstream commit
a5b0dc5a46c2 ("gcc-plugins: make it possible to disable
CONFIG_GCC_PLUGINS again"), but regardless we want to explicitly set
GGC_PLUGINS one way or the other because it's one of those annoying
Kconfig symbols that depends on the environment you prepare the
configuration in.

Cc: Jerry Snitselaar <jsnitsel@redhat.com>
Signed-off-by: Jeremy Cline <jcline@redhat.com>