[redhat] New configs in drivers/gpio
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 CONFIG_GPIO_SIFIVE:

 Say yes here to support the GPIO device on SiFive SoCs.

 Symbol: GPIO_SIFIVE [=n]
 Type  : bool
 Prompt: SiFive GPIO support
   Location:
     -> Device Drivers
       -> GPIO Support (GPIOLIB [=y])
         -> Memory mapped GPIO drivers
   Defined at drivers/gpio/Kconfig:482
   Depends on: GPIOLIB [=y] && HAS_IOMEM [=y] && OF_GPIO [=y] && IRQ_DOMAIN_HIERARCHY [=y]
   Selects: GPIO_GENERIC [=n] && GPIOLIB_IRQCHIP [=y] && REGMAP_MMIO [=n]

---