[redhat] New configs in drivers/platform
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 CONFIG_INTEL_UNCORE_FREQ_CONTROL:

 This driver allows control of uncore frequency limits on
 supported server platforms.
 Uncore frequency controls RING/LLC (last-level cache) clocks.

 To compile this driver as a module, choose M here: the module
 will be called intel-uncore-frequency.

 Symbol: INTEL_UNCORE_FREQ_CONTROL [=n]
 Type  : tristate
 Prompt: Intel Uncore frequency control driver
   Location:
     -> Device Drivers
       -> X86 Platform Specific Device Drivers (X86_PLATFORM_DEVICES [=y])
   Defined at drivers/platform/x86/Kconfig:1339
   Depends on: X86 [=y] && X86_PLATFORM_DEVICES [=y] && X86_64 [=y]

Set to m at the request of Prarit.
---

Cc: David Arcari <darcari@redhat.com>
Cc: Prarit Bhargava <prarit@redhat.com>