[redhat] New configs in drivers/spi
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 CONFIG_SPI_HISI_SFC_V3XX:

 This enables support for HiSilicon v3xx SPI-NOR flash controller
 found in hi16xx chipsets.

 Symbol: SPI_HISI_SFC_V3XX [=n]
 Type  : tristate
 Prompt: HiSilicon SPI-NOR Flash Controller for Hi16XX chipsets
   Location:
     -> Device Drivers
       -> SPI support (SPI [=y])
   Defined at drivers/spi/Kconfig:284
   Depends on: SPI [=y] && SPI_MASTER [=y] && (ARM64 [=y] && ACPI [=y] || COMPILE_TEST [=n]) && HAS_IOMEM [=y]
   Selects: CONFIG_MTD_SPI_NOR

---