[redhat] New configs in sound/core
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 CONFIG_SND_CTL_VALIDATION:

 Say Y to enable the additional validation of each control element
 access, including sanity-checks like whether the values returned
 from the driver are in the proper ranges or the check of the invalid
 access at out-of-array areas.

 Symbol: SND_CTL_VALIDATION [=n]
 Type  : bool
 Prompt: Perform sanity-checks for each control element access
   Location:
     -> Device Drivers
       -> Sound card support (SOUND [=m])
         -> Advanced Linux Sound Architecture (SND [=m])
           -> Debug (SND_DEBUG [=y])
   Defined at sound/core/Kconfig:181
   Depends on: SOUND [=m] && !UML && SND [=m] && SND_DEBUG [=y]

---

This setting has been enabled on debug kernels at the request of
Jaroslav Kysela.

Cc: Jaroslav Kysela <jkysela@redhat.com>