[redhat] New configs in drivers/gpu
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 CONFIG_DRM_ANALOGIX_ANX6345:

 ANX6345 is an ultra-low Full-HD DisplayPort/eDP
 transmitter designed for portable devices. The
 ANX6345 transforms the LVTTL RGB output of an
 application processor to eDP or DisplayPort.

 Symbol: DRM_ANALOGIX_ANX6345 [=n]
 Type  : tristate
 Defined at drivers/gpu/drm/bridge/analogix/Kconfig:2
   Prompt: Analogix ANX6345 bridge
   Depends on: HAS_IOMEM [=y] && DRM [=m] && DRM_BRIDGE [=y] && OF [=y]
   Location:
     -> Device Drivers
       -> Graphics support
         -> Display Interface Bridges
 Selects: DRM_ANALOGIX_DP [=n] && DRM_KMS_HELPER [=m] && REGMAP_I2C [=m]

---

 CONFIG_DRM_LVDS_CODEC:

 Support for transparent LVDS encoders and decoders that don't
 require any configuration.

 Symbol: DRM_LVDS_CODEC [=n]
 Type  : tristate
 Defined at drivers/gpu/drm/bridge/Kconfig:38
   Prompt: Transparent LVDS encoders and decoders support
   Depends on: HAS_IOMEM [=y] && DRM [=m] && DRM_BRIDGE [=y] && OF [=y]
   Location:
     -> Device Drivers
       -> Graphics support
         -> Display Interface Bridges
 Selects: DRM_KMS_HELPER [=m] && DRM_PANEL_BRIDGE [=y]

---

Cc: David Airlie <airlied@redhat.com>
Cc: Adam Jackson <ajax@redhat.com>
Cc: Rob Clark <rclark@redhat.com>