[redhat] New configs in drivers/reset
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 CONFIG_RESET_BRCMSTB_RESCAL:

 This enables the RESCAL reset controller for SATA, PCIe0, or PCIe1 on
 BCM7216.

 Symbol: RESET_BRCMSTB_RESCAL [=n]
 Type  : bool
 Defined at drivers/reset/Kconfig:52
   Prompt: Broadcom STB RESCAL reset controller
   Depends on: RESET_CONTROLLER [=y]
   Location:
     -> Device Drivers
       -> Reset Controller Support (RESET_CONTROLLER [=y])

---

 CONFIG_RESET_INTEL_GW:

 This enables the reset controller driver for Intel Gateway SoCs.
 Say Y to control the reset signals provided by reset controller.
 Otherwise, say N.

 Symbol: RESET_INTEL_GW [=n]
 Type  : bool
 Defined at drivers/reset/Kconfig:74
   Prompt: Intel Reset Controller Driver
   Depends on: RESET_CONTROLLER [=y] && OF [=y]
   Location:
     -> Device Drivers
       -> Reset Controller Support (RESET_CONTROLLER [=y])
 Selects: REGMAP_MMIO [=y]

---