[redhat] New configs in init/Kconfig
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 CONFIG_BOOT_CONFIG:

 Extra boot config allows system admin to pass a config file as
 complemental extension of kernel cmdline when booting.
 The boot config file must be attached at the end of initramfs
 with checksum and size.
 See <file:Documentation/admin-guide/bootconfig.rst> for details.

 If unsure, say Y.

 Symbol: BOOT_CONFIG [=y]
 Type  : bool
 Defined at init/Kconfig:1227
   Prompt: Boot config support
   Depends on: BLK_DEV_INITRD [=y]
   Location:
     -> General setup
       -> Initial RAM filesystem and RAM disk (initramfs/initrd) support (BLK_DEV_INITRD [=y])
 Selects: LIBXBC [=y]

---