[redhat] New configs in mm/Kconfig.debug
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 CONFIG_PTDUMP_DEBUGFS:

 Say Y here if you want to show the kernel pagetable layout in a
 debugfs file. This information is only useful for kernel developers
 who are working in architecture specific areas of the kernel.
 It is probably not a good idea to enable this feature in a production
 kernel.

 If in doubt, say N.

 Symbol: PTDUMP_DEBUGFS [=n]
 Type  : bool
 Defined at mm/Kconfig.debug:127
   Prompt: Export kernel pagetable layout to userspace via debugfs
   Depends on: DEBUG_KERNEL [=y] && DEBUG_FS [=y] && GENERIC_PTDUMP [=y]
   Location:
     -> Kernel hacking
       -> Memory Debugging
 Selects: PTDUMP_CORE [=y]

---

At the request of Waiman Long, this has been enabled for debug kernels.

Cc: Waiman Long <longman@redhat.com>
Cc: Larry Woodman <lwoodman@redhat.com>