[redhat] Include Fedora's merge.pl script
Fedora includes a script in its dist-git that allows users to drop
config settings into the "kernel-local" file which is merged into the
final configuration. This is purely for users and is empty in the
dist-git tree so it has no effect.

A number of folks use this script, so in order to minimize the
disruption of ARK and Fedora this includes it and the associated
specfile changes.

Cc: Don Zickus <dzickus@redhat.com>
Cc: Clark Williams <williams@redhat.com>
Cc: Jan Stancek <jstancek@redhat.com>
Cc: Augusto Caringi <acaringi@redhat.com>
Cc: Frantisek Hrbata <fhrbata@redhat.com>
Cc: Luis Claudio Goncalves <lgoncalv@redhat.com>
Cc: Juri Lelli <jlelli@redhat.com>
Cc: Rado Vrbovsky <rvrbovsk@redhat.com>
Cc: Herton R. Krzesinski <herton@redhat.com>
Cc: Denys Vlasenko <dvlasenk@redhat.com>
Cc: Bruno Meneguele <bmeneg@redhat.com>
Signed-off-by: Jeremy Cline <jcline@redhat.com>