Build the srpm for internal branch CI using the vanilla tree
We've got one oddball setting which we adjust the Kconfig default in a
patch in ark-patches. The setting is, therefore, invalid on vanilla
trees, which led to jumping through hoops to get the downstream patches
applied to the internal branch before testing.

This was painful because we needed to fetch the ark-patches branch to
some depth along with the master branch, figure out the merge base, and
yank all the patches out for application. It's more reasonable to just
set the one setting back to its upstream default and be careful about
further tweaks to Kconfigs in the downstream patch set.

Signed-off-by: Jeremy Cline <jcline@redhat.com>