[redhat] New configs in drivers/platform
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 CONFIG_SURFACE_3_POWER_OPREGION:

 This driver provides support for ACPI operation
 region of the Surface 3 battery platform driver.

 Symbol: SURFACE_3_POWER_OPREGION [=n]
 Type  : tristate
 Defined at drivers/platform/x86/Kconfig:858
   Prompt: Surface 3 battery platform operation region support
   Depends on: X86 [=y] && X86_PLATFORM_DEVICES [=y] && ACPI [=y] && I2C [=y]
   Location:
     -> Device Drivers
       -> X86 Platform Specific Device Drivers (X86_PLATFORM_DEVICES [=y])

---