[redhat] New configs in drivers/spi
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 CONFIG_SPI_MUX:

 This adds support for SPI multiplexers. Each SPI mux will be
 accessible as a SPI controller, the devices behind the mux will appear
 to be chip selects on this controller. It is still necessary to
 select one or more specific mux-controller drivers.

 Symbol: SPI_MUX [=n]
 Type  : tristate
 Defined at drivers/spi/Kconfig:919
   Prompt: SPI multiplexer support
   Depends on: SPI [=y] && SPI_MASTER [=y]
   Location:
     -> Device Drivers
       -> SPI support (SPI [=y])
 Selects: MULTIPLEXER [=n]

---