[redhat] New configs in fs/exfat
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 CONFIG_EXFAT_FS:

 This allows you to mount devices formatted with the exFAT file system.
 exFAT is typically used on SD-Cards or USB sticks.

 To compile this as a module, choose M here: the module will be called
 exfat.

 Symbol: EXFAT_FS [=n]
 Type  : tristate
 Defined at fs/exfat/Kconfig:3
   Prompt: exFAT filesystem support
   Depends on: BLOCK [=y]
   Location:
     -> File systems
       -> DOS/FAT/EXFAT/NT Filesystems
 Selects: NLS [=y]

---