[redhat] New configs in lib/Kconfig.debug
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 CONFIG_TEST_LOCKUP:

 This builds the "test_lockup" module that helps to make sure
 that watchdogs and lockup detectors are working properly.

 Depending on module parameters it could emulate soft or hard
 lockup, "hung task", or locking arbitrary lock for a long time.
 Also it could generate series of lockups with cooling-down periods.

 If unsure, say N.

 Symbol: TEST_LOCKUP [=n]
 Type  : tristate
 Defined at lib/Kconfig.debug:979
   Prompt: Test module to generate lockups
   Location:
     -> Kernel hacking
       -> Debug Oops, Lockups and Hangs

---