[redhat] New configs in drivers/virtio
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 CONFIG_VIRTIO_VDPA:

 This driver provides support for virtio based paravirtual
 device driver over vDPA bus. For this to be useful, you need
 an appropriate vDPA device implementation that operates on a
 physical device to allow the datapath of virtio to be
 offloaded to hardware.

 If unsure, say M.

 Symbol: VIRTIO_VDPA [=n]
 Type  : tristate
 Defined at drivers/virtio/Kconfig:46
   Prompt: vDPA driver for virtio devices
   Depends on: VIRTIO_MENU [=y]
   Location:
     -> Device Drivers
       -> Virtio drivers (VIRTIO_MENU [=y])
 Selects: VDPA [=n] && VIRTIO [=y]

Enabled at the request of Michael S. Tsirkin. Disabled on zfcpdump as it
seems unnecessary and introduces a number of dependencies.

---

Cc: "Michael S. Tsirkin" <mst@redhat.com>
Cc: Jason Wang <jasowang@redhat.com>
Cc: rhvirt-patches@redhat.com