Sign off generated configuration patches
Have the script sign off the configuration patches it generates.
Although it seems somewhat meaningless to have a script agree to the
terms of the DCO, the authors of the script authoring the patches have
and this makes running a CI job to check all patches for Signed-off-by
tags simpler.

Signed-off-by: Jeremy Cline <jcline@redhat.com>