Introduce a Sphinx documentation project
Although the GitLab wiki is pretty nice for getting started, it doesn't
offer a great way for folks to contribute or provide reviews for
changes. Convert the current Wiki to RST and build it with Sphinx. This
should be nearly identical to the content of the wiki with the exception
of a new section on contributing to the documentation.

When this is accepted, a new CI job can be added to automatically lint,
build, and host the Sphinx documentation on GitLab pages[0].

[0] https://docs.gitlab.com/ee/user/project/pages/

Signed-off-by: Jeremy Cline <jcline@redhat.com>