New configs in drivers/firmware
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 CONFIG_EFI_GENERIC_STUB_INITRD_CMDLINE_LOADER:

 Select this config option to add support for the initrd= command
 line parameter, allowing an initrd that resides on the same volume
 as the kernel image to be loaded into memory.

 This method is deprecated.

 Symbol: EFI_GENERIC_STUB_INITRD_CMDLINE_LOADER [=y]
 Type  : bool
 Defined at drivers/firmware/efi/Kconfig:127
   Prompt: Enable the command line initrd loader
   Depends on: EFI [=y] && EFI_STUB [=y] && (EFI_GENERIC_STUB [=y] || X86)
   Visible if: EFI [=y] && EFI_STUB [=y] && (EFI_GENERIC_STUB [=y] || X86) && !X86
   Location:
     -> Firmware Drivers
       -> EFI (Extensible Firmware Interface) Support

Cc: Lenny Szubowicz <lszubowi@redhat.com>
Cc: David Arcari <darcari@redhat.com>
Cc: Mark Langsdorf <mlangsdo@redhat.com>