[redhat] New configs in drivers/mfd
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 CONFIG_MFD_GATEWORKS_GSC:

 Enable support for the Gateworks System Controller (GSC) found
 on Gateworks Single Board Computers supporting system functions
 such as push-button monitor, multiple ADC's for voltage and
 temperature monitoring, fan controller and watchdog monitor.
 This driver provides common support for accessing the device.
 Additional drivers must be enabled in order to use the
 functionality of the device.

 Symbol: MFD_GATEWORKS_GSC [=n]
 Type  : tristate
 Defined at drivers/mfd/Kconfig:410
   Prompt: Gateworks System Controller
   Depends on: HAS_IOMEM [=y] && I2C [=y] && OF [=y]
   Location:
     -> Device Drivers
       -> Multifunction device drivers
 Selects: MFD_CORE [=m] && REGMAP_I2C [=m] && REGMAP_IRQ [=n]

---

Signed-off-by: CKI@GitLab <cki-project@redhat.com>