[redhat] New configs in block/Kconfig
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 CONFIG_BLK_INLINE_ENCRYPTION:

 Build the blk-crypto subsystem. Enabling this lets the
 block layer handle encryption, so users can take
 advantage of inline encryption hardware if present.

 Symbol: BLK_INLINE_ENCRYPTION [=n]
 Type  : bool
 Defined at block/Kconfig:189
   Prompt: Enable inline encryption support in block layer
   Depends on: BLOCK [=y]
   Location:
     -> Enable the block layer (BLOCK [=y])

---

Cc: Ming Lei <minlei@redhat.com>
Cc: Jeff Moyer <jmoyer@redhat.com>
Cc: Vivek Goyal <vgoyal@redhat.com>
Signed-off-by: CKI@GitLab <cki-project@redhat.com>