[redhat] New configs in net/bridge
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 CONFIG_BRIDGE_MRP:

 If you say Y here, then the Ethernet bridge will be able to run MRP
 protocol to detect loops

 Say N to exclude this support and reduce the binary size.

 If unsure, say N.

 Symbol: BRIDGE_MRP [=n]
 Type  : bool
 Defined at net/bridge/Kconfig:65
   Prompt: MRP protocol
   Depends on: NET [=y] && BRIDGE [=m]
   Location:
     -> Networking support (NET [=y])
       -> Networking options
         -> 802.1d Ethernet Bridging (BRIDGE [=m])

---

Signed-off-by: Justin M. Forbes <jforbes@fedoraproject.org>