New configs in drivers/power
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 CONFIG_POWER_RESET_VEXPRESS:

 Power off and reset support for the ARM Ltd. Versatile
 Express boards.

 Symbol: POWER_RESET_VEXPRESS [=n]
 Type  : bool
 Defined at drivers/power/reset/Kconfig:184
   Prompt: ARM Versatile Express power-off and reset driver
   Depends on: POWER_RESET [=y] && (ARM || ARM64 [=y]) && VEXPRESS_CONFIG [=y]=y
   Location:
     -> Device Drivers
       -> Board level reset or power off (POWER_RESET [=y])

Cc: Lenny Szubowicz <lszubowi@redhat.com>