[redhat] New configs in lib/Kconfig.debug
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 CONFIG_DEBUG_INFO_COMPRESSED:

 Compress the debug information using zlib.  Requires GCC 5.0+ or Clang
 5.0+, binutils 2.26+, and zlib.

 Users of dpkg-deb via scripts/package/builddeb may find an increase in
 size of their debug .deb packages with this config set, due to the
 debug info being compressed with zlib, then the object files being
 recompressed with a different compression scheme. But this is still
 preferable to setting $KDEB_COMPRESS to "none" which would be even
 larger.

 Symbol: DEBUG_INFO_COMPRESSED [=n]
 Type  : bool
 Defined at lib/Kconfig.debug:216
   Prompt: Compressed debugging information
   Depends on: DEBUG_INFO [=y]
   Location:
     -> Kernel hacking
       -> Compile-time checks and compiler options
         -> Compile the kernel with debug info (DEBUG_INFO [=y])

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>