[redhat] New configs in drivers/usb
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 CONFIG_USB_XHCI_PCI_RENESAS:

 Say 'Y' to enable the support for the Renesas xHCI controller with
 firwmare. Make sure you have the firwmare for the device and
 installed on your system for this device to work.
 If unsure, say 'N'.

 Symbol: USB_XHCI_PCI_RENESAS [=n]
 Type  : tristate
 Defined at drivers/usb/host/Kconfig:46
   Prompt: Support for additional Renesas xHCI controller with firwmare
   Depends on: USB_SUPPORT [=y] && USB [=y] && USB_XHCI_HCD [=y]
   Location:
     -> Device Drivers
       -> USB support (USB_SUPPORT [=y])
         -> xHCI HCD (USB 3.0) support (USB_XHCI_HCD [=y])

---

Cc: Torez Smith <torez@redhat.com>
Cc: Don Zickus <dzickus@redhat.com>
Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>