Fix PATCHLEVEL for merge window
When building during the merge window, we need to bump PATCHLEVEL in the
Makefile because Linus doesn't bump it until rc1. Something similar has
been in the Fedora spec forever.  Using 5.8 as an example, without this
change we build a kernel rpm version
kernel-5.8.0-0.rc0.20200608gitaf7b4801030c.1.fc33 but with a uname of
kernel-5.7.0-0.rc0.20200608gitaf7b4801030c.1.fc33 which places the
modules in the rpm versioned directory where the kernel does not know to
look for them.  With this change, both the rpm and the uname would be
the kernel-5.8 version.

While it might be cleaner to only run this replacement if we are in the
merge window, this is much simpler than making changes across 2
additional Makefiles, and the end result should always be the same.

Signed-off-by: Justin M. Forbes <jforbes@fedoraproject.org>