[redhat] New configs in drivers/power
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 CONFIG_BATTERY_CW2015:

 Say Y here to enable support for the cellwise cw2015
 battery fuel gauge (used in the Pinebook Pro & others)

 This driver can also be built as a module. If so, the module will be
 called cw2015_battery.

 Symbol: BATTERY_CW2015 [=n]
 Type  : tristate
 Defined at drivers/power/supply/Kconfig:119
   Prompt: CW2015 Battery driver
   Depends on: POWER_SUPPLY [=y] && I2C [=y]
   Location:
     -> Device Drivers
       -> Power supply class support (POWER_SUPPLY [=y])
 Selects: REGMAP_I2C [=m]

---

 CONFIG_CHARGER_BD99954:

 Say Y here to enable support for getting battery and charger
 information and altering charger configurations from the ROHM
 BD99954 charger IC.

 Symbol: CHARGER_BD99954 [=n]
 Type  : tristate
 Defined at drivers/power/supply/Kconfig:724
   Prompt: ROHM bd99954 charger driver
   Depends on: POWER_SUPPLY [=y] && I2C [=y]
   Location:
     -> Device Drivers
       -> Power supply class support (POWER_SUPPLY [=y])
 Selects: LINEAR_RANGES [=n]

---

Signed-off-by: Justin M. Forbes <jforbes@fedoraproject.org>