New configs in drivers/mailbox
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 CONFIG_QCOM_IPCC:

 Qualcomm Technologies, Inc. Inter-Processor Communication Controller
 (IPCC) driver for MSM devices. The driver provides mailbox support for
 sending interrupts to the clients. On the other hand, the driver also
 acts as an interrupt controller for receiving interrupts from clients.
 Say Y here if you want to build this driver.

 Symbol: QCOM_IPCC [=n]
 Type  : bool
 Defined at drivers/mailbox/Kconfig:247
   Prompt: Qualcomm Technologies, Inc. IPCC driver
   Depends on: MAILBOX [=y] && (ARCH_QCOM [=y] || COMPILE_TEST [=n])
   Location:
     -> Device Drivers
       -> Mailbox Hardware Support (MAILBOX [=y])

Cc: Al Stone <ahs3@redhat.com>