[redhat] New configs in drivers/i2c
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 CONFIG_I2C_QCOM_CCI:

 If you say yes to this option, support will be included for the
 built-in camera control interface on the Qualcomm SoCs.

 This driver can also be built as a module.  If so, the module
 will be called i2c-qcom-cci.

 Symbol: I2C_QCOM_CCI [=n]
 Type  : tristate
 Defined at drivers/i2c/busses/Kconfig:900
   Prompt: Qualcomm Camera Control Interface
   Depends on: I2C [=y] && HAS_IOMEM [=y] && (ARCH_QCOM [=y] || COMPILE_TEST [=n])
   Location:
     -> Device Drivers
       -> I2C support
         -> I2C support (I2C [=y])
           -> I2C Hardware Bus support

---

Cc: Gopal Tiwari <gtiwari@redhat.com>
Cc: David Arcari <darcari@redhat.com>
Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>