genspec.sh: 'touch' empty Patchlist file for single tarball
I was playing around with SINGLE_TARBALL=1 to workaround an issue and
noticed the patch for breaking out individual patches did not account
for that scenario.

The srpm fails to generate due to the missing Patchlist file, which
is used to track the individual patch files.

Fix this by creating a empty file to use or 'touch $SOURCES/Patchlist'.

This allows the srpm to package all the necessary files.  I tested the
'prep' stage of rpmbuild and this empty file has no side effects.

Signed-off-by: Don Zickus <dzickus@redhat.com>