Add ability to sync upstream through Makefile
Currently all the magic to sync upstream with os-build and ark-patches
is done through scripts in redhat/scripts/ci and .gitlab-ci.yml.

Make this easier by enabling this routine through a redhat/Makefile.

This allows:
* gitlab-ci.yml and maintainer to use same script
* allows maintainer to run local test runs before committing
* developers to hack on things a little more easily

This change also adds the ability to 'git push' the changes.  This is
safe because os-build and ark-patches are considered 'protected'
branches meaning only the developers in the 'Maintainers' group can
push to those branches (which is a tiny group right now).

Because I can't atomically commit this change and the change to
kernel-ark-ci's gitlab-ci.yml file, I need this change to go in first.

Then I can modify the gitlab-ci.yml file.

Functionally, this change should be no different than what currently
runs today.