Update secureboot signing for dual keys
As part of the transition for the boothole vulnerability, we are signing
the kernel with both new keys and the old keys.  These are the spec
changes to make that happen.  We have actually been building with this
for some time, but didn't want to push the changes until it was public.

Signed-off-by: Justin M. Forbes <jforbes@fedoraproject.org>