[redhat] New configs in drivers/char
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 CONFIG_HW_RANDOM_BA431:

 This driver provides kernel-side support for the Random Number
 Generator hardware based on Silex Insight BA431 IP.

 To compile this driver as a module, choose M here: the
 module will be called ba431-rng.

 Symbol: HW_RANDOM_BA431 [=n]
 Type  : tristate
 Defined at drivers/char/hw_random/Kconfig:77
   Prompt: Silex Insight BA431 Random Number Generator support
   Depends on: HW_RANDOM [=y] && HAS_IOMEM [=y]
   Location:
     -> Device Drivers
       -> Character devices
         -> Hardware Random Number Generator Core support (HW_RANDOM [=y])

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>