Makefile.common rename TAG to _TAG
I would like to use the TAG variable as input to the make merge and
release targets.  This allows the maintainer to control where the
merge or release starts from.

The internal variable TAG conflicts with this when the external TAG
is empty.  The Makefile accidentally chooses the internal TAG variable.

Resolve this by renaming the internal TAG variable to _TAG.  This
allows the scripts to pass in an external TAG variable with no
conflict.

The internal TAG variable is not used anywhere outside
Makefile.common.

No expected changes, purely cosmetic.

Signed-off-by: Don Zickus <dzickus@redhat.com>