New configs in arch/powerpc
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 CONFIG_CMDLINE:

 On some platforms, there is currently no way for the boot loader to
 pass arguments to the kernel. For these platforms, you can supply
 some command-line options at build time by entering them here.  In
 most cases you will need to specify the root device here.

 Symbol: CMDLINE [=]
 Type  : string
 Defined at arch/powerpc/Kconfig:882
   Prompt: Initial kernel command string
   Location:
     -> Kernel options

Cc: kernel-patches@redhat.com