[redhat] New configs in drivers/pci
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 CONFIG_PCI_J721E_HOST:

 Say Y here if you want to support the TI J721E PCIe platform
 controller in host mode. TI J721E PCIe controller uses Cadence PCIe
 core.

 Symbol: PCI_J721E_HOST [=n]
 Type  : bool
 Defined at drivers/pci/controller/cadence/Kconfig:48
   Prompt: TI J721E PCIe platform host controller
   Depends on: PCI [=y] && OF [=y]
   Location:
     -> Device Drivers
       -> PCI support (PCI [=y])
         -> PCI controller drivers
           -> Cadence PCIe controllers support
 Selects: PCIE_CADENCE_HOST [=n] && PCI_J721E [=n]

---

Cc: Myron Stowe <myron.stowe@redhat.com>
Cc: Prarit Bhargava <prarit@redhat.com>
Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>