[redhat] New configs in drivers/tty
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 CONFIG_SERIAL_IMX_EARLYCON:

 If you have enabled the earlycon on the Freescale IMX
 CPU you can make it the earlycon by answering Y to this option.

 Symbol: SERIAL_IMX_EARLYCON [=n]
 Type  : bool
 Defined at drivers/tty/serial/Kconfig:518
   Prompt: Earlycon on IMX serial port
   Depends on: TTY [=y] && HAS_IOMEM [=y] && OF [=y]
   Location:
     -> Device Drivers
       -> Character devices
         -> Enable TTY (TTY [=y])
           -> Serial drivers
 Selects: SERIAL_EARLYCON [=y]

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>