[redhat] New configs in drivers/usb
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 CONFIG_USB_OTG_PRODUCTLIST:

 If you say Y here, the "otg_productlist.h" file will be used as a
 product list, so USB peripherals not listed there will be
 rejected during enumeration.  This behavior is required by the
 USB OTG and EH specification for all devices not on your product's
 "Targeted Peripherals List".  "Embedded Hosts" are likewise
 allowed to support only a limited number of peripherals.

 Symbol: USB_OTG_PRODUCTLIST [=n]
 Type  : bool
 Defined at drivers/usb/core/Kconfig:58
   Prompt: Rely on OTG and EH Targeted Peripherals List
   Depends on: USB_SUPPORT [=y] && USB [=y]
   Location:
     -> Device Drivers
       -> USB support (USB_SUPPORT [=y])

---

Cc: Torez Smith <torez@redhat.com>
Cc: Don Zickus <dzickus@redhat.com>
Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>