[redhat] New configs in drivers/xen
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 CONFIG_XEN_UNPOPULATED_ALLOC:

 Use unpopulated memory ranges in order to create mappings for guest
 memory regions, including grant maps and foreign pages. This avoids
 having to balloon out RAM regions in order to obtain physical memory
 space to create such mappings.

 Symbol: XEN_UNPOPULATED_ALLOC [=n]
 Type  : bool
 Defined at drivers/xen/Kconfig:328
   Prompt: Use unpopulated memory ranges for guest mappings
   Depends on: XEN [=y] && X86 [=y] && ZONE_DEVICE [=y]
   Location:
     -> Device Drivers
       -> Xen driver support

---

Cc: Vitaly Kuznetsov <vkuznets@redhat.com>
Cc: Igor Mammedov <imammedo@redhat.com>
Cc: Andrew Jones <drjones@redhat.com>
Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>