Merge Upstream script fixes
As part of splitting the merge-upstream and release stages apart,
the merge upstream wasn't working quite right.  These are cleanup
fixes to fix that.

The main change is to drop the ark rebase patches script.  Running
that through a Makefile through gitlab CI creates a funky environment
where the next script (update-configs.sh) no longer exists because
of the git checkout changes.  Keep rebase-patches in the gitlab-ci
script for now.

That will be removed soon enough.  No effect on normal workflow.

The other fixes are self-explanatory.

Signed-off-by: Don Zickus <dzickus@redhat.com>