Create Patchlist.changelog file
A source git tree's workflow and Fedora's separation of upstream and distro
contributions don't overlap well.

A source git tree is always merging upstream content and downstream content
while avoiding a rebase.  By separating the tarballs and patches out, this
effectively causes a rebase.  This 'rebase' causes patches to not apply
correctly as git took care of it previously.

This patch creates a Patchlist.changelog file that lists all the SHAs that
are added to the upstream tree.

The goal is to create a useful file such that the source git tree can
create a single tarball and use the Patchlist.changelog file to adhere
to the spirit of the Fedora's separation rules.

No technical changes, just a new file added to the srpm.

Signed-off-by: Don Zickus <dzickus@redhat.com>