Partial revert: Add master merge check
I created an optimization to speed up the automated scripts when there was
nothing to merge in redhat/scripts/ci/ark-update-configs.sh under
2d1d129bbe310da5a8751e8e8ff4dc2209337d22

The thought was, if 'master' wasn't updated or needed merging, how can there be
any new configs to add?

Well it turns out that when merge conflicts arise, the merge of master into
os-build has to be done by hand and pushed manually.  This usually skips the
new config generation.

The above optimization then skips the config generation on later cron jobs
because the merge already happened (until the master is actually updated).

There is nothing wrong with re-running the config generation script even if
there is nothing to do.  Just undo the optimization.

Signed-off-by: Don Zickus <dzickus@redhat.com>