[redhat] New configs in drivers/acpi
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 CONFIG_DPTF_PCH_FIVR:

 This driver adds support for Dynamic Platform and Thermal Framework
 (DPTF) PCH FIVR Participant device support. This driver allows to
 switch PCH FIVR (Fully Integrated Voltage Regulator) frequency.
 This participant is responsible for exposing:
       freq_mhz_low_clock
       freq_mhz_high_clock

 To compile this driver as a module, choose M here:
 the module will be called dptf_pch_fivr.

 Symbol: DPTF_PCH_FIVR [=n]
 Type  : tristate
 Defined at drivers/acpi/dptf/Kconfig:18
   Prompt: DPTF PCH FIVR Participant
   Depends on: ACPI [=y] && X86 [=y]
   Location:
     -> Power management and ACPI options
       -> ACPI (Advanced Configuration and Power Interface) Support (ACPI [=y])

---

Cc: Al Stone <ahs3@redhat.com>
Cc: Lenny Szubowicz <lszubowi@redhat.com>
Cc: Prarit Bhargava <prarit@redhat.com>
Cc: David Arcari <darcari@redhat.com>
Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>