[redhat] New configs in drivers/misc
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 CONFIG_HISI_HIKEY_USB:

 If you say yes here this adds support for the on-board USB GPIO hub
 found on HiKey 960/970 boards, which is necessary to support
 switching between the dual-role USB-C port and the USB-A host ports
 using only one USB controller.

 Symbol: HISI_HIKEY_USB [=n]
 Type  : tristate
 Defined at drivers/misc/Kconfig:459
   Prompt: USB GPIO Hub on HiSilicon Hikey 960/970 Platform
   Depends on: (OF [=y] && GPIOLIB [=y] || COMPILE_TEST [=n]) && USB_ROLE_SWITCH [=m]
   Location:
     -> Device Drivers
       -> Misc devices

---

 CONFIG_INTEL_MEI_VIRTIO:

 This module implements mei hw emulation over virtio transport.
 The module will be called mei_virtio.
 Enable this if your virtual machine supports virtual mei
 device over virtio.

 Symbol: INTEL_MEI_VIRTIO [=n]
 Type  : tristate
 Defined at drivers/misc/mei/Kconfig:49
   Prompt: Intel MEI interface emulation with virtio framework
   Depends on: X86 [=y] && PCI [=y] && VIRTIO_PCI [=y]
   Location:
     -> Device Drivers
       -> Misc devices
 Selects: INTEL_MEI [=m]

---

Cc: John Linville <linville@redhat.com>
Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>