[redhat] New configs in drivers/pinctrl
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 CONFIG_PINCTRL_MSM8226:

 This is the pinctrl, pinmux, pinconf and gpiolib driver for the
 Qualcomm Technologies Inc TLMM block found on the Qualcomm
 Technologies Inc MSM8226 platform.

 Symbol: PINCTRL_MSM8226 [=n]
 Type  : tristate
 Defined at drivers/pinctrl/qcom/Kconfig:65
   Prompt: Qualcomm 8226 pin controller driver
   Depends on: PINCTRL [=y] && (ARCH_QCOM [=y] || COMPILE_TEST [=n]) && GPIOLIB [=y] && OF [=y]
   Location:
     -> Device Drivers
       -> Pin controllers (PINCTRL [=y])
 Selects: PINCTRL_MSM [=y]

---

Cc: David Arcari <darcari@redhat.com>
Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>