[redhat] New configs in drivers/video
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 CONFIG_BACKLIGHT_KTD253:

 Say y to enabled the backlight driver for the Kinetic KTD253
 which is a 1-wire GPIO-controlled backlight found in some mobile
 phones.

 Symbol: BACKLIGHT_KTD253 [=n]
 Type  : tristate
 Defined at drivers/video/backlight/Kconfig:185
   Prompt: Backlight Driver for Kinetic KTD253
   Depends on: HAS_IOMEM [=y] && BACKLIGHT_CLASS_DEVICE [=y] && (GPIOLIB [=y] || COMPILE_TEST [=n])
   Location:
     -> Device Drivers
       -> Graphics support
         -> Backlight & LCD device support
           -> Lowlevel Backlight controls (BACKLIGHT_CLASS_DEVICE [=y])

---

Cc: David Airlie <airlied@redhat.com>
Cc: Adam Jackson <ajax@redhat.com>
Cc: Lyude Paul <lyude@redhat.com>
Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>