[redhat] New configs in kernel/bpf
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 CONFIG_BPF_PRELOAD:

 This builds kernel module with several embedded BPF programs that are
 pinned into BPF FS mount point as human readable files that are
 useful in debugging and introspection of BPF programs and maps.

 Symbol: BPF_PRELOAD [=n]
 Type  : bool
 Defined at kernel/bpf/preload/Kconfig:6
   Prompt: Preload BPF file system with kernel specific program and map iterators
   Depends on: BPF [=y] && !COMPILE_TEST [=n]
   Location:
     -> General setup
 Selects: USERMODE_DRIVER [=n]

---

Cc: Jiri Olsa <jolsa@redhat.com>
Cc: Yauheni Kaliuta <ykaliuta@redhat.com>
Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>