[redhat] New configs in lib/Kconfig.debug
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 CONFIG_TEST_FREE_PAGES:

 Test that a memory leak does not occur due to a race between
 freeing a block of pages and a speculative page reference.
 Loading this module is safe if your kernel has the bug fixed.
 If the bug is not fixed, it will leak gigabytes of memory and
 probably OOM your system.

 Symbol: TEST_FREE_PAGES [=n]
 Type  : tristate
 Defined at lib/Kconfig.debug:2370
   Prompt: Test freeing pages
   Depends on: RUNTIME_TESTING_MENU [=y]
   Location:
     -> Kernel hacking
       -> Kernel Testing and Coverage
         -> Runtime Testing (RUNTIME_TESTING_MENU [=y])

---

Cc: Prarit Bhargava <prarit@redhat.com>
Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>