[redhat] New configs in kernel/rcu
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 CONFIG_RCU_SCALE_TEST:

 This option provides a kernel module that runs performance
 tests on the RCU infrastructure.  The kernel module may be built
 after the fact on the running kernel to be tested, if desired.

 Say Y here if you want RCU performance tests to be built into
 the kernel.
 Say M if you want the RCU performance tests to build as a module.
 Say N if you are unsure.

 Symbol: RCU_SCALE_TEST [=n]
 Type  : tristate
 Defined at kernel/rcu/Kconfig.debug:26
   Prompt: performance tests for RCU
   Depends on: DEBUG_KERNEL [=y]
   Location:
     -> Kernel hacking
       -> RCU Debugging
 Selects: TORTURE_TEST [=n] && SRCU [=y] && TASKS_RCU [=n] && TASKS_RUDE_RCU [=y] && TASKS_TRACE_RCU [=y]

---

Cc: Waiman Long <longman@redhat.com>
Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>