[redhat] create-tarball.sh: Fix syntax flagged by shellcheck
SC2006: Use $(...) notation instead of legacy backticked `...`.
SC1083: This { is literal. Check expression (missing ;/\n?) or quote it.
SC2086: Double quote to prevent globbing and word splitting.

BUT:
Disable shellcheck on last step:

On the last step, we DEPEND on word splitting for correct interpretation
of XZ_OPTIONS and XZ_THREADS, so remove quotes and, just before the
xz step:

  shellcheck disable=SC2086

Signed-off-by: Ben Crocker <bcrocker@redhat.com>