[redhat] ark-rebase-patches.sh: Fix for shellcheck
Changes actually consist of disabling shellcheck on a few lines:
• SC2086: Double quote to prevent globbing and word splitting.
• SC2059: Don't use variables in the printf format string. Use printf...

For the former, in the two gitlab commands involved,
GITLAB_CONFIG_OPT DEPENDS on word splitting.
For the latter, the ShellCheck wiki actually recommend disabling the warning
because a multi-line pattern is in use and there is no good rewrite.

Signed-off-by: Ben Crocker <bcrocker@redhat.com>