[redhat] Avoid offset of 1 between release in Makefile and commit
When doing dist-release, the release commit will have the wrong release
number, eg.:

  commit db61196cc6aab3f6ece65027551ebfb2f5edea90
  ...
      [redhat] kernel-5.10.0-0.rc3.20201111giteccc87672492.69.test
  ...
  -RHEL_RELEASE = 69
  +RHEL_RELEASE = 70
  ...

While the new release is 70 in example above, the commit title has the wrong
value. That mismatch happens because the release number is a Makefile.rhelver
variable included and used by make rules, and make doesn't see the updated
value, you need to eg. reload it through a sub make process. This change adds
a new rule and new make invocation so the proper RHEL_RELEASE value is used when
doing the release commit.

Signed-off-by: Herton R. Krzesinski <herton@redhat.com>