[redhat] fix the detection of changed files in dist-release
The git diff-index has a catch which you must pay attention, if you eg.
do this:

touch redhat/Makefile
git diff-index HEAD

It'll show there is a change even if contents didn't change. This makes
the current logic to not work since we may not really have changed the
contents, but the redirection will act similar as a touch on those files.
We need to run update-index before it so diff-index check works properly.

Also, the "exit 0" was not working as intended, and the Makefile commands
after the check were still running, thus I changed the code to use an if
conditional to make it work.

Signed-off-by: Herton R. Krzesinski <herton@redhat.com>