docs: Update docs to reflect newer workflow.
The workflow has recently changed such that all development is done
on the 'os-build' branch.  Update the docs to show how easy it is
to make a change, commit it, generate the srpm and upload it to koji.

Also add it a build dep for making a srpm: patchutils (for filterdiff).

V2: Fix checkout command and setup master branch
V3: Use default upstream/master and git pull mechanism
V4: Use default origin/master.  Add text to setup mirroring

Cc: Bastien Nocera <bnocera@redhat.com>
Cc: Prarit Bhargava <prarit@redhat.com>
Cc: Justin Forbes <jforbes@redhat.com>
Signed-off-by: Don Zickus <dzickus@redhat.com>