[redhat] New configs in drivers/ptp
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 CONFIG_PTP_1588_CLOCK_OCP:

 This driver adds support for an OpenCompute time card.

 The OpenCompute time card is an atomic clock along with
 a GPS receiver that provides a Grandmaster clock source
 for a PTP enabled network.

 More information is available at http://www.timingcard.com/

 Symbol: PTP_1588_CLOCK_OCP [=n]
 Type  : tristate
 Defined at drivers/ptp/Kconfig:154
   Prompt: OpenCompute TimeCard as PTP clock
   Depends on: PTP_1588_CLOCK [=y] && HAS_IOMEM [=y] && PCI [=y]
   Location:
     -> Device Drivers
       -> PTP clock support

---

Cc: Jiri Benc <jbenc@redhat.com>
Cc: Prarit Bhargava <prarit@redhat.com>
Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>