[redhat] New configs in mm/Kconfig
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 CONFIG_GUP_TEST:

 Provides /sys/kernel/debug/gup_test, which in turn provides a way
 to make ioctl calls that can launch kernel-based unit tests for
 the get_user_pages*() and pin_user_pages*() family of API calls.

 These tests include benchmark testing of the _fast variants of
 get_user_pages*() and pin_user_pages*(), as well as smoke tests of
 the non-_fast variants.

 There is also a sub-test that allows running dump_page() on any
 of up to eight pages (selected by command line args) within the
 range of user-space addresses. These pages are either pinned via
 pin_user_pages*(), or pinned via get_user_pages*(), as specified
 by other command line arguments.

 See tools/testing/selftests/vm/gup_test.c

 Symbol: GUP_TEST [=n]
 Type  : bool
 Defined at mm/Kconfig:824
   Prompt: Enable infrastructure for get_user_pages()-related unit tests
   Depends on: DEBUG_FS [=y]
   Location:
     -> Memory Management options

---

Cc: Rafael Aquini <aquini@redhat.com>
Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>