[redhat] New configs in drivers/hid
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 CONFIG_AMD_SFH_HID:

 If you say yes to this option, support will be included for the
 AMD Sensor Fusion Hub.
 This driver will enable sensors functionality on AMD platforms
 starting from 17h family of RYZEN parts.

 This driver can also be built as a module. If so, the module will
 be called amd-sfh.
 Say Y or M here if you want to support AMD SFH. If unsure, say N.
 Symbol: AMD_SFH_HID [=n]
 Type  : tristate
 Defined at drivers/hid/amd-sfh-hid/Kconfig:7
   Prompt: AMD Sensor Fusion Hub
   Depends on: INPUT [=y] && (X86_64 [=y] || COMPILE_TEST [=n]) && PCI [=y] && HID [=y]
   Location:
     -> Device Drivers
       -> HID support
         -> AMD SFH HID Support

---

Cc: Benjamin Tissoires <benjamin.tissoires@redhat.com>
Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>