[redhat] New configs in drivers/soc
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 CONFIG_LITEX_SOC_CONTROLLER:

 This option enables the SoC Controller Driver which verifies
 LiteX CSR access and provides common litex_get_reg/litex_set_reg
 accessors.
 All drivers that use functions from litex.h must depend on
 LITEX.

 Symbol: LITEX_SOC_CONTROLLER [=n]
 Type  : tristate
 Defined at drivers/soc/litex/Kconfig:8
   Prompt: Enable LiteX SoC Controller driver
   Depends on: OF [=y] || COMPILE_TEST [=n]
   Location:
     -> Device Drivers
       -> SOC (System On Chip) specific Drivers
         -> Enable LiteX SoC Builder specific drivers
 Selects: LITEX [=n]

---

Cc: Mark Salter <msalter@redhat.com>
Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>