[redhat] New configs in drivers/pwm
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 CONFIG_PWM_ATMEL_TCB:

 Generic PWM framework driver for Atmel Timer Counter Block.

 A Timer Counter Block provides 6 PWM devices grouped by 2.
 Devices in a given group must have the same period.

 To compile this driver as a module, choose M here: the module
 will be called pwm-atmel-tcb.

 Symbol: PWM_ATMEL_TCB [=n]
 Type  : tristate
 Defined at drivers/pwm/Kconfig:76
   Prompt: Atmel TC Block PWM support
   Depends on: PWM [=y] && OF [=y]
   Location:
     -> Device Drivers
       -> Pulse-Width Modulation (PWM) Support (PWM [=y])
 Selects: REGMAP_MMIO [=n]

---

 CONFIG_PWM_DWC:

 PWM driver for Synopsys DWC PWM Controller attached to a PCI bus.

 To compile this driver as a module, choose M here: the module
 will be called pwm-dwc.

 Symbol: PWM_DWC [=n]
 Type  : tristate
 Defined at drivers/pwm/Kconfig:167
   Prompt: DesignWare PWM Controller
   Depends on: PWM [=y] && PCI [=y]
   Location:
     -> Device Drivers
       -> Pulse-Width Modulation (PWM) Support (PWM [=y])

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>