New configs in lib/Kconfig.debug
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 CONFIG_DEBUG_VM_PGTABLE:

 This option provides a debug method which can be used to test
 architecture page table helper functions on various platforms in
 verifying if they comply with expected generic MM semantics. This
 will help architecture code in making sure that any changes or
 new additions of these helpers still conform to expected
 semantics of the generic MM. Platforms will have to opt in for
 this through ARCH_HAS_DEBUG_VM_PGTABLE.

 If unsure, say N.

 Symbol: DEBUG_VM_PGTABLE [=n]
 Type  : bool
 Defined at lib/Kconfig.debug:702
   Prompt: Debug arch page table for semantics compliance
   Depends on: MMU [=y] && ARCH_HAS_DEBUG_VM_PGTABLE [=y]
   Location:
     -> Kernel hacking
       -> Memory Debugging

Cc: Prarit Bhargava <prarit@redhat.com>