[redhat] New configs in sound/soc
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 CONFIG_SND_SOC_SOF_INTEL_SOUNDWIRE_LINK:

 This adds support for SoundWire with Sound Open Firmware
 for Intel(R) platforms.
 Say Y if you want to enable SoundWire links with SOF.
 If unsure select "N".

 Symbol: SND_SOC_SOF_INTEL_SOUNDWIRE_LINK [=n]
 Type  : bool
 Defined at sound/soc/sof/intel/Kconfig:356
   Prompt: SOF support for SoundWire
   Depends on: SOUND [=m] && !UML && SND [=m] && SND_SOC [=m] && SND_SOC_SOF_TOPLEVEL [=y] && SND_SOC_SOF_INTEL_TOPLEVEL [=y] && SND_SOC_SOF_INTEL_PCI [=m] && ACPI [=y]
   Location:
     -> Device Drivers
       -> Sound card support (SOUND [=m])
         -> Advanced Linux Sound Architecture (SND [=m])
           -> ALSA for SoC audio support (SND_SOC [=m])
             -> Sound Open Firmware Support (SND_SOC_SOF_TOPLEVEL [=y])
               -> SOF support for Intel audio DSPs (SND_SOC_SOF_INTEL_TOPLEVEL [=y])

---

Cc: Jaroslav Kysela <jkysela@redhat.com>
Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>

v2:
- CONFIG_SND_SOC_SOF_INTEL_SOUNDWIRE_LINK set to y for x86 only.

Signed-off-by: Patrick Talbert <ptalbert@redhat.com>