configs: clean up LSM configs
1. Set CONFIG_SECURITY_LOCKDOWN_LSM=y on both Fedora and ARK and move
   the associated configs from fedora/ to common/. On both this is
   required for proper UEFI secure boot support.
2. Remove ark/generic/CONFIG_SECURITY_SELINUX_BOOTPARAM_VALUE - this
   config has been removed upstream.
3. Deduplicate default value of CONFIG_LSM_MMAP_MIN_ADDR - set it to
   65536 under common/ and only override it in
   fedora/generic/arm/armv7/.
4. Trim LSMs that are not build-enabled from CONFIG_LSM on Fedora/ARK,
   which can now be unified under common/.

Note that this commit adds the Lockdown LSM to the default LSM list and
therefore effectively enables it on both Fedora (where it was enabled in
build, but disabled on boot) and ARK (where it wasn't even enabled at
build). According to Peter Robinson and Al Stone it should be enabled,
so hopefully this is the expected result.

Signed-off-by: Ondrej Mosnacek <omosnace@redhat.com>