Add a redhat/rebase-notes.txt file
As discussed on the Fedora kernel list add a redhat/rebase-notes.txt
file with notes about things to keep in mind when rebasing the
Fedora kernel packages to a new upstream release.

And populate this file with a note about disabling
SND_INTEL_BYT_PREFER_SOF for Fedora 33 and older.

Signed-off-by: Hans de Goede <hdegoede@redhat.com>