[redhat] New configs in drivers/perf
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 CONFIG_ARM_CMN:

 Support for PMU events monitoring on the Arm CMN-600 Coherent Mesh
 Network interconnect.

 Symbol: ARM_CMN [=n]
 Type  : tristate
 Defined at drivers/perf/Kconfig:44
   Prompt: Arm CMN-600 PMU support
   Depends on: PERF_EVENTS [=y] && (ARM64 [=y] || COMPILE_TEST [=n] && 64BIT [=y])
   Location:
     -> Device Drivers
       -> Performance monitor support

---

Cc: Mark Salter <msalter@redhat.com>
Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>