[redhat] New configs in drivers/char
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 CONFIG_TCG_TIS_I2C_CR50:

 This is a driver for the Google cr50 I2C TPM interface which is a
 custom microcontroller and requires a custom i2c protocol interface
 to handle the limitations of the hardware.  To compile this driver
 as a module, choose M here; the module will be called tcg_tis_i2c_cr50.

 Symbol: TCG_TIS_I2C_CR50 [=n]
 Type  : tristate
 Defined at drivers/char/tpm/Kconfig:89
   Prompt: TPM Interface Specification 2.0 Interface (I2C - CR50)
   Depends on: TCG_TPM [=y] && I2C [=y]
   Location:
     -> Device Drivers
       -> Character devices
         -> TPM Hardware Support (TCG_TPM [=y])
 Selects: TCG_CR50

---

Cc: John Linville <linville@redhat.com>
Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>