[redhat] New configs in drivers/powercap
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 CONFIG_DTPM:

 This enables support for the power capping for the dynamic
 thermal power management userspace engine.

 Symbol: DTPM [=n]
 Type  : bool
 Defined at drivers/powercap/Kconfig:47
   Prompt: Power capping for Dynamic Thermal Power Management
   Depends on: POWERCAP [=y]
   Location:
     -> Device Drivers
       -> Generic powercap sysfs driver (POWERCAP [=y])

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>