[redhat] New configs in drivers/ntb
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 CONFIG_NTB_EPF:

 This driver supports EPF NTB on configurable endpoint.
 If unsure, say N.

 Symbol: NTB_EPF [=n]
 Type  : tristate
 Defined at drivers/ntb/hw/epf/Kconfig:1
   Prompt: Generic EPF Non-Transparent Bridge support
   Depends on: NTB [=m] && m && MODULES [=y]
   Location:
     -> Device Drivers
       -> Non-Transparent Bridge support (NTB [=m])

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>