[redhat] New configs in lib/Kconfig.debug
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 CONFIG_DEBUG_INFO_DWARF_TOOLCHAIN_DEFAULT:

 The implicit default version of DWARF debug info produced by a
 toolchain changes over time.

 This can break consumers of the debug info that haven't upgraded to
 support newer revisions, and prevent testing newer versions, but
 those should be less common scenarios.

 If unsure, say Y.

 Symbol: DEBUG_INFO_DWARF_TOOLCHAIN_DEFAULT [=n]
 Type  : bool
 Defined at lib/Kconfig.debug:264
   Prompt: Rely on the toolchain's implicit default DWARF version
   Depends on: <choice>
   Location:
     -> Kernel hacking
       -> Compile-time checks and compiler options
         -> Compile the kernel with debug info (DEBUG_INFO [=y])
           -> DWARF version (<choice> [=y])

---

Cc: Prarit Bhargava <prarit@redhat.com>
Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>