[redhat] New configs in arch/arm64
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 CONFIG_CMDLINE_EXTEND:

 The command-line arguments provided by the boot loader will be
 appended to the default kernel command string.

 Symbol: CMDLINE_EXTEND [=n]
 Type  : bool
 Defined at arch/arm64/Kconfig:1859
   Prompt: Extend bootloader kernel arguments
   Depends on: <choice>
   Location:
     -> Boot options
       -> Default kernel command string (CMDLINE [=console=ttyAMA0])
         -> Kernel command line type (<choice> [=y])

---

 CONFIG_CMDLINE_FROM_BOOTLOADER:

 Uses the command-line options passed by the boot loader. If
 the boot loader doesn't provide any, the default kernel command
 string provided in CMDLINE will be used.

 Symbol: CMDLINE_FROM_BOOTLOADER [=y]
 Type  : bool
 Defined at arch/arm64/Kconfig:1852
   Prompt: Use bootloader kernel arguments if available
   Depends on: <choice>
   Location:
     -> Boot options
       -> Default kernel command string (CMDLINE [=console=ttyAMA0])
         -> Kernel command line type (<choice> [=y])

---

Cc: Mark Salter <msalter@redhat.com>
Cc: Mark Langsdorf <mlangsdo@redhat.com>
Cc: Jeremy Linton <jlinton@redhat.com>
Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>