[redhat] New configs in arch/x86
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 CONFIG_KVM_XEN:

 Provides KVM support for the hosting Xen HVM guests and
 passing Xen hypercalls to userspace.

 If in doubt, say "N".

 Symbol: KVM_XEN [=n]
 Type  : bool
 Defined at arch/x86/kvm/Kconfig:106
   Prompt: Support for Xen hypercall interface
   Depends on: VIRTUALIZATION [=y] && KVM [=m]
   Location:
     -> Virtualization (VIRTUALIZATION [=y])
       -> Kernel-based Virtual Machine (KVM) support (KVM [=m])

---

Cc: David Arcari <darcari@redhat.com>
Cc: Prarit Bhargava <prarit@redhat.com>
Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>