redhat: add initial support for centos stream dist-git sync on Makefiles
This change introduces some initial centos stream support, mainly what's
needed for a dist-git sync. The dist-git sync is initially proposed to
be a fork + merge request model, and thus we need some changes for the
Makefiles. I also dropped IS_FEDORA variable, and replace it with
DISTRO, which can have either "fedora" or "centos" (and could be set
later to "rhel" or anything else if needed through the code).

While at it, I also fixed the default RHPKG_BIN to rhpkg instead of
rhpkg-sha512 in the default RHEL case, the former was a wrapper for
the sha512 transition on dist-git at RHEL 8 time and is not needed
anymore, it is now deprecated and will be removed from rhpkg.

Example of a dist-git sync with new centos dist-git:
make DIST=".el9" GL_DISTGIT_USER=<gitlab user> BUILDID="" RHDISTGIT_BRANCH=c9s dist-git

v2: use PACKAGE_NAME in Makefile.rhpkg as suggested by Jan Stancek
v3: make the default DISTRO rhel in case DIST is not fc*, to not break
    folks building kernel-ark on rhel which are used to not have to
    specify an extra variable, as reported by Don Zickus

Signed-off-by: Herton R. Krzesinski <herton@redhat.com>