[redhat] New configs in sound/soc
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 CONFIG_SND_SOC_INTEL_CATPT:

 Enable support for Intel(R) Haswell and Broadwell platforms
 with I2S codec present. This is a recommended option.
 Say Y or m if you have such device.
 If unsure, say N.

 Symbol: SND_SOC_INTEL_CATPT [=n]
 Type  : tristate
 Defined at sound/soc/intel/Kconfig:37
   Prompt: Haswell and Broadwell
   Depends on: SOUND [=m] && !UML && SND [=m] && SND_SOC [=m] && SND_SOC_INTEL_SST_TOPLEVEL [=y] && (ACPI [=y] || COMPILE_TEST [=n]) && DMADEVICES [=y] && SND_DMA_SGBUF [=y]
   Location:
     -> Device Drivers
       -> Sound card support (SOUND [=m])
         -> Advanced Linux Sound Architecture (SND [=m])
           -> ALSA for SoC audio support (SND_SOC [=m])
             -> Intel ASoC SST drivers (SND_SOC_INTEL_SST_TOPLEVEL [=y])
 Selects: DW_DMAC_CORE [=y] && SND_SOC_ACPI_INTEL_MATCH [=m]
 Selected by [n]:
   - SND_SOC_INTEL_HASWELL [=n] && SOUND [=m] && !UML && SND [=m] && SND_SOC [=m] && SND_SOC_INTEL_SST_TOPLEVEL [=y]

---

Cc: Jaroslav Kysela <jkysela@redhat.com>
Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>