filter-modules.sh: Fix copy/paste error 'input'
Fix a copy/paste error of 'input' and use the obviously intended 'char'.
The effect of this fix is that the cm4000_cs, cm4040_cs and mwave
modules will now be shipped in Fedora's kernel-modules package, as was
apparently intended. Nothing should change for RHEL, which actually
doesn't build these modules.

Signed-off-by: Paul Bolle <pebolle@tiscali.nl>