[redhat] New configs in crypto/Kconfig
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 CONFIG_CRYPTO_ECDSA:

 Elliptic Curve Digital Signature Algorithm (NIST P192, P256 etc.)
 is A NIST cryptographic standard algorithm. Only signature verification
 is implemented.

 Symbol: CRYPTO_ECDSA [=n]
 Type  : tristate
 Defined at crypto/Kconfig:245
   Prompt: ECDSA (NIST P192, P256 etc.) algorithm
   Depends on: CRYPTO [=y]
   Location:
     -> Cryptographic API (CRYPTO [=y])
 Selects: CRYPTO_ECC [=m] && CRYPTO_AKCIPHER [=y] && ASN1 [=y]

---

Cc: Herbert Xu <herbert.xu@redhat.com>
Cc: "David S. Miller" <davem@redhat.com>
Cc: Ondrej Mosnacek <omosnace@redhat.com>
Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>

v2:
 - CONFIG_CRYPTO_ECDSA enable as a module.

Signed-off-by: Patrick Talbert <ptalbert@redhat.com>