Set date in package release from repository commit, not system clock
Packages can be rebuilt from tagged commits in the kernel-ark repository by
cloning it and running the packaging scripts. This should be reproducible:
packages built from identical sources should have the same NVR.

For builds based on upstream snapshots, the package release contains a date
that is meant to correspond to changes in the fedora-ark repository. (It is
not the date of the upstream snapshot.) Instead, the packaging scripts set
this date from the system clock. Change this so that the commit date in the
kernel-ark repository is used.

Signed-off-by: David Ward <david.ward@ll.mit.edu>