[redhat] New configs in drivers/net/ethernet/mellanox
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 CONFIG_MLX5_TC_SAMPLE:

 Say Y here if you want to support offloading sample rules via tc
 sample action.
 If set to N, will not be able to configure tc rules with sample
 action.

 If unsure, set to Y

 Symbol: MLX5_TC_SAMPLE [=y]
 Type  : bool
 Defined at drivers/net/ethernet/mellanox/mlx5/core/Kconfig:107
   Prompt: MLX5 TC sample offload support
   Depends on: NETDEVICES [=y] && ETHERNET [=y] && NET_VENDOR_MELLANOX [=y] && MLX5_CLS_ACT [=y]
   Location:
     -> Device Drivers
       -> Network device support (NETDEVICES [=y])
         -> Ethernet driver support (ETHERNET [=y])
           -> Mellanox devices (NET_VENDOR_MELLANOX [=y])
             -> Mellanox 5th generation network adapters (ConnectX series) Ethernet support (MLX5_CORE_EN [=y])
               -> Mellanox Technologies MLX5 SRIOV E-Switch support (MLX5_ESWITCH [=y])
                 -> MLX5 TC classifier action support (MLX5_CLS_ACT [=y])

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>