[redhat] New configs in drivers/leds
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 CONFIG_LEDS_RT4505:

 This option enables support for the RT4505 flash LED controller.
 RT4505 includes torch and flash functions with programmable current.
 And it's commonly used to compensate the illuminance for the camera
 inside the mobile product like as phones or tablets.

 Symbol: LEDS_RT4505 [=n]
 Type  : tristate
 Defined at drivers/leds/flash/Kconfig:5
   Prompt: LED support for RT4505 flashlight controller
   Depends on: NEW_LEDS [=y] && LEDS_CLASS_FLASH [=m] && I2C [=y] && OF [=y] && (V4L2_FLASH_LED_CLASS [=n] || !V4L2_FLASH_LED_CLASS [=n])
   Location:
     -> Device Drivers
       -> LED Support (NEW_LEDS [=y])
 Selects: REGMAP_I2C [=m]

---

Cc: Tony Camuso <tcamuso@redhat.com>
Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>