Enable CONFIG_DRM_AMDGPU_USERPTR for everyone
In the 5.12 kernel cycle, CONFIG_HSA_AMD was changed to select
DRM_AMDGPU_USERPTR. This creates a config mismatch for ppc64le as
HSA_AMD was enabled there, but DRM_AMDGPU_USERPTR was not. This change
enables DRM_AMDGPU_USERPTR for everything, it is dependent on
DRM_AMDGPU, so this will not change the current s390X configs.

Signed-off-by: Justin M. Forbes <jforbes@fedoraproject.org>