[redhat] New configs in mm/Kconfig
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 CONFIG_CMA_SYSFS:

 This option exposes some sysfs attributes to get information
 from CMA.

 Symbol: CMA_SYSFS [=n]
 Type  : bool
 Defined at mm/Kconfig:532
   Prompt: CMA information through sysfs interface
   Depends on: CMA [=y] && SYSFS [=y]
   Location:
     -> Memory Management options
       -> Contiguous Memory Allocator (CMA [=y])

---

Cc: Rafael Aquini <aquini@redhat.com>
Cc: Aristeu Rozanski <arozansk@redhat.com>
Cc: Don Dutile <ddutile@redhat.com>
Cc: Nico Pache <npache@redhat.com>
Cc: Waiman Long <longman@redhat.com>
Cc: mm-maint@redhat.com
Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>

v2:
 - CONFIG_CMA_SYSFS enabled.

Signed-off-by: Patrick Talbert <ptalbert@redhat.com>