redhat: add initial rpminspect configuration
CI has been doing rpminspect runs against built rpm packages.

We have to add specific configuration for rpminspect, that can vary
depending on the kernel branch. For example, older releases didn't
have the kernel meta-package, but today kernel have it and requires
kernel-core/kernel-modules which are the ones which contains files.
This configuration is supposed as a tweak only for the cases where
it makes sense having it here, to ship with dist-git, instead of
changing rpminspect-data-* packages.

Signed-off-by: Herton R. Krzesinski <herton@redhat.com>