[redhat] New configs in drivers/vdpa
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 CONFIG_VDPA_SIM_BLOCK:

 vDPA block device simulator which terminates IO request in a
 memory buffer.

 Symbol: VDPA_SIM_BLOCK [=n]
 Type  : tristate
 Defined at drivers/vdpa/Kconfig:29
   Prompt: vDPA simulator for block device
   Depends on: VDPA [=m] && VDPA_SIM [=m]
   Location:
     -> Device Drivers
       -> vDPA drivers (VDPA [=m])
         -> vDPA device simulator core (VDPA_SIM [=m])

---

 CONFIG_VDPA_SIM_NET:

 vDPA networking device simulator which loops TX traffic back to RX.

 Symbol: VDPA_SIM_NET [=n]
 Type  : tristate
 Defined at drivers/vdpa/Kconfig:22
   Prompt: vDPA simulator for networking device
   Depends on: VDPA [=m] && VDPA_SIM [=m]
   Location:
     -> Device Drivers
       -> vDPA drivers (VDPA [=m])
         -> vDPA device simulator core (VDPA_SIM [=m])
 Selects: GENERIC_NET_UTILS [=y]

---

Cc: "Michael S. Tsirkin" <mst@redhat.com>
Cc: Jason Wang <jasowang@redhat.com>
Cc: rhvirt-patches@redhat.com
Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>

v3:
 - rebase
 - CONFIG_VDPA_SIM_NET remove matching fedora config
v2:
 - CONFIG_VDPA_SIM_BLOCK enabled as a module
 - CONFIG_VDPA_SIM_NET enabled as a module

Signed-off-by: Patrick Talbert <ptalbert@redhat.com>