rpmspec: revert/drop content hash for kernel-headers
This was a change done in the RHEL 8 kernel package, with the purpose
of providing means to other tools to determine if the kernel-headers
changed between package updates. For example, linux containers that
only install kernel-headers would not be flagged for the need of CVE
updates if kernel-headers didn't change, in case of kernel CVEs that
didn't affect the headers.

But this content hash ended up never being used and AFAIK there
were no users. The need to avoid kernel CVEs flagging containers needing
updates only because of kernel-headers installed but not affected,
was handled through changes made in Red Hat's errata tool, and the
content hash here has no use anymore.

Signed-off-by: Herton R. Krzesinski <herton@redhat.com>