[redhat] Flip some diag modules from m to y
Bugzilla: https://bugzilla.redhat.com/show_bug.cgi?id=1948340
Upstream Status: RHEL only
Tested: Instructions in the first comment in bz1948340.

Diag modules are required when investigating sockets with the `ss` tool,
and are loaded on-demand. This is also what happens when using the
`sosreport` tool. Recent discussions pointed out it was not always
possible to load extra modules on a production system, for various
reasons, and the `sosreport` tool was updated to avoid modules to be
loaded by default. Links to the discussions are available in the bz.

The issue is with this policy the `ss` command is now skipped, except if
the administrator explicitly allows it to do so or if diag modules are
already loaded. This results in a loss of valuable information when
generating an `sosreport`, impacting the following investigation.

This patch enables a set of selected DIAG modules to be built-in, so
that `ss` and `sosreport` can be used without loading additional
modules (only diag modules not depending on another loadable module were
considered):

- CONFIG_INET_DIAG
- CONFIG_INET_RAW_DIAG
- CONFIG_INET_TCP_DIAG
- CONFIG_INET_UDP_DIAG
- CONFIG_INET_MPTCP_DIAG
- CONFIG_NETLINK_DIAG
- CONFIG_PACKET_DIAG
- CONFIG_UNIX_DIAG

The generated configuration do not enable any additional non-listed
options with this change. The kernel image increase in size is very
minimal.

Signed-off-by: Antoine Tenart <atenart@redhat.com>