[redhat] New configs in lib/Kconfig.debug
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 CONFIG_TEST_CLOCKSOURCE_WATCHDOG:

 Enable this option to create a kernel module that will trigger
 a test of the clocksource watchdog.  This module may be loaded
 via modprobe or insmod in which case it will run upon being
 loaded, or it may be built in, in which case it will run
 shortly after boot.

 If unsure, say N.

 Symbol: TEST_CLOCKSOURCE_WATCHDOG [=n]
 Type  : tristate
 Defined at lib/Kconfig.debug:2591
   Prompt: Test clocksource watchdog in kernel space
   Depends on: RUNTIME_TESTING_MENU [=y] && CLOCKSOURCE_WATCHDOG [=y]
   Location:
     -> Kernel hacking
       -> Kernel Testing and Coverage
         -> Runtime Testing (RUNTIME_TESTING_MENU [=y])

---

Cc: Prarit Bhargava <prarit@redhat.com>
Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>