[redhat] New configs in drivers/net/wwan
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 CONFIG_IOSM:

 This driver enables Intel M.2 WWAN Device communication.

 If you have one of those Intel M.2 WWAN Modules and wish to use it in
 Linux say Y/M here.

 If unsure, say N.

 Symbol: IOSM [=n]
 Type  : tristate
 Defined at drivers/net/wwan/Kconfig:59
   Prompt: IOSM Driver for Intel M.2 WWAN Device
   Depends on: NETDEVICES [=y] && WWAN [=y] && INTEL_IOMMU [=y]
   Location:
     -> Device Drivers
       -> Network device support (NETDEVICES [=y])
         -> Wireless WAN
           -> WWAN Driver Core (WWAN [=y])

---

 CONFIG_WWAN_HWSIM:

 This driver is a developer testing tool that can be used to test WWAN
 framework.

 To compile this driver as a module, choose M here: the module will be
 called wwan_hwsim.  If unsure, say N.

 Symbol: WWAN_HWSIM [=n]
 Type  : tristate
 Defined at drivers/net/wwan/Kconfig:19
   Prompt: Simulated WWAN device
   Depends on: NETDEVICES [=y] && WWAN [=y]
   Location:
     -> Device Drivers
       -> Network device support (NETDEVICES [=y])
         -> Wireless WAN
           -> WWAN Driver Core (WWAN [=y])

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>