[redhat] New configs in drivers/clk
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 CONFIG_COMMON_CLK_HI3559A:

 Build the clock driver for hi3559a.

 Symbol: COMMON_CLK_HI3559A [=y]
 Type  : bool
 Defined at drivers/clk/hisilicon/Kconfig:18
   Prompt: Hi3559A Clock Driver
   Depends on: COMMON_CLK [=y] && (ARCH_HISI [=y] || COMPILE_TEST [=n])
   Location:
     -> Device Drivers
       -> Common Clock Framework (COMMON_CLK [=y])

---

 CONFIG_LMK04832:

 Say yes here to build support for Texas Instruments' LMK04832 Ultra
 Low-Noise JESD204B Compliant Clock Jitter Cleaner With Dual Loop PLLs

 Symbol: LMK04832 [=n]
 Type  : tristate
 Defined at drivers/clk/Kconfig:54
   Prompt: Ti LMK04832 JESD204B Compliant Clock Jitter Cleaner
   Depends on: COMMON_CLK [=y] && SPI [=y]
   Location:
     -> Device Drivers
       -> Common Clock Framework (COMMON_CLK [=y])
 Selects: REGMAP_SPI [=m]

---

Cc: Mark Salter <msalter@redhat.com>
Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>

v3:
 - CONFIG_COMMON_CLK_HI3559A enabled for fedora.
v2:
 - CONFIG_COMMON_CLK_HI3559A moved to arm/aarch64 and disabled.

Signed-off-by: Patrick Talbert <ptalbert@redhat.com>