[redhat] New configs in drivers/mtd
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 CONFIG_MTD_MCHP48L640:

 This enables access to Microchip 48L640 EERAM chips, using SPI.

 Symbol: MTD_MCHP48L640 [=n]
 Type  : tristate
 Defined at drivers/mtd/devices/Kconfig:92
   Prompt: Microchip 48L640 EERAM
   Depends on: MTD [=m] && HAS_IOMEM [=y] && SPI_MASTER [=y]
   Location:
     -> Device Drivers
       -> Memory Technology Device (MTD) support (MTD [=m])
         -> Self-contained MTD device drivers

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>