[redhat] New configs in drivers/power
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 CONFIG_BATTERY_RT5033:

 This adds support for battery fuel gauge in Richtek RT5033 PMIC.
 The fuelgauge calculates and determines the battery state of charge
 according to battery open circuit voltage.

 Symbol: BATTERY_RT5033 [=n]
 Type  : tristate
 Defined at drivers/power/supply/Kconfig:713
   Prompt: RT5033 fuel gauge support
   Depends on: POWER_SUPPLY [=y] && I2C [=y]
   Location:
     -> Device Drivers
       -> Power supply class support (POWER_SUPPLY [=y])
 Selects: REGMAP_I2C [=m]

---

Cc: Lenny Szubowicz <lszubowi@redhat.com>
Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>