[redhat] New configs in drivers/watchdog
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 CONFIG_WATCHDOG_HRTIMER_PRETIMEOUT:

 Enable this if you want to use a hrtimer timer based pretimeout for
 watchdogs that do not natively support pretimeout support. Be aware
 that because this pretimeout functionality uses hrtimers, it may not
 be able to fire before the actual watchdog fires in some situations.

 Symbol: WATCHDOG_HRTIMER_PRETIMEOUT [=n]
 Type  : bool
 Defined at drivers/watchdog/Kconfig:76
   Prompt: Enable watchdog hrtimer-based pretimeouts
   Depends on: WATCHDOG [=y]
   Location:
     -> Device Drivers
       -> Watchdog Timer Support (WATCHDOG [=y])

---

Cc: Jonathan Toppins <jtoppins@redhat.com>
Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>