[redhat] New configs in lib/Kconfig.debug
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 CONFIG_STACKTRACE_BUILD_ID:

 Selecting this option adds build ID information for symbols in
 stacktraces printed with the printk format '%p[SR]b'.

 This option is intended for distros where debuginfo is not easily
 accessible but can be downloaded given the build ID of the vmlinux or
 kernel module where the function is located.

 Symbol: STACKTRACE_BUILD_ID [=n]
 Type  : bool
 Defined at lib/Kconfig.debug:38
   Prompt: Show build ID information in stacktraces
   Depends on: PRINTK [=y]
   Location:
     -> Kernel hacking
       -> printk and dmesg options

---

Cc: Prarit Bhargava <prarit@redhat.com>
Cc: Aristeu Rozanski <arozansk@redhat.com>
Cc: Christoph von Recklinghausen <crecklin@redhat.com>
Cc: Eugene Syromiatnikov <esyromia@redhat.com>
Cc: "Herton R. Krzesinski" <herton@redhat.com>
Cc: Nico Pache <npache@redhat.com>
Cc: Oleg Nesterov <onestero@redhat.com>
Cc: Rafael Aquini <aquini@redhat.com>
Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>