[redhat] New configs in sound/soc
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: SND_SOC_SSM2518 [=n]
 Type  : tristate
 Defined at sound/soc/codecs/Kconfig:1327
   Prompt: Analog Devices SSM2518 Class-D Amplifier
   Depends on: SOUND [=m] && !UML && SND [=m] && SND_SOC [=m] && I2C [=y]
   Location:
     -> Device Drivers
       -> Sound card support (SOUND [=m])
         -> Advanced Linux Sound Architecture (SND [=m])
           -> ALSA for SoC audio support (SND_SOC [=m])
             -> CODEC drivers
 Implied by [n]:
   - SND_SOC_ALL_CODECS [=n] && SOUND [=m] && !UML && SND [=m] && SND_SOC [=m] && COMPILE_TEST [=n]

---

Cc: Jaroslav Kysela <jkysela@redhat.com>
Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>