redhat/configs: Move CHACHA and POLY1305 to core kernel to allow BIG_KEYS=y
Bugzilla: https://bugzilla.redhat.com/show_bug.cgi?id=1983298
Upstream: Config only
Testing: keyutils testsuite

The big-key key type is a bool as it's expected to be built into the core
kernel, however it requires ChaCha20Poly1305 crypto to function.
Currently, the crypto bits are built as modules, so BIG_KEYS doesn't even
get enabled.  BIG_KEYS does a depends-on in Kconfig because using select is
non-transitive and all the dependencies would need selecting - but that's a
variable list, depending on arch.

Fix this by setting the arch-specific ChaCha20 and Poly1305 deps to =y as
well as the direct crypto library modules required by BIG_KEYS.  BIG_KEYS
is already set to =y so that then becomes available automatically.

This was detected by the keyutils testsuite trying to create a big key.

Signed-off-by: David Howells <dhowells@redhat.com>