redhat: add gating configuration for centos stream/rhel9
Right now when we do the dist-git sync, it overrides/deletes everything
which is not on the src.rpm, which might delete any gating configuration
on dist-git. Make sure we always copy the gating configuration for
centos/rhel9. This might need some tweak later if Fedora or future RHEL
uses the same gating file but with different config.

Signed-off-by: Herton R. Krzesinski <herton@redhat.com>