redhat: configs: Enable CONFIG_WIRELESS_HOTKEY
WIRELESS_HOTKEY is not new, it is a rename of the HP_WIRELESS
Kconfig symbol because the same firmware interface is now also used
on laptops from other vendors.

CONFIG_HP_WIRELESS has always been enabled in RHEL kernels and it
was in fact still enabled in 5.13.0-1.el9. So enabling
CONFIG_WIRELESS_HOTKEY basically restores the old status-quo.

Signed-off-by: Hans de Goede <hdegoede@redhat.com>