Disable liquidio driver on ark/rhel
Bugzilla: https://bugzilla.redhat.com/show_bug.cgi?id=1993393

liquidio driver was already disabled on Fedora, but was still being
enabled on RHEL/ark. This was enabled back in rhel8 but we got a request
to revert that later in rhel8. So there is no reason to keep it enabled
anymore in RHEL/ark too.

Disabling the driver also allowed some simplification/removal of now
uneeded config files and unification of them. While at it, I also
moved CAVIUM_PTP to m in rhel/ark as is already the case in Fedora
(afaik there is no apparent reason to keep it builtin, and was just the
result of default choice of this option looking at the option history),
and I also disabled LIQUIDIO_VF following the liquidio disabling (this
affects Fedora too where it was still enabled).

Signed-off-by: Herton R. Krzesinski <herton@redhat.com>