arm64: use common CONFIG_MAX_ZONEORDER for arm kernel
Now that RHEL9 is using 4K pagesize, MAX_ZONEORDER is defaulting
to 11. Fedora uses an out of tree patch to default to 13 when
building for server class machines. RHEL9 should also be using 13,
so make the MAX_ZONEORDER config and the out of tree patch common
between RHEL9 and Fedora.

Signed-off-by: Mark Salter <msalter@redhat.com>