[redhat] New configs in certs/Kconfig
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: MODULE_SIG_KEY_TYPE_ECDSA [=n]
 Type  : bool
 Defined at certs/Kconfig:31
   Prompt: ECDSA
   Depends on: <choice> && (MODULE_SIG [=y] || IMA_APPRAISE_MODSIG [=y] && MODULES [=y])
   Location:
     -> Cryptographic API (CRYPTO [=y])
       -> Certificates for signature checking
         -> Type of module signing key to be generated (<choice> [=y])
 Selects: CRYPTO_ECDSA [=m]

---

 Symbol: MODULE_SIG_KEY_TYPE_RSA [=y]
 Type  : bool
 Defined at certs/Kconfig:25
   Prompt: RSA
   Depends on: <choice> && (MODULE_SIG [=y] || IMA_APPRAISE_MODSIG [=y] && MODULES [=y])
   Location:
     -> Cryptographic API (CRYPTO [=y])
       -> Certificates for signature checking
         -> Type of module signing key to be generated (<choice> [=y])

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>