New configs in drivers/net/ethernet/litex
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: LITEX_LITEETH [=n]
 Type  : tristate
 Defined at drivers/net/ethernet/litex/Kconfig:18
   Prompt: LiteX Ethernet support
   Depends on: NETDEVICES [=y] && ETHERNET [=y] && NET_VENDOR_LITEX [=y] && OF_NET [=y]
   Location:
     -> Device Drivers
       -> Network device support (NETDEVICES [=y])
         -> Ethernet driver support (ETHERNET [=y])
           -> LiteX devices (NET_VENDOR_LITEX [=y])

v2:
 - NET_VENDOR_LITEX disabled

 Signed-off-by: Patrick Talbert <ptalbert@redhat.com>