New configs in fs/ksmbd
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: SMB_SERVER [=n]
 Type  : tristate
 Defined at fs/ksmbd/Kconfig:1
   Prompt: SMB3 server support (EXPERIMENTAL)
   Depends on: NETWORK_FILESYSTEMS [=y] && INET [=y] && MULTIUSER [=y] && FILE_LOCKING [=y]
   Location:
     -> File systems
       -> Network File Systems (NETWORK_FILESYSTEMS [=y])
 Selects: NLS [=y] && NLS_UTF8 [=m] && CRYPTO [=y] && CRYPTO_MD4 [=m] && CRYPTO_MD5 [=y] && CRYPTO_HMAC [=y] && CRYPTO_ECB [=y] && CRYPTO_LIB_DES [=m] && CRYPTO_SHA256 [=y] && CRYPTO_CMAC [=y] && CRYPTO_SHA512 [=y] && CRYPTO_AEAD2 [=y] && CRYPTO_CCM [=y] && CRYPTO_GCM [=y] && ASN1 [=y] && OID_REGISTRY [=y]