[redhat] New configs in drivers/iio
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: AD5110 [=n]
 Type  : tristate
 Defined at drivers/iio/potentiometer/Kconfig:9
   Prompt: Analog Devices AD5110 and similar Digital Potentiometer driver
   Depends on: IIO [=m] && I2C [=y]
   Location:
     -> Device Drivers
       -> Industrial I/O support (IIO [=m])
         -> Digital potentiometers

---

 Symbol: SENSIRION_SGP40 [=n]
 Type  : tristate
 Defined at drivers/iio/chemical/Kconfig:134
   Prompt: Sensirion SGP40 gas sensor
   Depends on: IIO [=m] && I2C [=y]
   Location:
     -> Device Drivers
       -> Industrial I/O support (IIO [=m])
         -> Chemical Sensors
 Selects: CRC8 [=m]

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>