[redhat] New configs in sound/soc
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: SND_SOC_AMD_ACP5x [=n]
 Type  : tristate
 Defined at sound/soc/amd/Kconfig:56
   Prompt: AMD Audio Coprocessor-v5.x I2S support
   Depends on: SOUND [=m] && !UML && SND [=m] && SND_SOC [=m] && X86 [=y] && PCI [=y]
   Location:
     -> Device Drivers
       -> Sound card support (SOUND [=m])
         -> Advanced Linux Sound Architecture (SND [=m])
           -> ALSA for SoC audio support (SND_SOC [=m])

---

 Symbol: SND_SOC_MT8195 [=n]
 Type  : tristate
 Defined at sound/soc/mediatek/Kconfig:188
   Prompt: ASoC support for Mediatek MT8195 chip
   Depends on: SOUND [=m] && !UML && SND [=m] && SND_SOC [=m]
   Location:
     -> Device Drivers
       -> Sound card support (SOUND [=m])
         -> Advanced Linux Sound Architecture (SND [=m])
           -> ALSA for SoC audio support (SND_SOC [=m])
 Selects: SND_SOC_MEDIATEK [=n]

---

Cc: Jaroslav Kysela <jkysela@redhat.com>
Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>