[redhat] New configs in drivers/platform
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: INTEL_SAR_INT1092 [=n]
 Type  : tristate
 Defined at drivers/platform/x86/intel/int1092/Kconfig:1
   Prompt: Intel Specific Absorption Rate Driver
   Depends on: X86 [=y] && X86_PLATFORM_DEVICES [=y] && X86_PLATFORM_DRIVERS_INTEL [=y] && ACPI [=y]
   Location:
     -> Device Drivers
       -> X86 Platform Specific Device Drivers (X86_PLATFORM_DEVICES [=y])
         -> Intel x86 Platform Specific Device Drivers (X86_PLATFORM_DRIVERS_INTEL [=y])

---

 Symbol: MERAKI_MX100 [=n]
 Type  : tristate
 Defined at drivers/platform/x86/Kconfig:284
   Prompt: Cisco Meraki MX100 Platform Driver
   Depends on: X86 [=y] && X86_PLATFORM_DEVICES [=y] && GPIOLIB [=y] && GPIO_ICH [=m] && LEDS_CLASS [=y]
   Location:
     -> Device Drivers
       -> X86 Platform Specific Device Drivers (X86_PLATFORM_DEVICES [=y])
 Selects: LEDS_GPIO [=n]

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>