Build kernel-doc for Fedora
We have had a few requests to bring back the kernel-doc package for
Fedora.  While I wanted to bring it into kernel-tools so that it was
built less frequently, it is already being maintained in the kernel spec
so let's just add noarch to Fedora so that it builds there too.

Signed-off-by: Justin M. Forbes <jforbes@fedoraproject.org>