redhat: bump RHEL_MAJOR and add the changelog file for it
kernel-ark does not track anymore work for RHEL 9/CentOS Stream 9. We look
forward to the next major. While "8.99" was chosen as the version for
the work preceding RHEL 9, bump the major so the considered version now is
"9.99". I'm intentionally not removing redhat/kernel.changelog-8.99 yet
in this commit, to avoid any conflicts while this isn't merged. When
this change goes in, I'll propose a new merge request for the removal of
this old changelog file.

Also I'm not including/moving all old changelog entries into the new file,
to avoid conflicts in the next os-build rebase. So when this is included/merged
the first time, no old entries before last release commit will be in the
changelog. However, all changelog entries will be picked up with the first
release commit in the next os-build rebase.

Signed-off-by: Herton R. Krzesinski <herton@redhat.com>