redhat: rename usage of .rhel8git.mk to .rhpkg.mk
We do not need to use the rhel version in the file name to look for the
RHDISTGIT_CACHE value. Usually this file is used by the maintainers for
configuring the location of the dist-git cache (the dist-git repo can
be big so this saves time in the dist-git sync operation). But having
the version in the name means bumping/changing the major each release,
we do not need to do it: if the maintainer has to maintain different
majors in the same stream, he will use the same dist-git repo/cache for
it. In the case different dist-git repos are needed or the dist-git
repos differ (are not synced between them), the .rhpkg.mk can have if
statements to single out the repos.

Signed-off-by: Herton R. Krzesinski <herton@redhat.com>