[redhat] New configs in arch/arm64
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: ARM64_MTE [=y]
 Type  : bool
 Defined at arch/arm64/Kconfig:1647
   Prompt: Memory Tagging Extension support
   Depends on: ARM64_AS_HAS_MTE [=y] && ARM64_TAGGED_ADDR_ABI [=y] && AS_HAS_ARMV8_5 [=y] && AS_HAS_LSE_ATOMICS [=y] && ARM64_PAN [=y]
   Location:
     -> Kernel Features
       -> ARMv8.5 architectural features
 Selects: ARCH_USES_HIGH_VMA_FLAGS [=y]

---

Cc: Mark Salter <msalter@redhat.com>
Cc: Mark Langsdorf <mlangsdo@redhat.com>
Cc: Jeremy Linton <jlinton@redhat.com>
Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>