Add kfence_test to mod-internal.list
MR 1422 enabled a number of kunit tests in the config including
CONFIG_KFENCE_KUNIT_TEST, but this test does not build in  RHEL configs
because KFENCE itself it turned off. As a result, the module for the
test was not added to mod-internal.list causing depmod to fail on Fedora
builds.  This simply adds the module to the internal list so that it
filters properly.

Signed-off-by: Justin M. Forbes <jforbes@fedoraproject.org>