[redhat] New configs in arch/x86
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: SCHED_CLUSTER [=y]
 Type  : bool
 Defined at arch/x86/Kconfig:1004
   Prompt: Cluster scheduler support
   Depends on: SMP [=y]
   Location:
     -> Processor type and features

---

 Symbol: STRICT_SIGALTSTACK_SIZE [=n]
 Type  : bool
 Defined at arch/x86/Kconfig:2404
   Prompt: Enforce strict size checking for sigaltstack
   Depends on: DYNAMIC_SIGFRAME [=y]
   Location:
     -> Processor type and features

---

Cc: David Arcari <darcari@redhat.com>
Cc: Prarit Bhargava <prarit@redhat.com>
Cc: Dean Nelson <dnelson@redhat.com>
Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>