[redhat] New configs in lib/Kconfig.debug
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: DEBUG_PREEMPT [=y]
 Type  : bool
 Defined at lib/Kconfig.debug:1217
   Prompt: Debug preemptible kernel
   Depends on: DEBUG_KERNEL [=y] && PREEMPTION [=y] && TRACE_IRQFLAGS_SUPPORT [=y]
   Location:
     -> Kernel hacking

---

 Symbol: MEMCPY_KUNIT_TEST [=m]
 Type  : tristate
 Defined at lib/Kconfig.debug:2456
   Prompt: Test memcpy(), memmove(), and memset() functions at runtime
   Depends on: RUNTIME_TESTING_MENU [=y] && KUNIT [=m]
   Visible if: RUNTIME_TESTING_MENU [=y] && KUNIT [=m] && !KUNIT_ALL_TESTS [=m]
   Location:
     -> Kernel hacking
       -> Kernel Testing and Coverage
         -> Runtime Testing (RUNTIME_TESTING_MENU [=y])

---

Cc: Prarit Bhargava <prarit@redhat.com>
Cc: Aristeu Rozanski <arozansk@redhat.com>
Cc: Christoph von Recklinghausen <crecklin@redhat.com>
Cc: Eugene Syromiatnikov <esyromia@redhat.com>
Cc: "Herton R. Krzesinski" <herton@redhat.com>
Cc: Nico Pache <npache@redhat.com>
Cc: Oleg Nesterov <onestero@redhat.com>
Cc: Rafael Aquini <aquini@redhat.com>
Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>