[redhat] New configs in drivers/gpu
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: DRM_PANEL_EDP [=n]
 Type  : tristate
 Defined at drivers/gpu/drm/panel/Kconfig:91
   Prompt: support for simple Embedded DisplayPort panels
   Depends on: HAS_IOMEM [=y] && DRM [=m] && DRM_PANEL [=y] && OF [=y] && BACKLIGHT_CLASS_DEVICE [=y] && PM [=y]
   Location:
     -> Device Drivers
       -> Graphics support
         -> Display Panels
 Selects: VIDEOMODE_HELPERS [=n] && DRM_DP_AUX_BUS [=n]

---

 Symbol: DRM_PANEL_SAMSUNG_S6D27A1 [=n]
 Type  : tristate
 Defined at drivers/gpu/drm/panel/Kconfig:408
   Prompt: Samsung S6D27A1 DPI panel driver
   Depends on: HAS_IOMEM [=y] && DRM [=m] && DRM_PANEL [=y] && OF [=y] && SPI [=y] && GPIOLIB [=y]
   Location:
     -> Device Drivers
       -> Graphics support
         -> Display Panels
 Selects: DRM_MIPI_DBI [=n]

---

Cc: David Airlie <airlied@redhat.com>
Cc: Adam Jackson <ajax@redhat.com>
Cc: Lyude Paul <lyude@redhat.com>
Cc: "Michel Dänzer" <mdaenzer@redhat.com>
Cc: "Jérôme Glisse" <jglisse@redhat.com>
Cc: Karol Herbst <kherbst@redhat.com>
Cc: Ben Skeggs <bskeggs@redhat.com>
Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>