[redhat] New configs in lib/Kconfig.debug
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: DEBUG_INFO_DWARF5 [=n]
 Type  : bool
 Defined at lib/Kconfig.debug:296
   Prompt: Generate DWARF Version 5 debuginfo
   Depends on: <choice> && (!CC_IS_CLANG [=n] || CC_IS_CLANG [=n] && (AS_IS_LLVM [=n] || AS_IS_GNU [=y] && AS_VERSION [=25000]>=23502)) && !DEBUG_INFO_BTF [=n]
   Location:
     Main menu
       -> Kernel hacking
         -> Compile-time checks and compiler options
           -> Compile the kernel with debug info (DEBUG_INFO [=y])
             -> DWARF version (<choice> [=y])

---

Cc: Prarit Bhargava <prarit@redhat.com>
Cc: Aristeu Rozanski <arozansk@redhat.com>
Cc: Christoph von Recklinghausen <crecklin@redhat.com>
Cc: Eugene Syromiatnikov <esyromia@redhat.com>
Cc: "Herton R. Krzesinski" <herton@redhat.com>
Cc: Nico Pache <npache@redhat.com>
Cc: Oleg Nesterov <onestero@redhat.com>
Cc: Rafael Aquini <aquini@redhat.com>
Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>

v2:
- CONFIG_DEBUG_INFO_DWARF_TOOLCHAIN_DEFAULT enabled in common
- CONFIG_DEBUG_INFO_DWARF4 disabled in common
- CONFIG_DEBUG_INFO_DWARF5 disabled in common

Signed-off-by: Patrick Talbert <ptalbert@redhat.com>