redhat: disable upstream check for rpminspect
Bugzilla: INTERNAL
Upstream Status: RHEL-only

The upstream test on rpminspect checks eg. if we changed the upstream sources
contents in the src.rpm when we do updates. Sources in the spec file are checked
for changes. It would make sense to have it checked if we had stable sources
which mostly never change, but that's not the case, eg. the kernel tarball
always changes each release usually. The test is mostly noise right now,
so disable it.

Signed-off-by: Herton R. Krzesinski <herton@redhat.com>