[redhat] New configs in drivers/cpufreq
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: X86_AMD_PSTATE [=n]
 Type  : tristate
 Defined at drivers/cpufreq/Kconfig.x86:37
   Prompt: AMD Processor P-State driver
   Depends on: CPU_FREQ [=y] && X86 [=y] && ACPI [=y]
   Location:
     Main menu
       -> Power management and ACPI options
         -> CPU Frequency scaling
           -> CPU Frequency scaling (CPU_FREQ [=y])
 Selects: ACPI_PROCESSOR [=y] && ACPI_CPPC_LIB [=y] && CPU_FREQ_GOV_SCHEDUTIL [=y]

---

Cc: Al Stone <ahs3@redhat.com>
Cc: Lenny Szubowicz <lszubowi@redhat.com>
Cc: Prarit Bhargava <prarit@redhat.com>
Cc: David Arcari <darcari@redhat.com>
Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>

v2:
 CONFIG_X86_AMD_PSTATE: enabled and moved under x86.

Signed-off-by: Patrick Talbert <ptalbert@redhat.com>