[redhat] New configs in drivers/net/wireless
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: IWLMEI [=n]
 Type  : tristate
 Defined at drivers/net/wireless/intel/iwlwifi/Kconfig:151
   Prompt: Intel Management Engine communication over WLAN
   Depends on: NETDEVICES [=y] && WLAN [=y] && WLAN_VENDOR_INTEL [=y] && INTEL_MEI [=m] && PM [=y]
   Location:
     Main menu
       -> Device Drivers
         -> Network device support (NETDEVICES [=y])
           -> Wireless LAN (WLAN [=y])
             -> Intel devices (WLAN_VENDOR_INTEL [=y])

---

Cc: "Íñigo Huguet" <ihuguet@redhat.com>
Cc: Jarod Wilson <jarod@redhat.com>
Cc: John Linville <linville@redhat.com>
Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>