[redhat] New configs in drivers/platform
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: ASUS_TF103C_DOCK [=n]
 Type  : tristate
 Defined at drivers/platform/x86/Kconfig:312
   Prompt: Asus TF103C 2-in-1 keyboard dock
   Depends on: X86 [=y] && X86_PLATFORM_DEVICES [=y] && ACPI [=y] && I2C [=y] && INPUT [=y] && HID [=y] && GPIOLIB [=y]
   Location:
     Main menu
       -> Device Drivers
         -> X86 Platform Specific Device Drivers (X86_PLATFORM_DEVICES [=y])

---

 Symbol: SIEMENS_SIMATIC_IPC [=n]
 Type  : tristate
 Defined at drivers/platform/x86/Kconfig:1129
   Prompt: Siemens Simatic IPC Class driver
   Depends on: X86 [=y] && X86_PLATFORM_DEVICES [=y] && PCI [=y]
   Location:
     Main menu
       -> Device Drivers
         -> X86 Platform Specific Device Drivers (X86_PLATFORM_DEVICES [=y])

---

 Symbol: YOGABOOK_WMI [=n]
 Type  : tristate
 Defined at drivers/platform/x86/Kconfig:130
   Prompt: Lenovo Yoga Book tablet WMI key driver
   Depends on: X86 [=y] && X86_PLATFORM_DEVICES [=y] && ACPI_WMI [=m] && INPUT [=y]
   Location:
     Main menu
       -> Device Drivers
         -> X86 Platform Specific Device Drivers (X86_PLATFORM_DEVICES [=y])
 Selects: LEDS_CLASS [=y] && NEW_LEDS [=y]

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>