[redhat] New configs in drivers/clk
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: COMMON_CLK_LAN966X [=n]
 Type  : bool
 Defined at drivers/clk/Kconfig:232
   Prompt: Generic Clock Controller driver for LAN966X SoC
   Depends on: COMMON_CLK [=y]
   Location:
     Main menu
       -> Device Drivers
         -> Common Clock Framework (COMMON_CLK [=y])

---

Cc: Mark Salter <msalter@redhat.com>
Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>