[redhat] New configs in drivers/platform
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: INTEL_VSEC [=n]
 Type  : tristate
 Defined at drivers/platform/x86/intel/Kconfig:174
   Prompt: Intel Vendor Specific Extended Capabilities Driver
   Depends on: X86 [=y] && X86_PLATFORM_DEVICES [=y] && PCI [=y]
   Location:
     Main menu
       -> Device Drivers
         -> X86 Platform Specific Device Drivers (X86_PLATFORM_DEVICES [=y])
 Selects: AUXILIARY_BUS [=y]

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>

v2:
 CONFIG_INTEL_VSEC moved under x86 and enabled as a module.

Signed-off-by: Patrick Talbert <ptalbert@redhat.com>