[redhat] New configs in kernel/trace
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: FTRACE_SORT_STARTUP_TEST [=n]
 Type  : bool
 Defined at kernel/trace/Kconfig:918
   Prompt: Verify compile time sorting of ftrace functions
   Depends on: FTRACE [=y] && DYNAMIC_FTRACE [=y] && BUILDTIME_TABLE_SORT [=y]
   Location:
     Main menu
       -> Kernel hacking
         -> Tracers (FTRACE [=y])

---

Cc: Jerome Marchand <jmarchan@redhat.com>
Cc: Jiri Olsa <jolsa@redhat.com>
Cc: Joe Lawrence <joe.lawrence@redhat.com>
Cc: Josh Poimboeuf <jpoimboe@redhat.com>
Cc: Luis Claudio Goncalves <lgoncalv@redhat.com>
Cc: Waiman Long <longman@redhat.com>
Cc: Yannick Cote <ycote@redhat.com>
Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>