[redhat] New configs in drivers/video
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: FRAMEBUFFER_CONSOLE_LEGACY_ACCELERATION [=n]
 Type  : bool
 Defined at drivers/video/console/Kconfig:81
   Prompt: Enable legacy fbcon hardware acceleration code
   Depends on: VT [=y] && FRAMEBUFFER_CONSOLE [=y]
   Location:
     Main menu
       -> Device Drivers
         -> Graphics support
           -> Console display driver support
             -> Framebuffer Console support (FRAMEBUFFER_CONSOLE [=y])

---

Cc: David Airlie <airlied@redhat.com>
Cc: Adam Jackson <ajax@redhat.com>
Cc: Lyude Paul <lyude@redhat.com>
Cc: "Michel Dänzer" <mdaenzer@redhat.com>
Cc: "Jérôme Glisse" <jglisse@redhat.com>
Cc: Karol Herbst <kherbst@redhat.com>
Cc: Ben Skeggs <bskeggs@redhat.com>
Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>