redhat/configs: Disable CONFIG_SENSORS_NCT6683 in RHEL for arm/aarch64
Doing a 'modprobe nct6683' results in a panic on some aarch64 systems.
This hwmon driver should not be enabled for aarch64. And since it is already
disabled for everything else in redhat/configs/ark/generic, the proper way
to do this is to simply remove the config option from arm/aarch64.

Bugzilla: https://bugzilla.redhat.com/show_bug.cgi?id=2041186
Signed-off-by: Dean Nelson <dnelson@redhat.com>