redhat: configs: Disable team driver
Bugzilla: https://bugzilla.redhat.com/show_bug.cgi?id=1945477
Upstream Status: RHEL-only

This drivers have been deprecated with RHEL9, so with RHEL10 it is time
to remove it. Disable the NET_TEAM on ARK. No need to set the modes as
they depend on team driver.

Keep the config enabled in Fedora as Fedora user may still need this driver.

Signed-off-by: Hangbin Liu <haliu@redhat.com>