redhat/Makefile: Rename KVERSION to SPECVERSION
KVERSION is passed into the spec file and is easy to confuse with the
kernel KVERSION.  Change the name to SPECVERSION so that readers
understand it is the %Version field of the specfile.

As a result RPMKSUBLEVEL can be dropped from genspec.sh.

Rename KVERSION to SPECVERSION and use it in the specfile.

Signed-off-by: Prarit Bhargava <prarit@redhat.com>