[redhat] New configs in arch/Kconfig
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: SHADOW_CALL_STACK [=n]
 Type  : bool
 Defined at arch/Kconfig:606
   Prompt: Shadow Call Stack
   Depends on: ARCH_SUPPORTS_SHADOW_CALL_STACK [=y] && (DYNAMIC_FTRACE_WITH_REGS [=y] || !FUNCTION_GRAPH_TRACER [=y])
   Location:
     Main menu
       -> General architecture-dependent options

---

Cc: Prarit Bhargava <prarit@redhat.com>
Cc: David Arcari <darcari@redhat.com>
Cc: Mark Salter <msalter@redhat.com>
Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>