New configs in drivers/dax
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: DAX [=m]
 Type  : tristate
 Defined at drivers/dax/Kconfig:2
   Prompt: DAX: direct access to differentiated memory
   Location:
     Main menu
       -> Device Drivers
 Selects: SRCU [=y]
 Selected by [m]:
   - BLK_DEV_PMEM [=m] && LIBNVDIMM [=m]
 Selected by [n]:
   - DCSSBLK [=n] && BLK_DEV [=y] && S390 && BLOCK [=y]
   - FS_DAX [=n] && MMU [=y] && !ARM && !MIPS && !SPARC && (ZONE_DEVICE [=n] || FS_DAX_LIMITED [=n])