[redhat] New configs in drivers/firmware
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: ARM_SCMI_TRANSPORT_SMC_ATOMIC_ENABLE [=n]
 Type  : bool
 Defined at drivers/firmware/arm_scmi/Kconfig:81
   Prompt: Enable atomic mode support for SCMI SMC transport
   Depends on: ARM_SCMI_PROTOCOL [=y] && ARM_SCMI_TRANSPORT_SMC [=y]
   Location:
     Main menu
       -> Device Drivers
         -> Firmware Drivers
           -> ARM System Control and Management Interface Protocol
             -> ARM System Control and Management Interface (SCMI) Message Protocol (ARM_SCMI_PROTOCOL [=y])
               -> SCMI transport based on SMC (ARM_SCMI_TRANSPORT_SMC [=y])

---

Cc: Lenny Szubowicz <lszubowi@redhat.com>
Cc: David Arcari <darcari@redhat.com>
Cc: Mark Langsdorf <mlangsdo@redhat.com>
Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>