[redhat] New configs in drivers/cxl
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: CXL_PCI [=y]
 Type  : tristate
 Defined at drivers/cxl/Kconfig:16
   Prompt: PCI manageability
   Depends on: CXL_BUS [=y]
   Location:
     Main menu
       -> Device Drivers
         -> PCI support (PCI [=y])
           -> CXL (Compute Express Link) Devices Support (CXL_BUS [=y])

---

Cc: Myron Stowe <mstowe@redhat.com>
Cc: Aristeu Rozanski <arozansk@redhat.com>
Cc: John Linville <linville@redhat.com>
Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>