[redhat] New configs in drivers/net/wireless
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: MT7921U [=n]
 Type  : tristate
 Defined at drivers/net/wireless/mediatek/mt76/mt7921/Kconfig:28
   Prompt: MediaTek MT7921U (USB) support
   Depends on: NETDEVICES [=y] && WLAN [=y] && WLAN_VENDOR_MEDIATEK [=y] && MAC80211 [=m] && USB [=y]
   Location:
     Main menu
       -> Device Drivers
         -> Network device support (NETDEVICES [=y])
           -> Wireless LAN (WLAN [=y])
             -> MediaTek devices (WLAN_VENDOR_MEDIATEK [=y])
 Selects: MT76_USB [=m] && MT7921_COMMON [=m]

---

Cc: "Íñigo Huguet" <ihuguet@redhat.com>
Cc: Jarod Wilson <jarod@redhat.com>
Cc: John Linville <linville@redhat.com>
Cc: Michal Schmidt <mschmidt@redhat.com>
Cc: Petr Oros <poros@redhat.com>
Cc: Ken Cox <jkc@redhat.com>
Cc: "José Ignacio Tornos Martínez" <jtornosm@redhat.com>
Cc: Kamal Heib <kheib@redhat.com>
Cc: Stefan Assmann <sassmann@redhat.com>
Cc: Corinna Vinschen <vinschen@redhat.com>
Cc: Ivan Vecera <ivecera@redhat.com>
Cc: Jonathan Toppins <jtoppins@redhat.com>
Cc: Josef Oskera <joskera@redhat.com>
Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>