[redhat] New configs in drivers/phy
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: PHY_FSL_LYNX_28G [=n]
 Type  : tristate
 Defined at drivers/phy/freescale/Kconfig:26
   Prompt: Freescale Layerscape Lynx 28G SerDes PHY support
   Depends on: OF [=y]
   Location:
     Main menu
       -> Device Drivers
         -> PHY Subsystem
 Selects: GENERIC_PHY [=n]

---

Cc: Petr Oros <poros@redhat.com>
Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>