[redhat] New configs in arch/s390
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: EXPOLINE_EXTERN [=n]
 Type  : bool
 Defined at arch/s390/Kconfig:538
   Prompt: Generate expolines as extern functions.
   Depends on: EXPOLINE [=y] && CC_IS_GCC [=y] && GCC_VERSION [=200000]>=110200 && y
   Location:
     Main menu
       -> Processor type and features
         -> Avoid speculative indirect branches in the kernel (EXPOLINE [=y])

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>

v2: CONFIG_EXPOLINE_EXTERN enable and move under generic/s390x/.

Signed-off-by: Patrick Talbert <ptalbert@redhat.com>