[redhat] New configs in drivers/mfd
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: MFD_MAX77714 [=n]
 Type  : tristate
 Defined at drivers/mfd/Kconfig:852
   Prompt: Maxim Semiconductor MAX77714 PMIC Support
   Depends on: HAS_IOMEM [=y] && I2C [=y] && (OF [=y] || COMPILE_TEST [=n])
   Location:
     Main menu
       -> Device Drivers
         -> Multifunction device drivers
 Selects: MFD_CORE [=m] && REGMAP_I2C [=m]

---

 Symbol: MFD_SIMPLE_MFD_I2C [=n]
 Type  : tristate
 Defined at drivers/mfd/Kconfig:1204
   Prompt: Simple Multi-Functional Device support (I2C)
   Depends on: HAS_IOMEM [=y] && I2C [=y]
   Location:
     Main menu
       -> Device Drivers
         -> Multifunction device drivers
 Selects: MFD_CORE [=y] && REGMAP_I2C [=m]
 Selected by [n]:
   - MFD_SL28CPLD [=n] && HAS_IOMEM [=y] && I2C [=y] && (ARCH_LAYERSCAPE || COMPILE_TEST [=n])

---

Cc: Torez Smith <torez@redhat.com>
Cc: David Arcari <darcari@redhat.com>
Cc: Tony Camuso <tcamuso@redhat.com>
Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>