[redhat] New configs in drivers/power
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: BATTERY_SAMSUNG_SDI [=n]
 Type  : bool
 Defined at drivers/power/supply/Kconfig:192
   Prompt: Samsung SDI batteries
   Depends on: POWER_SUPPLY [=y]
   Location:
     Main menu
       -> Device Drivers
         -> Power supply class support (POWER_SUPPLY [=y])

---

 Symbol: BATTERY_UG3105 [=n]
 Type  : tristate
 Defined at drivers/power/supply/Kconfig:885
   Prompt: uPI uG3105 battery monitor driver
   Depends on: POWER_SUPPLY [=y] && I2C [=y]
   Location:
     Main menu
       -> Device Drivers
         -> Power supply class support (POWER_SUPPLY [=y])

---

 Symbol: IP5XXX_POWER [=n]
 Type  : tristate
 Defined at drivers/power/supply/Kconfig:54
   Prompt: Injoinic IP5xxx power bank IC driver
   Depends on: POWER_SUPPLY [=y] && I2C [=y]
   Location:
     Main menu
       -> Device Drivers
         -> Power supply class support (POWER_SUPPLY [=y])
 Selects: REGMAP_I2C [=m]

---

Cc: Lenny Szubowicz <lszubowi@redhat.com>
Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>