redhat/configs: Set CONFIG_X86_AMD_PSTATE built-in on Fedora
In order to use the amd-pstate driver on AMD systems, users must first
unload the acpi-cpufreq driver and then load the amd-pstate driver.
Another option is to denylist the acpi-cpufreq driver but this also
requires user modifications to the system.

In ARK we have decided to build the amd-pstate driver into the kernel so
that it loads before the acpi-cpufreq driver.  Do the same for Fedora.

Set CONFIG_X86_AMD_PSTATE to 'y' on Fedora.

Suggested-by: Joel Wirāmu Pauling <jwp@redhat.com>
Signed-off-by: Prarit Bhargava <prarit@redhat.com>