[redhat] New configs in arch/x86
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: X86_KERNEL_IBT [=n]
 Type  : bool
 Defined at arch/x86/Kconfig:1855
   Prompt: Indirect Branch Tracking
   Depends on: X86_64 [=y] && CC_HAS_IBT [=y] && STACK_VALIDATION [=y] && (!LD_IS_LLD [=n] || LLD_VERSION [=0]>=140000)
   Location:
     Main menu
       -> Processor type and features

---

 Symbol: X86_X32_ABI [=n]
 Type  : bool
 Defined at arch/x86/Kconfig:2848
   Prompt: x32 ABI for 64-bit mode
   Depends on: X86_64 [=y] && y
   Location:
     Main menu
       -> Binary Emulations

---

Cc: David Arcari <darcari@redhat.com>
Cc: Prarit Bhargava <prarit@redhat.com>
Cc: Dean Nelson <dnelson@redhat.com>
Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>