[redhat] New configs in drivers/misc
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: OPEN_DICE [=n]
 Type  : tristate
 Defined at drivers/misc/Kconfig:474
   Prompt: Open Profile for DICE driver
   Depends on: OF_RESERVED_MEM [=y]
   Location:
     Main menu
       -> Device Drivers
         -> Misc devices

---

Cc: Doug Ledford <dledford@redhat.com>
Cc: Prarit Bhargava <prarit@redhat.com>
Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>