[redhat] New configs in drivers/clk
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: CLK_GATE_KUNIT_TEST [=m]
 Type  : tristate
 Defined at drivers/clk/Kconfig:459
   Prompt: Basic gate type Kunit test
   Depends on: COMMON_CLK [=y] && KUNIT [=m]
   Visible if: COMMON_CLK [=y] && KUNIT [=m] && !KUNIT_ALL_TESTS [=m]
   Location:
     Main menu
       -> Device Drivers
         -> Common Clock Framework (COMMON_CLK [=y])

---

 Symbol: CLK_KUNIT_TEST [=m]
 Type  : tristate
 Defined at drivers/clk/Kconfig:452
   Prompt: Basic Clock Framework Kunit Tests
   Depends on: COMMON_CLK [=y] && KUNIT [=m]
   Visible if: COMMON_CLK [=y] && KUNIT [=m] && !KUNIT_ALL_TESTS [=m]
   Location:
     Main menu
       -> Device Drivers
         -> Common Clock Framework (COMMON_CLK [=y])

---

 Symbol: COMMON_CLK_RS9_PCIE [=n]
 Type  : tristate
 Defined at drivers/clk/Kconfig:353
   Prompt: Clock driver for Renesas 9-series PCIe clock generators
   Depends on: COMMON_CLK [=y] && I2C [=y] && OF [=y]
   Location:
     Main menu
       -> Device Drivers
         -> Common Clock Framework (COMMON_CLK [=y])
 Selects: REGMAP_I2C [=m]

---

Cc: Mark Salter <msalter@redhat.com>
Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>