Enabled INTEGRITY_MACHINE_KEYRING for all configs.
There's a new option to add a Linux kernel keyring called machine
to enable a MOK (Machine Operator Key) to be enrolled for use in the
kernel. This option allows users to enroll/trust a key for things
like self (or third party) signed kernel modules.

Further details in the following thread:
https://lkml.org/lkml/2022/1/25/1422

Note it requires shim 15.5 and a newer version of mokutil.

Signed-off-by: Peter Robinson <pbrobinson@redhat.com>