redhat: indicate HEAD state in tarball/rpm name
For the non-Fedora (RHEL) case, we are looking to move to a setup where we
include Merge Requests as soon as they are ready, and thus the head of
main may not always be a tagged build. This gives some indication to
someone basing a build on main that they're basing their work on the tree
at a state between tagged builds.

If HEAD is a tagged build, as usual, you should see:
  linux-5.18.0-0.rc3.30.test.tar.xz
  kernel-5.18.0-0.rc3.30.test.src.rpm

If HEAD has additional commits, with this change, you should see:
  linux-5.18.0-0.rc3.30.g0f62.test.tar.xz
  kernel-5.18.0-0.rc3.30.g0f62.test.src.rpm

The .gSHA field is the first four chars of the HEAD sha of the branch after
adding your local changes.

This is directly related to:
  https://gitlab.com/redhat/rhel/src/kernel/rhel-8/-/merge_requests/2405

Signed-off-by: Jarod Wilson <jarod@redhat.com>