Add system-sb-certs for RHEL-9
Copy the RHEL-9 signing requirement over to ARK.  This change is wrapped
with "{?rhel}{?centos} && !{?eln}" to allow Fedora and ELN to continue
using their signing technology.

The system-sb-certs is just a metapackage for either rhel-sb-certs or
centos-sb-certs, depending on what is installed on your system.

Tested by building under koji eln successfully.

Signed-off-by: Don Zickus <dzickus@redhat.com>