Revert variable switch for lasttag
MR 1757 cleaned up a lot of Makefile variables. In this case though, it
has us searching for a full tag with a release, after it has been
bumped.  As this will never match, the script assumes that this is the
first release and appends the entire history to the changelog every
time.  This change reverts that specific change, and cleans up several
thousand lines of changelog mess that it caused.

Signed-off-by: Justin M. Forbes <jforbes@fedoraproject.org>