[redhat] New configs in drivers/firmware
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: EFI_COCO_SECRET [=n]
 Type  : bool
 Defined at drivers/firmware/efi/Kconfig:315
   Prompt: EFI Confidential Computing Secret Area Support
   Depends on: EFI [=y]
   Location:
     Main menu
       -> Device Drivers
         -> Firmware Drivers
 Selected by [n]:
   - EFI_SECRET [=n] && VIRT_DRIVERS [=y] && EFI [=y] && X86_64 [=y]

---

 Symbol: EFI_DISABLE_RUNTIME [=n]
 Type  : bool
 Defined at drivers/firmware/efi/Kconfig:300
   Prompt: Disable EFI runtime services support by default
   Location:
     Main menu
       -> Device Drivers
         -> Firmware Drivers

---

Cc: Lenny Szubowicz <lszubowi@redhat.com>
Cc: David Arcari <darcari@redhat.com>
Cc: Mark Langsdorf <mlangsdo@redhat.com>
Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>

v2:
- CONFIG_EFI_COCO_SECRET enabled for ark.

Signed-off-by: Patrick Talbert <ptalbert@redhat.com>