[redhat] New configs in drivers/platform
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: INTEL_IFS [=n]
 Type  : tristate
 Defined at drivers/platform/x86/intel/ifs/Kconfig:1
   Prompt: Intel In Field Scan
   Depends on: X86_PLATFORM_DEVICES [=y] && X86 [=y] && CPU_SUP_INTEL [=y] && 64BIT [=y] && SMP [=y]
   Location:
     Main menu
       -> Device Drivers
         -> X86 Platform Specific Device Drivers (X86_PLATFORM_DEVICES [=y])
 Selects: INTEL_IFS_DEVICE

---

 Symbol: NVSW_SN2201 [=n]
 Type  : tristate
 Defined at drivers/platform/mellanox/Kconfig:81
   Prompt: Nvidia SN2201 platform driver support
   Depends on: MELLANOX_PLATFORM [=y] && REGMAP [=y] && HWMON [=y] && I2C [=y] && REGMAP_I2C [=m]
   Location:
     Main menu
       -> Device Drivers
         -> Platform support for Mellanox hardware (MELLANOX_PLATFORM [=y])

---

 Symbol: WINMATE_FM07_KEYS [=n]
 Type  : tristate
 Defined at drivers/platform/x86/Kconfig:1155
   Prompt: Winmate FM07/FM07P front-panel keys driver
   Depends on: X86 [=y] && X86_PLATFORM_DEVICES [=y] && INPUT [=y]
   Location:
     Main menu
       -> Device Drivers
         -> X86 Platform Specific Device Drivers (X86_PLATFORM_DEVICES [=y])

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>