[redhat] New configs in drivers/net/wwan
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: MTK_T7XX [=n]
 Type  : tristate
 Defined at drivers/net/wwan/Kconfig:108
   Prompt: MediaTek PCIe 5G WWAN modem T7xx device
   Depends on: NETDEVICES [=y] && WWAN [=y] && PCI [=y]
   Location:
     Main menu
       -> Device Drivers
         -> Network device support (NETDEVICES [=y])
           -> Wireless WAN
             -> WWAN Driver Core (WWAN [=y])

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>