[redhat] New configs in drivers/bus
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: QCOM_SSC_BLOCK_BUS [=n]
 Type  : bool
 Defined at drivers/bus/Kconfig:155
   Prompt: Qualcomm SSC Block Bus Init Driver
   Depends on: ARCH_QCOM [=y]
   Location:
     Main menu
       -> Device Drivers
         -> Bus devices

---

Cc: Myron Stowe <mstowe@redhat.com>
Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>