Fix up ZSMALLOC config for s390
Due to some dependency changes upstream, ZSMALLOC was causing a config
mismatch on s390.  It is set to =y in common for all arches, but the
only listed consumer used to be ZSWAP, which was disabled on s390. Now,
it is selected by ZRAM which is built as a module everywhere except for
s390 zfcpdump, where modules are not supported, so it is built in.  As
ZSWAP is not enabled on s390 at all, it seemed the best course to fix
the mismatch was to make ZSMALLOC match the config for ZRAM on s390.

Signed-off-by: Justin M. Forbes <jforbes@fedoraproject.org>