[redhat] New configs in arch/x86
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: MICROCODE_LATE_LOADING [=n]
 Type  : bool
 Defined at arch/x86/Kconfig:1361
   Prompt: Late microcode loading (DANGEROUS)
   Depends on: MICROCODE [=y]
   Location:
     Main menu
       -> Processor type and features
         -> CPU microcode loading support (MICROCODE [=y])

---

Cc: David Arcari <darcari@redhat.com>
Cc: Prarit Bhargava <prarit@redhat.com>
Cc: Dean Nelson <dnelson@redhat.com>
Cc: Lenny Szubowicz <lszubowi@redhat.com>
Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>