Move common/generic/s390x/zfcpdump/ configs to ark/generic/s390x/zfcpdump/
As Justin mentions, Fedora does not build zfcpdump flavor. common/ is for
things which are identical between Fedora and RHEL/ARK. The ark/ directory
is for things that are only set that way in RHEL/ARK. For the first year
or so of the new kernel workflow, all new RHEL zfcpdump entries were created
in common/ instead of ark/. That only has changed a few months ago. So fix
this now.

Signed-off-by: Vladis Dronov <vdronov@redhat.com>