[redhat] New configs in drivers/xen
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: XEN_VIRTIO [=n]
 Type  : bool
 Defined at drivers/xen/Kconfig:346
   Prompt: Xen virtio support
   Depends on: XEN [=y] && VIRTIO [=y]
   Location:
     Main menu
       -> Device Drivers
         -> Xen driver support
 Selects: XEN_GRANT_DMA_OPS [=n] && XEN_GRANT_DMA_IOMMU [=n]

---

Cc: Vitaly Kuznetsov <vkuznets@redhat.com>
Cc: Igor Mammedov <imammedo@redhat.com>
Cc: Andrew Jones <drjones@redhat.com>
Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>