[redhat] New configs in drivers/virtio
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: VIRTIO_HARDEN_NOTIFICATION [=n]
 Type  : bool
 Defined at drivers/virtio/Kconfig:32
   Prompt: Harden virtio notification
   Depends on: VIRTIO_MENU [=y]
   Location:
     Main menu
       -> Device Drivers
         -> Virtio drivers (VIRTIO_MENU [=y])

---

Cc: "Michael S. Tsirkin" <mst@redhat.com>
Cc: Jason Wang <jasowang@redhat.com>
Cc: rhvirt-patches@redhat.com
Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>