redhat: make kernel-zfcpdump-core to not provide kernel-core/kernel
Upstream Status: RHEL only

As reported on bug 2027654*, when doing dnf install kernel on s390x it
might pull kernel-zfcpdump-core which should not be a choice for a
generic kernel. Adjust the kernel_variant_package macro to add an
option allowing adding provides or not and use it for the zfcpdump
variant.

v2: - fix if condition to not hide kernel-<flavour>-core-uname-r which is
      required by the kernel-<flavour> meta package.
    - wrap kernel provides in kernel_reqprovconf macro as well
    - rename the macro option from -p to -o as suggested by Daniel Horak
v3: - fix kernel_reqprovconf expansion as noted by Bruno Goncalves/Jan Stancek
v4: - do not wrap installonlypkg since that should be only a hint to
      yum/dnf to allow multiple versions of same packages installed

This same change was applied to centos stream 9, and is being forwarded
ported now to ARK/Fedora, with a fix for the -m option of
kernel_variant_package, it doesn't have an argument so ':' can't
be added after it, for some reason rpm in centos didn't have issues with
it, but on ARK/Fedora we get a failure.

*https://bugzilla.redhat.com/show_bug.cgi?id=2027654

Signed-off-by: Herton R. Krzesinski <herton@redhat.com>