[redhat] New configs in crypto/Kconfig
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: CRYPTO_ARIA [=n]
 Type  : tristate
 Defined at crypto/Kconfig:1647
   Prompt: ARIA cipher algorithm
   Depends on: CRYPTO [=y]
   Location:
     Main menu
       -> Cryptographic API (CRYPTO [=y])
 Selects: CRYPTO_ALGAPI [=y]

---

 Symbol: CRYPTO_HCTR2 [=n]
 Type  : tristate
 Defined at crypto/Kconfig:557
   Prompt: HCTR2 support
   Depends on: CRYPTO [=y]
   Location:
     Main menu
       -> Cryptographic API (CRYPTO [=y])
 Selects: CRYPTO_XCTR [=n] && CRYPTO_POLYVAL [=n] && CRYPTO_MANAGER [=y]

---

 Symbol: CRYPTO_POLYVAL_CLMUL_NI [=n]
 Type  : tristate
 Defined at crypto/Kconfig:811
   Prompt: POLYVAL hash function (CLMUL-NI accelerated)
   Depends on: CRYPTO [=y] && X86 [=y] && 64BIT [=y]
   Location:
     Main menu
       -> Cryptographic API (CRYPTO [=y])
 Selects: CRYPTO_POLYVAL [=n]

---

Cc: Herbert Xu <herbert.xu@redhat.com>
Cc: "David S. Miller" <davem@redhat.com>
Cc: Vladis Dronov <vdronov@redhat.com>
Cc: Bruno Meneguele <bmeneg@redhat.com>
Cc: Ondrej Mosnacek <omosnace@redhat.com>
Cc: Phil Auld <pauld@redhat.com>
Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>