[redhat] New configs in drivers/clocksource
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: TEGRA186_TIMER [=n]
 Type  : bool
 Defined at drivers/clocksource/Kconfig:160
   Prompt: NVIDIA Tegra186 timer driver
   Depends on: GENERIC_CLOCKEVENTS [=y] && (ARCH_TEGRA [=y] || COMPILE_TEST [=n]) && WATCHDOG [=y] && WATCHDOG_CORE [=y]
   Location:
     Main menu
       -> Device Drivers
         -> Clock Source drivers

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>