[redhat] New configs in drivers/perf
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: HNS3_PMU [=n]
 Type  : tristate
 Defined at drivers/perf/hisilicon/Kconfig:18
   Prompt: HNS3 PERF PMU
   Depends on: PERF_EVENTS [=y] && (ARM64 [=y] || COMPILE_TEST [=n]) && PCI [=y]
   Location:
     Main menu
       -> Device Drivers
         -> Performance monitor support

---

Cc: Mark Salter <msalter@redhat.com>
Cc: Mark Langsdorf <mlangsdo@redhat.com>
Cc: Jeremy Linton <jlinton@redhat.com>
Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>