[redhat] New configs in kernel/time
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: CONTEXT_TRACKING_USER_FORCE [=n]
 Type  : bool
 Defined at kernel/time/Kconfig:158
   Prompt: Force user context tracking
   Depends on: GENERIC_CLOCKEVENTS [=y] && CONTEXT_TRACKING_USER [=y]
   Location:
     Main menu
       -> General setup
         -> Timers subsystem

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>

v2:
- CONFIG_CONTEXT_TRACKING_FORCE renamed to
  CONFIG_CONTEXT_TRACKING_USER_FORCE.

v3:
- CONFIG_CONTEXT_TRACKING_USER_FORCE fix file content.

Signed-off-by: Patrick Talbert <ptalbert@redhat.com>