[redhat] New configs in drivers/hid
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: I2C_HID_OF_ELAN [=n]
 Type  : tristate
 Defined at drivers/hid/i2c-hid/Kconfig:35
   Prompt: Driver for Elan hid-i2c based devices on OF systems
   Depends on: I2C [=y] && INPUT [=y] && OF [=y]
   Location:
     Main menu
       -> Device Drivers
         -> HID support
           -> I2C HID support

---

Cc: Benjamin Tissoires <benjamin.tissoires@redhat.com>
Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>