[redhat] New configs in drivers/infiniband
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: INFINIBAND_ERDMA [=n]
 Type  : tristate
 Defined at drivers/infiniband/hw/erdma/Kconfig:2
   Prompt: Alibaba Elastic RDMA Adapter (ERDMA) support
   Depends on: INFINIBAND [=m] && (INFINIBAND_USER_ACCESS [=m] || !INFINIBAND_USER_ACCESS [=m]) && PCI_MSI [=y] && 64BIT [=y] && INFINIBAND_ADDR_TRANS [=y] && INFINIBAND_USER_ACCESS [=m]
   Location:
     Main menu
       -> Device Drivers
         -> InfiniBand support (INFINIBAND [=m])

---

Cc: Michal Schmidt <mschmidt@redhat.com>
Cc: Petr Oros <poros@redhat.com>
Cc: Ken Cox <jkc@redhat.com>
Cc: "José Ignacio Tornos Martínez" <jtornosm@redhat.com>
Cc: Kamal Heib <kheib@redhat.com>
Cc: Stefan Assmann <sassmann@redhat.com>
Cc: Corinna Vinschen <vinschen@redhat.com>
Cc: Ivan Vecera <ivecera@redhat.com>
Cc: "Íñigo Huguet" <ihuguet@redhat.com>
Cc: Jonathan Toppins <jtoppins@redhat.com>
Cc: Josef Oskera <joskera@redhat.com>
Cc: Izabela Bakollari <ibakolla@redhat.com>
Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>