[redhat] New configs in kernel/power
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: PM_USERSPACE_AUTOSLEEP [=n]
 Type  : bool
 Defined at kernel/power/Kconfig:146
   Prompt: Userspace opportunistic sleep
   Depends on: PM_SLEEP [=y]
   Location:
     Main menu
       -> Power management and ACPI options

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>