[redhat] New configs in drivers/char
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: TCG_TIS_I2C [=n]
 Type  : tristate
 Defined at drivers/char/tpm/Kconfig:77
   Prompt: TPM Interface Specification 1.3 Interface / TPM 2.0 FIFO Interface - (I2C - generic)
   Depends on: TCG_TPM [=y] && I2C [=y]
   Location:
     Main menu
       -> Device Drivers
         -> Character devices
           -> TPM Hardware Support (TCG_TPM [=y])
 Selects: CRC_CCITT [=y] && TCG_TIS_CORE [=y]

---

Cc: Doug Ledford <dledford@redhat.com>
Cc: Prarit Bhargava <prarit@redhat.com>
Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>

v2:
- CONFIG_TCG_TIS_I2C enabled.
- CONFIG_TCG_TIS_SPI enabled.

v3:
- CONFIG_TCG_TIS_SPI_CR50 disabled.
- CONFIG_TCG_TIS_I2C enabled as a module for s390x.

Signed-off-by: Patrick Talbert <ptalbert@redhat.com>