[redhat] New configs in drivers/iommu
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: ARM_SMMU_QCOM_DEBUG [=n]
 Type  : bool
 Defined at drivers/iommu/Kconfig:366
   Prompt: ARM SMMU QCOM implementation defined debug support
   Depends on: IOMMU_SUPPORT [=y] && ARM_SMMU_QCOM [=y]
   Location:
     Main menu
       -> Device Drivers
         -> IOMMU Hardware Support (IOMMU_SUPPORT [=y])
           -> ARM Ltd. System MMU (SMMU) Support (ARM_SMMU [=y])

---

Cc: Jerry Snitselaar <jsnitsel@redhat.com>
Cc: Myron Stowe <myron.stowe@redhat.com>
Cc: Alex Williamson <alex.williamson@redhat.com>
Cc: Don Dutile <ddutile@redhat.com>
Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>