[redhat] New configs in drivers/power
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: POWER_MLXBF [=n]
 Type  : tristate
 Defined at drivers/power/reset/Kconfig:300
   Prompt: Mellanox BlueField power handling driver
   Depends on: POWER_RESET [=y] && GPIO_MLXBF2 [=m] && ACPI [=y]
   Location:
     Main menu
       -> Device Drivers
         -> Board level reset or power off (POWER_RESET [=y])

---

Cc: Lenny Szubowicz <lszubowi@redhat.com>
Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>