[redhat] New configs in drivers/rtc
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: RTC_DRV_NCT3018Y [=n]
 Type  : tristate
 Defined at drivers/rtc/Kconfig:386
   Prompt: Nuvoton NCT3018Y
   Depends on: RTC_CLASS [=y] && I2C [=y] && OF [=y]
   Location:
     Main menu
       -> Device Drivers
         -> Real Time Clock (RTC_CLASS [=y])

---

Cc: Waiman Long <longman@redhat.com>
Cc: Prarit Bhargava <prarit@redhat.com>
Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>