[redhat] New configs in arch/arm64
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: ARM64_ERRATUM_2457168 [=y]
 Type  : bool
 Defined at arch/arm64/Kconfig:920
   Prompt: Cortex-A510: 2457168: workaround for AMEVCNTR01 incrementing incorrectly
   Depends on: ARM64_AMU_EXTN [=y]
   Location:
     Main menu
       -> Kernel Features
         -> ARM errata workarounds via the alternatives framework

---

Cc: Mark Salter <msalter@redhat.com>
Cc: Mark Langsdorf <mlangsdo@redhat.com>
Cc: Jeremy Linton <jlinton@redhat.com>
Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>