redhat: kernel.spec: selftests: abort on build failure
When a selftest does not compile, abort the whole build. We want to know
that we broke selftests by a backport.

This commit is forward ported from RHEL 8. It could not have been forward
ported to ARK: with rebases to the upstream the bpf selftests break all the
time. Now it's the first opportunity when we can take it.

Signed-off-by: Jiri Benc <jbenc@redhat.com>
Signed-off-by: Prarit Bhargava <prarit@redhat.com>