redhat/Makefile: add dist-assert-tree-clean target
Add dist-assert-tree-clean target that fails whenever staged/unstaged
changes or untracked files are present in the git tree.

Manual invocation of this target is not expected, thus left out of
dist-help. It's intended to act as a guard against any temporary changes
prior to running tooling that could change or revert them.

No checks are performed whenever $FORCE is non-empty.

Signed-off-by: Čestmír Kalina <ckalina@redhat.com>
Signed-off-by: Prarit Bhargava <prarit@redhat.com>