kabi: add stablelist helpers
Add helper scripts to update the kabi stablelist checksums and to diff
symtypes output.

redhat/kabi/diff-kabi computes the current symtypes files and compares
it to the reference symtypes stored in the stablelist.

redhat/kabi/update-kabi computes the current checksums and reference
symtypes files and updates the stablelist contents. It does not stage
any changes.

redhat/kabi/symtypes computes symtypes image, preimage or a diff for
a symtypes node.

Signed-off-by: Čestmír Kalina <ckalina@redhat.com>
Signed-off-by: Prarit Bhargava <prarit@redhat.com>