[redhat] New configs in drivers/i2c
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: I2C_PCI1XXXX [=n]
 Type  : tristate
 Defined at drivers/i2c/busses/Kconfig:1270
   Prompt: PCI1XXXX I2C Host Adapter
   Depends on: I2C [=y] && HAS_IOMEM [=y] && PCI [=y]
   Location:
     Main menu
       -> Device Drivers
         -> I2C support
           -> I2C support (I2C [=y])
             -> I2C Hardware Bus support

---

Cc: Gopal Tiwari <gtiwari@redhat.com>
Cc: David Arcari <darcari@redhat.com>
Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>