[redhat] New configs in drivers/spi
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: SPI_MICROCHIP_CORE_QSPI [=n]
 Type  : tristate
 Defined at drivers/spi/Kconfig:594
   Prompt: Microchip FPGA QSPI controllers
   Depends on: SPI [=y] && SPI_MASTER [=y]
   Location:
     Main menu
       -> Device Drivers
         -> SPI support (SPI [=y])

---

Cc: Al Stone <ahs3@redhat.com>
Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>