[redhat] New configs in drivers/platform
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: AMD_PMF [=n]
 Type  : tristate
 Defined at drivers/platform/x86/amd/pmf/Kconfig:6
   Prompt: AMD Platform Management Framework
   Depends on: X86_PLATFORM_DEVICES [=y] && ACPI [=y] && PCI [=y]
   Location:
     Main menu
       -> Device Drivers
         -> X86 Platform Specific Device Drivers (X86_PLATFORM_DEVICES [=y])
 Selects: ACPI_PLATFORM_PROFILE [=m]

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>