redhat/configs: Disable fbdev drivers and use simpledrm everywhere
Bugzilla: https://bugzilla.redhat.com/show_bug.cgi?id=1986223

Replace the fbdev drivers with the latest simpledrm driver and the DRM
fbdev emulation layer.

This change has been in Fedora for some time but it is something that's
also desirable for ARK. Change the config options in common so they are
shared across all flavors.

The only exception is s390x since that arch doesn't have a firmware that
is able to setup a framebuffer to be used by generic graphics drivers to
scanout. But let's keep DRM built as a module since that could be useful.

Signed-off-by: Javier Martinez Canillas <javierm@redhat.com>