[redhat] New configs in arch/arm64
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: ARCH_BCM [=n]
 Type  : bool
 Defined at arch/arm64/Kconfig.platforms:36
   Prompt: Broadcom SoC Support
   Location:
     Main menu
       -> Platform selection

---

 Symbol: ARCH_NXP [=n]
 Type  : bool
 Defined at arch/arm64/Kconfig.platforms:196
   Prompt: NXP SoC support
   Location:
     Main menu
       -> Platform selection

---

 Symbol: ARM64_ERRATUM_2658417 [=y]
 Type  : bool
 Defined at arch/arm64/Kconfig:738
   Prompt: Cortex-A510: 2658417: remove BF16 support due to incorrect result
   Location:
     Main menu
       -> Kernel Features
         -> ARM errata workarounds via the alternatives framework

---

Cc: Mark Salter <msalter@redhat.com>
Cc: Mark Langsdorf <mlangsdo@redhat.com>
Cc: Jeremy Linton <jlinton@redhat.com>
Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>