[redhat] New configs in lib/kunit
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: KUNIT_DEFAULT_ENABLED [=y]
 Type  : bool
 Defined at lib/kunit/Kconfig:62
   Prompt: Default value of kunit.enable
   Depends on: KUNIT [=m]
   Location:
     Main menu
       -> Kernel hacking
         -> Kernel Testing and Coverage
           -> KUnit - Enable support for unit tests (KUNIT [=m])

---

Cc: Prarit Bhargava <prarit@redhat.com>
Cc: Aristeu Rozanski <arozansk@redhat.com>
Cc: Christoph von Recklinghausen <crecklin@redhat.com>
Cc: Eugene Syromiatnikov <esyromia@redhat.com>
Cc: "Herton R. Krzesinski" <herton@redhat.com>
Cc: Nico Pache <npache@redhat.com>
Cc: Oleg Nesterov <onestero@redhat.com>
Cc: Rafael Aquini <aquini@redhat.com>
Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>