[redhat] New configs in drivers/ata
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: AHCI_DWC [=n]
 Type  : tristate
 Defined at drivers/ata/Kconfig:179
   Prompt: Synopsys DWC AHCI SATA support
   Depends on: ATA [=m] && HAS_DMA [=y]
   Location:
     Main menu
       -> Device Drivers
         -> Serial ATA and Parallel ATA drivers (libata) (ATA [=m])
 Selects: SATA_HOST [=y] && MFD_SYSCON [=n]

---

 Symbol: PATA_OF_PLATFORM [=n]
 Type  : tristate
 Defined at drivers/ata/Kconfig:1122
   Prompt: OpenFirmware platform device PATA support
   Depends on: ATA [=y] && ATA_SFF [=y] && OF [=y]
   Location:
     Main menu
       -> Device Drivers
         -> Serial ATA and Parallel ATA drivers (libata) (ATA [=y])
           -> ATA SFF support (for legacy IDE and PATA) (ATA_SFF [=y])
 Selects: PATA_PLATFORM [=n]

---

Cc: Tomas Henzl <thenzl@redhat.com>
Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>