[redhat] New configs in drivers/media
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: VIDEO_ZORAN [=n]
 Type  : tristate
 Defined at drivers/media/pci/zoran/Kconfig:1
   Prompt: Zoran ZR36057/36067 Video For Linux (Deprecated)
   Depends on: MEDIA_SUPPORT [=m] && MEDIA_PCI_SUPPORT [=y] && MEDIA_CAMERA_SUPPORT [=y] && PCI [=y] && I2C_ALGOBIT [=m] && VIDEO_DEV [=m] && !ALPHA && DEBUG_FS [=y]
   Location:
     Main menu
       -> Device Drivers
         -> Multimedia support (MEDIA_SUPPORT [=m])
           -> Media drivers
             -> Media PCI Adapters (MEDIA_PCI_SUPPORT [=y])
 Selects: VIDEOBUF2_DMA_CONTIG [=n] && VIDEO_ADV7170 [=n] && VIDEO_ADV7175 [=n] && VIDEO_BT819 [=n] && VIDEO_BT856 [=n] && VIDEO_BT866 [=n] && VIDEO_KS0127 [=n] && VIDEO_SAA711X [=n] && VIDEO_SAA7110 [=n] && VIDEO_SAA7185 [=n] && VIDEO_VPX3220 [=n]

---

Cc: Dean Nelson <dnelson@redhat.com>
Cc: Jarod Wilson <jarod@redhat.com>
Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>