[redhat] New configs in drivers/clk
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: COMMON_CLK_VC7 [=n]
 Type  : tristate
 Defined at drivers/clk/Kconfig:380
   Prompt: Clock driver for Renesas Versaclock 7 devices
   Depends on: COMMON_CLK [=y] && I2C [=y] && OF [=y]
   Location:
     Main menu
       -> Device Drivers
         -> Common Clock Framework (COMMON_CLK [=y])
 Selects: REGMAP_I2C [=m]

---

 Symbol: COMMON_CLK_XLNX_CLKWZRD [=n]
 Type  : tristate
 Defined at drivers/clk/xilinx/Kconfig:20
   Prompt: Xilinx Clocking Wizard
   Depends on: COMMON_CLK [=y] && OF [=y] && HAS_IOMEM [=y]
   Location:
     Main menu
       -> Device Drivers
         -> Common Clock Framework (COMMON_CLK [=y])

---

Cc: Mark Salter <msalter@redhat.com>
Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>