[redhat] New configs in drivers/nvmem
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: NVMEM_QCOM_QFPROM [=n]
 Type  : tristate
 Defined at drivers/nvmem/Kconfig:204
   Prompt: QCOM QFPROM Support
   Depends on: NVMEM [=y] && (ARCH_QCOM [=y] || COMPILE_TEST [=n]) && HAS_IOMEM [=y]
   Location:
     Main menu
       -> Device Drivers
         -> NVMEM Support (NVMEM [=y])

---

 Symbol: NVMEM_U_BOOT_ENV [=n]
 Type  : tristate
 Defined at drivers/nvmem/Kconfig:325
   Prompt: U-Boot environment variables support
   Depends on: NVMEM [=y] && OF [=y] && MTD [=m]
   Location:
     Main menu
       -> Device Drivers
         -> NVMEM Support (NVMEM [=y])
 Selects: CRC32 [=y]

---

Cc: Mark Salter <msalter@redhat.com>
Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>