[redhat] New configs in arch/powerpc
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: POWER10_CPU [=n]
 Type  : bool
 Defined at arch/powerpc/platforms/Kconfig.cputype:175
   Prompt: POWER10
   Depends on: <choice> && PPC_BOOK3S_64 [=y]
   Location:
     Main menu
       -> Processor support
         -> CPU selection (<choice> [=y])
 Selects: ARCH_HAS_FAST_MULTIPLIER [=y]

---

Cc: Steve Best <sbest@redhat.com>
Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>