[redhat] New configs in lib/Kconfig
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: FORCE_NR_CPUS [=n]
 Type  : bool
 Defined at lib/Kconfig:531
   Prompt: NR_CPUS is set to an actual number of CPUs
   Depends on: SMP [=y]
   Location:
     Main menu
       -> Library routines

---

Cc: Prarit Bhargava <prarit@redhat.com>
Cc: Aristeu Rozanski <arozansk@redhat.com>
Cc: Christoph von Recklinghausen <crecklin@redhat.com>
Cc: Eugene Syromiatnikov <esyromia@redhat.com>
Cc: "Herton R. Krzesinski" <herton@redhat.com>
Cc: Nico Pache <npache@redhat.com>
Cc: Oleg Nesterov <onestero@redhat.com>
Cc: Rafael Aquini <aquini@redhat.com>
Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>