[redhat] New configs in lib/Kconfig.debug
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: DEBUG_MAPLE_TREE [=n]
 Type  : bool
 Defined at lib/Kconfig.debug:1643
   Prompt: Debug maple trees
   Depends on: DEBUG_KERNEL [=y]
   Location:
     Main menu
       -> Kernel hacking
         -> Debug kernel data structures
 Selected by [n]:
   - DEBUG_VM_MAPLE_TREE [=n] && DEBUG_VM [=n]

---

 Symbol: DEBUG_VM_MAPLE_TREE [=n]
 Type  : bool
 Defined at lib/Kconfig.debug:820
   Prompt: Debug VM maple trees
   Depends on: DEBUG_VM [=y]
   Location:
     Main menu
       -> Kernel hacking
         -> Memory Debugging
           -> Debug VM (DEBUG_VM [=y])
 Selects: DEBUG_MAPLE_TREE [=n]

---

Cc: Prarit Bhargava <prarit@redhat.com>
Cc: Aristeu Rozanski <arozansk@redhat.com>
Cc: Christoph von Recklinghausen <crecklin@redhat.com>
Cc: Eugene Syromiatnikov <esyromia@redhat.com>
Cc: "Herton R. Krzesinski" <herton@redhat.com>
Cc: Nico Pache <npache@redhat.com>
Cc: Oleg Nesterov <onestero@redhat.com>
Cc: Rafael Aquini <aquini@redhat.com>
Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>