spec: fix path to `installing_core` stamp file for subpackages
The other scriptlets which inspect or delete this file expect it to have
the `+SUBPKG` suffix. Looking at the original patch introducing this
line, I think that was the intent, but here we need to use `%{?-v}`
instead of `%{?1}` since unlike in the other places the file is
referenced, the subpackage is a flagged macro parameter rather than
positional.

Fixes e815131473e4 ("redhat: spec: trigger dracut when modules are
installed separately").

Fixes: https://github.com/coreos/rpm-ostree/issues/4057

Signed-off-by: Jonathan Lebon <jonathan@jlebon.com>