redhat/docs/index.rst: Add FLAVOR information to generate configs for local builds
If you use `make dist-configs` the RHEL flavor is used by default, this
can be a bit confusing because you don't really know which is the
default, instead modify the command to explicitly specify the flavor
wanted.

While here, also fix the name for the generated config that needs to be
copied.

Signed-off-by: Enric Balletbo i Serra <eballetbo@redhat.com>