redhat: fix selftest git command so it picks the right commit
The current git command used to fetch the latest release string may end
up finding the latest release string of a different branch such as
ark-latest and then comparing it to the version string in the
changelog-9.99 file of the current branch. These two strings won't
necessarily match so the test fails.

Fix this by dropping '--all' from the git command so we end up finding
the latest release commit of the current branch and comparing it to the
changelog-9.99 file's string in the same current branch. These should be
the same.

Signed-off-by: Patrick Talbert <ptalbert@redhat.com>