redhat/configs: enable CONFIG_GNSS for RHEL
CONFIG_GNSS is enabled only in Fedora. The ice driver will soon make use
the GNSS subsystem in upstream. Let's enable GNSS in common, but disable
it in s390x-zfcpdump, which ought to be minimal with no networking.

We do not need any GNSS hardware driver modules in RHEL. Those remain
enabled only in Fedora.

Signed-off-by: Michal Schmidt <mschmidt@redhat.com>