[redhat] New configs in drivers/acpi
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: ACPI_FFH [=n]
 Type  : bool
 Defined at drivers/acpi/Kconfig:567
   Prompt: ACPI FFH Address Space
   Depends on: ACPI [=y]
   Location:
     -> Power management and ACPI options
       -> ACPI (Advanced Configuration and Power Interface) Support (ACPI [=y])
         -> ACPI FFH Address Space (ACPI_FFH [=n])

---

Cc: Mark Langsdorf <mlangsdo@redhat.com>
Cc: Lenny Szubowicz <lszubowi@redhat.com>
Cc: Prarit Bhargava <prarit@redhat.com>
Cc: David Arcari <darcari@redhat.com>
Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>

v2:
- CONFIG_ACPI_FFH enabled for ark.

Signed-off-by: Patrick Talbert <ptalbert@redhat.com>