[redhat] New configs in drivers/nvdimm
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: NVDIMM_SECURITY_TEST [=n]
 Type  : bool
 Defined at drivers/nvdimm/Kconfig:117
   Prompt: Enable NVDIMM security unit tests
   Depends on: LIBNVDIMM [=m] && NVDIMM_KEYS [=y]
   Location:
     -> Device Drivers
       -> NVDIMM (Non-Volatile Memory Device) Support (LIBNVDIMM [=m])
         -> Enable NVDIMM security unit tests (NVDIMM_SECURITY_TEST [=n])

---

Cc: Jeff Moyer <jmoyer@redhat.com>
Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>