[redhat] New configs in drivers/perf
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: ARM_CORESIGHT_PMU_ARCH_SYSTEM_PMU [=n]
 Type  : tristate
 Defined at drivers/perf/arm_cspmu/Kconfig:5
   Prompt: ARM Coresight Architecture PMU
   Depends on: PERF_EVENTS [=y] && ARM64 [=y] && ACPI [=y] && (ACPI_APMT [=y] || COMPILE_TEST [=n])
   Location:
     -> Device Drivers
       -> Performance monitor support
         -> ARM Coresight Architecture PMU (ARM_CORESIGHT_PMU_ARCH_SYSTEM_PMU [=n])

---

Cc: Mark Salter <msalter@redhat.com>
Cc: Mark Langsdorf <mlangsdo@redhat.com>
Cc: Jeremy Linton <jlinton@redhat.com>
Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>

v2:
- CONFIG_ARM_CORESIGHT_PMU_ARCH_SYSTEM_PMU enabled for ark under
  aarch64.

Signed-off-by: Patrick Talbert <ptalbert@redhat.com>