[redhat] New configs in arch/x86
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: EFI_HANDOVER_PROTOCOL [=y]
 Type  : bool
 Defined at arch/x86/Kconfig:1984
   Prompt: EFI handover protocol (DEPRECATED)
   Depends on: EFI_STUB [=y]
   Location:
     -> Processor type and features
       -> EFI runtime service support (EFI [=y])
         -> EFI stub support (EFI_STUB [=y])
           -> EFI handover protocol (DEPRECATED) (EFI_HANDOVER_PROTOCOL [=y])

---

Cc: David Arcari <darcari@redhat.com>
Cc: Prarit Bhargava <prarit@redhat.com>
Cc: Dean Nelson <dnelson@redhat.com>
Cc: Lenny Szubowicz <lszubowi@redhat.com>
Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>