[redhat] New configs in drivers/accel
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: DRM_ACCEL [=n]
 Type  : bool
 Defined at drivers/accel/Kconfig:9
   Prompt: Compute Acceleration Framework
   Depends on: DRM [=y]
   Location:
     -> Device Drivers
       -> Compute Acceleration Framework (DRM_ACCEL [=n])

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>