[redhat] New configs in drivers/block
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: ZRAM_MULTI_COMP [=n]
 Type  : bool
 Defined at drivers/block/zram/Kconfig:82
   Prompt: Enable multiple compression streams
   Depends on: BLK_DEV [=y] && ZRAM [=m]
   Location:
     -> Device Drivers
       -> Block devices (BLK_DEV [=y])
         -> Enable multiple compression streams (ZRAM_MULTI_COMP [=n])

---

Cc: Ming Lei <ming.lei@redhat.com>
Cc: Jeff Moyer <jmoyer@redhat.com>
Cc: Vivek Goyal <vgoyal@redhat.com>
Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>