[redhat] New configs in drivers/bluetooth
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: BT_HCIBCM4377 [=n]
 Type  : tristate
 Defined at drivers/bluetooth/Kconfig:289
   Prompt: HCI BCM4377/4378/4387 PCIe driver
   Depends on: NET [=y] && BT [=m] && PCI [=y]
   Location:
     -> Networking support (NET [=y])
       -> Bluetooth subsystem support (BT [=m])
         -> Bluetooth device drivers
           -> HCI BCM4377/4378/4387 PCIe driver (BT_HCIBCM4377 [=n])
 Selects: FW_LOADER [=y]

---

 Symbol: BT_HCIBTUSB_POLL_SYNC [=y]
 Type  : bool
 Defined at drivers/bluetooth/Kconfig:48
   Prompt: Enable USB poll_sync for Bluetooth USB devices by default
   Depends on: NET [=y] && BT [=m] && BT_HCIBTUSB [=m]
   Location:
     -> Networking support (NET [=y])
       -> Bluetooth subsystem support (BT [=m])
         -> Bluetooth device drivers
           -> HCI USB driver (BT_HCIBTUSB [=m])
             -> Enable USB poll_sync for Bluetooth USB devices by default (BT_HCIBTUSB_POLL_SYNC [=y])

---

Cc: Gopal Tiwari <gtiwari@redhat.com>
Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>