[redhat] New configs in drivers/net/ethernet/netronome
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: NFP_NET_IPSEC [=y]
 Type  : bool
 Defined at drivers/net/ethernet/netronome/Kconfig:57
   Prompt: NFP IPsec crypto offload support
   Depends on: NETDEVICES [=y] && ETHERNET [=y] && NET_VENDOR_NETRONOME [=y] && NFP [=m] && XFRM_OFFLOAD [=y]
   Location:
     -> Device Drivers
       -> Network device support (NETDEVICES [=y])
         -> Ethernet driver support (ETHERNET [=y])
           -> Netronome(R) devices (NET_VENDOR_NETRONOME [=y])
             -> Netronome(R) NFP4000/NFP6000 NIC driver (NFP [=m])
               -> NFP IPsec crypto offload support (NFP_NET_IPSEC [=y])

---

Cc: Michal Schmidt <mschmidt@redhat.com>
Cc: Petr Oros <poros@redhat.com>
Cc: Ken Cox <jkc@redhat.com>
Cc: "José Ignacio Tornos Martínez" <jtornosm@redhat.com>
Cc: Kamal Heib <kheib@redhat.com>
Cc: Stefan Assmann <sassmann@redhat.com>
Cc: Corinna Vinschen <vinschen@redhat.com>
Cc: Ivan Vecera <ivecera@redhat.com>
Cc: "Íñigo Huguet" <ihuguet@redhat.com>
Cc: Jonathan Toppins <jtoppins@redhat.com>
Cc: Josef Oskera <joskera@redhat.com>
Cc: Izabela Bakollari <ibakolla@redhat.com>
Cc: Jarod Wilson <jarod@redhat.com>
Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>