[redhat] New configs in drivers/watchdog
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: ADVANTECH_EC_WDT [=n]
 Type  : tristate
 Defined at drivers/watchdog/Kconfig:1058
   Prompt: Advantech Embedded Controller Watchdog Timer
   Depends on: WATCHDOG [=y] && X86 [=y]
   Location:
     -> Device Drivers
       -> Watchdog Timer Support (WATCHDOG [=y])
         -> Advantech Embedded Controller Watchdog Timer (ADVANTECH_EC_WDT [=n])

---

Cc: Prarit Bhargava <prarit@redhat.com>
Cc: David Arcari <darcari@redhat.com>
Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>