[redhat] New configs in sound/soc
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: SND_SOC_INTEL_AVS_MACH_PROBE [=n]
 Type  : tristate
 Defined at sound/soc/intel/avs/boards/Kconfig:80
   Prompt: Probing (data) board
   Depends on: SOUND [=m] && !UML && SND [=m] && SND_SOC [=m] && SND_SOC_INTEL_AVS [=m] && DEBUG_FS [=y]
   Location:
     -> Device Drivers
       -> Sound card support (SOUND [=m])
         -> Advanced Linux Sound Architecture (SND [=m])
           -> ALSA for SoC audio support (SND_SOC [=m])
             -> Intel AVS driver (SND_SOC_INTEL_AVS [=m])
               -> Intel AVS Machine drivers
                 -> Probing (data) board (SND_SOC_INTEL_AVS_MACH_PROBE [=n])
 Selects: SND_HWDEP [=m]

---

Cc: Jaroslav Kysela <jkysela@redhat.com>
Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>