Switch input into config checkoptions
In preparation for checking removed config options, make the change
simple by swapping the inputs into the awk script.  Read the generated
config first and compare it to the merged config.  This makes it easy to
see if a merged config was removed from the generated config.

Update the failure langauge to reflect the change.

Nothing changes with this commit.

Signed-off-by: Don Zickus <dzickus@redhat.com>