Turn on debug builds for aarch64 Fedora
Fedora only did debug builds for x86 in the past. This was largely due
to slow build times on other arches in koji, coupled with much smaller
userbases.  Let's turn on aarch64 now as those builders are much faster
and the userbase is considerably larger these days.  Replacing i686 here
makes sense as we no longer build for that arch.

Signed-off-by: Justin M. Forbes <jforbes@fedoraproject.org>