[redhat] New configs in lib/Kconfig.debug
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: MEMCPY_SLOW_KUNIT_TEST [=y]
 Type  : bool
 Defined at lib/Kconfig.debug:2569
   Prompt: Include exhaustive memcpy tests
   Depends on: RUNTIME_TESTING_MENU [=y] && MEMCPY_KUNIT_TEST [=m]
   Location:
     -> Kernel hacking
       -> Kernel Testing and Coverage
         -> Runtime Testing (RUNTIME_TESTING_MENU [=y])
           -> Test memcpy(), memmove(), and memset() functions at runtime (MEMCPY_KUNIT_TEST [=m])
             -> Include exhaustive memcpy tests (MEMCPY_SLOW_KUNIT_TEST [=y])

---

Cc: Prarit Bhargava <prarit@redhat.com>
Cc: Wander Lairson Costa <wander@redhat.com>
Cc: Aristeu Rozanski <arozansk@redhat.com>
Cc: Christoph von Recklinghausen <crecklin@redhat.com>
Cc: Eugene Syromiatnikov <esyromia@redhat.com>
Cc: "Herton R. Krzesinski" <herton@redhat.com>
Cc: Nico Pache <npache@redhat.com>
Cc: Oleg Nesterov <onestero@redhat.com>
Cc: Rafael Aquini <aquini@redhat.com>
Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>