Fix up SQUASHFS decompression configs
The 6.2 merge window changed the way that squashfs decompression
threading worked. It now allows the number of threads to be set as mount
options if all options are compiled.  Until a decicion is made on this
functionality, this MR factors in the new options and gives us a built
kernel config similar to what has been supported in previous releases.
We select CONFIG_SQUASHFS_DECOMP_MULTI_PERCPU as the only option. If a
different config is desired, I am happy to change this, but our current
defaults were not working as it forced a config mismatch, and I did not
want the state to get lost.

Signed-off-by: Justin M. Forbes <jforbes@fedoraproject.org>