Trim Changelog for 6.3 series
I am not entirely sure why, but when we rebase for a new upstream
release, the changelog resets itself, but it leaves all of the old
entries as well, resulting in double  entries. A reset at least makes
sure we don't have duplicates for everything, but the date is still
wrong.

Signed-off-by: Justin M. Forbes <jforbes@fedoraproject.org>