[redhat] New configs in kernel/rcu
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: RCU_CPU_STALL_CPUTIME [=n]
 Type  : bool
 Defined at kernel/rcu/Kconfig.debug:95
   Prompt: Provide additional RCU stall debug information
   Depends on: RCU_STALL_COMMON [=y]
   Location:
     -> Kernel hacking
       -> RCU Debugging
         -> Provide additional RCU stall debug information (RCU_CPU_STALL_CPUTIME [=n])

---

Cc: Waiman Long <longman@redhat.com>
Cc: Artem Savkov <asavkov@redhat.com>
Cc: Clark Williams <williams@redhat.com>
Cc: Phil Auld <pauld@redhat.com>
Cc: Wander Lairson Costa <wander@redhat.com>
Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>