[redhat] New configs in init/Kconfig
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: BOOT_CONFIG_FORCE [=n]
 Type  : bool
 Defined at init/Kconfig:1387
   Prompt: Force unconditional bootconfig processing
   Depends on: BOOT_CONFIG [=y]
   Location:
     -> General setup
       -> Boot config support (BOOT_CONFIG [=y])
         -> Force unconditional bootconfig processing (BOOT_CONFIG_FORCE [=n])

---

Cc: Prarit Bhargava <prarit@redhat.com>
Cc: Aristeu Rozanski <arozansk@redhat.com>
Cc: Artem Savkov <asavkov@redhat.com>
Cc: Clark Williams <williams@redhat.com>
Cc: "Herton R. Krzesinski" <herton@redhat.com>
Cc: Jan Stancek <jstancek@redhat.com>
Cc: Josh Poimboeuf <jpoimboe@kernel.org>
Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>