[redhat] New configs in mm/Kconfig
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: ZSMALLOC_CHAIN_SIZE [=8]
 Type  : integer
 Range : [4 16]
 Defined at mm/Kconfig:194
   Prompt: Maximum number of physical pages per-zspage
   Depends on: ZSMALLOC [=y]
   Location:
     -> Memory Management options
       -> Support for paging of anonymous memory (swap) (SWAP [=y])
         -> Compressed cache for swap pages (ZSWAP [=y])
           -> N:1 compression allocator (zsmalloc) (ZSMALLOC [=y])
             -> Maximum number of physical pages per-zspage (ZSMALLOC_CHAIN_SIZE [=8])

---

Cc: Rafael Aquini <aquini@redhat.com>
Cc: Aristeu Rozanski <arozansk@redhat.com>
Cc: Artem Savkov <asavkov@redhat.com>
Cc: Baoquan He <bhe@redhat.com>
Cc: Christoph von Recklinghausen <crecklin@redhat.com>
Cc: Donald Dutile <ddutile@redhat.com>
Cc: "Herton R. Krzesinski" <herton@redhat.com>
Cc: Leah Leshchinsky <lleshchi@redhat.com>
Cc: Nico Pache <npache@redhat.com>
Cc: Waiman Long <longman@redhat.com>
Cc: Ricardo Robaina <rrobaina@redhat.com>
Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>