ark: enable wifi on aarch64
Enable wireless core and drivers for aarch64 in RHEL, needed for a lot
of Edge hardware. We're enabling the same options and drivers that are
enabled for x86_64, which was the only architecture where wifi was
supported.

CONFIG_MAC80211_DEBUGFS and CONFIG_MAC80211_RC_DEFAULT had values in
rhel/generic that were unused, being overwritten in x86. The values to be
used for all arches, both for RHEL and Fedora, are the same, so moving
them to common/generic.

Comparing what CONFIGs are enabled by default with CONFIG_WIRELESS and
CONFIG_WLAN in x86_64 and aarch64, there are 2 CONFIGs that are aarch64
specific, and we can set them to "not set" for the moment:
  CONFIG_ATH10K_AHB  <- configured to "not set" here
  CONFIG_PWRSEQ_SD8787  <- it was already configured to "not set"

Suggested-by: Peter Robinson <pbrobinson@redhat.com>
Signed-off-by: Íñigo Huguet <ihuguet@redhat.com>