[redhat] New configs in drivers/phy
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: PHY_QCOM_EUSB2_REPEATER [=n]
 Type  : tristate
 Defined at drivers/phy/qualcomm/Kconfig:124
   Prompt: Qualcomm SNPS eUSB2 Repeater Driver
   Depends on: OF [=y] && (ARCH_QCOM [=y] || COMPILE_TEST [=n])
   Location:
     -> Device Drivers
       -> PHY Subsystem
         -> Qualcomm SNPS eUSB2 Repeater Driver (PHY_QCOM_EUSB2_REPEATER [=n])
 Selects: GENERIC_PHY [=y]

---

 Symbol: PHY_QCOM_SNPS_EUSB2 [=n]
 Type  : tristate
 Defined at drivers/phy/qualcomm/Kconfig:115
   Prompt: Qualcomm SNPS eUSB2 PHY Driver
   Depends on: OF [=y] && (ARCH_QCOM [=y] || COMPILE_TEST [=n])
   Location:
     -> Device Drivers
       -> PHY Subsystem
         -> Qualcomm SNPS eUSB2 PHY Driver (PHY_QCOM_SNPS_EUSB2 [=n])
 Selects: GENERIC_PHY [=y]

---

Cc: Petr Oros <poros@redhat.com>
Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>