[redhat] New configs in drivers/tty
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: SERIAL_8250_PCI1XXXX [=y]
 Type  : tristate
 Defined at drivers/tty/serial/8250/Kconfig:298
   Prompt: Microchip 8250 based serial port
   Depends on: TTY [=y] && HAS_IOMEM [=y] && SERIAL_8250 [=y] && PCI [=y]
   Location:
     -> Device Drivers
       -> Character devices
         -> Enable TTY (TTY [=y])
           -> Serial drivers
             -> Microchip 8250 based serial port (SERIAL_8250_PCI1XXXX [=y])
 Selects: SERIAL_8250_PCILIB [=y]

---

 Symbol: SERIAL_EARLYCON_SEMIHOST [=n]
 Type  : bool
 Defined at drivers/tty/serial/Kconfig:76
   Prompt: Early console using Arm compatible semihosting
   Depends on: TTY [=y] && HAS_IOMEM [=y] && (ARM64 [=y] || ARM || RISCV)
   Location:
     -> Device Drivers
       -> Character devices
         -> Enable TTY (TTY [=y])
           -> Serial drivers
             -> Early console using Arm compatible semihosting (SERIAL_EARLYCON_SEMIHOST [=n])
 Selects: SERIAL_CORE [=y] && SERIAL_CORE_CONSOLE [=y] && SERIAL_EARLYCON [=y]

---

Cc: John Linville <linville@redhat.com>
Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>