[redhat] New configs in drivers/cxl
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: CXL_REGION [=y]
 Type  : bool
 Defined at drivers/cxl/Kconfig:106
   Prompt: CXL: Region Support
   Depends on: CXL_BUS [=m] && SPARSEMEM [=y]
   Location:
     -> Device Drivers
       -> PCI support (PCI [=y])
         -> CXL (Compute Express Link) Devices Support (CXL_BUS [=m])
           -> CXL: Region Support (CXL_REGION [=y])
 Selects: MEMREGION [=y] && GET_FREE_REGION [=y]

---

Cc: Myron Stowe <mstowe@redhat.com>
Cc: John Linville <linville@redhat.com>
Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>