[redhat] New configs in drivers/vdpa
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: MLX5_VDPA_STEERING_DEBUG [=n]
 Type  : bool
 Defined at drivers/vdpa/Kconfig:74
   Prompt: expose steering counters on debugfs
   Depends on: VDPA [=m]
   Location:
     -> Device Drivers
       -> vDPA drivers (VDPA [=m])
         -> expose steering counters on debugfs (MLX5_VDPA_STEERING_DEBUG [=n])
 Selects: MLX5_VDPA [=y]

---

 Symbol: SNET_VDPA [=n]
 Type  : tristate
 Defined at drivers/vdpa/Kconfig:101
   Prompt: SolidRun's vDPA driver for SolidNET
   Depends on: VDPA [=m] && PCI_MSI [=y] && PCI_IOV [=y] && (HWMON [=y] || HWMON [=y]=n)
   Location:
     -> Device Drivers
       -> vDPA drivers (VDPA [=m])
         -> SolidRun's vDPA driver for SolidNET (SNET_VDPA [=n])

---

Cc: "Michael S. Tsirkin" <mst@redhat.com>
Cc: Jason Wang <jasowang@redhat.com>
Cc: rhvirt-patches@redhat.com
Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>