redhat/Makefile: fix default values for dist-brew's DISTRO and DIST
In dist-brew target, target-specific variables to change the default
value of DISTRO and DIST to 'rhel' and '.el9', but they're not useful
because the rest of variables that depends on them, such us BUILD_TARGET,
has been calculated yet.

The most visible error is that `make dist-brew` fails with 'brew: error:
No such build target: rawhide'. Rawhide is the default BUILD_TARGET for
Fedora, not for RHEL.

Fix it by recursively calling `make` so the sub-make has the correct DISTRO
and DIST values since the beginning, thus calculating the rest of values
correctly. Variables BUILD_TARGET and BUILD_PROFILE need to be cleared
so they're recalculated in the 2nd level sub-make.

However, variables listed in Makefile.variables file must be allowed to
be customizable by the user, so we can't override the ones defined by
the user. Save them in variables *_USERDEF so we know if they were
defined by user or by the Makefile's logic.

Also, fix a missing closing quote in `ifeq ("$(DIST)", ".eln)`.

Signed-off-by: Íñigo Huguet <ihuguet@redhat.com>