generic: common: Parport and paride/ata cleanups
The parallel IDE interface has bee replace with PATA_PARPORT
using the ATA stack, this is mostly irrelevant as PARIDE was
already disabled everywhere (RHEL and Fedora) so we just move
this new option to common and leave it disabled and clean up
the old PARIDE configs.

We also remove the PARPORT_AX88796 dead option while we're
at it as it no longer exists upstream.

Signed-off-by: Peter Robinson <pbrobinson@redhat.com>