kernel.spec: Fix UKI naming to comply with BLS
The Boot Loader Specification (BLS) [1] suggests that type #2 entries (UKIs)
shall have the same names like type #1 entries with the exception of
the suffix being .efi instead of .conf. In particular this means that
the name for installed UKIs should be
<entry-token-or-machine-id>-<version>.efi. The same naming convention is
used by kernel-installs 90-uki-copy.install [2] in systemd v253.

Adjust the workaround to install UKIs in the kernel.spec to follow the
same naming convention in order to prevent breaking user space tools
that rely on it.

[1] https://uapi-group.org/specifications/specs/boot_loader_specification/
[2] https://github.com/systemd/systemd/blob/v253/src/kernel-install/90-uki-copy.install

Signed-off-by: Philipp Rudo <prudo@redhat.com>