signal: Add proper comment about the preempt-disable in ptrace_stop().
Commit 53da1d9456fe7 ("fix ptrace slowness") added a preempt-disable section
between read_unlock() and the following schedule() invocation without
explaining why it is needed.

Replace the comment with an explanation why this is needed. Clarify that
it is needed for correctness but for performance reasons.

Signed-off-by: Sebastian Andrzej Siewior <bigeasy@linutronix.de>
Link: https://lore.kernel.org/20230406205713.1843072-2-bigeasy@linutronix.de
(cherry picked from commit bc9467a9d39a25bd4eeaa0a960ece26dc7e7ec28)
Signed-off-by: Clark Williams <williams@redhat.com>