Fix up the RHEL configs for xtables and ipset
At some point, we had an MR which disabled xtables and ipset.  This put
configs to turn them off in the ark/generic directory.  Quite a bit
later, it was decided to rever that change, and MR2267 was created to do
that. In the meantime, the ark directory was renamed to rhel.  As that
MR was deleting files in ark, and they had already been deleted, this
was not raised as a problem, but we did not get the desired result
because those files do still exist in the rhel directory.  This change
simply deletes the files in the rhel/generic directory making things
work as MR2267 intended.

Link:https://gitlab.com/cki-project/kernel-ark/-/issues/105
Cc: Phil Sutter <psutter@redhat.com>
Signed-off-by: Justin M. Forbes <jforbes@fedoraproject.org>