redhat/Makefile: Add target to print DISTRELEASETAG
Add a make target that prints the git tag that dist-release-tag would
set. This allows querying the tag in external scripts, for example to
check if a tag already exists before running dist-release-tag, as the
latter will rightfully fail if it already exists.

For this to work DISTLOCALVERSION must not be shown, which is currently
printed on every invocation. Follow the upstream kernel's lead for
targets like kernelversion and prevent DISTLOCALVERSION from being
shown when the make flag -s or --quiet is used.

Signed-off-by: Thorsten Leemhuis <fedora@leemhuis.info>