fedora: wifi: hermes: disable 802.11b driver
The old Hermes driver is one of the original 11b chipsets used
on Orinoco, Prism and a few other devices. It's the last driver
using the old WEXT interface which we officially retired in F-36.

It was mostly used on old laptops and PCMCIA (not cardbus) on
old PPC Macs and 32 bit x86 devices. It's been shipped in the
kernel extra modules for ever. Let's disable it as it's unlikely
to have any users.

Signed-off-by: Peter Robinson <pbrobinson@redhat.com>