[redhat] New configs in drivers/gpu
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: DRM_SAMSUNG_DSIM [=n]
 Type  : tristate
 Defined at drivers/gpu/drm/bridge/Kconfig:223
   Prompt: Samsung MIPI DSIM bridge driver
   Depends on: DRM [=y] && DRM_BRIDGE [=y] && COMMON_CLK [=y] && OF [=y] && HAS_IOMEM [=y]
   Location:
     -> Device Drivers
       -> Graphics support
         -> Display Interface Bridges
           -> Samsung MIPI DSIM bridge driver (DRM_SAMSUNG_DSIM [=n])
 Selects: DRM_KMS_HELPER [=y] && DRM_MIPI_DSI [=n] && DRM_PANEL_BRIDGE [=y]
 Selected by [n]:
   - DRM_EXYNOS_DSI [=n] && HAS_IOMEM [=y] && DRM_EXYNOS [=n] && (DRM_EXYNOS_FIMD [=n] || DRM_EXYNOS5433_DECON [=n] || DRM_EXYNOS7_DECON [=n])

---

 Symbol: DRM_VIRTIO_GPU_KMS [=y]
 Type  : bool
 Defined at drivers/gpu/drm/virtio/Kconfig:15
   Prompt: Virtio GPU driver modesetting support
   Depends on: HAS_IOMEM [=y] && DRM_VIRTIO_GPU [=m]
   Location:
     -> Device Drivers
       -> Graphics support
         -> Virtio GPU driver (DRM_VIRTIO_GPU [=m])
           -> Virtio GPU driver modesetting support (DRM_VIRTIO_GPU_KMS [=y])

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>