[redhat] New configs in drivers/platform
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: LENOVO_YMC [=n]
 Type  : tristate
 Defined at drivers/platform/x86/Kconfig:464
   Prompt: Lenovo Yoga Tablet Mode Control
   Depends on: X86_PLATFORM_DEVICES [=y] && ACPI_WMI [=m] && INPUT [=y]
   Location:
     -> Device Drivers
       -> X86 Platform Specific Device Drivers (X86_PLATFORM_DEVICES [=y])
         -> Lenovo Yoga Tablet Mode Control (LENOVO_YMC [=n])
 Selects: INPUT_SPARSEKMAP [=m]

---

 Symbol: MSI_EC [=n]
 Type  : tristate
 Defined at drivers/platform/x86/Kconfig:647
   Prompt: MSI EC Extras
   Depends on: X86_PLATFORM_DEVICES [=y] && ACPI [=y] && ACPI_BATTERY [=y]
   Location:
     -> Device Drivers
       -> X86 Platform Specific Device Drivers (X86_PLATFORM_DEVICES [=y])
         -> MSI EC Extras (MSI_EC [=n])

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>