[redhat] New configs in drivers/net/ethernet/amd
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: PDS_CORE [=n]
 Type  : tristate
 Defined at drivers/net/ethernet/amd/Kconfig:189
   Prompt: AMD/Pensando Data Systems Core Device Support
   Depends on: NETDEVICES [=y] && ETHERNET [=y] && NET_VENDOR_AMD [=y] && 64BIT [=y] && PCI [=y]
   Location:
     -> Device Drivers
       -> Network device support (NETDEVICES [=y])
         -> Ethernet driver support (ETHERNET [=y])
           -> AMD devices (NET_VENDOR_AMD [=y])
             -> AMD/Pensando Data Systems Core Device Support (PDS_CORE [=n])

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>