[redhat] New configs in net/Kconfig
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: MAX_SKB_FRAGS [=17]
 Type  : integer
 Range : [17 45]
 Defined at net/Kconfig:274
   Prompt: Maximum number of fragments per skb_shared_info
   Depends on: NET [=y]
   Location:
     -> Networking support (NET [=y])
       -> Networking options
         -> Maximum number of fragments per skb_shared_info (MAX_SKB_FRAGS [=17])

---

 Symbol: NET_HANDSHAKE_KUNIT_TEST [=m]
 Type  : tristate
 Defined at net/Kconfig:76
   Prompt: KUnit tests for the handshake upcall mechanism
   Depends on: NET [=y] && KUNIT [=m]
   Visible if: NET [=y] && KUNIT [=m] && !KUNIT_ALL_TESTS [=m]
   Location:
     -> Networking support (NET [=y])
       -> Networking options
         -> KUnit tests for the handshake upcall mechanism (NET_HANDSHAKE_KUNIT_TEST [=m])

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>