[redhat] New configs in drivers/nvmem
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: NVMEM_LAYOUT_ONIE_TLV [=n]
 Type  : tristate
 Defined at drivers/nvmem/layouts/Kconfig:14
   Prompt: ONIE tlv support
   Depends on: NVMEM [=y]
   Location:
     -> Device Drivers
       -> NVMEM Support (NVMEM [=y])
         -> Layout Types
           -> ONIE tlv support (NVMEM_LAYOUT_ONIE_TLV [=n])
 Selects: CRC32 [=y]

---

 Symbol: NVMEM_LAYOUT_SL28_VPD [=n]
 Type  : tristate
 Defined at drivers/nvmem/layouts/Kconfig:5
   Prompt: Kontron sl28 VPD layout support
   Depends on: NVMEM [=y]
   Location:
     -> Device Drivers
       -> NVMEM Support (NVMEM [=y])
         -> Layout Types
           -> Kontron sl28 VPD layout support (NVMEM_LAYOUT_SL28_VPD [=n])
 Selects: CRC8 [=m]

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>