[redhat] New configs in samples/Kconfig
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: SAMPLE_KMEMLEAK [=n]
 Type  : tristate
 Defined at samples/Kconfig:276
   Prompt: Simple test for the kernel memory leak detector
   Depends on: SAMPLES [=y] && DEBUG_KMEMLEAK [=y] && m && MODULES [=y]
   Location:
     -> Kernel hacking
       -> Sample kernel code (SAMPLES [=y])
         -> Simple test for the kernel memory leak detector (SAMPLE_KMEMLEAK [=n])

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>