[redhat] New configs in fs/xfs
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: XFS_SUPPORT_ASCII_CI [=y]
 Type  : bool
 Defined at fs/xfs/Kconfig:50
   Prompt: Support deprecated case-insensitive ascii (ascii-ci=1) format
   Depends on: BLOCK [=y] && XFS_FS [=m]
   Location:
     -> File systems
       -> XFS filesystem support (XFS_FS [=m])
         -> Support deprecated case-insensitive ascii (ascii-ci=1) format (XFS_SUPPORT_ASCII_CI [=y])

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>

v2:
- CONFIG_XFS_SUPPORT_ASCII_CI disabled.

Signed-off-by: Patrick Talbert <ptalbert@redhat.com>