[redhat] New configs in drivers/leds
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: LEDS_BD2606MVV [=n]
 Type  : tristate
 Defined at drivers/leds/Kconfig:554
   Prompt: LED driver for BD2606MVV
   Depends on: NEW_LEDS [=y] && LEDS_CLASS [=y] && I2C [=y]
   Location:
     -> Device Drivers
       -> LED Support (NEW_LEDS [=y])
         -> LED driver for BD2606MVV (LEDS_BD2606MVV [=n])
 Selects: REGMAP_I2C [=m]

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>