[redhat] New configs in drivers/soundwire
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: SOUNDWIRE_AMD [=n]
 Type  : tristate
 Defined at drivers/soundwire/Kconfig:21
   Prompt: AMD SoundWire Manager driver
   Depends on: SOUNDWIRE [=m] && ACPI [=y] && SND_SOC [=m]
   Location:
     -> Device Drivers
       -> SoundWire support (SOUNDWIRE [=m])
         -> AMD SoundWire Manager driver (SOUNDWIRE_AMD [=n])
 Selects: SOUNDWIRE_GENERIC_ALLOCATION [=m]

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>