Enable IO_URING for RHEL
While RHEL 9 has been running with IO_URING disabled, upstream did not
want to remove the config EXPERT gating because "in the future it is
assumed that IO_URING is enabled everywhere".  It seems that future has
come, and with 6.4, we can not build with IO_URING disabled. The easy
solution is to turn IO_URING back on for RHEL.   If this is not
acceptable, someone would need to fix the build issues with 6.4 and
IO_URING disabled.

Signed-off-by: Justin M. Forbes <jforbes@fedoraproject.org>