redhat/configs: increase CONFIG_FRAME_WARN for Fedora on aarch64
Currently aarch64 has CONFIG_FRAME_WARN set to 1024, which is too small
and causes warnings to be emited during build. Looking at the defaults
in lib/Kconfig.debug shows that it recommends using 1024 for 32-bit
systems and 2048 for 64-bit systems. RHEL is already set to use 2048
for all supported architectures. Fedora no longer supports arm32. Let's
just set this to 2048 for all variants.

Signed-off-by: Brian Masney <bmasney@redhat.com>