Remove CONFIG_ARCH_FORCE_MAX_ORDER for aarch64
Upstream commit f3c3762178fad5cf5dc108fb061ba010818323b4 took away the
prompt so this is no longer configurable. The default is 10. While this
is different from what we have been running, it was determined that the
older ThunderX platform was the only reason that this value should be
higher.

Signed-off-by: Justin M. Forbes <jforbes@fedoraproject.org>