CI: sync ark-create-release push commands with output
A user can run the ci scripts and either auto push the changes or not.
If the changes are not pushed, a message is printed explaining the steps
needed to properly push the changes.

Over time the real push and the output got out of sync.  Just wrap them
in variables and use eval to execute them.  This should keep things in
sync for a good while.

Also, notice some cruft was modified.  The output did not match the
reality inside the .gitlab-ci.yml file.  And the RELEASE stuff was never
used.

The only things being pushed are: BRANCH(os-build), release tag,
ark-latest, and ark-infra.

No behaviour changes.

Signed-off-by: Don Zickus <dzickus@redhat.com>