CI: ark-update-configs: sync push command and output
A rather large change that just moves everything into variables.  The
end result is the same except the push command is also echo'd before
being eval'd.  This allows the script to keep the command and output in
sync.

The script relies on the trap feature to cleanup the tempfile if
anything goes wrong.

No functional changes.

Signed-off-by: Don Zickus <dzickus@redhat.com>