[redhat] kernel/rh_messages.c: move hardware tables to rh_messages.h
This moves the static message strings and hardware tables that were
in rh_messages.c into an auto-generated header file (rh_messages.h) that
contains the lists of unmaintained, deprecated and disabled drivers and
pci devices.

For kernel-ark, there is no need to maintain driver and device lists so
they are left empty. Note that for Gemini, there would be two separate
lists depending on if we build RHEL 9 or 10.

Signed-off-by: Scott Weaver <scweaver@redhat.com>