[redhat] New configs in arch/arm64
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: ARCH_FORCE_MAX_ORDER [=10]
 Type  : integer
 Defined at arch/arm64/Kconfig:1519
   Prompt: Order of maximal physically contiguous allocations
   Visible if: ARM64_4K_PAGES [=y] || ARM64_16K_PAGES [=n]
   Location:
     -> Kernel Features
       -> Order of maximal physically contiguous allocations (ARCH_FORCE_MAX_ORDER [=10])

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>