[redhat] New configs in drivers/cpufreq
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: X86_AMD_PSTATE_DEFAULT_MODE [=3]
 Type  : integer
 Range : [1 4]
 Defined at drivers/cpufreq/Kconfig.x86:54
   Prompt: AMD Processor P-State default mode
   Depends on: CPU_FREQ [=y] && X86 [=y] && X86_AMD_PSTATE [=y]
   Location:
     -> Power management and ACPI options
       -> CPU Frequency scaling
         -> CPU Frequency scaling (CPU_FREQ [=y])
           -> AMD Processor P-State driver (X86_AMD_PSTATE [=y])
             -> AMD Processor P-State default mode (X86_AMD_PSTATE_DEFAULT_MODE [=3])

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>