[redhat] New configs in lib/Kconfig.debug
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: HARDLOCKUP_DETECTOR_PREFER_BUDDY [=n]
 Type  : bool
 Defined at lib/Kconfig.debug:1074
   Prompt: Prefer the buddy CPU hardlockup detector
   Depends on: HARDLOCKUP_DETECTOR [=y] && HAVE_HARDLOCKUP_DETECTOR_PERF [=y] && HAVE_HARDLOCKUP_DETECTOR_BUDDY [=y] && !HAVE_HARDLOCKUP_DETECTOR_ARCH [=n]
   Location:
     -> Kernel hacking
       -> Debug Oops, Lockups and Hangs
         -> Detect Hard Lockups (HARDLOCKUP_DETECTOR [=y])
           -> Prefer the buddy CPU hardlockup detector (HARDLOCKUP_DETECTOR_PREFER_BUDDY [=n])

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>