[redhat] New configs in drivers/gpio
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: GPIO_MLXBF3 [=n]
 Type  : tristate
 Defined at drivers/gpio/Kconfig:1602
   Prompt: Mellanox BlueField 3 SoC GPIO
   Depends on: GPIOLIB [=y] && PCI [=y] && (MELLANOX_PLATFORM [=y] && ARM64 [=y] || COMPILE_TEST [=n])
   Location:
     -> Device Drivers
       -> GPIO Support (GPIOLIB [=y])
         -> PCI GPIO expanders
           -> Mellanox BlueField 3 SoC GPIO (GPIO_MLXBF3 [=n])
 Selects: GPIO_GENERIC [=m] && GPIOLIB_IRQCHIP [=y]
 Selected by [n]:
   - PINCTRL_MLXBF3 [=n] && PINCTRL [=y] && (MELLANOX_PLATFORM [=y] && ARM64 [=y] || COMPILE_TEST [=n])

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>