[redhat] New configs in arch/arm64
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: ARCH_MA35 [=n]
 Type  : bool
 Defined at arch/arm64/Kconfig.platforms:228
   Prompt: Nuvoton MA35 Architecture
   Location:
     -> Platform selection
       -> Nuvoton MA35 Architecture (ARCH_MA35 [=n])
 Selects: GPIOLIB [=y] && PINCTRL [=y] && RESET_CONTROLLER [=y]

---

 Symbol: ARCH_STM32 [=n]
 Type  : bool
 Defined at arch/arm64/Kconfig.platforms:288
   Prompt: STMicroelectronics STM32 SoC Family
   Location:
     -> Platform selection
       -> STMicroelectronics STM32 SoC Family (ARCH_STM32 [=n])
 Selects: GPIOLIB [=y] && PINCTRL [=y] && PINCTRL_STM32MP257 [=n] && ARM_SMC_MBOX && ARM_SCMI_PROTOCOL [=y] && COMMON_CLK_SCMI [=m]

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>