[redhat] New configs in drivers/hwmon
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: MAX31827 [=n]
 Type  : tristate
 Defined at drivers/hwmon/Kconfig:1101
   Prompt: MAX31827 low-power temperature switch and similar devices
   Depends on: HWMON [=y] && I2C [=y]
   Location:
     -> Device Drivers
       -> Hardware Monitoring support (HWMON [=y])
         -> MAX31827 low-power temperature switch and similar devices (MAX31827 [=n])
 Selects: REGMAP_I2C [=m]

---

 Symbol: SENSORS_HP_WMI [=n]
 Type  : tristate
 Defined at drivers/hwmon/Kconfig:2424
   Prompt: HP WMI Sensors
   Depends on: HWMON [=y] && ACPI [=y] && ACPI_WMI [=m]
   Location:
     -> Device Drivers
       -> Hardware Monitoring support (HWMON [=y])
         -> HP WMI Sensors (SENSORS_HP_WMI [=n])

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>