[redhat] New configs in sound/drivers
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: SND_PCMTEST [=n]
 Type  : tristate
 Defined at sound/drivers/Kconfig:112
   Prompt: Virtual PCM test driver
   Depends on: SOUND [=m] && !UML && SND [=m] && SND_DRIVERS [=y]
   Location:
     -> Device Drivers
       -> Sound card support (SOUND [=m])
         -> Advanced Linux Sound Architecture (SND [=m])
           -> Generic sound devices (SND_DRIVERS [=y])
             -> Virtual PCM test driver (SND_PCMTEST [=n])
 Selects: SND_PCM [=m]

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>