[redhat] New configs in sound/core
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: SND_SEQ_UMP [=n]
 Type  : bool
 Defined at sound/core/seq/Kconfig:63
   Prompt: Support for UMP events
   Depends on: SOUND [=m] && !UML && SND [=m] && SND_SEQUENCER [=m]
   Location:
     -> Device Drivers
       -> Sound card support (SOUND [=m])
         -> Advanced Linux Sound Architecture (SND [=m])
           -> Sequencer support (SND_SEQUENCER [=m])
             -> Support for UMP events (SND_SEQ_UMP [=n])

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>

v2:
- CONFIG_SND_SEQ_UMP enabled for rhel.

Signed-off-by: Patrick Talbert <ptalbert@redhat.com>