[redhat] New configs in sound/usb
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: SND_USB_AUDIO_MIDI_V2 [=n]
 Type  : bool
 Defined at sound/usb/Kconfig:28
   Prompt: MIDI 2.0 support by USB Audio driver
   Depends on: SOUND [=m] && !UML && SND [=m] && SND_USB [=y] && USB [=y] && SND_USB_AUDIO [=m]
   Location:
     -> Device Drivers
       -> Sound card support (SOUND [=m])
         -> Advanced Linux Sound Architecture (SND [=m])
           -> USB sound devices (SND_USB [=y])
             -> USB Audio/MIDI driver (SND_USB_AUDIO [=m])
               -> MIDI 2.0 support by USB Audio driver (SND_USB_AUDIO_MIDI_V2 [=n])

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>

v2:
- CONFIG_SND_USB_AUDIO_MIDI_V2 enabled.
- CONFIG_SND_UMP_LEGACY_RAWMIDI enabled.

Signed-off-by: Patrick Talbert <ptalbert@redhat.com>