[redhat] New configs in drivers/char
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: HW_RANDOM_HISTB [=y]
 Type  : tristate
 Defined at drivers/char/hw_random/Kconfig:338
   Prompt: Hisilicon STB Random Number Generator support
   Depends on: HW_RANDOM [=y] && (ARCH_HISI [=y] || COMPILE_TEST [=n])
   Location:
     -> Device Drivers
       -> Character devices
         -> Hardware Random Number Generator Core support (HW_RANDOM [=y])
           -> Hisilicon STB Random Number Generator support (HW_RANDOM_HISTB [=y])

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>

v2:
- CONFIG_HW_RANDOM_HISTB is disabled.

Signed-off-by: Patrick Talbert <ptalbert@redhat.com>