[redhat] New configs in drivers/mfd
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: MFD_MAX5970 [=n]
 Type  : tristate
 Defined at drivers/mfd/Kconfig:269
   Prompt: Maxim 5970/5978 power switch and monitor
   Depends on: HAS_IOMEM [=y] && I2C [=y] && OF [=y]
   Location:
     -> Device Drivers
       -> Multifunction device drivers
         -> Maxim 5970/5978 power switch and monitor (MFD_MAX5970 [=n])
 Selects: MFD_SIMPLE_MFD_I2C [=n]

---

 Symbol: MFD_MAX77541 [=n]
 Type  : tristate
 Defined at drivers/mfd/Kconfig:787
   Prompt: Analog Devices MAX77541/77540 PMIC Support
   Depends on: HAS_IOMEM [=y] && I2C [=y]=y
   Location:
     -> Device Drivers
       -> Multifunction device drivers
         -> Analog Devices MAX77541/77540 PMIC Support (MFD_MAX77541 [=n])
 Selects: MFD_CORE [=y] && REGMAP_I2C [=m] && REGMAP_IRQ [=y]

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>