[redhat] New configs in drivers/leds
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: LEDS_AW200XX [=n]
 Type  : tristate
 Defined at drivers/leds/Kconfig:97
   Prompt: LED support for Awinic AW20036/AW20054/AW20072
   Depends on: NEW_LEDS [=y] && LEDS_CLASS [=y] && I2C [=y]
   Location:
     -> Device Drivers
       -> LED Support (NEW_LEDS [=y])
         -> LED support for Awinic AW20036/AW20054/AW20072 (LEDS_AW200XX [=n])

---

 Symbol: LEDS_LM3697 [=n]
 Type  : tristate
 Defined at drivers/leds/Kconfig:841
   Prompt: LED driver for LM3697
   Depends on: NEW_LEDS [=y] && LEDS_CLASS [=y] && I2C [=y] && OF [=y]
   Location:
     -> Device Drivers
       -> LED Support (NEW_LEDS [=y])
         -> LED driver for LM3697 (LEDS_LM3697 [=n])
 Selects: LEDS_TI_LMU_COMMON [=n]

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>