[redhat] New configs in samples/Kconfig
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: SAMPLE_TPS6594_PFSM [=n]
 Type  : bool
 Defined at samples/Kconfig:256
   Prompt: Build example program working with TPS6594 PFSM driver
   Depends on: SAMPLES [=y] && HEADERS_INSTALL [=y] && CC_CAN_LINK [=y]
   Location:
     -> Kernel hacking
       -> Sample kernel code (SAMPLES [=y])
         -> Build example program working with TPS6594 PFSM driver (SAMPLE_TPS6594_PFSM [=n])

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>