redhat: fix the 'eln BUILD_TARGET' self-test
This test is supposed to confirm that when DIST is set to ".eln" the
BUILD_TARGET will be set to "eln" but the check is inversed. Fix it.

Also, since this test is typically invoked by make, the BUILD_TARGET,
DIST, & DISTRO environment vars will already be set. Then when the test
calls `make dist-dump-variables` it finds these existing values instead
of deriving new ones. So if the test is run from a fedora system it
will fail because setting DIST doesn't have the expected effect when
BUILD_TARGET & DISTRO are already set.

Fix this by having the test first unset BUILD_TARGET and DISTRO.

Signed-off-by: Patrick Talbert <ptalbert@redhat.com>