[redhat] New configs in drivers/firewire
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: FIREWIRE_KUNIT_UAPI_TEST [=m]
 Type  : tristate
 Defined at drivers/firewire/Kconfig:21
   Prompt: KUnit tests for layout of structure in UAPI
   Depends on: (PCI [=y] || COMPILE_TEST [=n]) && FIREWIRE [=m] && KUNIT [=m]
   Visible if: (PCI [=y] || COMPILE_TEST [=n]) && FIREWIRE [=m] && KUNIT [=m] && !KUNIT_ALL_TESTS [=m]
   Location:
     -> Device Drivers
       -> IEEE 1394 (FireWire) support
         -> FireWire driver stack (FIREWIRE [=m])
           -> KUnit tests for layout of structure in UAPI (FIREWIRE_KUNIT_UAPI_TEST [=m])

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>

v2:
- CONFIG_FIREWIRE_KUNIT_UAPI_TEST is disabled for rhel.

Signed-off-by: Patrick Talbert <ptalbert@redhat.com>