[redhat] New configs in drivers/media
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: VIDEO_OV01A10 [=n]
 Type  : tristate
 Defined at drivers/media/i2c/Kconfig:341
   Prompt: OmniVision OV01A10 sensor support
   Depends on: MEDIA_SUPPORT [=m] && VIDEO_DEV [=m] && I2C [=y]
   Visible if: MEDIA_SUPPORT [=m] && VIDEO_DEV [=m] && I2C [=y] && MEDIA_CAMERA_SUPPORT [=y]
   Location:
     -> Device Drivers
       -> Multimedia support (MEDIA_SUPPORT [=m])
         -> Media ancillary drivers
           -> Camera sensor devices
             -> OmniVision OV01A10 sensor support (VIDEO_OV01A10 [=n])
 Selects: MEDIA_CONTROLLER [=y] && VIDEO_V4L2_SUBDEV_API [=n] && V4L2_FWNODE [=n]

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>