[redhat] New configs in fs/autofs
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: AUTOFS_FS [=n]
 Type  : tristate
 Defined at fs/autofs/Kconfig:2
   Prompt: Kernel automounter support (supports v3, v4 and v5)
   Location:
     -> File systems
       -> Kernel automounter support (supports v3, v4 and v5) (AUTOFS_FS [=n])

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>

v2:
- CONFIG_AUTOFS_FS enable in common.
- CONFIG_AUTOFS4_FS delete files as this config no longer exists.

Signed-off-by: Patrick Talbert <ptalbert@redhat.com>