redhat/configs: Remove `CONFIG_HZ_1000 is not set` for aarch64
Historically, the value of CONFIG_HZ has been set to 100 on aarch64
platforms. However, this is not a good value to use for interactive devices.
1000 Hz is the preferred choice for systems requiring fast interactive responses
to events. Apparently, there is no reason why this is not set for aarch64
like it is for x86_64, so, unless someone have a reason to maintain the
100 value by default for Fedora kernels, remove the `CONFIG_HZ_1000 is not set`
option for aarch64 and use the common option.

Signed-off-by: Enric Balletbo i Serra <eballetbo@redhat.com>