redhat/self-test: Remove rpmlint test
rpmlint reports errors and warnings on syntax and variable usage in spec
files.  Hundreds of errors are reported when rpmlint is run on the
kernel.spec.  These errors are for everything from RPM_BUILD_ROOT usage to
listing unversioned package dependencies.  These errors are all caused by
explicit design choices by the kernel maintainers and spec authors.

For example, the use of RPM_BUILD_ROOT is required to move kernel compile
artificats from one directory to another in the rpm build phase which
seems to be an exception called out in rpmlint (but is still reported as
an error).

These errors can be blocked but only at a global level; that is, if error
reporting for these issues were blocked then new issues would not be
reported.

The rpmlint test failures don't offer new insight into the kernel spec.
For now, remove the rpmlint test from the self-tests.

Signed-off-by: Prarit Bhargava <prarit@redhat.com>