Remove the armv7 config directory from Fedora again
While the MR to remove armv7 got rid of everything, I had to create a
directory temporarily to allow the CONFIG_EXPERT merge to succeed
without forcing a rebase. It is in now, so we can delete the armv7
directory again, and all of the config items that CONFIG_EXPERT put in
there.

Signed-off-by: Justin M. Forbes <jforbes@fedoraproject.org>