Add clang config framework
Compiling for clang is getting trickier to do because the configs are
different than gcc in some cases.  Mimic the kgcov framework (which uses
priority files) to setup clang to be easier to update their configs.

This leads to strange cases where clang debug configs need different
settings than clang non-debug configs.  The priority.snip can handle the
variety of partial snippets for variants and arches well, the
kernel.spec needed to be updated to apply it reasonably.

I created a generic partial snippet loop that focuses on arch and
non-debug vs debug.  This required adding debug options to
priority.snip.  Once implemented it is a matter of collecting the
snippets and applying them based on arch and debug.  This should provide
a variety of flexibility for teams like clang and make it easier to roll
in config changes based on which compiler is used.

Signed-off-by: Don Zickus <dzickus@redhat.com>