Remove unpackaged kgcov config files
The kgcov option is applied as a partial snippet in the spec file.  The
generated kernel-*-kgcov.config are not packaged and therefore unused.

My motivation here is because they conflict with the CLANG options when
I overlay the CLANG options on the gcov config file.  Instead of special
casing them, let's just drop them.

Signed-off-by: Don Zickus <dzickus@redhat.com>