[redhat] New configs in arch/s390
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: CERT_STORE [=n]
 Type  : bool
 Defined at arch/s390/Kconfig:516
   Prompt: Get user certificates via DIAG320
   Depends on: KEYS [=y]
   Location:
     -> Processor type and features
       -> Get user certificates via DIAG320 (CERT_STORE [=n])
 Selects: CRYPTO_LIB_SHA256 [=y]

---

 Symbol: S390_HYPFS [=y]
 Type  : bool
 Defined at arch/s390/Kconfig:882
   Prompt: s390 hypervisor information
   Location:
     -> Virtualization
       -> s390 hypervisor information (S390_HYPFS [=y])

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>

v2:
- CONFIG_CERT_STORE enabled and moved under s390x
- CONFIG_S390_HYPFS moved under s390x

Signed-off-by: Patrick Talbert <ptalbert@redhat.com>