[redhat] New configs in fs/pstore
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: PSTORE_COMPRESS [=y]
 Type  : bool
 Defined at fs/pstore/Kconfig:24
   Prompt: Pstore compression (deflate)
   Depends on: MISC_FILESYSTEMS [=y] && PSTORE [=y]
   Location:
     -> File systems
       -> Miscellaneous filesystems (MISC_FILESYSTEMS [=y])
         -> Persistent store support (PSTORE [=y])
           -> Pstore compression (deflate) (PSTORE_COMPRESS [=y])
 Selects: ZLIB_INFLATE [=y] && ZLIB_DEFLATE [=y]

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>