[redhat] New configs in fs/Kconfig
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: TMPFS_QUOTA [=n]
 Type  : bool
 Defined at fs/Kconfig:236
   Prompt: Tmpfs quota support
   Depends on: TMPFS [=y]
   Location:
     -> File systems
       -> Pseudo filesystems
         -> Tmpfs virtual memory file system support (former shm fs) (TMPFS [=y])
           -> Tmpfs quota support (TMPFS_QUOTA [=n])
 Selects: QUOTA [=y]

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>