[redhat] New configs in arch/powerpc
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: CRYPTO_CHACHA20_P10 [=n]
 Type  : tristate
 Defined at arch/powerpc/crypto/Kconfig:114
   Prompt: Ciphers: ChaCha20, XChacha20, XChacha12 (P10 or later)
   Depends on: CRYPTO [=y] && !KMSAN [=n] && PPC [=y] && PPC64 [=y] && CPU_LITTLE_ENDIAN [=y]
   Location:
     -> Cryptographic API (CRYPTO [=y])
       -> Accelerated Cryptographic Algorithms for CPU (powerpc)
         -> Ciphers: ChaCha20, XChacha20, XChacha12 (P10 or later) (CRYPTO_CHACHA20_P10 [=n])
 Selects: CRYPTO_SKCIPHER [=y] && CRYPTO_LIB_CHACHA_GENERIC [=y] && CRYPTO_ARCH_HAVE_LIB_CHACHA [=n]

---

 Symbol: CRYPTO_POLY1305_P10 [=n]
 Type  : tristate
 Defined at arch/powerpc/crypto/Kconfig:128
   Prompt: Hash functions: Poly1305 (P10 or later)
   Depends on: CRYPTO [=y] && !KMSAN [=n] && PPC [=y] && PPC64 [=y] && CPU_LITTLE_ENDIAN [=y]
   Location:
     -> Cryptographic API (CRYPTO [=y])
       -> Accelerated Cryptographic Algorithms for CPU (powerpc)
         -> Hash functions: Poly1305 (P10 or later) (CRYPTO_POLY1305_P10 [=n])
 Selects: CRYPTO_HASH [=y] && CRYPTO_LIB_POLY1305_GENERIC [=y]

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>