[redhat] New configs in drivers/base
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: REGMAP_BUILD [=n]
 Type  : bool
 Defined at drivers/base/regmap/Kconfig:26
   Prompt: Enable regmap build
   Depends on: KUNIT [=m]
   Location:
     -> Device Drivers
       -> Generic Driver Options
         -> Enable regmap build (REGMAP_BUILD [=n])
 Selects: REGMAP [=y]

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>