[redhat] New configs in drivers/md
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: MD_BITMAP_FILE [=y]
 Type  : bool
 Defined at drivers/md/Kconfig:54
   Prompt: MD bitmap file support (deprecated)
   Depends on: MD [=y]
   Location:
     -> Device Drivers
       -> Multiple devices driver support (RAID and LVM) (MD [=y])
         -> MD bitmap file support (deprecated) (MD_BITMAP_FILE [=y])

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>