[redhat] New configs in drivers/net/ethernet/mellanox
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: MLX5_MACSEC [=n]
 Type  : bool
 Defined at drivers/net/ethernet/mellanox/mlx5/core/Kconfig:143
   Prompt: Connect-X support for MACSec offload
   Depends on: NETDEVICES [=y] && ETHERNET [=y] && NET_VENDOR_MELLANOX [=y] && MLX5_CORE_EN [=y] && MACSEC [=m]
   Location:
     -> Device Drivers
       -> Network device support (NETDEVICES [=y])
         -> Ethernet driver support (ETHERNET [=y])
           -> Mellanox devices (NET_VENDOR_MELLANOX [=y])
             -> Mellanox 5th generation network adapters (ConnectX series) core driver (MLX5_CORE [=m])
               -> Mellanox 5th generation network adapters (ConnectX series) Ethernet support (MLX5_CORE_EN [=y])
                 -> Connect-X support for MACSec offload (MLX5_MACSEC [=n])

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>

v2:
- CONFIG_MLX5_MACSEC enable for rhel

Signed-off-by: Patrick Talbert <ptalbert@redhat.com>