[redhat] New configs in drivers/net/Kconfig
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: NETCONSOLE_EXTENDED_LOG [=n]
 Type  : bool
 Defined at drivers/net/Kconfig:335
   Prompt: Set kernel extended message by default
   Depends on: NETDEVICES [=y] && NET_CORE [=y] && NETCONSOLE [=m]
   Location:
     -> Device Drivers
       -> Network device support (NETDEVICES [=y])
         -> Network core driver support (NET_CORE [=y])
           -> Network console logging support (NETCONSOLE [=m])
             -> Set kernel extended message by default (NETCONSOLE_EXTENDED_LOG [=n])

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>