[redhat] New configs in kernel/Kconfig.kexec
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: CRASH_HOTPLUG [=y]
 Type  : bool
 Defined at kernel/Kconfig.kexec:115
   Prompt: Update the crash elfcorehdr on system configuration changes
   Depends on: CRASH_DUMP [=y] && (HOTPLUG_CPU [=y] || MEMORY_HOTPLUG [=y]) && ARCH_SUPPORTS_CRASH_HOTPLUG [=y]
   Location:
     -> General setup
       -> Kexec and crash features
         -> kernel crash dumps (CRASH_DUMP [=y])
           -> Update the crash elfcorehdr on system configuration changes (CRASH_HOTPLUG [=y])

---

 Symbol: CRASH_MAX_MEMORY_RANGES [=8192]
 Type  : integer
 Defined at kernel/Kconfig.kexec:129
   Prompt: Specify the maximum number of memory regions for the elfcorehdr
   Depends on: CRASH_HOTPLUG [=y]
   Location:
     -> General setup
       -> Kexec and crash features
         -> kernel crash dumps (CRASH_DUMP [=y])
           -> Update the crash elfcorehdr on system configuration changes (CRASH_HOTPLUG [=y])
             -> Specify the maximum number of memory regions for the elfcorehdr (CRASH_MAX_MEMORY_RANGES [=8192])

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>

v2:
- CONFIG_CRASH_HOTPLUG moved under x86.
- CONFIG_CRASH_MAX_MEMORY_RANGES moved under x86.

Signed-off-by: Patrick Talbert <ptalbert@redhat.com>