[redhat] New configs in drivers/clk
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: COMMON_CLK_VC3 [=n]
 Type  : tristate
 Defined at drivers/clk/Kconfig:381
   Prompt: Clock driver for Renesas VersaClock 3 devices
   Depends on: COMMON_CLK [=y] && I2C [=y] && OF [=y]
   Location:
     -> Device Drivers
       -> Common Clock Framework (COMMON_CLK [=y])
         -> Clock driver for Renesas VersaClock 3 devices (COMMON_CLK_VC3 [=n])
 Selects: REGMAP_I2C [=m]

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>