[redhat] New configs in drivers/video
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: FB_DEVICE [=y]
 Type  : bool
 Defined at drivers/video/fbdev/core/Kconfig:28
   Prompt: Provide legacy /dev/fb* device
   Depends on: HAS_IOMEM [=y] && FB_CORE [=y]
   Location:
     -> Device Drivers
       -> Graphics support
         -> Frame buffer Devices
           -> Provide legacy /dev/fb* device (FB_DEVICE [=y])

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>