[redhat] New configs in fs/overlayfs
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: OVERLAY_FS_DEBUG [=n]
 Type  : bool
 Defined at fs/overlayfs/Kconfig:128
   Prompt: Overlayfs: turn on extra debugging checks
   Depends on: OVERLAY_FS [=m]
   Location:
     -> File systems
       -> Overlay filesystem support (OVERLAY_FS [=m])
         -> Overlayfs: turn on extra debugging checks (OVERLAY_FS_DEBUG [=n])

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>

v2:
- CONFIG_OVERLAY_FS_DEBUG enable for rhel debug.

Signed-off-by: Patrick Talbert <ptalbert@redhat.com>