[redhat] New configs in arch/x86
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: X86_USER_SHADOW_STACK [=n]
 Type  : bool
 Defined at arch/x86/Kconfig:1924
   Prompt: X86 userspace shadow stack
   Depends on: AS_WRUSS [=y] && X86_64 [=y]
   Location:
     -> Processor type and features
       -> X86 userspace shadow stack (X86_USER_SHADOW_STACK [=n])
 Selects: ARCH_USES_HIGH_VMA_FLAGS [=y] && X86_CET [=y]

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>