[redhat] New configs in drivers/platform
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: SEL3350_PLATFORM [=n]
 Type  : tristate
 Defined at drivers/platform/x86/Kconfig:1087
   Prompt: SEL-3350 LEDs and power supplies
   Depends on: X86_PLATFORM_DEVICES [=y] && ACPI [=y] && GPIOLIB [=y] && PINCTRL_BROXTON [=m]
   Location:
     -> Device Drivers
       -> X86 Platform Specific Device Drivers (X86_PLATFORM_DEVICES [=y])
         -> SEL-3350 LEDs and power supplies (SEL3350_PLATFORM [=n])
 Selects: POWER_SUPPLY [=y] && NEW_LEDS [=y] && LEDS_CLASS [=y] && LEDS_GPIO [=n]

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>