[redhat] New configs in drivers/phy
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: PHY_QCOM_M31_USB [=n]
 Type  : tristate
 Defined at drivers/phy/qualcomm/Kconfig:146
   Prompt: Qualcomm M31 HS PHY driver support
   Depends on: USB [=y] && (ARCH_QCOM [=y] || COMPILE_TEST [=n])
   Location:
     -> Device Drivers
       -> PHY Subsystem
         -> Qualcomm M31 HS PHY driver support (PHY_QCOM_M31_USB [=n])
 Selects: GENERIC_PHY [=y]

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>