[redhat] New configs in drivers/leds
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: LEDS_PCA995X [=n]
 Type  : tristate
 Defined at drivers/leds/Kconfig:524
   Prompt: LED Support for PCA995x I2C chips
   Depends on: NEW_LEDS [=y] && LEDS_CLASS [=y] && I2C [=y]
   Location:
     -> Device Drivers
       -> LED Support (NEW_LEDS [=y])
         -> LED Support for PCA995x I2C chips (LEDS_PCA995X [=n])

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>