[redhat] New configs in fs/xfs
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: XFS_ASSERT_FATAL [=y]
 Type  : bool
 Defined at fs/xfs/Kconfig:207
   Prompt: XFS fatal asserts
   Depends on: BLOCK [=y] && XFS_FS [=m] && XFS_DEBUG [=y]
   Location:
     -> File systems
       -> XFS Debugging support (XFS_DEBUG [=y])
         -> XFS fatal asserts (XFS_ASSERT_FATAL [=y])

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>