redhat: remove pending-rhel CONFIG_XFS_ASSERT_FATAL file
The CONFIG_XFS_ASSERT_FATAL config came up because the new XFS_ONLINE_SCRUB_STATS
config item defaulted to being enabled and it selects FS_DEBUG.

We've updated MR !2686 so XFS_ONLINE_SCRUB_STATS is no longer going to
be enabled, so FS_DEBUG isn't going to be enabled and CONFIG_XFS_ASSERT_FATAL
becomes a moot point.

So for now we remove the pending file and do not need to set this.

Signed-off-by: Patrick Talbert <ptalbert@redhat.com>