[redhat] New configs in arch/arm64
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: ARM64_ERRATUM_2966298 [=y]
 Type  : bool
 Defined at arch/arm64/Kconfig:1040
   Prompt: Cortex-A520: 2966298: workaround for speculatively executed unprivileged load
   Location:
     -> Kernel Features
       -> ARM errata workarounds via the alternatives framework
         -> Cortex-A520: 2966298: workaround for speculatively executed unprivileged load (ARM64_ERRATUM_2966298 [=y])

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>