[redhat] New configs in arch/x86
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: IA32_EMULATION_DEFAULT_DISABLED [=n]
 Type  : bool
 Defined at arch/x86/Kconfig:2958
   Prompt: IA32 emulation disabled by default
   Depends on: IA32_EMULATION [=y]
   Location:
     -> Binary Emulations
       -> IA32 Emulation (IA32_EMULATION [=y])
         -> IA32 emulation disabled by default (IA32_EMULATION_DEFAULT_DISABLED [=n])

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>