Trim changelog after version bump
For some reason, after an upstream version bump, the changelog decides
to reset itself and add an entry for every commit since the repo was
started. I have to then manually delete the old entries as we start a
new cycle.  At some point, we should really fix the changelog scripts.

Signed-off-by: Justin M. Forbes <jforbes@fedoraproject.org>