[redhat] New configs in drivers/hwmon
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: SENSORS_LTC2991 [=n]
 Type  : tristate
 Defined at drivers/hwmon/Kconfig:945
   Prompt: Analog Devices LTC2991
   Depends on: HWMON [=y] && I2C [=y]
   Location:
     -> Device Drivers
       -> Hardware Monitoring support (HWMON [=y])
         -> Analog Devices LTC2991 (SENSORS_LTC2991 [=n])

---

 Symbol: SENSORS_POWERZ [=n]
 Type  : tristate
 Defined at drivers/hwmon/Kconfig:842
   Prompt: ChargerLAB POWER-Z USB-C tester
   Depends on: HWMON [=y] && USB [=y]
   Location:
     -> Device Drivers
       -> Hardware Monitoring support (HWMON [=y])
         -> ChargerLAB POWER-Z USB-C tester (SENSORS_POWERZ [=n])

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>