[redhat] New configs in drivers/net/ethernet/intel
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: IDPF [=n]
 Type  : tristate
 Defined at drivers/net/ethernet/intel/Kconfig:360
   Prompt: Intel(R) Infrastructure Data Path Function Support
   Depends on: NETDEVICES [=y] && ETHERNET [=y] && NET_VENDOR_INTEL [=y] && PCI_MSI [=y]
   Location:
     -> Device Drivers
       -> Network device support (NETDEVICES [=y])
         -> Ethernet driver support (ETHERNET [=y])
           -> Intel devices (NET_VENDOR_INTEL [=y])
             -> Intel(R) Infrastructure Data Path Function Support (IDPF [=n])
 Selects: DIMLIB [=y] && PAGE_POOL [=y] && PAGE_POOL_STATS [=y]

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>

v2:
- CONFIG_IDPF enable as a module.

Signed-off-by: Patrick Talbert <ptalbert@redhat.com>