[redhat] New configs in drivers/regulator
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: REGULATOR_MAX77503 [=n]
 Type  : tristate
 Defined at drivers/regulator/Kconfig:567
   Prompt: Analog Devices MAX77503 Regulator
   Depends on: REGULATOR [=y] && I2C [=y]
   Location:
     -> Device Drivers
       -> Voltage and Current Regulator Support (REGULATOR [=y])
         -> Analog Devices MAX77503 Regulator (REGULATOR_MAX77503 [=n])
 Selects: REGMAP_I2C [=y]

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>