[redhat] New configs in net/wireless
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: CFG80211_KUNIT_TEST [=m]
 Type  : tristate
 Defined at net/wireless/Kconfig:204
   Prompt: KUnit tests for cfg80211
   Depends on: NET [=y] && WIRELESS [=y] && KUNIT [=m] && CFG80211 [=m] && !KERNEL_6_2
   Visible if: NET [=y] && WIRELESS [=y] && KUNIT [=m] && CFG80211 [=m] && !KERNEL_6_2 && !KUNIT_ALL_TESTS [=m]
   Location:
     -> Networking support (NET [=y])
       -> Wireless (WIRELESS [=y])
         -> cfg80211 - wireless configuration API (CFG80211 [=m])
           -> KUnit tests for cfg80211 (CFG80211_KUNIT_TEST [=m])

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>