[redhat] New configs in arch/arm64
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: ARCH_PENSANDO [=n]
 Type  : bool
 Defined at arch/arm64/Kconfig.platforms:247
   Prompt: AMD Pensando Platforms
   Location:
     -> Platform selection
       -> AMD Pensando Platforms (ARCH_PENSANDO [=n])

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>

v2:
- CONFIG_ARCH_PENSANDO move to arm/aarch64 and enable.

Signed-off-by: Patrick Talbert <ptalbert@redhat.com>