[redhat] New configs in arch/x86
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: KVM_MAX_NR_VCPUS [=4096]
 Type  : integer
 Range : [1024 4096]
 Defined at arch/x86/kvm/Kconfig:157
   Prompt: Maximum number of vCPUs per KVM guest
   Depends on: VIRTUALIZATION [=y] && KVM [=m]
   Location:
     -> Virtualization (VIRTUALIZATION [=y])
       -> Maximum number of vCPUs per KVM guest (KVM_MAX_NR_VCPUS [=4096])

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>