[redhat] New configs in drivers/pci
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: HOTPLUG_PCI_ACPI_AMPERE_ALTRA [=n]
 Type  : tristate
 Defined at drivers/pci/hotplug/Kconfig:64
   Prompt: ACPI PCI Hotplug driver Ampere Altra extensions
   Depends on: PCI [=y] && HOTPLUG_PCI [=y] && HOTPLUG_PCI_ACPI [=y] && HAVE_ARM_SMCCC_DISCOVERY [=y]
   Location:
     -> Device Drivers
       -> PCI support (PCI [=y])
         -> Support for PCI Hotplug (HOTPLUG_PCI [=y])
           -> ACPI PCI Hotplug driver (HOTPLUG_PCI_ACPI [=y])
             -> ACPI PCI Hotplug driver Ampere Altra extensions (HOTPLUG_PCI_ACPI_AMPERE_ALTRA [=n])

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>