[redhat] New configs in drivers/pinctrl
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: PINCTRL_CS42L43 [=n]
 Type  : tristate
 Defined at drivers/pinctrl/cirrus/Kconfig:2
   Prompt: Cirrus Logic CS42L43 Pinctrl Driver
   Depends on: PINCTRL [=y] && MFD_CS42L43 [=m]
   Location:
     -> Device Drivers
       -> Pin controllers (PINCTRL [=y])
         -> Cirrus Logic CS42L43 Pinctrl Driver (PINCTRL_CS42L43 [=n])
 Selects: GPIOLIB [=y] && PINMUX [=y] && PINCONF [=y] && GENERIC_PINCONF [=y]

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>