[redhat] New configs in mm/Kconfig
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: PCP_BATCH_SCALE_MAX [=5]
 Type  : integer
 Range : [0 6]
 Defined at mm/Kconfig:708
   Prompt: Maximum scale factor of PCP (Per-CPU pageset) batch allocate/free
   Location:
     -> Memory Management options
       -> Maximum scale factor of PCP (Per-CPU pageset) batch allocate/free (PCP_BATCH_SCALE_MAX [=5])

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>