[redhat] New configs in arch/powerpc
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: GUEST_STATE_BUFFER_TEST [=n]
 Type  : tristate
 Defined at arch/powerpc/Kconfig.debug:85
   Prompt: Enable Guest State Buffer unit tests
   Depends on: KUNIT [=m] && KVM_BOOK3S_HV_POSSIBLE [=y]
   Location:
     -> Kernel hacking
       -> powerpc Debugging
         -> Enable Guest State Buffer unit tests (GUEST_STATE_BUFFER_TEST [=n])

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>