[redhat] New configs in drivers/iio
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: LTC2309 [=n]
 Type  : tristate
 Defined at drivers/iio/adc/Kconfig:610
   Prompt: Linear Technology LTC2309 ADC driver
   Depends on: IIO [=m] && I2C [=y]
   Location:
     -> Device Drivers
       -> Industrial I/O support (IIO [=m])
         -> Analog to digital converters
           -> Linear Technology LTC2309 ADC driver (LTC2309 [=n])

---

 Symbol: MCP3564 [=n]
 Type  : tristate
 Defined at drivers/iio/adc/Kconfig:792
   Prompt: Microchip Technology MCP3461/2/4/R, MCP3561/2/4/R driver
   Depends on: SPI [=y] && IIO [=m]
   Location:
     -> Device Drivers
       -> Industrial I/O support (IIO [=m])
         -> Analog to digital converters
           -> Microchip Technology MCP3461/2/4/R, MCP3561/2/4/R driver (MCP3564 [=n])

---

 Symbol: ROHM_BM1390 [=n]
 Type  : tristate
 Defined at drivers/iio/pressure/Kconfig:19
   Prompt: ROHM BM1390GLV-Z pressure sensor driver
   Depends on: IIO [=m] && I2C [=y]
   Location:
     -> Device Drivers
       -> Industrial I/O support (IIO [=m])
         -> Pressure sensors
           -> ROHM BM1390GLV-Z pressure sensor driver (ROHM_BM1390 [=n])

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>