redhat: rh_kabi: introduce RH_KABI_EXCLUDE_WITH_SIZE
This macro is similar to RH_KABI_EXTEND_WITH_SIZE, but with a
different intention.  While RH_KABI_EXTEND_WITH_SIZE is meant to add
new elements that may grow in the future to kABI-protected structures,
RH_KABI_EXCLUDE_WITH_SIZE should be used to hide existing elements
from kABI checksum computation, while allowing them to be extended in
the future without silently breaking binary compatibility of any field
that follows it.

The implementation differs from RH_KABI_EXTEND_WITH_SIZE so that _size
is included in the kABI checksum. That way, we can't accidentally grow
the reserved space, as that would change the offsets of all the
following fields.

The size check guarantees that binary compatibility is preserved.
Until now, __RH_KABI_CHECK_SIZE was only used in the !__GENKSYMS__
versions of macros, so it wasn't defined in the __GENKSYMS__
case. We need to define it to nothing so that _new is ignored by
checksum computation.

Signed-off-by: Sabrina Dubroca <sdubroca@redhat.com>