Flip CONFIG_NVME_AUTH to m in pending
Upstream commit d680063482885 changed quite a bit about how the
NVME_AUTH Kconfig works.  At this point, it seems that NVME_AUTH must be
a module, or all of the crypto options it depends upon must be inline.
I will open a proper MR to address this outside of pending.

Signed-off-by: Justin M. Forbes <jforbes@fedoraproject.org>