redhat/self-test: Remove --all from git query
It was reported that executing this test on an older tag failed.  This
occurs because the git query to find the most recent "[redhat] kernel"
commit looks at the entire git history instead of the branch's history.

Removing the '--all' from the "[redhat] kernel" lookup resolves this
problem.

Testing:

There are 3 cases that I considered which all succeed with this change:

1.  running on os-build or ark-latest

        git co origin/os-build
        RHEL_MAJOR=9 RHEL_MINOR=99 bats redhat/self-test/2001-dist-release.bats

2.  a new branch based on os-build/ark-latest

        git co origin/ark-latest
        git checkout -b junk
        /# make some changes
        RHEL_MAJOR=9 RHEL_MINOR=99 bats redhat/self-test/2001-dist-release.bats

3.  an older branch based on a tag

        git co kernel-6.6.0-0.rc7.56567a20b22b.59
        RHEL_MAJOR=9 RHEL_MINOR=99 bats redhat/self-test/2001-dist-release.bats

Signed-off-by: Prarit Bhargava <prarit@redhat.com>