[redhat] New configs in block/Kconfig
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: BLK_DEV_WRITE_MOUNTED [=y]
 Type  : bool
 Defined at block/Kconfig:81
   Prompt: Allow writing to mounted block devices
   Depends on: BLOCK [=y]
   Location:
     -> Enable the block layer (BLOCK [=y])
       -> Allow writing to mounted block devices (BLK_DEV_WRITE_MOUNTED [=y])

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>