[redhat] New configs in lib/Kconfig
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: STACKDEPOT_MAX_FRAMES [=64]
 Type  : integer
 Range : [1 256]
 Defined at lib/Kconfig:725
   Prompt: Maximum number of frames in trace saved in stack depot
   Depends on: STACKDEPOT [=y]
   Location:
     -> Library routines
       -> Maximum number of frames in trace saved in stack depot (STACKDEPOT_MAX_FRAMES [=64])

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>