[redhat] New configs in lib/Kconfig.kasan
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: KASAN_EXTRA_INFO [=n]
 Type  : bool
 Defined at lib/Kconfig.kasan:205
   Prompt: Record and report more information
   Depends on: KASAN [=y]
   Location:
     -> Kernel hacking
       -> Memory Debugging
         -> KASAN: dynamic memory safety error detector (KASAN [=y])
           -> Record and report more information (KASAN_EXTRA_INFO [=n])

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>