[redhat] New configs in samples/Kconfig
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: SAMPLE_CGROUP [=n]
 Type  : bool
 Defined at samples/Kconfig:288
   Prompt: Build cgroup sample code
   Depends on: SAMPLES [=y] && CGROUPS [=y] && CC_CAN_LINK [=y] && HEADERS_INSTALL [=y]
   Location:
     -> Kernel hacking
       -> Sample kernel code (SAMPLES [=y])
         -> Build cgroup sample code (SAMPLE_CGROUP [=n])

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>