[redhat] New configs in drivers/net/ethernet/intel
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: ICE_HWMON [=y]
 Type  : bool
 Defined at drivers/net/ethernet/intel/Kconfig:302
   Prompt: Intel(R) Ethernet Connection E800 Series Support HWMON support
   Depends on: NETDEVICES [=y] && ETHERNET [=y] && NET_VENDOR_INTEL [=y] && ICE [=m] && HWMON [=y] && (ICE [=m]!=y || HWMON [=y]!=m)
   Location:
     -> Device Drivers
       -> Network device support (NETDEVICES [=y])
         -> Ethernet driver support (ETHERNET [=y])
           -> Intel devices (NET_VENDOR_INTEL [=y])
             -> Intel(R) Ethernet Connection E800 Series Support (ICE [=m])
               -> Intel(R) Ethernet Connection E800 Series Support HWMON support (ICE_HWMON [=y])

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>