[redhat] New configs in drivers/net/wan
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: FRAMER [=n]
 Type  : tristate
 Defined at drivers/net/wan/framer/Kconfig:6
   Prompt: Framer Subsystem
   Depends on: NETDEVICES [=y] && WAN [=y]
   Location:
     -> Device Drivers
       -> Network device support (NETDEVICES [=y])
         -> Wan interfaces support (WAN [=y])
           -> Framer Subsystem (FRAMER [=n])

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>