Fix up manual merge error in security.h
When doing the manual merge fix up I seem to have deleted a closing }.
This commit corrects that error, and brings the "security: lockdown:
expose a hook to lock the kernel down" change back to what it was
supposed to be. In rebase, this can be fixed correctly and this commit
dropped.

Reported-by: Scott Weaver <scweaver@redhat.com>
Signed-off-by: Justin M. Forbes <jforbes@fedoraproject.org>