[redhat] New configs in arch/x86
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: KVM_HYPERV [=y]
 Type  : bool
 Defined at arch/x86/kvm/Kconfig:141
   Prompt: Support for Microsoft Hyper-V emulation
   Depends on: VIRTUALIZATION [=y] && KVM [=m]
   Location:
     -> Virtualization (VIRTUALIZATION [=y])
       -> Kernel-based Virtual Machine (KVM) support (KVM [=m])
         -> Support for Microsoft Hyper-V emulation (KVM_HYPERV [=y])

---

 Symbol: KVM_SW_PROTECTED_VM [=n]
 Type  : bool
 Defined at arch/x86/kvm/Kconfig:77
   Prompt: Enable support for KVM software-protected VMs
   Depends on: VIRTUALIZATION [=y] && EXPERT [=y] && KVM [=m] && X86_64 [=y]
   Location:
     -> Virtualization (VIRTUALIZATION [=y])
       -> Kernel-based Virtual Machine (KVM) support (KVM [=m])
         -> Enable support for KVM software-protected VMs (KVM_SW_PROTECTED_VM [=n])
 Selects: KVM_GENERIC_PRIVATE_MEM [=n]

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>