[redhat] New configs in drivers/misc
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: INTEL_MEI_VSC_HW [=n]
 Type  : tristate
 Defined at drivers/misc/mei/Kconfig:61
   Prompt: Intel visual sensing controller device transport driver
   Depends on: INTEL_MEI [=m] && ACPI [=y] && SPI [=y] && (GPIOLIB [=y] || COMPILE_TEST [=n])
   Location:
     -> Device Drivers
       -> Misc devices
         -> Intel Management Engine Interface (INTEL_MEI [=m])
           -> Intel visual sensing controller device transport driver (INTEL_MEI_VSC_HW [=n])

---

 Symbol: NSM [=n]
 Type  : tristate
 Defined at drivers/misc/Kconfig:565
   Prompt: Nitro (Enclaves) Security Module support
   Depends on: VIRTIO [=y]
   Location:
     -> Device Drivers
       -> Misc devices
         -> Nitro (Enclaves) Security Module support (NSM [=n])
 Selects: HW_RANDOM [=y]

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>