[redhat] New configs in drivers/pinctrl
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: PINCTRL_INTEL_PLATFORM [=n]
 Type  : tristate
 Defined at drivers/pinctrl/intel/Kconfig:40
   Prompt: Intel pinctrl and GPIO platform driver
   Depends on: PINCTRL [=y] && (ACPI [=y] && X86 [=y] || COMPILE_TEST [=n]) && ACPI [=y]
   Location:
     -> Device Drivers
       -> Pin controllers (PINCTRL [=y])
         -> Intel pinctrl drivers
           -> Intel pinctrl and GPIO platform driver (PINCTRL_INTEL_PLATFORM [=n])
 Selects: PINCTRL_INTEL [=y]

---

 Symbol: PINCTRL_METEORPOINT [=n]
 Type  : tristate
 Defined at drivers/pinctrl/intel/Kconfig:141
   Prompt: Intel Meteor Point pinctrl and GPIO driver
   Depends on: PINCTRL [=y] && (ACPI [=y] && X86 [=y] || COMPILE_TEST [=n]) && ACPI [=y]
   Location:
     -> Device Drivers
       -> Pin controllers (PINCTRL [=y])
         -> Intel pinctrl drivers
           -> Intel Meteor Point pinctrl and GPIO driver (PINCTRL_METEORPOINT [=n])
 Selects: PINCTRL_INTEL [=y]

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>