[redhat] New configs in drivers/vfio
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: VFIO_DEBUGFS [=n]
 Type  : bool
 Defined at drivers/vfio/Kconfig:83
   Prompt: Export VFIO internals in DebugFS
   Depends on: VFIO [=m] && DEBUG_FS [=y]
   Location:
     -> Device Drivers
       -> VFIO Non-Privileged userspace driver framework (VFIO [=m])
         -> Export VFIO internals in DebugFS (VFIO_DEBUGFS [=n])

---

 Symbol: VIRTIO_VFIO_PCI [=n]
 Type  : tristate
 Defined at drivers/vfio/pci/virtio/Kconfig:2
   Prompt: VFIO support for VIRTIO NET PCI devices
   Depends on: VFIO [=m] && PCI [=y] && MMU [=y] && VIRTIO_PCI [=y] && VIRTIO_PCI_ADMIN_LEGACY [=y]
   Location:
     -> Device Drivers
       -> VFIO Non-Privileged userspace driver framework (VFIO [=m])
         -> VFIO support for PCI devices
           -> VFIO support for VIRTIO NET PCI devices (VIRTIO_VFIO_PCI [=n])
 Selects: VFIO_PCI_CORE [=m]

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>