Enable merge-rt pipeline
Add the gitlab-ci.yaml changes to enable a scheduled pipeline.

Also modify the ark-merge-rt.sh script to use OS_BUILD_BASE as the base
for the rt and automotive branches if they don't already exist.  This
avoids the giant config creation output if we use upstream-rt as the
base and merge in os-build.  The problem with that output is it
overflows the gitlab ci output buffer and you can't see the full output.
This is only seen the first time it is created or when run locally for
the first time.

Signed-off-by: Don Zickus <dzickus@redhat.com>