[redhat] New configs in drivers/iio
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: HDC3020 [=n]
 Type  : tristate
 Defined at drivers/iio/humidity/Kconfig:51
   Prompt: TI HDC3020 relative humidity and temperature sensor
   Depends on: IIO [=m] && I2C [=y]
   Location:
     -> Device Drivers
       -> Industrial I/O support (IIO [=m])
         -> Humidity sensors
           -> TI HDC3020 relative humidity and temperature sensor (HDC3020 [=n])
 Selects: CRC8 [=m]

 Commit: f1dfcbaa7b9d (iio: humidity: hdc3020: Add Makefile, Kconfig and MAINTAINERS entry)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>