[redhat] New configs in drivers/gpu
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: DRM_NOUVEAU_GSP_DEFAULT [=n]
 Type  : bool
 Defined at drivers/gpu/drm/nouveau/Kconfig:104
   Prompt: Use GSP firmware for Turing/Ampere (needs firmware installed)
   Depends on: HAS_IOMEM [=y] && DRM_NOUVEAU [=m]
   Location:
     -> Device Drivers
       -> Graphics support
         -> Nouveau (NVIDIA) cards (DRM_NOUVEAU [=m])
           -> Use GSP firmware for Turing/Ampere (needs firmware installed) (DRM_NOUVEAU_GSP_DEFAULT [=n])

 Commit: 1d492944d3d0 (nouveau/gsp: add kconfig option to enable GSP paths by default)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>