Revert "net: bump CONFIG_MAX_SKB_FRAGS to 45"
This reverts commit 0b16af2cf5bc660974be148795d8f79cad78cb4f.

There are multiple issues triggered by this, this change is not
essencial and the issues are not easy to troubleshoot. Lets just revert
this for now while we debug what went wrong.

One of the issues was reported here:
https://bugzilla.redhat.com/show_bug.cgi?id=2265320

Another was noticed by Sabrina and is related to mlx5, as it triggers this
WARN:

int mlx5e_attach_netdev(struct mlx5e_priv *priv)
{
	...

        /* Validate the max_wqe_size_sq capability. */
        if (WARN_ON_ONCE(mlx5e_get_max_sq_wqebbs(priv->mdev) < MLX5E_MAX_TX_WQEBBS)) {
                mlx5_core_warn(priv->mdev, "MLX5E: Max SQ WQEBBs firmware capability: %u, needed %u\n",
                               mlx5e_get_max_sq_wqebbs(priv->mdev), (unsigned int)MLX5E_MAX_TX_WQEBBS);

[   10.420789] mlx5_core 0000:05:00.0: mlx5e_attach_netdev:5797:(pid 689): MLX5E: Max SQ WQEBBs firmware capability: 16, needed 23

Signed-off-by: Marcelo Ricardo Leitner <mleitner@redhat.com>
Signed-off-by: Sabrina Dubroca <sdubroca@redhat.com>