[redhat] New configs in drivers/bus
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: TEGRA_ACONNECT [=n]
 Type  : tristate
 Defined at drivers/bus/Kconfig:187
   Prompt: Tegra ACONNECT Bus Driver
   Depends on: ARCH_TEGRA [=y] && OF [=y] && PM [=y]
   Location:
     -> Device Drivers
       -> Bus devices
         -> Tegra ACONNECT Bus Driver (TEGRA_ACONNECT [=n])

 Commit: 46a88534afb5 (bus: Add support for Tegra ACONNECT)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>