[redhat] New configs in drivers/net/ethernet/marvell
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: OCTEON_EP_VF [=n]
 Type  : tristate
 Defined at drivers/net/ethernet/marvell/octeon_ep_vf/Kconfig:6
   Prompt: Marvell Octeon PCI Endpoint NIC VF Driver
   Depends on: NETDEVICES [=y] && ETHERNET [=y] && NET_VENDOR_MARVELL [=y] && 64BIT [=y] && PCI [=y]
   Location:
     -> Device Drivers
       -> Network device support (NETDEVICES [=y])
         -> Ethernet driver support (ETHERNET [=y])
           -> Marvell devices (NET_VENDOR_MARVELL [=y])
             -> Marvell Octeon PCI Endpoint NIC VF Driver (OCTEON_EP_VF [=n])

 Commit: cb7dd712189f (octeon_ep_vf: Add driver framework and device initialization)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>