[redhat] New configs in drivers/soc
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: ARCH_TEGRA_241_SOC [=n]
 Type  : bool
 Defined at drivers/soc/tegra/Kconfig:136
   Prompt: NVIDIA Tegra241 SoC
   Depends on: ARCH_TEGRA [=y] && ARM64 [=y]
   Location:
     -> Device Drivers
       -> SOC (System On Chip) specific Drivers
         -> NVIDIA Tegra241 SoC (ARCH_TEGRA_241_SOC [=n])

 Commit: 8402074f3023 (soc/tegra: fuse: Add support for Tegra241)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>

v2:
- CONFIG_ARCH_TEGRA_241_SOC: move under arm/aarch64.

Signed-off-by: Patrick Talbert <ptalbert@redhat.com>