[redhat] New configs in drivers/ptp
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: PTP_1588_CLOCK_FC3W [=n]
 Type  : tristate
 Defined at drivers/ptp/Kconfig:158
   Prompt: RENESAS FemtoClock3 Wireless as PTP clock
   Depends on: PTP_1588_CLOCK [=y] && I2C [=y]
   Location:
     -> Device Drivers
       -> PTP clock support
         -> RENESAS FemtoClock3 Wireless as PTP clock (PTP_1588_CLOCK_FC3W [=n])

 Commit: 1ddfecafabf7 (ptp: add FemtoClock3 Wireless as ptp hardware clock)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>