[redhat] New configs in lib/Kconfig.debug
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: STRING_HELPERS_KUNIT_TEST [=m]
 Type  : tristate
 Defined at lib/Kconfig.debug:2361
   Prompt: KUnit test string helpers at runtime
   Depends on: RUNTIME_TESTING_MENU [=y] && KUNIT [=m]
   Visible if: RUNTIME_TESTING_MENU [=y] && KUNIT [=m] && !KUNIT_ALL_TESTS [=m]
   Location:
     -> Kernel hacking
       -> Kernel Testing and Coverage
         -> Runtime Testing (RUNTIME_TESTING_MENU [=y])
           -> KUnit test string helpers at runtime (STRING_HELPERS_KUNIT_TEST [=m])

 Commit: fb57550fcbd8 (string: Convert helpers selftest to KUnit)

---

 Symbol: STRING_KUNIT_TEST [=m]
 Type  : tristate
 Defined at lib/Kconfig.debug:2356
   Prompt: KUnit test string functions at runtime
   Depends on: RUNTIME_TESTING_MENU [=y] && KUNIT [=m]
   Visible if: RUNTIME_TESTING_MENU [=y] && KUNIT [=m] && !KUNIT_ALL_TESTS [=m]
   Location:
     -> Kernel hacking
       -> Kernel Testing and Coverage
         -> Runtime Testing (RUNTIME_TESTING_MENU [=y])
           -> KUnit test string functions at runtime (STRING_KUNIT_TEST [=m])

 Commit: 29d8568849fe (string: Convert selftest to KUnit)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>