[redhat] New configs in drivers/mmc
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: MMC_DW_HI3798MV200 [=n]
 Type  : tristate
 Defined at drivers/mmc/host/Kconfig:801
   Prompt: Hi3798MV200 specific extensions for Synopsys DW Memory Card Interface
   Depends on: MMC [=m] && MMC_DW [=m]
   Location:
     -> Device Drivers
       -> MMC/SD/SDIO card support (MMC [=m])
         -> Synopsys DesignWare Memory Card Interface (MMC_DW [=m])
           -> Hi3798MV200 specific extensions for Synopsys DW Memory Card Interface (MMC_DW_HI3798MV200 [=n])
 Selects: MMC_DW_PLTFM [=m]

 Commit: 25d043841db1 (mmc: dw_mmc: add support for hi3798mv200)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>