[redhat] New configs in drivers/md
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: DM_VDO [=n]
 Type  : tristate
 Defined at drivers/md/dm-vdo/Kconfig:3
   Prompt: VDO: deduplication and compression target
   Depends on: MD [=y] && 64BIT [=y] && BLK_DEV_DM [=m]
   Location:
     -> Device Drivers
       -> Multiple devices driver support (RAID and LVM) (MD [=y])
         -> Device mapper support (BLK_DEV_DM [=m])
           -> VDO: deduplication and compression target (DM_VDO [=n])
 Selects: DM_BUFIO [=m] && LZ4_COMPRESS [=m] && LZ4_DECOMPRESS [=y]

 Commit: f36b1d3ba533 (dm vdo: use a proper Makefile for dm-vdo)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>