[redhat] New configs in drivers/pinctrl
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: PINCTRL_AW9523 [=n]
 Type  : tristate
 Defined at drivers/pinctrl/Kconfig:130
   Prompt: Awinic AW9523/AW9523B I2C GPIO expander pinctrl driver
   Depends on: PINCTRL [=y] && OF [=y] && I2C [=y]
   Location:
     -> Device Drivers
       -> Pin controllers (PINCTRL [=y])
         -> Awinic AW9523/AW9523B I2C GPIO expander pinctrl driver (PINCTRL_AW9523 [=n])
 Selects: PINMUX [=y] && PINCONF [=y] && GENERIC_PINCONF [=y] && GPIOLIB [=y] && GPIOLIB_IRQCHIP [=y] && REGMAP [=y] && REGMAP_I2C [=y]

 Commit: 576623d70661 (pinctrl: Add driver for Awinic AW9523/B I2C GPIO Expander)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>