[redhat] New configs in sound/core
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: SND_CORE_TEST [=m]
 Type  : tristate
 Defined at sound/core/Kconfig:42
   Prompt: Sound core KUnit test
   Depends on: SOUND [=m] && SND [=m] && KUNIT [=m]
   Location:
     -> Device Drivers
       -> Sound card support (SOUND [=m])
         -> Advanced Linux Sound Architecture (SND [=m])
           -> Sound core KUnit test (SND_CORE_TEST [=m])
 Selects: SND_PCM [=m]

 Commit: 3e39acf56ede (ALSA: core: Add sound core KUnit test)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>