[redhat] New configs in drivers/dma
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: TEGRA210_ADMA [=n]
 Type  : tristate
 Defined at drivers/dma/Kconfig:644
   Prompt: NVIDIA Tegra210 ADMA support
   Depends on: DMADEVICES [=y] && (ARCH_TEGRA [=y] || COMPILE_TEST [=n])
   Location:
     -> Device Drivers
       -> DMA Engine support (DMADEVICES [=y])
         -> NVIDIA Tegra210 ADMA support (TEGRA210_ADMA [=n])
 Selects: DMA_ENGINE [=y] && DMA_VIRTUAL_CHANNELS [=y]

 Commit: f46b195799b5 (dmaengine: tegra-adma: Add support for Tegra210 ADMA)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>