[redhat] New configs in drivers/vfio
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: NVGRACE_GPU_VFIO_PCI [=n]
 Type  : tristate
 Defined at drivers/vfio/pci/nvgrace-gpu/Kconfig:2
   Prompt: VFIO support for the GPU in the NVIDIA Grace Hopper Superchip
   Depends on: VFIO [=m] && PCI [=y] && MMU [=y] && (ARM64 [=y] || COMPILE_TEST [=n] && 64BIT [=y])
   Location:
     -> Device Drivers
       -> VFIO Non-Privileged userspace driver framework (VFIO [=m])
         -> VFIO support for PCI devices
           -> VFIO support for the GPU in the NVIDIA Grace Hopper Superchip (NVGRACE_GPU_VFIO_PCI [=n])
 Selects: VFIO_PCI_CORE [=m]

 Commit: 701ab935859f (vfio/nvgrace-gpu: Add vfio pci variant module for grace hopper)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>