[redhat] New configs in fs/fuse
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: FUSE_PASSTHROUGH [=y]
 Type  : bool
 Defined at fs/fuse/Kconfig:56
   Prompt: FUSE passthrough operations support
   Depends on: FUSE_FS [=m]
   Location:
     -> File systems
       -> FUSE (Filesystem in Userspace) support (FUSE_FS [=m])
         -> FUSE passthrough operations support (FUSE_PASSTHROUGH [=y])
 Selects: FS_STACK [=y]

 Commit: 7dc4e97a4f9a (fuse: introduce FUSE_PASSTHROUGH capability)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>