[redhat] New configs in drivers/mtd
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: MTD_UBI_NVMEM [=n]
 Type  : tristate
 Defined at drivers/mtd/ubi/Kconfig:117
   Prompt: UBI virtual NVMEM
   Depends on: MTD [=m] && MTD_UBI [=m] && NVMEM [=y]
   Location:
     -> Device Drivers
       -> Memory Technology Device (MTD) support (MTD [=m])
         -> Enable UBI - Unsorted block images (MTD_UBI [=m])
           -> UBI virtual NVMEM (MTD_UBI_NVMEM [=n])

 Commit: 3ce485803da1 (mtd: ubi: provide NVMEM layer over UBI volumes)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>