[redhat] New configs in arch/x86
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: SPECTRE_BHI_OFF [=n]
 Type  : bool
 Defined at arch/x86/Kconfig:2650
   Prompt: off
   Depends on: <choice>
   Location:
     -> Mitigations for speculative execution vulnerabilities (SPECULATION_MITIGATIONS [=y])
       -> Clear branch history (<choice> [=y])
         -> off (SPECTRE_BHI_OFF [=n])

 Commit: ec9404e40e8f (x86/bhi: Add BHI mitigation knob)

---

 Symbol: SPECTRE_BHI_ON [=y]
 Type  : bool
 Defined at arch/x86/Kconfig:2646
   Prompt: on
   Depends on: <choice>
   Location:
     -> Mitigations for speculative execution vulnerabilities (SPECULATION_MITIGATIONS [=y])
       -> Clear branch history (<choice> [=y])
         -> on (SPECTRE_BHI_ON [=y])

 Commit: ec9404e40e8f (x86/bhi: Add BHI mitigation knob)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>