[redhat] New configs in arch/x86
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: MITIGATION_SPECTRE_BHI [=y]
 Type  : bool
 Defined at arch/x86/Kconfig:2636
   Prompt: Mitigate Spectre-BHB (Branch History Injection)
   Depends on: SPECULATION_MITIGATIONS [=y] && CPU_SUP_INTEL [=y]
   Location:
     -> Mitigations for speculative execution vulnerabilities (SPECULATION_MITIGATIONS [=y])
       -> Mitigate Spectre-BHB (Branch History Injection) (MITIGATION_SPECTRE_BHI [=y])

 Commit: 4f511739c54b (x86/bugs: Replace CONFIG_SPECTRE_BHI_{ON,OFF} with CONFIG_MITIGATION_SPECTRE_BHI)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>