[redhat] New configs in arch/Kconfig
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: CPU_MITIGATIONS [=y]
 Type  : bool
 Defined at arch/x86/Kconfig:2492
   Prompt: Mitigations for CPU vulnerabilities
   Location:
     -> Mitigations for CPU vulnerabilities (CPU_MITIGATIONS [=y])
 Defined at arch/Kconfig:16
   Depends on: !ARCH_CONFIGURES_CPU_MITIGATIONS [=y]

 Commit: fe42754b94a4 (cpu: Re-enable CPU mitigations by default for !X86 architectures)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>