[redhat] New configs in lib/Kconfig.debug
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: DEBUG_INFO_BTF_MODULES [=y]
 Type  : bool
 Defined at lib/Kconfig.debug:410
   Prompt: Generate BTF type information for kernel modules
   Depends on: DEBUG_INFO [=y] && DEBUG_INFO_BTF [=y] && MODULES [=y] && PAHOLE_HAS_SPLIT_BTF [=y]
   Location:
     -> Kernel hacking
       -> Compile-time checks and compiler options
         -> Generate BTF type information for kernel modules (DEBUG_INFO_BTF_MODULES [=y])

 Commit: 5f9ae91f7c0d (kbuild: Build kernel module BTFs if BTF is enabled and pahole supports it)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>