Support the first day after a rebase
It is possible the day of a rebase, there is no new merges from
upstream.  This means os-build looks like a GA release, but it is
post-rebase.  That will confuse the rt branch selection logic.  Treat
the post-rebase os-build as the starting point of an rc0.  This will
force the rt-devel logic to use the pre-rebase branch as the branch to
continue with until rt-devel upstream moves forward.

Signed-off-by: Don Zickus <dzickus@redhat.com>