[redhat] New configs in arch/arm64
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: ARCH_AIROHA [=n]
 Type  : bool
 Defined at arch/arm64/Kconfig.platforms:11
   Prompt: Airoha SoC Support
   Location:
     -> Platform selection
       -> Airoha SoC Support (ARCH_AIROHA [=n])
 Selects: ARM_PSCI && HAVE_ARM_ARCH_TIMER

 Commit: 428ae88ef519 (arm64: add Airoha EN7581 platform)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>