[redhat] New configs in arch/s390
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: AP [=y]
 Type  : tristate
 Defined at arch/s390/Kconfig:737
   Prompt: Support for Adjunct Processors (ap)
   Location:
     -> I/O subsystem
       -> Support for Adjunct Processors (ap) (AP [=y])

 Commit: 123760841a2e (s390/ap: modularize ap bus)

---

 Symbol: AP_DEBUG [=n]
 Type  : bool
 Defined at arch/s390/Kconfig:750
   Prompt: Enable debug features for Adjunct Processor (ap) devices
   Depends on: AP [=y]
   Location:
     -> I/O subsystem
       -> Support for Adjunct Processors (ap) (AP [=y])
         -> Enable debug features for Adjunct Processor (ap) devices (AP_DEBUG [=n])

 Commit: b3840c8bfc27 (s390/ap: rename ap debug configuration option)

---

 Symbol: KERNEL_IMAGE_BASE [=0x3FFE0000000]
 Type  : hex
 Range : [0x100000 0x1FFFFFE0000000]
 Defined at arch/s390/Kconfig:605
   Prompt: Kernel image base address
   Location:
     -> Processor type and features
       -> Kernel image base address (KERNEL_IMAGE_BASE [=0x3FFE0000000])

 Commit: 54f2ecc3188f (s390: Map kernel at fixed location when KASLR is disabled)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>