[redhat] New configs in drivers/char
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: TCG_TPM2_HMAC [=y]
 Type  : bool
 Defined at drivers/char/tpm/Kconfig:30
   Prompt: Use HMAC and encrypted transactions on the TPM bus
   Depends on: TCG_TPM [=y]
   Location:
     -> Device Drivers
       -> Character devices
         -> TPM Hardware Support (TCG_TPM [=y])
           -> Use HMAC and encrypted transactions on the TPM bus (TCG_TPM2_HMAC [=y])
 Selects: CRYPTO_ECDH [=y] && CRYPTO_LIB_AESCFB [=y] && CRYPTO_LIB_SHA256 [=y]

 Commit: d2add27cf2b8 (tpm: Add NULL primary creation)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>