[redhat] New configs in drivers/pinctrl
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: PINCTRL_SCMI [=n]
 Type  : tristate
 Defined at drivers/pinctrl/Kconfig:453
   Prompt: Pinctrl driver using SCMI protocol interface
   Depends on: PINCTRL [=y] && (ARM_SCMI_PROTOCOL [=y] || COMPILE_TEST [=n])
   Location:
     -> Device Drivers
       -> Pin controllers (PINCTRL [=y])
         -> Pinctrl driver using SCMI protocol interface (PINCTRL_SCMI [=n])
 Selects: PINMUX [=y] && GENERIC_PINCONF [=y]

 Commit: eb524cb651b7 (pinctrl: Implementation of the generic scmi-pinctrl driver)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>