[redhat] New configs in drivers/gpio
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: GPIO_GRANITERAPIDS [=n]
 Type  : tristate
 Defined at drivers/gpio/Kconfig:324
   Prompt: Intel Granite Rapids-D vGPIO support
   Depends on: GPIOLIB [=y] && HAS_IOMEM [=y] && (X86 [=y] || COMPILE_TEST [=n])
   Location:
     -> Device Drivers
       -> GPIO Support (GPIOLIB [=y])
         -> Memory mapped GPIO drivers
           -> Intel Granite Rapids-D vGPIO support (GPIO_GRANITERAPIDS [=n])
 Selects: GPIOLIB_IRQCHIP [=y]

 Commit: ecc4b1418e23 (gpio: Add Intel Granite Rapids-D vGPIO driver)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>