[redhat] New configs in drivers/net/phy
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: AIR_EN8811H_PHY [=n]
 Type  : tristate
 Defined at drivers/net/phy/Kconfig:79
   Prompt: Airoha EN8811H 2.5 Gigabit PHY
   Depends on: NETDEVICES [=y] && PHYLIB [=y]
   Location:
     -> Device Drivers
       -> Network device support (NETDEVICES [=y])
         -> PHY Device support and infrastructure (PHYLIB [=y])
           -> Airoha EN8811H 2.5 Gigabit PHY (AIR_EN8811H_PHY [=n])

 Commit: 71e79430117d (net: phy: air_en8811h: Add the Airoha EN8811H PHY driver)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>