[redhat] New configs in net/smc
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: SMC_LO [=n]
 Type  : bool
 Defined at net/smc/Kconfig:24
   Prompt: SMC intra-OS shortcut with loopback-ism
   Depends on: NET [=y] && SMC [=m]
   Location:
     -> Networking support (NET [=y])
       -> Networking options
         -> SMC socket protocol family (SMC [=m])
           -> SMC intra-OS shortcut with loopback-ism (SMC_LO [=n])

 Commit: 46ac64419ded (net/smc: introduce loopback-ism for SMC intra-OS shortcut)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>