[redhat] New configs in arch/x86
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: KVM_INTEL_PROVE_VE [=y]
 Type  : bool
 Defined at arch/x86/kvm/Kconfig:98
   Prompt: Check that guests do not receive #VE exceptions
   Depends on: VIRTUALIZATION [=y] && KVM_INTEL [=m]
   Location:
     -> Virtualization (VIRTUALIZATION [=y])
       -> Kernel-based Virtual Machine (KVM) support (KVM [=m])
         -> KVM for Intel (and compatible) processors support (KVM_INTEL [=m])
           -> Check that guests do not receive #VE exceptions (KVM_INTEL_PROVE_VE [=y])

 Commit: 8131cf5b4fd8 (KVM: VMX: Introduce test mode related to EPT violation VE)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>