[redhat] New configs in drivers/media
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: VIDEO_INTEL_IPU6 [=n]
 Type  : tristate
 Defined at drivers/media/pci/intel/ipu6/Kconfig:1
   Prompt: Intel IPU6 driver
   Depends on: MEDIA_SUPPORT [=m] && PCI [=y] && MEDIA_PCI_SUPPORT [=y] && (ACPI [=y] || COMPILE_TEST [=n]) && VIDEO_DEV [=m] && X86 [=y] && X86_64 [=y] && HAS_DMA [=y]
   Location:
     -> Device Drivers
       -> Multimedia support (MEDIA_SUPPORT [=m])
         -> Media drivers
           -> Media PCI Adapters (MEDIA_PCI_SUPPORT [=y])
             -> Intel IPU6 driver (VIDEO_INTEL_IPU6 [=n])
 Selects: DMA_OPS [=y] && IOMMU_IOVA [=y] && VIDEO_V4L2_SUBDEV_API [=y] && MEDIA_CONTROLLER [=y] && VIDEOBUF2_DMA_CONTIG [=n] && V4L2_FWNODE [=m] && IPU_BRIDGE [=n]

 Commit: c70281cc83d6 (media: intel/ipu6: add Kconfig and Makefile)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>