[redhat] New configs in drivers/iommu
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: ARM_SMMU_V3_KUNIT_TEST [=m]
 Type  : tristate
 Defined at drivers/iommu/Kconfig:417
   Prompt: KUnit tests for arm-smmu-v3 driver
   Depends on: IOMMU_SUPPORT [=y] && ARM_SMMU_V3 [=y] && KUNIT [=m] && ARM_SMMU_V3_SVA [=y]
   Visible if: IOMMU_SUPPORT [=y] && ARM_SMMU_V3 [=y] && KUNIT [=m] && ARM_SMMU_V3_SVA [=y] && !KUNIT_ALL_TESTS [=m]
   Location:
     -> Device Drivers
       -> IOMMU Hardware Support (IOMMU_SUPPORT [=y])
         -> ARM Ltd. System MMU Version 3 (SMMUv3) Support (ARM_SMMU_V3 [=y])
           -> Shared Virtual Addressing support for the ARM SMMUv3 (ARM_SMMU_V3_SVA [=y])
             -> KUnit tests for arm-smmu-v3 driver (ARM_SMMU_V3_KUNIT_TEST [=m])

 Commit: 56e1a4cc2588 (iommu/arm-smmu-v3: Add unit tests for arm_smmu_write_entry)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>

v2:
- CONFIG_ARM_SMMU_V3_KUNIT_TEST: move under arm/aarch64/.

Signed-off-by: Patrick Talbert <ptalbert@redhat.com>