[redhat] New configs in drivers/phy
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: PHY_FSL_SAMSUNG_HDMI_PHY [=n]
 Type  : tristate
 Defined at drivers/phy/freescale/Kconfig:38
   Prompt: Samsung HDMI PHY support
   Depends on: (ARCH_MXC [=y] && ARM64 [=y] || COMPILE_TEST [=n]) && OF [=y] && HAS_IOMEM [=y] && COMMON_CLK [=y]
   Location:
     -> Device Drivers
       -> PHY Subsystem
         -> Samsung HDMI PHY support (PHY_FSL_SAMSUNG_HDMI_PHY [=n])
 Implied by [n]:
   - DRM_IMX8MP_DW_HDMI_BRIDGE [=n] && HAS_IOMEM [=y] && DRM [=y] && DRM_BRIDGE [=y] && (ARCH_MXC [=y] || COMPILE_TEST [=n]) && OF [=y] && COMMON_CLK [=y]

 Commit: 6ad082bee902 (phy: freescale: add Samsung HDMI PHY)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>