[redhat] New configs in arch/arm64
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: ARM64_ERRATUM_3194386 [=y]
 Type  : bool
 Defined at arch/arm64/Kconfig:1073
   Prompt: Cortex-X4: 3194386: workaround for MSR SSBS not self-synchronizing
   Location:
     -> Kernel Features
       -> ARM errata workarounds via the alternatives framework
         -> Cortex-X4: 3194386: workaround for MSR SSBS not self-synchronizing (ARM64_ERRATUM_3194386 [=y])
 Selects: ARM64_WORKAROUND_SPECULATIVE_SSBS [=y]

 Commit: 7187bb7d0b5c (arm64: errata: Add workaround for Arm errata 3194386 and 3312417)

---

 Symbol: ARM64_ERRATUM_3312417 [=y]
 Type  : bool
 Defined at arch/arm64/Kconfig:1092
   Prompt: Neoverse-V3: 3312417: workaround for MSR SSBS not self-synchronizing
   Location:
     -> Kernel Features
       -> ARM errata workarounds via the alternatives framework
         -> Neoverse-V3: 3312417: workaround for MSR SSBS not self-synchronizing (ARM64_ERRATUM_3312417 [=y])
 Selects: ARM64_WORKAROUND_SPECULATIVE_SSBS [=y]

 Commit: 7187bb7d0b5c (arm64: errata: Add workaround for Arm errata 3194386 and 3312417)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>