[redhat] New configs in drivers/rtc
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: RTC_DRV_RX8111 [=n]
 Type  : tristate
 Defined at drivers/rtc/Kconfig:667
   Prompt: Epson RX8111
   Depends on: RTC_CLASS [=y] && I2C [=y]
   Location:
     -> Device Drivers
       -> Real Time Clock (RTC_CLASS [=y])
         -> Epson RX8111 (RTC_DRV_RX8111 [=n])
 Selects: REGMAP_I2C [=m]

 Commit: f8c81d15f4bb (rtc: Add driver for Epson RX8111)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>