redhat: rhdocs: Remove the rhdocs directory
As the rhdocs git-subtree has been in a broken state and essentially
unmaintained for a long time, let us remove it from the kernel
repository, and replace it with a direct link to the CentOS docs
project && the rendered docs.

In the case of kernel-ark, remove the need for incremental updates of
the RHMAINTAINERS file.

Signed-off-by: Derek Barbosa <debarbos@redhat.com>