gitlab-ci: merge ark-latest before tagging cki-gating
When running in a periodic pipeline on ark-latest, all changes from
ark-latest are already included in the tagged commit.

This is not the case for MR pipelines, where the CKI pipeline merge
stage takes care of this for the kernel build. During tagging, add the
ark-latest merging so that gating pipelines run with a similar code base
as in the pipeline used for tagging.

Signed-off-by: Michael Hofmann <mhofmann@redhat.com>