redhat/kernel.spec: fix run of mod-denylist
In commit 1854c03a97ee I updated the directory for extra modules,
but I failed to notice that running of mod-denylist is conditioned
on that same variable, so mod-denylist stopped running for extra
modules, generating no /etc/*blacklist.conf files. Fix that and
run mod-denylist for all packages except for modules and module-core,
which was the behaviour before a3325a8ee6b0 ("spec: rework filter-mods
and mod-denylist").

Fixes: 1854c03a97ee ("redhat/kernel.spec: keep extra modules in original directories")
Signed-off-by: Jan Stancek <jstancek@redhat.com>