[redhat] gitlab-ci: Use cki-tools builder image for scheduled jobs
Instead of having dnf install commands as part of the daily scheduled
jobs, bake the necessary packages into the image used by the job.

In cki-project/cki-tools!1301 we add the necessary bits to the cki-tools
image so now we can use that image and drop the dnf commands from the CI
definition.

Fixes: #159

Signed-off-by: Patrick Talbert <ptalbert@redhat.com>