fedora: arm: Enable basic support for S32G-VNP-RDB3 board
The S32G-VNP-RDB3 is a compact, highly optimized and integrated board
engineered for vehicle computer. This patches enable the configs to be
able to boot with this board and get serial console for it. While here,
we modified SCMI clocks to be built-in because clocks are needed very early
in the boot process of this board.

Signed-off-by: Enric Balletbo i Serra <eballetbo@redhat.com>