Remove new for GITLAB_TOKEN
The GITLAB_TOKEN was necessary to create gitlab issues if we ran into
merge conflicts.  That only worked if the env var REPORT_BUGS was set,
which it hasn't in a long while.  To simplify things, remove the filing
issue logic and thus the need to pass around GITLAB_TOKEN.

The other need for python-gitlab is to fetch MRs for "Include in
release", otherwise the whole python-gitlab requirement could disappear.

Signed-off-by: Don Zickus <dzickus@redhat.com>