[redhat] New configs in drivers/gpio
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: GPIO_SLOPPY_LOGIC_ANALYZER [=n]
 Type  : tristate
 Defined at drivers/gpio/Kconfig:1896
   Prompt: Sloppy GPIO logic analyzer
   Depends on: GPIOLIB [=y] && (GPIOLIB [=y] || COMPILE_TEST [=n]) && CPUSETS [=y] && DEBUG_FS [=y] && EXPERT [=y]
   Location:
     -> Device Drivers
       -> GPIO Support (GPIOLIB [=y])
         -> GPIO Debugging utilities
           -> Sloppy GPIO logic analyzer (GPIO_SLOPPY_LOGIC_ANALYZER [=n])

 Commit: 7828b7bbbf20 (gpio: add sloppy logic analyzer using polling)

---

 Symbol: GPIO_VIRTUSER [=n]
 Type  : tristate
 Defined at drivers/gpio/Kconfig:1911
   Prompt: GPIO Virtual User Testing Module
   Depends on: GPIOLIB [=y]
   Location:
     -> Device Drivers
       -> GPIO Support (GPIOLIB [=y])
         -> GPIO Debugging utilities
           -> GPIO Virtual User Testing Module (GPIO_VIRTUSER [=n])
 Selects: DEBUG_FS [=y] && CONFIGFS_FS [=y] && IRQ_WORK [=y]

 Commit: 91581c4b3f29 (gpio: virtuser: new virtual testing driver for the GPIO API)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>