[redhat] New configs in drivers/nvme
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: NVME_TARGET_DEBUGFS [=n]
 Type  : bool
 Defined at drivers/nvme/target/Kconfig:20
   Prompt: NVMe Target debugfs support
   Depends on: NVME_TARGET [=m]
   Location:
     -> Device Drivers
       -> NVME Support
         -> NVMe Target support (NVME_TARGET [=m])
           -> NVMe Target debugfs support (NVME_TARGET_DEBUGFS [=n])

 Commit: 649fd41420a8 (nvmet: add debugfs support)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>