[redhat] New configs in drivers/power
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: POWER_SEQUENCING [=n]
 Type  : tristate
 Defined at drivers/power/sequencing/Kconfig:3
   Prompt: Power Sequencing support
   Location:
     -> Device Drivers
       -> Power Sequencing support (POWER_SEQUENCING [=n])
 Selected by [n]:
   - PCI_PWRCTL_PWRSEQ [=n] && PCI [=y]

 Commit: 249ebf3f65f8 (power: sequencing: implement the pwrseq core)

---

 Symbol: POWER_SEQUENCING_QCOM_WCN [=m]
 Type  : tristate
 Defined at drivers/power/sequencing/Kconfig:16
   Prompt: Qualcomm WCN family PMU driver
   Depends on: POWER_SEQUENCING [=m]
   Location:
     -> Device Drivers
       -> Power Sequencing support (POWER_SEQUENCING [=m])
         -> Qualcomm WCN family PMU driver (POWER_SEQUENCING_QCOM_WCN [=m])

 Commit: 2f1630f437df (power: pwrseq: add a driver for the PMU module on the QCom WCN chipsets)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>