[redhat] New configs in drivers/memory
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: FSL_IFC [=n]
 Type  : bool
 Defined at drivers/memory/Kconfig:169
   Prompt: Freescale IFC driver
   Depends on: MEMORY [=y] && (FSL_SOC || ARCH_LAYERSCAPE [=y] || SOC_LS1021A || COMPILE_TEST [=n]) && HAS_IOMEM [=y]
   Location:
     -> Device Drivers
       -> Memory Controller drivers (MEMORY [=y])
         -> Freescale IFC driver (FSL_IFC [=n])

 Commit: 42d87b184f03 (powerpc: select MEMORY for FSL_IFC to not break existing .config files)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>