[redhat] New configs in drivers/net/wireless
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: RTL8192DU [=n]
 Type  : tristate
 Defined at drivers/net/wireless/realtek/rtlwifi/Kconfig:122
   Prompt: Realtek RTL8192DU USB Wireless Network Adapter
   Depends on: NETDEVICES [=y] && WLAN [=y] && WLAN_VENDOR_REALTEK [=y] && RTL_CARDS [=m] && USB [=y]
   Location:
     -> Device Drivers
       -> Network device support (NETDEVICES [=y])
         -> Wireless LAN (WLAN [=y])
           -> Realtek devices (WLAN_VENDOR_REALTEK [=y])
             -> Realtek rtlwifi family of devices (RTL_CARDS [=m])
               -> Realtek RTL8192DU USB Wireless Network Adapter (RTL8192DU [=n])
 Selects: RTLWIFI [=m] && RTLWIFI_USB [=m] && RTL8192D_COMMON [=m]

 Commit: 59ea089dcba3 (wifi: rtlwifi: Enable the new rtl8192du driver)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>