[redhat] New configs in drivers/soc
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: QCOM_PD_MAPPER [=n]
 Type  : tristate
 Defined at drivers/soc/qcom/Kconfig:75
   Prompt: Qualcomm Protection Domain Mapper
   Depends on: NET [=y] && QRTR [=m]
   Location:
     -> Device Drivers
       -> SOC (System On Chip) specific Drivers
         -> Qualcomm SoC drivers
           -> Qualcomm Protection Domain Mapper (QCOM_PD_MAPPER [=n])
 Selects: QCOM_QMI_HELPERS [=m] && QCOM_PDR_MSG [=n] && AUXILIARY_BUS [=y]

 Commit: 1ebcde047c54 (soc: qcom: add pd-mapper implementation)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>