[redhat] New configs in drivers/leds
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: LEDS_SY7802 [=n]
 Type  : tristate
 Defined at drivers/leds/flash/Kconfig:124
   Prompt: LED support for the Silergy SY7802
   Depends on: NEW_LEDS [=y] && LEDS_CLASS_FLASH [=m] && I2C [=y] && OF [=y] && GPIOLIB [=y]
   Location:
     -> Device Drivers
       -> LED Support (NEW_LEDS [=y])
         -> LED support for the Silergy SY7802 (LEDS_SY7802 [=n])
 Selects: REGMAP_I2C [=y]

 Commit: c581f17a66b6 (leds: sy7802: Add support for Silergy SY7802 flash LED controller)

---

 Symbol: LEDS_TRIGGER_INPUT_EVENTS [=n]
 Type  : tristate
 Defined at drivers/leds/trigger/Kconfig:148
   Prompt: LED Input events trigger
   Depends on: NEW_LEDS [=y] && LEDS_TRIGGERS [=y] && INPUT [=y]
   Location:
     -> Device Drivers
       -> LED Support (NEW_LEDS [=y])
         -> LED Trigger support (LEDS_TRIGGERS [=y])
           -> LED Input events trigger (LEDS_TRIGGER_INPUT_EVENTS [=n])

 Commit: 6f963a20eed7 (leds: trigger: Add new LED Input events trigger)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>