[redhat] New configs in drivers/video
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: BACKLIGHT_LM3509 [=n]
 Type  : tristate
 Defined at drivers/video/backlight/Kconfig:376
   Prompt: Backlight Driver for LM3509
   Depends on: HAS_IOMEM [=y] && BACKLIGHT_CLASS_DEVICE [=y] && I2C [=y]
   Location:
     -> Device Drivers
       -> Graphics support
         -> Backlight & LCD device support
           -> Lowlevel Backlight controls (BACKLIGHT_CLASS_DEVICE [=y])
             -> Backlight Driver for LM3509 (BACKLIGHT_LM3509 [=n])
 Selects: REGMAP_I2C [=m]

 Commit: b72755f5b577 (backlight: Add new lm3509 backlight driver)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>