[redhat] New configs in mm/Kconfig
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: SLAB_BUCKETS [=y]
 Type  : bool
 Defined at mm/Kconfig:276
   Prompt: Support allocation from separate kmalloc buckets
   Depends on: !SLUB_TINY [=n]
   Location:
     -> Memory Management options
       -> Slab allocator options
         -> Support allocation from separate kmalloc buckets (SLAB_BUCKETS [=y])

 Commit: 67f2df3b82d0 (mm/slab: Plumb kmem_buckets into __do_kmalloc_node())

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>