[redhat] New configs in arch/powerpc
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: CRYPTO_CURVE25519_PPC64 [=n]
 Type  : tristate
 Defined at arch/powerpc/crypto/Kconfig:5
   Prompt: Public key crypto: Curve25519 (PowerPC64)
   Depends on: CRYPTO [=y] && !KMSAN [=n] && PPC [=y] && PPC64 [=y] && CPU_LITTLE_ENDIAN [=y]
   Location:
     -> Cryptographic API (CRYPTO [=y])
       -> Accelerated Cryptographic Algorithms for CPU (powerpc)
         -> Public key crypto: Curve25519 (PowerPC64) (CRYPTO_CURVE25519_PPC64 [=n])
 Selects: CRYPTO_LIB_CURVE25519_GENERIC [=m] && CRYPTO_ARCH_HAVE_LIB_CURVE25519 [=n]

 Commit: b42519dbba83 (crypto: ppc/curve25519 - Update Kconfig and Makefile for ppc64le)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>

v2:
- CONFIG_CRYPTO_CURVE25519_PPC64 enabled as a module for powerpc.

Signed-off-by: Patrick Talbert <ptalbert@redhat.com>