[redhat] New configs in drivers/md
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: DM_VERITY_VERIFY_ROOTHASH_SIG_PLATFORM_KEYRING [=y]
 Type  : bool
 Defined at drivers/md/Kconfig:543
   Prompt: Verity data device root hash signature verification with platform keyring
   Depends on: MD [=y] && DM_VERITY_VERIFY_ROOTHASH_SIG [=y] && INTEGRITY_PLATFORM_KEYRING [=y]
   Location:
     -> Device Drivers
       -> Multiple devices driver support (RAID and LVM) (MD [=y])
         -> Device mapper support (BLK_DEV_DM [=m])
           -> Verity target support (DM_VERITY [=m])
             -> Verity data device root hash signature verification support (DM_VERITY_VERIFY_ROOTHASH_SIG [=y])
               -> Verity data device root hash signature verification with platform keyring (DM_VERITY_VERIFY_ROOTHASH_SIG_PLATFORM_KEYRING [=y])

 Commit: 6fce1f40e951 (dm verity: add support for signature verification with platform keyring)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>