[redhat] New configs in drivers/pinctrl
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: PINCTRL_IMX91 [=n]
 Type  : tristate
 Defined at drivers/pinctrl/freescale/Kconfig:198
   Prompt: IMX91 pinctrl driver
   Depends on: PINCTRL [=y] && ARCH_MXC [=y]
   Location:
     -> Device Drivers
       -> Pin controllers (PINCTRL [=y])
         -> IMX91 pinctrl driver (PINCTRL_IMX91 [=n])
 Selects: PINCTRL_IMX [=y]

 Commit: 993e2fc48fa5 (pinctrl: imx: Add pinctrl driver support for i.MX91)

---

 Symbol: PINCTRL_IMX_SCMI [=n]
 Type  : tristate
 Defined at drivers/pinctrl/freescale/Kconfig:10
   Prompt: i.MX95 pinctrl driver using SCMI protocol interface
   Depends on: PINCTRL [=y] && (ARM_SCMI_PROTOCOL [=y] && OF [=y] || COMPILE_TEST [=n])
   Location:
     -> Device Drivers
       -> Pin controllers (PINCTRL [=y])
         -> i.MX95 pinctrl driver using SCMI protocol interface (PINCTRL_IMX_SCMI [=n])
 Selects: PINMUX [=y] && GENERIC_PINCONF [=y] && GENERIC_PINCTRL_GROUPS [=y] && GENERIC_PINMUX_FUNCTIONS [=y]

 Commit: b755521fd6eb (pinctrl: imx: support SCMI pinctrl protocol for i.MX95)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>