[redhat] New configs in drivers/misc
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: KEBA_CP500 [=n]
 Type  : tristate
 Defined at drivers/misc/keba/Kconfig:2
   Prompt: KEBA CP500 system FPGA support
   Depends on: PCI [=y]
   Location:
     -> Device Drivers
       -> Misc devices
         -> KEBA CP500 system FPGA support (KEBA_CP500 [=n])
 Selects: AUXILIARY_BUS [=y]

 Commit: a1944676767e (misc: keba: Add basic KEBA CP500 system FPGA support)

---

 Symbol: MARVELL_CN10K_DPI [=n]
 Type  : tristate
 Defined at drivers/misc/Kconfig:588
   Prompt: Octeon CN10K DPI driver
   Depends on: PCI [=y] && (ARCH_THUNDER [=y] || COMPILE_TEST [=n] && 64BIT [=y])
   Location:
     -> Device Drivers
       -> Misc devices
         -> Octeon CN10K DPI driver (MARVELL_CN10K_DPI [=n])

 Commit: 5f67eef6dff3 (misc: mrvl-cn10k-dpi: add Octeon CN10K DPI administrative driver)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>