[redhat] New configs in sound/pci
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: SND_HDA_CODEC_SENARYTECH [=n]
 Type  : tristate
 Defined at sound/pci/hda/Kconfig:297
   Prompt: Build Senarytech HD-audio codec support
   Depends on: SOUND [=m] && SND [=m] && SND_HDA [=m]
   Location:
     -> Device Drivers
       -> Sound card support (SOUND [=m])
         -> Advanced Linux Sound Architecture (SND [=m])
           -> HD-Audio
             -> Build Senarytech HD-audio codec support (SND_HDA_CODEC_SENARYTECH [=n])
 Selects: SND_HDA_GENERIC [=m] && SND_HDA_GENERIC_LEDS [=y]

 Commit: eb882afcfa83 (ALSA: hda/senarytech: add senarytech codec support)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>

v2:
- SND_HDA_CODEC_SENARYTECH enabled as a module.

Signed-off-by: Patrick Talbert <ptalbert@redhat.com>