[redhat] New configs in arch/arm64
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: COMPRESSED_INSTALL [=n]
 Type  : bool
 Defined at arch/arm64/Kconfig:2340
   Prompt: Install compressed image by default
   Location:
     -> Boot options
       -> Install compressed image by default (COMPRESSED_INSTALL [=n])

 Commit: 4c7be57f2772 (arm64: allow installing compressed image by default)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>