kernel: config: enable erofs lzma compression
Jira: https://issues.redhat.com/browse/RHEL-36449
Upstream Status: RHEL-only

We are planning on adding the erofs file system to the list of supported
file systems in RHEL 10.

It's used by composefs and, in time, as a replacement for squashfs.

To start with we want to enable lzma compression along with the
dependencies it needs.

Signed-off-by: Ian Kent <ikent@redhat.com>