[redhat] New configs in fs/netfs
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: NETFS_DEBUG [=n]
 Type  : bool
 Defined at fs/netfs/Kconfig:25
   Prompt: Enable dynamic debugging netfslib and FS-Cache
   Depends on: NETFS_SUPPORT [=m]
   Location:
     -> File systems
       -> Caches
         -> Enable dynamic debugging netfslib and FS-Cache (NETFS_DEBUG [=n])

 Commit: fcad93360df4 (netfs: Rename CONFIG_FSCACHE_DEBUG to CONFIG_NETFS_DEBUG)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>