[redhat] New configs in arch/s390
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: RANDOMIZE_IDENTITY_BASE [=n]
 Type  : bool
 Defined at arch/s390/Kconfig:607
   Prompt: Randomize the address of the identity mapping base
   Depends on: RANDOMIZE_BASE [=y]
   Location:
     -> Processor type and features
       -> Randomize the address of the kernel image (KASLR) (RANDOMIZE_BASE [=y])
         -> Randomize the address of the identity mapping base (RANDOMIZE_IDENTITY_BASE [=n])

 Commit: 32db401965f1 (s390/mm: Pin identity mapping base to zero)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>