redhat: Drop bpftool from kernel spec
We are going to build bpftool as a separate package. There are 2 main
reasons for this change: (1) bpftool version is independent of the
kernel and having it in the kernel spec causes all kinds of problems and
(2) bpftool is loosely coupled to the libbpf library which is also
packaged separately.

Note that bpftool is still built during kernel build in 2 places:
- Bootstrap bpftool for generating vmlinux.h which goes into
  kernel-devel.
- As a part of kernel BPF selftests build. Here, we opt to use the
  bpftool from kernel sources rather than the system one because
  upstream selftests sometimes require unreleased bpftool features to
  build successfully.

Signed-off-by: Viktor Malik <vmalik@redhat.com>