[redhat] New configs in drivers/net/ethernet/Kconfig
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: OA_TC6 [=n]
 Type  : tristate
 Defined at drivers/net/ethernet/Kconfig:161
   Prompt: OPEN Alliance TC6 10BASE-T1x MAC-PHY support
   Depends on: NETDEVICES [=y] && ETHERNET [=y] && SPI [=y]
   Location:
     -> Device Drivers
       -> Network device support (NETDEVICES [=y])
         -> Ethernet driver support (ETHERNET [=y])
           -> OPEN Alliance TC6 10BASE-T1x MAC-PHY support (OA_TC6 [=n])
 Selects: PHYLIB [=y]
 Selected by [n]:
   - LAN865X [=n] && NETDEVICES [=y] && ETHERNET [=y] && NET_VENDOR_MICROCHIP [=y] && SPI [=y]

 Commit: aa58bec064ab (net: ethernet: oa_tc6: implement register write operation)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>