[redhat] New configs in drivers/thermal
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: THERMAL_CORE_TESTING [=n]
 Type  : tristate
 Defined at drivers/thermal/Kconfig:43
   Prompt: Thermal core testing facility
   Depends on: THERMAL [=y] && DEBUG_FS [=y]
   Location:
     -> Device Drivers
       -> Thermal drivers (THERMAL [=y])
         -> Thermal core testing facility (THERMAL_CORE_TESTING [=n])

 Commit: f6a034f2df42 (thermal: Introduce a debugfs-based testing facility)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>