[redhat] New configs in drivers/virt
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: ARM_PKVM_GUEST [=n]
 Type  : bool
 Defined at drivers/virt/coco/pkvm-guest/Kconfig:1
   Prompt: Arm pKVM protected guest driver
   Depends on: VIRT_DRIVERS [=y] && ARM64 [=y]
   Location:
     -> Device Drivers
       -> Virtualization drivers (VIRT_DRIVERS [=y])
         -> Arm pKVM protected guest driver (ARM_PKVM_GUEST [=n])

 Commit: a06c3fad49a5 (drivers/virt: pkvm: Add initial support for running as a protected guest)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>