redhat: change schedule jobs image from cki-tools to builder-rawhide
The cki-tools image has minimal build requirements for the various
redhat scripts and some build targets. This causes a problem because the
environment does not match that of the builder containers where the
kernels are actually built. The big difference for now is the lack of a
rust toolchain in the cki-tools image.

The result is a 'dist-configs-check' in the cki-tools image environment
shows all rust related config items being disabled so new rust config
items are not considered relevant. But the same check in a builder
container environment will give different results as the rust toolchain
is present.

Fix this by changing the daily scheduled jobs to use the builder-rawhide
image.

Signed-off-by: Patrick Talbert <ptalbert@redhat.com>