[redhat] New configs in drivers/rtc
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: RTC_DRV_IMX_BBM_SCMI [=y]
 Type  : tristate
 Defined at drivers/rtc/Kconfig:1830
   Prompt: NXP i.MX BBM SCMI RTC support
   Depends on: RTC_CLASS [=y] && (IMX_SCMI_BBM_EXT [=y] || COMPILE_TEST [=n])
   Location:
     -> Device Drivers
       -> Real Time Clock (RTC_CLASS [=y])
         -> NXP i.MX BBM SCMI RTC support (RTC_DRV_IMX_BBM_SCMI [=y])

 Commit: 3008598ef3b0 (rtc: support i.MX95 BBM RTC)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>