[redhat] New configs in security/ipe
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: SECURITY_IPE [=n]
 Type  : bool
 Defined at security/ipe/Kconfig:6
   Prompt: Integrity Policy Enforcement (IPE)
   Depends on: SECURITY [=y] && SECURITYFS [=y] && AUDIT [=y] && AUDITSYSCALL [=y]
   Location:
     -> Security options
       -> Integrity Policy Enforcement (IPE) (SECURITY_IPE [=n])
 Selects: PKCS7_MESSAGE_PARSER [=y] && SYSTEM_DATA_VERIFICATION [=y] && IPE_PROP_DM_VERITY [=n] && IPE_PROP_DM_VERITY_SIGNATURE [=n] && IPE_PROP_FS_VERITY [=n] && IPE_PROP_FS_VERITY_BUILTIN_SIG [=n]

 Commit: 0311507792b5 (lsm: add IPE lsm)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>