[redhat] New configs in drivers/iommu
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: TEGRA241_CMDQV [=n]
 Type  : bool
 Defined at drivers/iommu/Kconfig:428
   Prompt: NVIDIA Tegra241 CMDQ-V extension support for ARM SMMUv3
   Depends on: IOMMU_SUPPORT [=y] && ARM_SMMU_V3 [=y] && ACPI [=y]
   Location:
     -> Device Drivers
       -> IOMMU Hardware Support (IOMMU_SUPPORT [=y])
         -> ARM Ltd. System MMU Version 3 (SMMUv3) Support (ARM_SMMU_V3 [=y])
           -> NVIDIA Tegra241 CMDQ-V extension support for ARM SMMUv3 (TEGRA241_CMDQV [=n])

 Commit: 918eb5c856f6 (iommu/arm-smmu-v3: Add in-kernel support for NVIDIA Tegra241 (Grace) CMDQV)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>

v2:
- TEGRA241_CMDQV enabled in common/arm/aarch64/.

Signed-off-by: Patrick Talbert <ptalbert@redhat.com>