[redhat] New configs in drivers/misc
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: RPMB [=n]
 Type  : tristate
 Defined at drivers/misc/Kconfig:107
   Prompt: RPMB partition interface
   Depends on: MMC [=m]
   Location:
     -> Device Drivers
       -> Misc devices
         -> RPMB partition interface (RPMB [=n])

 Commit: 1e9046e3a154 (rpmb: add Replay Protected Memory Block (RPMB) subsystem)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>