[redhat] New configs in drivers/mfd
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: MFD_ADP5585 [=n]
 Type  : tristate
 Defined at drivers/mfd/Kconfig:23
   Prompt: Analog Devices ADP5585 keypad decoder and I/O expander driver
   Depends on: HAS_IOMEM [=y] && I2C [=y] && (OF [=y] || COMPILE_TEST [=n])
   Location:
     -> Device Drivers
       -> Multifunction device drivers
         -> Analog Devices ADP5585 keypad decoder and I/O expander driver (MFD_ADP5585 [=n])
 Selects: MFD_CORE [=m] && REGMAP_I2C [=m]

 Commit: 480a8ad683d7 (mfd: adp5585: Add Analog Devices ADP5585 core support)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>