[redhat] New configs in lib/Kconfig.debug
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: INT_POW_TEST [=m]
 Type  : tristate
 Defined at lib/Kconfig.debug:3064
   Prompt: Integer exponentiation (int_pow) test
   Depends on: KUNIT [=m]
   Visible if: KUNIT [=m] && !KUNIT_ALL_TESTS [=m]
   Location:
     -> Integer exponentiation (int_pow) test (INT_POW_TEST [=m])

 Commit: 7fcc9b53216c (lib/math: Add int_pow test suite)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>