[redhat] New configs in sound/core
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: SND_UTIMER [=n]
 Type  : bool
 Defined at sound/core/Kconfig:245
   Prompt: Enable support for userspace-controlled virtual timers
   Depends on: SOUND [=m] && SND [=m] && SND_TIMER [=m]
   Location:
     -> Device Drivers
       -> Sound card support (SOUND [=m])
         -> Advanced Linux Sound Architecture (SND [=m])
           -> Enable support for userspace-controlled virtual timers (SND_UTIMER [=n])

 Commit: 37745918e0e7 (ALSA: timer: Introduce virtual userspace-driven timers)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>