[redhat] New configs in drivers/hid
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: HID_GOODIX_SPI [=n]
 Type  : tristate
 Defined at drivers/hid/Kconfig:407
   Prompt: Goodix GT7986U SPI HID touchscreen
   Depends on: HID_SUPPORT [=y] && HID [=y] && SPI_MASTER [=y]
   Location:
     -> Device Drivers
       -> HID bus support (HID_SUPPORT [=y])
         -> HID bus core support (HID [=y])
           -> Special HID drivers
             -> Goodix GT7986U SPI HID touchscreen (HID_GOODIX_SPI [=n])

 Commit: 75e16c8ce283 (HID: hid-goodix: Add Goodix HID-over-SPI driver)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>