[redhat] New configs in kernel/Kconfig.preempt
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: PREEMPT_RT [=n]
 Type  : bool
 Defined at kernel/Kconfig.preempt:70
   Prompt: Fully Preemptible Kernel (Real-Time)
   Depends on: EXPERT [=y] && ARCH_SUPPORTS_RT [=y]
   Location:
     -> General setup
       -> Preemption Model (<choice> [=n])
         -> Fully Preemptible Kernel (Real-Time) (PREEMPT_RT [=n])
 Selects: PREEMPTION [=y]

 Commit: a50a3f4b6a31 (sched/rt, Kconfig: Introduce CONFIG_PREEMPT_RT)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>