[redhat] New configs in kernel/Kconfig.preempt
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: SCHED_CLASS_EXT [=n]
 Type  : bool
 Defined at kernel/Kconfig.preempt:136
   Prompt: Extensible Scheduling Class
   Depends on: BPF_SYSCALL [=y] && BPF_JIT [=y] && DEBUG_INFO_BTF [=y]
   Location:
     -> General setup
       -> Extensible Scheduling Class (SCHED_CLASS_EXT [=n])
 Selects: STACKTRACE [=y]

 Commit: f0e1a0643a59 (sched_ext: Implement BPF extensible scheduler class)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>