[redhat] New configs in drivers/of
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: OF_OVERLAY_KUNIT_TEST [=m]
 Type  : tristate
 Defined at drivers/of/Kconfig:110
   Prompt: Device Tree overlay KUnit tests
   Depends on: OF [=y] && KUNIT [=m]
   Visible if: OF [=y] && KUNIT [=m] && !KUNIT_ALL_TESTS [=m]
   Location:
     -> Device Drivers
       -> Device Tree and Open Firmware support (OF [=y])
         -> Device Tree overlay KUnit tests (OF_OVERLAY_KUNIT_TEST [=m])
 Selects: OF_OVERLAY [=y]

 Commit: 5c9dd72d8385 (of: Add a KUnit test for overlays and test managed APIs)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>