[redhat] New configs in fs/Kconfig
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: NFS_LOCALIO [=n]
 Type  : bool
 Defined at fs/Kconfig:396
   Prompt: NFS client and server support for LOCALIO auxiliary protocol
   Depends on: NETWORK_FILESYSTEMS [=y] && NFSD [=m] && NFS_FS [=m]
   Location:
     -> File systems
       -> Network File Systems (NETWORK_FILESYSTEMS [=y])
         -> NFS client and server support for LOCALIO auxiliary protocol (NFS_LOCALIO [=n])
 Selects: NFS_COMMON_LOCALIO_SUPPORT [=n]

 Commit: 2a33a85be451 (nfs_common: add NFS LOCALIO auxiliary protocol enablement)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>

v2:
- CONFIG_NFS_LOCALIO enable as a built-in for rhel.
- CONFIG_NFS_COMMON_LOCALIO_SUPPORT enable as a module for rhel.

Signed-off-by: Patrick Talbert <ptalbert@redhat.com>