[redhat] New configs in drivers/rtc
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: RTC_DRV_SD2405AL [=n]
 Type  : tristate
 Defined at drivers/rtc/Kconfig:746
   Prompt: DFRobot SD2405AL
   Depends on: RTC_CLASS [=y] && I2C [=y]
   Location:
     -> Device Drivers
       -> Real Time Clock (RTC_CLASS [=y])
         -> DFRobot SD2405AL (RTC_DRV_SD2405AL [=n])
 Selects: REGMAP_I2C [=m]

 Commit: 5af858acea22 (rtc: Add driver for SD2405AL)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>