[redhat] configs: Sync aarch64 install behavior
The aarch64 'make install' behavior should match other arches
and follow the kernel configuration whether make install is
installing a compressed or uncompressed kernel. Previously on
arm in order to actually install the PE that self decompresses
it was necessarily to use 'make zinstall'. This should largely
now be transparent to users, except that their kernel images
are smaller if they weren't explicitly requesting this behavior
before.

Signed-off-by: Jeremy Linton <jlinton@redhat.com>