CONFIG_OF_OVERLAY: enable for aarch64 and powerpc
CONFIG_OF_OVERLAY is now selected by some KUnit tests where CONFIG_OF is
enabled.  Because these tests are enabled for aarch64 and powerpc we
also need to enable CONFIG_OF_OVERLAY for these arches.

See:
5776526beb95 ('clk: Add KUnit tests for clk fixed rate basic type')
274aff8711b2 ('clk: Add KUnit tests for clks registered with struct clk_parent_data')
5c9dd72d8385 (of: Add a KUnit test for overlays and test managed APIs)

This consolidates CONFIG_OF_OVERLAY now that we're enabling it for all
arches that support it.

Signed-off-by: Scott Weaver <scweaver@redhat.com>