Enable CONFIG_SECURITY_IPE for Fedora
IPE is an upcoming security module that was developed with the purpose of
building systems with immutable and signed partitions and making sure that the
code that is executed has appropriate properties. See e.g.
https://git.kernel.org/pub/scm/linux/kernel/git/torvalds/linux.git/commit/?id=ac6731870ed9
for details.

Commit 02e2f9aa33e461468de02e35ad977bd7233960ae makes IPE useful for users
by allowing the policies to be signed locally.

The secondary options are set to their defaults.

Signed-off-by: Zbigniew Jędrzejewski-Szmek <zbyszek@in.waw.pl>