redhat/configs: add bootconfig to kernel-tools package
Upstream Status: RHEL Only

Add the bootconfig binary to the kernel-tools package.
CONFIG_BOOT_CONFIG is already enabled in RHEL and this allows having a
separate file with the kernel parameters and is described at
https://docs.kernel.org/admin-guide/bootconfig.html. We intend to use
this for automotive for testing purposes since some embedded platforms
limit the kernel command line paramters to 511 bytes.

Signed-off-by: Brian Masney <bmasney@redhat.com>