Don't ignore gitkeep files for ark-infra
It was pointed out the the riscv kabi-module directory was missing
because it only had a .gitkeep file which is ignored as a '.*' file in
.gitignore.  Add an entry to redhat/.gitignore to not ignore .gitkeep
files.

While at it, cleanup redhat/scripts/ci/ark-create-release.sh to include
Kconfig.redhat in the ark-infra branch and remove the gitkeep note as it
is fixed now.

I updated ark-infra branch manually, so it is already resolved.  This
helps make sure it doesn't break again.

Suggested-by: Ondrej Mosnacek <omosnace@redhat.com>

Signed-off-by: Don Zickus <dzickus@redhat.com>