[redhat] New configs in arch/arm64
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: ARM64_HAFT [=y]
 Type  : bool
 Defined at arch/arm64/Kconfig:2190
   Prompt: Support for Hardware managed Access Flag for Table Descriptors
   Depends on: ARM64_HW_AFDBM [=y]
   Location:
     -> Kernel Features
       -> ARMv8.9 architectural features
         -> Support for Hardware managed Access Flag for Table Descriptors (ARM64_HAFT [=y])

 Commit: efe72541355d (arm64: Add support for FEAT_HAFT)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>