[redhat] New configs in block/partitions
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: OF_PARTITION [=n]
 Type  : bool
 Defined at block/partitions/Kconfig:273
   Prompt: Device Tree partition support
   Depends on: BLOCK [=y] && OF [=y]
   Visible if: BLOCK [=y] && OF [=y] && PARTITION_ADVANCED [=y]
   Location:
     -> Enable the block layer (BLOCK [=y])
       -> Partition Types
         -> Advanced partition selection (PARTITION_ADVANCED [=y])
           -> Device Tree partition support (OF_PARTITION [=n])

 Commit: 2e3a191e89f9 (block: add support for partition table defined in OF)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>