[redhat] New configs in drivers/crypto
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: PKEY_UV [=n]
 Type  : tristate
 Defined at drivers/crypto/Kconfig:158
   Prompt: PKEY UV support handler
   Depends on: CRYPTO [=y] && CRYPTO_HW [=y] && PKEY [=m] && S390_UV_UAPI [=y]
   Location:
     -> Cryptographic API (CRYPTO [=y])
       -> Hardware crypto devices (CRYPTO_HW [=y])
         -> Kernel API for protected key handling (PKEY [=m])
           -> PKEY UV support handler (PKEY_UV [=n])

 Commit: 73dfc79c6b04 (s390/pkey: Add new pkey handler module pkey-uv)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>