[redhat] New configs in drivers/perf
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: MARVELL_PEM_PMU [=n]
 Type  : tristate
 Defined at drivers/perf/Kconfig:287
   Prompt: MARVELL PEM PMU Support
   Depends on: PERF_EVENTS [=y] && (ARCH_THUNDER [=y] || COMPILE_TEST [=n] && 64BIT [=y])
   Location:
     -> Device Drivers
       -> Performance monitor support
         -> MARVELL PEM PMU Support (MARVELL_PEM_PMU [=n])

 Commit: e1dce56443a4 (perf/marvell: Marvell PEM performance monitor support)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>