[redhat] New configs in drivers/virtio
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: VIRTIO_MEM [=n]
 Type  : tristate
 Defined at drivers/virtio/Kconfig:123
   Prompt: Virtio mem driver
   Depends on: VIRTIO_MENU [=y] && (X86_64 || ARM64 || RISCV || S390 [=y]) && VIRTIO [=y] && MEMORY_HOTPLUG [=y] && MEMORY_HOTREMOVE [=y] && CONTIG_ALLOC [=y] && EXCLUSIVE_SYSTEM_RAM [=y]
   Location:
     -> Device Drivers
       -> Virtio drivers (VIRTIO_MENU [=y])
         -> Virtio mem driver (VIRTIO_MEM [=n])

 Commit: 5f1f79bbc9e2 (virtio-mem: Paravirtualized memory hotplug)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>