[redhat] New configs in drivers/virt
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: ARM_CCA_GUEST [=m]
 Type  : tristate
 Defined at drivers/virt/coco/arm-cca-guest/Kconfig:1
   Prompt: Arm CCA Guest driver
   Depends on: VIRT_DRIVERS [=y] && ARM64 [=y]
   Location:
     -> Device Drivers
       -> Virtualization drivers (VIRT_DRIVERS [=y])
         -> Arm CCA Guest driver (ARM_CCA_GUEST [=m])
 Selects: TSM_REPORTS [=m]

 Commit: 7999edc484ca (virt: arm-cca-guest: TSM_REPORT support for realms)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>