[redhat] New configs in drivers/acpi
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: ACPI_EC [=y]
 Type  : bool
 Defined at drivers/acpi/Kconfig:135
   Prompt: Embedded Controller
   Depends on: ACPI [=y] && HAS_IOPORT [=y]
   Location:
     -> Power management and ACPI options
       -> ACPI (Advanced Configuration and Power Interface) Support (ACPI [=y])
         -> Embedded Controller (ACPI_EC [=y])

 Commit: a6021aa24f64 (ACPI: EC: make EC support compile-time conditional)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>