[redhat] New configs in drivers/hwmon
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: SENSORS_ISL28022 [=n]
 Type  : tristate
 Defined at drivers/hwmon/Kconfig:858
   Prompt: Renesas ISL28022
   Depends on: HWMON [=y] && I2C [=y]
   Location:
     -> Device Drivers
       -> Hardware Monitoring support (HWMON [=y])
         -> Renesas ISL28022 (SENSORS_ISL28022 [=n])
 Selects: REGMAP_I2C [=m]

 Commit: 39671a14df4f (hwmon: (isl28022) new driver for ISL28022 power monitor)

---

 Symbol: SENSORS_NCT7363 [=n]
 Type  : tristate
 Defined at drivers/hwmon/Kconfig:1684
   Prompt: Nuvoton NCT7363Y
   Depends on: HWMON [=y] && I2C [=y]
   Location:
     -> Device Drivers
       -> Hardware Monitoring support (HWMON [=y])
         -> Nuvoton NCT7363Y (SENSORS_NCT7363 [=n])
 Selects: REGMAP_I2C [=m]

 Commit: 46b94c485ed1 (hwmon: Add driver for I2C chip Nuvoton NCT7363Y)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>