[redhat] New configs in drivers/mfd
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: MFD_CGBC [=n]
 Type  : tristate
 Defined at drivers/mfd/Kconfig:239
   Prompt: Congatec Board Controller
   Depends on: HAS_IOMEM [=y] && X86 [=y]
   Location:
     -> Device Drivers
       -> Multifunction device drivers
         -> Congatec Board Controller (MFD_CGBC [=n])
 Selects: MFD_CORE [=y]

 Commit: 6f1067cfbee7 (mfd: Add Congatec Board Controller driver)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>