[redhat] New configs in drivers/net/ethernet/marvell
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: RVU_ESWITCH [=m]
 Type  : tristate
 Defined at drivers/net/ethernet/marvell/octeontx2/Kconfig:50
   Prompt: Marvell RVU E-Switch support
   Depends on: NETDEVICES [=y] && ETHERNET [=y] && NET_VENDOR_MARVELL [=y] && OCTEONTX2_PF [=m]
   Location:
     -> Device Drivers
       -> Network device support (NETDEVICES [=y])
         -> Ethernet driver support (ETHERNET [=y])
           -> Marvell devices (NET_VENDOR_MARVELL [=y])
             -> Marvell OcteonTX2 NIC Physical Function driver (OCTEONTX2_PF [=m])
               -> Marvell RVU E-Switch support (RVU_ESWITCH [=m])

 Commit: 222a4eea9c6b (octeontx2-pf: RVU representor driver)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>