[redhat] New configs in drivers/iommu
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: ARM_SMMU_V3_IOMMUFD [=n]
 Type  : bool
 Defined at drivers/iommu/Kconfig:418
   Prompt: Enable IOMMUFD features for ARM SMMUv3 (EXPERIMENTAL)
   Depends on: IOMMU_SUPPORT [=y] && ARM_SMMU_V3 [=y] && IOMMUFD [=m]
   Location:
     -> Device Drivers
       -> IOMMU Hardware Support (IOMMU_SUPPORT [=y])
         -> ARM Ltd. System MMU Version 3 (SMMUv3) Support (ARM_SMMU_V3 [=y])
           -> Enable IOMMUFD features for ARM SMMUv3 (EXPERIMENTAL) (ARM_SMMU_V3_IOMMUFD [=n])

 Commit: 6912ec91828b (iommu/arm-smmu-v3: Support IOMMU_GET_HW_INFO via struct arm_smmu_hw_info)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>