[redhat] New configs in drivers/cpufreq
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: CPUFREQ_VIRT [=n]
 Type  : tristate
 Defined at drivers/cpufreq/Kconfig:220
   Prompt: Virtual cpufreq driver
   Depends on: CPU_FREQ [=y] && GENERIC_ARCH_TOPOLOGY [=y]
   Location:
     -> CPU Power Management
       -> CPU Frequency scaling
         -> CPU Frequency scaling (CPU_FREQ [=y])
           -> Virtual cpufreq driver (CPUFREQ_VIRT [=n])

 Commit: 4fd06a5358e0 (cpufreq: add virtual-cpufreq driver)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>