[redhat] New configs in lib/Kconfig.debug
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: CRC16_KUNIT_TEST [=m]
 Type  : tristate
 Defined at lib/Kconfig.debug:2861
   Prompt: KUnit tests for CRC16
   Depends on: RUNTIME_TESTING_MENU [=y] && KUNIT [=m]
   Location:
     -> Kernel hacking
       -> Kernel Testing and Coverage
         -> Runtime Testing (RUNTIME_TESTING_MENU [=y])
           -> KUnit tests for CRC16 (CRC16_KUNIT_TEST [=m])
 Selects: CRC16 [=y]

 Commit: 5d042707089f (lib/crc16_kunit.c: add KUnit tests for crc16)

---

 Symbol: UTIL_MACROS_KUNIT [=m]
 Type  : tristate
 Defined at lib/Kconfig.debug:2623
   Prompt: KUnit test util_macros.h functions at runtime
   Depends on: RUNTIME_TESTING_MENU [=y] && KUNIT [=m]
   Visible if: RUNTIME_TESTING_MENU [=y] && KUNIT [=m] && !KUNIT_ALL_TESTS [=m]
   Location:
     -> Kernel hacking
       -> Kernel Testing and Coverage
         -> Runtime Testing (RUNTIME_TESTING_MENU [=y])
           -> KUnit test util_macros.h functions at runtime (UTIL_MACROS_KUNIT [=m])

 Commit: 111314157f78 (lib: util_macros_kunit: add kunit test for util_macros.h)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>