[redhat] New configs in drivers/phy
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: PHY_NXP_PTN3222 [=n]
 Type  : tristate
 Defined at drivers/phy/Kconfig:85
   Prompt: NXP PTN3222 1-port eUSB2 to USB2 redriver
   Depends on: I2C [=y] && OF [=y]
   Location:
     -> Device Drivers
       -> PHY Subsystem
         -> NXP PTN3222 1-port eUSB2 to USB2 redriver (PHY_NXP_PTN3222 [=n])
 Selects: GENERIC_PHY [=n]

 Commit: c9be539e11f0 (phy: add NXP PTN3222 eUSB2 to USB2 redriver)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>