[redhat] New configs in drivers/rtc
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: RTC_DRV_RENESAS_RTCA3 [=n]
 Type  : tristate
 Defined at drivers/rtc/Kconfig:2019
   Prompt: Renesas RTCA-3 RTC
   Depends on: RTC_CLASS [=y] && ARCH_RENESAS [=y]
   Location:
     -> Device Drivers
       -> Real Time Clock (RTC_CLASS [=y])
         -> Renesas RTCA-3 RTC (RTC_DRV_RENESAS_RTCA3 [=n])

 Commit: d4488377609e (rtc: renesas-rtca3: Add driver for RTCA-3 available on Renesas RZ/G3S SoC)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>