redhat/Makefile: Add new dist-help functionality
The description of a Makefile target and the target itself are separated
in some cases by 100s of lines of code.  This is because the target
description is in the dist-full-help target, which is close to the bottom
of the Makefile.  This makes it difficult to remember to update a target's
help description when changing the functionality of a target, and it also
makes it difficult to remember to update the help targets when adding new
targets.

There is also too much text in the current output scheme for the
dist-full-help target.  It is just a wall of difficult to parse white
text, and it can't be easily subdivided into, for example, build vs kabi
targets.

The descriptions of each target listed in help have been moved to the
declaration of the target, with a "##<type>" style comment.  Functionality
for the dist-help and dist-full-help remain in place, albeit with a
cyan/light blue coloring for the target names which makes it easier to
read.

Additional targets for dist-help-HELPTYPE where HELPTYPE is one of
{cleaning build kabi configuration misc} have also been introduced.

Note: Some of the target descriptions are incorrect and may have other
issues (grammar, punctuation, etc.).  I have NOT changed these but will do
so in a follow up changeset.

Note: I chose '26' characters because it looks nice on both 80 and 100
column output.

Additional fix: Don't output the LOCALVERSION message when executing
dist-help commands.

v2: jstancek -- move descriptions to their own lines.

Signed-off-by: Prarit Bhargava <prarit@redhat.com>