[redhat] New configs in drivers/nvme
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: NVME_TARGET_PCI_EPF [=n]
 Type  : tristate
 Defined at drivers/nvme/target/Kconfig:119
   Prompt: NVMe PCI Endpoint Function target support
   Depends on: NVME_TARGET [=m] && PCI_ENDPOINT [=y] && (NVME_CORE [=m]=y [=y] || NVME_CORE [=m]=NVME_TARGET [=m])
   Location:
     -> Device Drivers
       -> NVME Support
         -> NVMe Target support (NVME_TARGET [=m])
           -> NVMe PCI Endpoint Function target support (NVME_TARGET_PCI_EPF [=n])

 Commit: 0faa0fe6f90e (nvmet: New NVMe PCI endpoint function target driver)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>