[redhat] New configs in drivers/irqchip
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: TI_SCI_INTA_IRQCHIP [=n]
 Type  : tristate
 Defined at drivers/irqchip/Kconfig:547
   Prompt: TI SCI INTA Interrupt Controller
   Depends on: TI_SCI_PROTOCOL [=y] && (ARCH_K3 [=y] || COMPILE_TEST [=n] && ARM64 [=y])
   Location:
     -> Device Drivers
       -> IRQ chip support
         -> TI SCI INTA Interrupt Controller (TI_SCI_INTA_IRQCHIP [=n])
 Selects: IRQ_DOMAIN_HIERARCHY [=y] && TI_SCI_INTA_MSI_DOMAIN [=n]

 Commit: 9f1463b86c13 (irqchip/ti-sci-inta: Add support for Interrupt Aggregator driver)

---

 Symbol: TI_SCI_INTR_IRQCHIP [=n]
 Type  : tristate
 Defined at drivers/irqchip/Kconfig:536
   Prompt: TI SCI INTR Interrupt Controller
   Depends on: TI_SCI_PROTOCOL [=y] && (ARCH_K3 [=y] || COMPILE_TEST [=n])
   Location:
     -> Device Drivers
       -> IRQ chip support
         -> TI SCI INTR Interrupt Controller (TI_SCI_INTR_IRQCHIP [=n])
 Selects: IRQ_DOMAIN_HIERARCHY [=y]

 Commit: cd844b0715ce (irqchip/ti-sci-intr: Add support for Interrupt Router driver)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>