[redhat] New configs in init/Kconfig
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: CGROUP_DMEM [=n]
 Type  : bool
 Defined at init/Kconfig:1143
   Prompt: Device memory controller (DMEM)
   Depends on: CGROUPS [=y]
   Location:
     -> General setup
       -> Control Group support (CGROUPS [=y])
         -> Device memory controller (DMEM) (CGROUP_DMEM [=n])
 Selects: PAGE_COUNTER [=y]

 Commit: b168ed458dde (kernel/cgroup: Add "dmem" memory accounting cgroup)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>

Signed-off-by: Scott Weaver <scweaver@redhat.com>