[redhat] New configs in kernel/rcu
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: RCU_TORTURE_TEST_CHK_RDR_STATE [=n]
 Type  : tristate
 Defined at kernel/rcu/Kconfig.debug:56
   Prompt: Check rcutorture reader state
   Depends on: RCU_TORTURE_TEST [=m]
   Location:
     -> Kernel hacking
       -> RCU Debugging
         -> torture tests for RCU (RCU_TORTURE_TEST [=m])
           -> Check rcutorture reader state (RCU_TORTURE_TEST_CHK_RDR_STATE [=n])

 Commit: ec9d6356bfda (rcutorture: Make rcutorture_one_extend() check reader state)

---

 Symbol: RCU_TORTURE_TEST_LOG_CPU [=n]
 Type  : tristate
 Defined at kernel/rcu/Kconfig.debug:72
   Prompt: Log CPU for rcutorture failures
   Depends on: RCU_TORTURE_TEST [=m]
   Location:
     -> Kernel hacking
       -> RCU Debugging
         -> torture tests for RCU (RCU_TORTURE_TEST [=m])
           -> Log CPU for rcutorture failures (RCU_TORTURE_TEST_LOG_CPU [=n])

 Commit: 579a05da40a4 (rcutorture: Decorate failing reader segments with CPU ID)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>