[redhat] New configs in crypto/Kconfig
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: CRYPTO_CRCT10DIF [=n]
 Type  : tristate
 Defined at crypto/Kconfig:1102
   Prompt: CRCT10DIF
   Depends on: CRYPTO [=y]
   Location:
     -> Cryptographic API (CRYPTO [=y])
       -> CRCs (cyclic redundancy checks)
         -> CRCT10DIF (CRYPTO_CRCT10DIF [=n])
 Selects: CRYPTO_HASH [=y] && CRC_T10DIF [=y]

 Commit: f1f142ad4348 (crypto: Kconfig - add submenus)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>