[redhat] New configs in drivers/hid
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: INTEL_THC_HID [=n]
 Type  : tristate
 Defined at drivers/hid/intel-thc-hid/Kconfig:7
   Prompt: Intel Touch Host Controller
   Depends on: HID_SUPPORT [=y] && X86_64 [=y] && PCI [=y] && ACPI [=y]
   Location:
     -> Device Drivers
       -> HID bus support (HID_SUPPORT [=y])
         -> Intel THC HID Support
           -> Intel Touch Host Controller (INTEL_THC_HID [=n])
 Selects: HID [=y]

 Commit: 1b2d05384c29 (HID: intel-thc-hid: Add basic THC driver skeleton)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>

v2:
- CONFIG_INTEL_THC_HID enabled as a module.
- CONFIG_INTEL_QUICKSPI enabled as a module.
- CONFIG_INTEL_QUICKI2C enabled as a module.

Signed-off-by: Patrick Talbert <ptalbert@redhat.com>