[redhat] New configs in drivers/hwmon
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: SENSORS_CRPS [=n]
 Type  : tristate
 Defined at drivers/hwmon/pmbus/Kconfig:88
   Prompt: Intel Common Redundant Power Supply
   Depends on: HWMON [=y] && PMBUS [=m]
   Location:
     -> Device Drivers
       -> Hardware Monitoring support (HWMON [=y])
         -> PMBus support (PMBUS [=m])
           -> Intel Common Redundant Power Supply (SENSORS_CRPS [=n])

 Commit: de076198d1e4 (hwmon: (pmbus/crps) Add Intel CRPS185 power supply)

---

 Symbol: SENSORS_TPS25990 [=n]
 Type  : tristate
 Defined at drivers/hwmon/pmbus/Kconfig:522
   Prompt: TI TPS25990
   Depends on: HWMON [=y] && PMBUS [=m]
   Location:
     -> Device Drivers
       -> Hardware Monitoring support (HWMON [=y])
         -> PMBus support (PMBUS [=m])
           -> TI TPS25990 (SENSORS_TPS25990 [=n])

 Commit: 339bca4ffdaa (hwmon: (pmbus/tps25990) Add initial support)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>