[redhat] New configs in drivers/mfd
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: MFD_QNAP_MCU [=n]
 Type  : tristate
 Defined at drivers/mfd/Kconfig:2389
   Prompt: QNAP microcontroller unit core driver
   Depends on: HAS_IOMEM [=y] && SERIAL_DEV_BUS [=y]
   Location:
     -> Device Drivers
       -> Multifunction device drivers
         -> QNAP microcontroller unit core driver (MFD_QNAP_MCU [=n])
 Selects: MFD_CORE [=y]

 Commit: 998f70d1806b (mfd: Add base driver for qnap-mcu devices)

---

 Symbol: MFD_UPBOARD_FPGA [=n]
 Type  : tristate
 Defined at drivers/mfd/Kconfig:2430
   Prompt: Support for the AAeon UP board FPGA
   Depends on: HAS_IOMEM [=y] && X86 [=y] && ACPI [=y]
   Location:
     -> Device Drivers
       -> Multifunction device drivers
         -> Support for the AAeon UP board FPGA (MFD_UPBOARD_FPGA [=n])
 Selects: MFD_CORE [=y]

 Commit: c2b148f3bc94 (mfd: Add support for AAEON UP board FPGA)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>