[redhat] New configs in lib/Kconfig.debug
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: CRC_BENCHMARK [=n]
 Type  : bool
 Defined at lib/Kconfig.debug:2855
   Prompt: Benchmark for the CRC functions
   Depends on: RUNTIME_TESTING_MENU [=y] && CRC_KUNIT_TEST [=m]
   Location:
     -> Kernel hacking
       -> Kernel Testing and Coverage
         -> Runtime Testing (RUNTIME_TESTING_MENU [=y])
           -> KUnit tests for CRC functions (CRC_KUNIT_TEST [=m])
             -> Benchmark for the CRC functions (CRC_BENCHMARK [=n])

 Commit: e47d9b1a76ed (lib/crc_kunit.c: add KUnit test suite for CRC library functions)

---

 Symbol: CRC_KUNIT_TEST [=m]
 Type  : tristate
 Defined at lib/Kconfig.debug:2841
   Prompt: KUnit tests for CRC functions
   Depends on: RUNTIME_TESTING_MENU [=y] && KUNIT [=m]
   Visible if: RUNTIME_TESTING_MENU [=y] && KUNIT [=m] && !KUNIT_ALL_TESTS [=m]
   Location:
     -> Kernel hacking
       -> Kernel Testing and Coverage
         -> Runtime Testing (RUNTIME_TESTING_MENU [=y])
           -> KUnit tests for CRC functions (CRC_KUNIT_TEST [=m])
 Selects: CRC16 [=y] && CRC_T10DIF [=y] && CRC32 [=y] && CRC64 [=y]

 Commit: e47d9b1a76ed (lib/crc_kunit.c: add KUnit test suite for CRC library functions)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>