[redhat] New configs in net/xfrm
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: XFRM_IPTFS [=n]
 Type  : tristate
 Defined at net/xfrm/Kconfig:138
   Prompt: IPsec IP-TFS/AGGFRAG (RFC 9347) encapsulation support
   Depends on: NET [=y] && INET [=y] && XFRM [=y]
   Location:
     -> Networking support (NET [=y])
       -> Networking options
         -> IPsec IP-TFS/AGGFRAG (RFC 9347) encapsulation support (XFRM_IPTFS [=n])

 Commit: ffa794846bf7 (xfrm: config: add CONFIG_XFRM_IPTFS)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>