redhat/kernel.spec: work around find-debuginfo aborting cross builds
debugedit > 5.0-16 now aborts cross builds. It was already attempting
to use host binaries on cross compiled ones, but with recent update
it also strictly checks for errors and aborts:

  objcopy: Unable to recognise the format of the input file `.../internal/drivers/gpu/drm/tests/drm_exec_test.ko'
  *** ERROR:: GDB exited with exit status 1 during index generation

This starts with commit:
  https://sourceware.org/git/?p=debugedit.git;a=commit;h=dfe1f7ff30f4e0be538835fca1e6348723ea7aa7

There's RFE opened to support overriding of host tools with one for
cross-compiling, RHEL-21797. But for now, undefine _include_gdb_index
to skip this code to allow cross builds to continue.

Signed-off-by: Jan Stancek <jstancek@redhat.com>