[redhat] New configs in arch/arm64
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: ARCH_BLAIZE [=n]
 Type  : bool
 Defined at arch/arm64/Kconfig.platforms:104
   Prompt: Blaize SoC Platforms
   Location:
     -> Platform selection
       -> Blaize SoC Platforms (ARCH_BLAIZE [=n])

 Commit: c0b454a51755 (arm64: Add Blaize BLZP1600 SoC family)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>