[redhat] New configs in drivers/firmware
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: FW_CS_DSP_KUNIT_TEST [=m]
 Type  : tristate
 Defined at drivers/firmware/cirrus/Kconfig:13
   Prompt: KUnit tests for Cirrus Logic cs_dsp
   Depends on: KUNIT [=m]
   Visible if: KUNIT [=m] && !KUNIT_ALL_TESTS [=m]
   Location:
     -> Device Drivers
       -> Firmware Drivers
         -> KUnit tests for Cirrus Logic cs_dsp (FW_CS_DSP_KUNIT_TEST [=m])
 Selects: REGMAP [=y] && FW_CS_DSP [=m] && FW_CS_DSP_KUNIT_TEST_UTILS [=m]

 Commit: dd0b6b1f29b9 (firmware: cs_dsp: Add KUnit testing of bin file download)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>