[redhat] New configs in drivers/i3c
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: MIPI_I3C_HCI_PCI [=n]
 Type  : tristate
 Defined at drivers/i3c/master/Kconfig:61
   Prompt: MIPI I3C Host Controller Interface PCI support
   Depends on: I3C [=m] && MIPI_I3C_HCI [=m] && PCI [=y]
   Location:
     -> Device Drivers
       -> I3C support (I3C [=m])
         -> MIPI I3C Host Controller Interface driver (EXPERIMENTAL) (MIPI_I3C_HCI [=m])
           -> MIPI I3C Host Controller Interface PCI support (MIPI_I3C_HCI_PCI [=n])

 Commit: 30bb1ce71215 (i3c: mipi-i3c-hci: Add support for MIPI I3C HCI on PCI bus)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>