[redhat] New configs in drivers/soc
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: ARCH_R9A09G047 [=n]
 Type  : bool
 Defined at drivers/soc/renesas/Kconfig:348
   Prompt: ARM64 Platform support for RZ/G3E
   Depends on: SOC_RENESAS [=y] && ARM64 [=y]
   Location:
     -> Device Drivers
       -> SOC (System On Chip) specific Drivers
         -> Renesas SoC driver support (SOC_RENESAS [=y])
           -> ARM64 Platform support for RZ/G3E (ARCH_R9A09G047 [=n])

 Commit: 12e0bd600e3c (soc: renesas: Add RZ/G3E (R9A09G047) config option)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>