[redhat] New configs in sound/pci
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: SND_HDA_SCODEC_TAS2781_SPI [=n]
 Type  : tristate
 Defined at sound/pci/hda/Kconfig:209
   Prompt: Build TAS2781 HD-audio side codec support for SPI Bus
   Depends on: SOUND [=m] && SND [=m] && SND_HDA [=m] && SPI_MASTER [=y] && ACPI [=y] && EFI [=y] && SND_SOC [=m]
   Location:
     -> Device Drivers
       -> Sound card support (SOUND [=m])
         -> Advanced Linux Sound Architecture (SND [=m])
           -> HD-Audio
             -> Build TAS2781 HD-audio side codec support for SPI Bus (SND_HDA_SCODEC_TAS2781_SPI [=n])
 Selects: CRC32 [=y]

 Commit: bb5f86ea50ff (ALSA: hda/tas2781: Add tas2781 hda SPI driver)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>