[redhat] New configs in samples/damon
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: SAMPLE_DAMON_PRCL [=n]
 Type  : bool
 Defined at samples/damon/Kconfig:17
   Prompt: DAMON sameple module for access-aware proactive reclamation
   Depends on: SAMPLES [=y] && DAMON [=y] && DAMON_VADDR [=y]
   Location:
     -> Kernel hacking
       -> Sample kernel code (SAMPLES [=y])
         -> DAMON Samples
           -> DAMON sameple module for access-aware proactive reclamation (SAMPLE_DAMON_PRCL [=n])

 Commit: 2aca254620a8 (samples/damon: introduce a skeleton of a smaple DAMON module for proactive reclamation)

---

 Symbol: SAMPLE_DAMON_WSSE [=n]
 Type  : bool
 Defined at samples/damon/Kconfig:5
   Prompt: DAMON sameple module for working set size estimation
   Depends on: SAMPLES [=y] && DAMON [=y] && DAMON_VADDR [=y]
   Location:
     -> Kernel hacking
       -> Sample kernel code (SAMPLES [=y])
         -> DAMON Samples
           -> DAMON sameple module for working set size estimation (SAMPLE_DAMON_WSSE [=n])

 Commit: 19d7c3adfdd4 (samples: add a skeleton of a sample DAMON module for working set size estimation)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>