[redhat] New configs in drivers/power
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: FUEL_GAUGE_STC3117 [=n]
 Type  : tristate
 Defined at drivers/power/supply/Kconfig:921
   Prompt: STMicroelectronics STC3117 fuel gauge driver
   Depends on: POWER_SUPPLY [=y] && CRC8 [=m] && I2C [=y]
   Location:
     -> Device Drivers
       -> Power supply class support (POWER_SUPPLY [=y])
         -> STMicroelectronics STC3117 fuel gauge driver (FUEL_GAUGE_STC3117 [=n])
 Selects: REGMAP_I2C [=m]

 Commit: 74e3f6208523 (power: supply: Add STC3117 fuel gauge unit driver)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>