[redhat] New configs in drivers/pps
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: PPS_GENERATOR [=n]
 Type  : tristate
 Defined at drivers/pps/generators/Kconfig:6
   Prompt: PPS generators support
   Depends on: PPS [=y]
   Location:
     -> Device Drivers
       -> PPS support (PPS [=y])
         -> PPS generators support (PPS_GENERATOR [=n])

 Commit: 86b525bed275 (drivers pps: add PPS generators support)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>