[redhat] New configs in drivers/usb
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: TYPEC_TBT_ALTMODE [=n]
 Type  : tristate
 Defined at drivers/usb/typec/altmodes/Kconfig:26
   Prompt: Thunderbolt3 Alternate Mode driver
   Depends on: USB_SUPPORT [=y] && TYPEC [=y]
   Location:
     -> Device Drivers
       -> USB support (USB_SUPPORT [=y])
         -> USB Type-C Support (TYPEC [=y])
           -> USB Type-C Alternate Mode drivers
             -> Thunderbolt3 Alternate Mode driver (TYPEC_TBT_ALTMODE [=n])

 Commit: 100e25738659 (usb: typec: Add driver for Thunderbolt 3 Alternate Mode)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>