redhat/configs: disable CONFIG_AF_UNIX_OOB on RHEL
This commit pretty much reverts
https://gitlab.com/cki-project/kernel-ark/-/merge_requests/3635
This was already on because of the lack of option, but now that we have
it, we actually didn't have any request to enable this on RHEL 9, which
doens't have this feature. With that, we also don't have a reason to
have it enabled on RHEL 10. So lets disable it while we can.

Signed-off-by: Marcelo Ricardo Leitner <mleitner@redhat.com>