redhat/configs: Move CONFIG_PROVE_RAW_LOCK_NESTING to common/debug
CONFIG_PROVE_RAW_LOCK_NESTING is a debug option that helps debugging RT
locking issue in !RT kernel. This option is only enabled on debug kernel
with CONFIG_PROVE_LOCKING on. Move it to common.

As the common/generic version isn't really needed, just remove it.

This kernel config option is forced on for arches that set
ARCH_SUPPORTS_RT (x86 & arm64), though it can be turned off for other
arches. For now, just turn it on for all and we can disable specific
arch later if needed.

Signed-off-by: Waiman Long <longman@redhat.com>