redhat/configs: enable in kernel compression/decompression of compressed modules
Our packaging for some time has shipped modules compressed with xz.
These are decompressed by modprobe/kmod on userspace, then provided to the
kernel.

However, the kernel for some time now supports in-kernel decompression since
the change b1ae6dc41eaa ("module: add in-kernel support for decompressing").
Doing in kernel decompression is better as pointed out by Andrew Zaborowski in
RHEL-79278: it allows kmod to use the finit_module() syscall, which allows the
kernel to better detect concurrent loading of modules preventing spikes in usage
of memory/vmalloc memory. The prevention of memory spikes/usage in concurrent
module loading was handled by Linus in commit 9b9879fc0327 ("modules: catch
concurrent module loads, treat them as idempotent"). But we are not able to take
advantage of this fix if we don't have in kernel decompression of modules enabled,
since kmod will not use finit_module() as the module is compressed and it
fallbacks to init_module().

Thus, make sure we enable in-kernel decompression of modules, which requires to
enable MODULE_COMPRESS: the key here is to keep MODULE_COMPRESS_ALL disabled,
which allows us to keep our current build infra-structure/spec file untouched,
since we build and compress the modules by ourselves and in the proper order
required due module signing. At the same time, it should allow external modules
to continue to be built without interference of enabling compress/decompress.

Signed-off-by: Herton R. Krzesinski <herton@redhat.com>