redhat: change compression flags for xz
The macro compression_flags is used in the compression of kernel modules.
The default xz integrity check used is crc64. In the next change, we
will switch to in-kernel decompression of modules. The problem is that
it doesn't have support for the crc64 check, so we must manually switch
to crc32. While at it, also add --lzma2=dict=1MiB which is also used by
the kernel at scripts/Makefile.modinst and in general follows
recommendation/dictionary sizes explained in the kernel documentation
at Documentation/staging/xz.rst.

Signed-off-by: Herton R. Krzesinski <herton@redhat.com>