kernel.spec: add missing tools-libs on s390x
tools-libs-devel requires tools-libs. Previously on s390x they would
be both skipped by %cpupowerarches check. But after adding ynl
headers and library to tools-libs-devel, it is now built also on s390x.
But tools-libs is still skipped and creates a missing dependency.

We could conditionally skip the Requires on s390x, but that adds extra
burden to keep such condition up to date as more artefacts get introduced
to tools-libs-devel in future. Simplest is to just create tools-libs on
s390x, even if it ends up empty for now.

Fixes: e9f967afa161 ("redhat: kernel.spec: add ynl to kernel-tools")

Reported-by: Bruno Goncalves <bgoncalv@redhat.com>
Signed-off-by: Jan Stancek <jstancek@redhat.com>