redhat: Consolidate CONFIG_VIRTIO_MEM and enable it also for s390x
s390x gained support for virtio-mem recently, see the ticket
https://issues.redhat.com/browse/RHEL-65048 for details. So
we want to enable CONFIG_VIRTIO_MEM now for all platforms that
support it, i.e. move the setting to the generic common folder
(just keep it disabled in the s390x zfcpdump kernel since that
should be kept as small as possible and likely does not need
memory hot-plugging.

Signed-off-by: Thomas Huth <thuth@redhat.com>