[redhat] New configs in drivers/bluetooth
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: BT_HCIBTUSB_AUTO_ISOC_ALT [=n]
 Type  : bool
 Defined at drivers/bluetooth/Kconfig:59
   Prompt: Automatically adjust alternate setting for Isoc endpoints
   Depends on: NET [=y] && BT [=m] && BT_HCIBTUSB [=m]
   Location:
     -> Networking support (NET [=y])
       -> Bluetooth subsystem support (BT [=m])
         -> Bluetooth device drivers
           -> HCI USB driver (BT_HCIBTUSB [=m])
             -> Automatically adjust alternate setting for Isoc endpoints (BT_HCIBTUSB_AUTO_ISOC_ALT [=n])

 Commit: 75ddcd5ad40e (Bluetooth: btusb: Configure altsetting for HCI_USER_CHANNEL)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>