redhat: introduce modules-extra-matched meta package
There are packages that require kmods from modules-extra to function
correctly. These however have no concept of what kernel variants
and versions are installed. Add a meta package that can act as
a single point of reference, that these other packages can Require
and leave it to kernel to pull in correct variant and version
of modules-extra package.

Signed-off-by: Jan Stancek <jstancek@redhat.com>