[redhat] New configs in kernel/rcu
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: FORCE_NEED_SRCU_NMI_SAFE [=n]
 Type  : bool
 Defined at kernel/rcu/Kconfig:68
   Prompt: Force selection of NEED_SRCU_NMI_SAFE
   Depends on: !TINY_SRCU [=n]
   Location:
     -> General setup
       -> RCU Subsystem
         -> Force selection of NEED_SRCU_NMI_SAFE (FORCE_NEED_SRCU_NMI_SAFE [=n])
 Selects: NEED_SRCU_NMI_SAFE [=n]

 Commit: 536e8b9b80bc (srcu: Add FORCE_NEED_SRCU_NMI_SAFE Kconfig for testing)

---

 Symbol: RCU_TORTURE_TEST_LOG_GP [=n]
 Type  : bool
 Defined at kernel/rcu/Kconfig.debug:87
   Prompt: Log grace-period numbers for rcutorture failures
   Depends on: RCU_TORTURE_TEST [=m]
   Location:
     -> Kernel hacking
       -> RCU Debugging
         -> torture tests for RCU (RCU_TORTURE_TEST [=m])
           -> Log grace-period numbers for rcutorture failures (RCU_TORTURE_TEST_LOG_GP [=n])

 Commit: 84ae91018af5 (rcutorture: Include grace-period sequence numbers in failure/close-call)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>