[redhat] New configs in security/Kconfig.hardening
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: HARDENED_USERCOPY_DEFAULT_ON [=y]
 Type  : bool
 Defined at security/Kconfig.hardening:306
   Prompt: Harden memory copies by default
   Depends on: HARDENED_USERCOPY [=y]
   Location:
     -> Security options
       -> Kernel hardening options
         -> Bounds checking
           -> Harden memory copies between kernel and userspace (HARDENED_USERCOPY [=y])
             -> Harden memory copies by default (HARDENED_USERCOPY_DEFAULT_ON [=y])

 Commit: d2132f453e33 (mm: security: Allow default HARDENED_USERCOPY to be set at compile time)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>