[redhat] New configs in drivers/cpufreq
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: CPUFREQ_ARCH_CUR_FREQ [=y]
 Type  : bool
 Defined at drivers/cpufreq/Kconfig.x86:343
   Prompt: Current frequency derived from HW provided feedback
   Depends on: CPU_FREQ [=y] && X86 [=y]
   Location:
     -> Power management and ACPI options
       -> CPU Frequency scaling
         -> CPU Frequency scaling (CPU_FREQ [=y])
           -> Current frequency derived from HW provided feedback (CPUFREQ_ARCH_CUR_FREQ [=y])

 Commit: fbb4a4759b54 (cpufreq: Introduce an optional cpuinfo_avg_freq sysfs entry)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>