[redhat] New configs in arch/arm64
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: ROCKCHIP_ERRATUM_3568002 [=y]
 Type  : bool
 Defined at arch/arm64/Kconfig:1307
   Prompt: Rockchip 3568002: GIC600 can not access physical addresses higher than 4GB
   Location:
     -> Kernel Features
       -> ARM errata workarounds via the alternatives framework
         -> Rockchip 3568002: GIC600 can not access physical addresses higher than 4GB (ROCKCHIP_ERRATUM_3568002 [=y])

 Commit: 2d81e1bb6252 (irqchip/gic-v3: Add Rockchip 3568002 erratum workaround)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>