[redhat] New configs in drivers/hv
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: MSHV_ROOT [=n]
 Type  : tristate
 Defined at drivers/hv/Kconfig:58
   Prompt: Microsoft Hyper-V root partition support
   Depends on: HYPERV [=m] && (X86_64 [=y] || ARM64) && !HYPERV_VTL_MODE [=n] && PAGE_SIZE_4KB [=y]
   Location:
     -> Device Drivers
       -> Microsoft Hyper-V guest support
         -> Microsoft Hyper-V root partition support (MSHV_ROOT [=n])
 Selects: EVENTFD [=y]

 Commit: 461fbbd036b1 (hyperv: Add CONFIG_MSHV_ROOT to gate root partition support)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>