[redhat] New configs in drivers/spi
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: SPI_QPIC_SNAND [=n]
 Type  : bool
 Defined at drivers/spi/Kconfig:939
   Prompt: QPIC SNAND controller
   Depends on: SPI [=y] && SPI_MASTER [=y] && (ARCH_QCOM [=y] || COMPILE_TEST [=n])
   Location:
     -> Device Drivers
       -> SPI support (SPI [=y])
         -> QPIC SNAND controller (SPI_QPIC_SNAND [=n])
 Selects: MTD [=m]

 Commit: 7304d1909080 (spi: spi-qpic: add driver for QCOM SPI NAND flash Interface)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>