[redhat] New configs in drivers/net/ethernet/realtek
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: R8169_LEDS [=n]
 Type  : bool
 Defined at drivers/net/ethernet/realtek/Kconfig:116
   Prompt: Support for controlling the NIC LEDs
   Depends on: NETDEVICES [=y] && ETHERNET [=y] && NET_VENDOR_REALTEK [=y] && R8169 [=m] && LEDS_TRIGGER_NETDEV [=m] && (R8169 [=m]!=y [=y] || LEDS_CLASS [=y]!=m [=m])
   Location:
     -> Device Drivers
       -> Network device support (NETDEVICES [=y])
         -> Ethernet driver support (ETHERNET [=y])
           -> Realtek devices (NET_VENDOR_REALTEK [=y])
             -> Realtek 8169/8168/8101/8125 ethernet support (R8169 [=m])
               -> Support for controlling the NIC LEDs (R8169_LEDS [=n])

 Commit: a2634a5ffcaf (r8169: fix building with CONFIG_LEDS_CLASS=m)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>