[redhat] New configs in arch/powerpc
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: HTMDUMP [=m]
 Type  : tristate
 Defined at arch/powerpc/platforms/pseries/Kconfig:132
   Prompt: PowerVM data dumper
   Depends on: PPC_PSERIES [=y] && DEBUG_FS [=y]
   Location:
     -> Platform support
       -> PowerVM data dumper (HTMDUMP [=m])

 Commit: 81c3d637c30f (powerpc/pseries: Export hardware trace macro dump via debugfs)

---

 Symbol: PPC_MICROWATT [=n]
 Type  : bool
 Defined at arch/powerpc/platforms/microwatt/Kconfig:2
   Prompt: Microwatt SoC platform
   Depends on: PPC_BOOK3S_64 [=y]
   Location:
     -> Platform support
       -> Microwatt SoC platform (PPC_MICROWATT [=n])
 Selects: PPC_XICS [=y] && PPC_ICS_NATIVE [=n] && PPC_ICP_NATIVE [=y] && PPC_UDBG_16550 [=y] && COMMON_CLK [=y]

 Commit: 53d143fe08c2 (powerpc: Add Microwatt platform)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>