[redhat] New configs in kernel/trace
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: RV_MON_SCHED [=n]
 Type  : bool
 Defined at kernel/trace/rv/monitors/sched/Kconfig:3
   Prompt: sched monitor
   Depends on: FTRACE [=y] && RV [=y]
   Location:
     -> Kernel hacking
       -> Tracers (FTRACE [=y])
         -> Runtime Verification (RV [=y])
           -> sched monitor (RV_MON_SCHED [=n])

 Commit: cb85c660fcd4 (rv: Add option for nested monitors and include sched)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>