[redhat] New configs in lib/Kconfig.debug
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: LONGEST_SYM_KUNIT_TEST [=m]
 Type  : tristate
 Defined at lib/Kconfig.debug:2869
   Prompt: Test the longest symbol possible
   Depends on: RUNTIME_TESTING_MENU [=y] && KUNIT [=m] && KPROBES [=y]
   Visible if: RUNTIME_TESTING_MENU [=y] && KUNIT [=m] && KPROBES [=y] && !KUNIT_ALL_TESTS [=m]
   Location:
     -> Kernel hacking
       -> Kernel Testing and Coverage
         -> Runtime Testing (RUNTIME_TESTING_MENU [=y])
           -> Test the longest symbol possible (LONGEST_SYM_KUNIT_TEST [=m])

 Commit: c104c16073b7 (Kunit to check the longest symbol length)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>