redhat: bump RHEL_MAJOR
Similar to 93f658c7b8a8 ("redhat: bump RHEL_MAJOR and add the changelog file for it"),
bump RHEL_MAJOR. kernel-ark does not track anymore work for RHEL 10/CentOS Stream 10.
We look forward to the next major. While "9.99" was chosen as the version for
the work preceding RHEL 10, bump the major so the considered version now is
"10.99".

Changelog will be created automatically on next release. Old entries are not moved,
to avoid conflicts in the next os-build rebase. All changelog entries will be picked
up with the first release commit in the next os-build rebase.

Signed-off-by: Jan Stancek <jstancek@redhat.com>