[redhat] New configs in crypto/Kconfig
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: CRYPTO_KRB5ENC [=n]
 Type  : tristate
 Defined at crypto/Kconfig:237
   Prompt: Kerberos 5 combined hash+cipher support
   Depends on: CRYPTO [=y]
   Location:
     -> Cryptographic API (CRYPTO [=y])
       -> Crypto core or helper
         -> Kerberos 5 combined hash+cipher support (CRYPTO_KRB5ENC [=n])
 Selects: CRYPTO_AEAD [=y] && CRYPTO_SKCIPHER [=y] && CRYPTO_MANAGER [=y] && CRYPTO_HASH [=y] && CRYPTO_NULL [=y]
 Selected by [n]:
   - CRYPTO_KRB5 [=n] && CRYPTO [=y]

 Commit: d1775a177f7f (crypto: Add 'krb5enc' hash and cipher AEAD algorithm)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>