[redhat] New configs in arch/s390
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: STRICT_MM_TYPECHECKS [=n]
 Type  : bool
 Defined at arch/s390/Kconfig.debug:16
   Prompt: Strict Memory Management Type Checks
   Depends on: DEBUG_KERNEL [=y]
   Location:
     -> Kernel hacking
       -> s390 Debugging
         -> Strict Memory Management Type Checks (STRICT_MM_TYPECHECKS [=n])

 Commit: 03544866df1b (s390/mm: Add configurable STRICT_MM_TYPECHECKS)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>