[redhat] New configs in drivers/char
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: TCG_ARM_CRB_FFA [=m]
 Type  : tristate
 Defined at drivers/char/tpm/Kconfig:213
   Prompt: TPM CRB over Arm FF-A Transport
   Depends on: TCG_TPM [=y] && ARM_FFA_TRANSPORT [=m] && TCG_CRB [=y]
   Location:
     -> Device Drivers
       -> Character devices
         -> TPM Hardware Support (TCG_TPM [=y])
           -> TPM 2.0 CRB Interface (TCG_CRB [=y])
             -> TPM CRB over Arm FF-A Transport (TCG_ARM_CRB_FFA [=m])

 Commit: eb93f0734ef1 (tpm_crb: ffa_tpm: Implement driver compliant to CRB over FF-A)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>