[redhat] New configs in drivers/fwctl
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: FWCTL [=n]
 Type  : tristate
 Defined at drivers/fwctl/Kconfig:2
   Prompt: fwctl device firmware access framework
   Location:
     -> Device Drivers
       -> fwctl device firmware access framework (FWCTL [=n])
 Selected by [n]:
   - CXL_BUS [=m] && PCI [=y] && CXL_FEATURES [=n]

 Commit: 2e4986cf2d52 (fwctl: Add basic structure for a class subsystem with a cdev)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>