[redhat] New configs in drivers/media
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: CEC_NXP_TDA9950 [=n]
 Type  : tristate
 Defined at drivers/media/cec/i2c/Kconfig:17
   Prompt: NXP Semiconductors TDA9950/TDA998X HDMI CEC
   Depends on: MEDIA_CEC_SUPPORT [=y]
   Location:
     -> Device Drivers
       -> CEC support
         -> HDMI CEC drivers (MEDIA_CEC_SUPPORT [=y])
           -> NXP Semiconductors TDA9950/TDA998X HDMI CEC (CEC_NXP_TDA9950 [=n])
 Selects: CEC_NOTIFIER [=y] && CEC_CORE [=m]

 Commit: caa6f4a75e9f (media: cec: move driver for TDA9950 from drm/i2c)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>