[redhat] New configs in fs/nfsd
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: NFSD_V4_DELEG_TIMESTAMPS [=n]
 Type  : bool
 Defined at fs/nfsd/Kconfig:179
   Prompt: Support delegated timestamps
   Depends on: NETWORK_FILESYSTEMS [=y] && NFSD_V4 [=y]
   Location:
     -> File systems
       -> Network File Systems (NETWORK_FILESYSTEMS [=y])
         -> Support delegated timestamps (NFSD_V4_DELEG_TIMESTAMPS [=n])

 Commit: 26a80762153b (NFSD: Add a Kconfig setting to enable delegated timestamps)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>