[redhat] New configs in drivers/hwtracing
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: CORESIGHT_CTCU [=n]
 Type  : tristate
 Defined at drivers/hwtracing/coresight/Kconfig:136
   Prompt: CoreSight TMC Control Unit driver
   Depends on: CORESIGHT [=m] && CORESIGHT_LINK_AND_SINK_TMC [=m]
   Location:
     -> Kernel hacking
       -> arm64 Debugging
         -> CoreSight Tracing Support (CORESIGHT [=m])
           -> CoreSight TMC Control Unit driver (CORESIGHT_CTCU [=n])

 Commit: f78d206f3d73 (Coresight: Add Coresight TMC Control Unit driver)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>