[redhat] New configs in drivers/phy
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: PHY_QCOM_UNIPHY_PCIE_28LP [=n]
 Type  : bool
 Defined at drivers/phy/qualcomm/Kconfig:157
   Prompt: PCIE UNIPHY 28LP PHY driver
   Depends on: ARCH_QCOM [=y] && COMMON_CLK [=y] && HAS_IOMEM [=y] && OF [=y]
   Location:
     -> Device Drivers
       -> PHY Subsystem
         -> PCIE UNIPHY 28LP PHY driver (PHY_QCOM_UNIPHY_PCIE_28LP [=n])
 Selects: GENERIC_PHY [=y]

 Commit: 74badb8b0b14 (phy: qcom: Introduce PCIe UNIPHY 28LP driver)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>