[redhat] New configs in drivers/watchdog
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: LENOVO_SE30_WDT [=n]
 Type  : tristate
 Defined at drivers/watchdog/Kconfig:282
   Prompt: Lenovo SE30 Watchdog
   Depends on: WATCHDOG [=y] && (X86 [=y] && DMI [=y] || COMPILE_TEST [=n]) && HAS_IOPORT [=y]
   Location:
     -> Device Drivers
       -> Watchdog Timer Support (WATCHDOG [=y])
         -> Lenovo SE30 Watchdog (LENOVO_SE30_WDT [=n])
 Selects: WATCHDOG_CORE [=y]

 Commit: c284153a2c55 (watchdog: lenovo_se30_wdt: Watchdog driver for Lenovo SE30 platform)

---

 Symbol: RENESAS_RZV2HWDT [=n]
 Type  : tristate
 Defined at drivers/watchdog/Kconfig:977
   Prompt: Renesas RZ/{G3E,V2H(P)} WDT Watchdog
   Depends on: WATCHDOG [=y] && (ARCH_RENESAS [=y] || COMPILE_TEST [=n]) && (PM [=y] || COMPILE_TEST [=n])
   Location:
     -> Device Drivers
       -> Watchdog Timer Support (WATCHDOG [=y])
         -> Renesas RZ/{G3E,V2H(P)} WDT Watchdog (RENESAS_RZV2HWDT [=n])
 Selects: WATCHDOG_CORE [=m]

 Commit: f6febd0a30b6 (watchdog: Add Watchdog Timer driver for RZ/V2H(P))

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>