[redhat] New configs in lib/Kconfig.debug
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: DETECT_HUNG_TASK_BLOCKER [=y]
 Type  : bool
 Defined at lib/Kconfig.debug:1283
   Prompt: Dump Hung Tasks Blocker
   Depends on: DETECT_HUNG_TASK [=y] && !PREEMPT_RT [=n]
   Location:
     -> Kernel hacking
       -> Debug Oops, Lockups and Hangs
         -> Detect Hung Tasks (DETECT_HUNG_TASK [=y])
           -> Dump Hung Tasks Blocker (DETECT_HUNG_TASK_BLOCKER [=y])

 Commit: 3cf67d61ff98 (hung_task: show the blocker task if the task is hung on mutex)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>