[redhat] New configs in drivers/tty
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: SERIAL_8250_NI [=n]
 Type  : tristate
 Defined at drivers/tty/serial/8250/Kconfig:572
   Prompt: NI 16550 based serial port
   Depends on: TTY [=y] && HAS_IOMEM [=y] && SERIAL_8250 [=y] && (X86 [=y] && ACPI [=y] || COMPILE_TEST [=n])
   Location:
     -> Device Drivers
       -> Character devices
         -> Enable TTY (TTY [=y])
           -> Serial drivers
             -> NI 16550 based serial port (SERIAL_8250_NI [=n])

 Commit: 3d5390f4dbe6 (serial: 8250: add driver for NI UARTs)

---

 Symbol: SERIAL_TEGRA_UTC [=n]
 Type  : tristate
 Defined at drivers/tty/serial/Kconfig:290
   Prompt: NVIDIA Tegra UART Trace Controller
   Depends on: TTY [=y] && HAS_IOMEM [=y] && (ARCH_TEGRA [=y] || COMPILE_TEST [=n])
   Location:
     -> Device Drivers
       -> Character devices
         -> Enable TTY (TTY [=y])
           -> Serial drivers
             -> NVIDIA Tegra UART Trace Controller (SERIAL_TEGRA_UTC [=n])
 Selects: SERIAL_CORE [=y]

 Commit: eb07e3a94679 (serial: tegra-utc: Add driver for Tegra UART Trace Controller (UTC))

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>