[redhat] New configs in drivers/usb
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: TYPEC_MUX_PS883X [=n]
 Type  : tristate
 Defined at drivers/usb/typec/mux/Kconfig:59
   Prompt: Parade PS883x Type-C retimer driver
   Depends on: USB_SUPPORT [=y] && TYPEC [=y] && I2C [=y] && (DRM [=y] || DRM [=y]=n [=n])
   Location:
     -> Device Drivers
       -> USB support (USB_SUPPORT [=y])
         -> USB Type-C Support (TYPEC [=y])
           -> USB Type-C Multiplexer/DeMultiplexer Switch support
             -> Parade PS883x Type-C retimer driver (TYPEC_MUX_PS883X [=n])
 Selects: DRM_AUX_BRIDGE [=n] && REGMAP_I2C [=m]

 Commit: 257a087c8b52 (usb: typec: Add support for Parade PS8830 Type-C Retimer)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>