[redhat] New configs in security/Kconfig
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: MSEAL_SYSTEM_MAPPINGS [=n]
 Type  : bool
 Defined at security/Kconfig:54
   Prompt: mseal system mappings
   Depends on: 64BIT [=y] && ARCH_SUPPORTS_MSEAL_SYSTEM_MAPPINGS [=y] && !CHECKPOINT_RESTORE [=n]
   Location:
     -> Security options
       -> mseal system mappings (MSEAL_SYSTEM_MAPPINGS [=n])

 Commit: 5796d3967c09 (mseal sysmap: kernel config and header change)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>