[redhat] New configs in arch/s390
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: MARCH_Z17 [=n]
 Type  : bool
 Defined at arch/s390/Kconfig:404
   Prompt: IBM z17
   Location:
     -> Processor type and features
       -> Processor type (<choice> [=n])
         -> IBM z17 (MARCH_Z17 [=n])
 Selects: HAVE_MARCH_Z17_FEATURES [=n]

 Commit: c51ea9888e88 (s390: Allow to compile with z17 optimizations)

---

 Symbol: TUNE_Z17 [=n]
 Type  : bool
 Defined at arch/s390/Kconfig:482
   Prompt: IBM z17
   Location:
     -> Processor type and features
       -> Tune code generation (<choice> [=n])
         -> IBM z17 (TUNE_Z17 [=n])

 Commit: c51ea9888e88 (s390: Allow to compile with z17 optimizations)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>