redhat: add namespace selftests to kernel-modules-internal package
Upstream Status: RHEL-only
JIRA: https://issues.redhat.com/browse/RHEL-88635

As the time of this commit there is no targeted testing in CKI for the
namespace subsystem. The addition of the two extant namespace kselftests
to CKI will pave the way for us to trigger these tests when the
underlying tested files are modified.

Signed-off-by: Joel Savitz <jsavitz@redhat.com>