[redhat] New configs in arch/x86
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: MITIGATION_ITS [=y]
 Type  : bool
 Defined at arch/x86/Kconfig:2714
   Prompt: Enable Indirect Target Selection mitigation
   Depends on: CPU_MITIGATIONS [=y] && CPU_SUP_INTEL [=y] && X86_64 [=y] && MITIGATION_RETPOLINE [=y] && MITIGATION_RETHUNK [=y]
   Location:
     -> Mitigations for CPU vulnerabilities (CPU_MITIGATIONS [=y])
       -> Enable Indirect Target Selection mitigation (MITIGATION_ITS [=y])
 Selects: EXECMEM [=y]

 Commit: 8754e67ad4ac (x86/its: Add support for ITS-safe indirect thunk)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>