redhat/Makefile: introduce AUTOMOTIVE_BUILD
This adds the automotive specific targets dist-automotive-configs and
dist-cbs, and introduces the AUTOMOTIVE_BUILD variable that can be used
to repurpose the existing stock kernel make targets for
kernel-automotive.

For example:
  make AUTOMOTIVE_BUILD=1 dist-srpm

In order to treat kernel-automotive as a separate package rather than a
standard variant of the stock kernel, this generates the configs based
on a temporary version of the priority.rhel file. In this way, the
automotive configs are generated as kernel-automotive instead of having
a strange looking name that implies we are generating the automotive
variant of kernel-automotive (which would happen if we only changed the
package name). This has the added benefit of only needing to build the
automotive configs rather than all of the stock kernel configs.

The kernel-automotive package is not a CentOS Stream package, but its
equivalent is the AutoSD distribution maintained by the Automotive SIG.
Packages for the SIG are built in CBS and thus the dist-cbs target was
added specifically for the purpose of making kernel-automotive scratch
builds.

Signed-off-by: Scott Weaver <scweaver@redhat.com>