kernel.spec: updates for automotive-only build
When we generate the kernel-automotive srpm, we only want to include
sources that will be required for that build. For automotive the
INCLUDE_FEDORA_FILES, INCLUDE_RHEL_FILES and INCLUDE_RT_FILES will be
disabled which means we need to arrange the sources so that
kernel-automotive will have all the files it requires.

When building automotive only, to generate kernel-automotive packages,
it doesn't make sense to create the automotive variant of automotive
(kernel-automotive-automotive) which would be the result if we only
changed the package name. Therefore, with automotive-only, we generate
'stock' kernel-automotive packages while keeping the ability to also
generate the automotive variant of the stock kernel if that would ever
be needed.

Signed-off-by: Scott Weaver <scweaver@redhat.com>