Enable building libcpupower bindings for ELN/Rawhide
Python bindings for libcpupower were introduced in Linux 6.12 that are
needed for tools like Tuna to manage c-states (the power states of a cpu).
6.17-rc1 introduced changes to make it possible for distributions like
Fedora Rawhide and ELN to build the bindings were merged in.

The MR associated with this issue will allow the bindings to build. With
consultation on the Fedora side kernel-tools-lib-devel was requested to be
the package to add the two files that are needed by the bindings. SWIG is
used to generate the bindings.

Signed-off-by: John B. Wyatt IV <jwyatt@redhat.com>
Signed-off-by: John B. Wyatt IV <sageofredondo@gmail.com>