[redhat] New configs in arch/x86
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: X86_NATIVE_CPU [=n]
 Type  : bool
 Defined at arch/x86/Kconfig.cpu:258
   Prompt: Build and optimize for local/native CPU
   Depends on: X86_64 [=y] && CC_HAS_MARCH_NATIVE [=y]
   Location:
     -> Processor type and features
       -> Build and optimize for local/native CPU (X86_NATIVE_CPU [=n])

 Commit: ea1dcca1de12 (x86/kbuild/64: Add the CONFIG_X86_NATIVE_CPU option to locally optimize the kernel with '-march=native')

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>