[redhat] New configs in crypto/Kconfig: CRYPTO_BENCHMARK
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: CRYPTO_BENCHMARK [=n]
 Type  : tristate
 Defined at crypto/Kconfig:240
   Prompt: Crypto benchmarking module
   Depends on: CRYPTO [=y] && (m [=m] && MODULES [=y] || EXPERT [=y])
   Location:
     -> Cryptographic API (CRYPTO [=y])
       -> Crypto core or helper
         -> Crypto benchmarking module (CRYPTO_BENCHMARK [=n])
 Selects: CRYPTO_MANAGER [=y]

 Commit: 3357b6c94569 (crypto: tcrypt - rename CRYPTO_TEST to CRYPTO_BENCHMARK)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>
Signed-off-by: Vladis Dronov <vdronov@redhat.com>