[redhat] New configs in crypto/Kconfig: CRYPTO_SELFTESTS
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: CRYPTO_SELFTESTS [=n]
 Type  : bool
 Defined at crypto/Kconfig:177
   Prompt: Enable cryptographic self-tests
   Depends on: CRYPTO [=y] && DEBUG_KERNEL [=y]
   Location:
     -> Cryptographic API (CRYPTO [=y])
       -> Crypto core or helper
         -> Enable cryptographic self-tests (CRYPTO_SELFTESTS [=n])

 Commit: 40b9969796bf (crypto: testmgr - replace CRYPTO_MANAGER_DISABLE_TESTS with CRYPTO_SELFTESTS)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>
Signed-off-by: Vladis Dronov <vdronov@redhat.com>