[redhat] New configs in drivers/block
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: BLK_DEV_ZONED_LOOP [=n]
 Type  : tristate
 Defined at drivers/block/Kconfig:410
   Prompt: Zoned loopback device support
   Depends on: BLK_DEV [=y] && BLK_DEV_ZONED [=y]
   Location:
     -> Device Drivers
       -> Block devices (BLK_DEV [=y])
         -> Zoned loopback device support (BLK_DEV_ZONED_LOOP [=n])

 Commit: eb0570c7df23 (block: new zoned loop block device driver)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>