[redhat] New configs in drivers/crypto
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: CRYPTO_DEV_QAT_6XXX [=n]
 Type  : tristate
 Defined at drivers/crypto/intel/qat/Kconfig:73
   Prompt: Support for Intel(R) QuickAssist Technology QAT_6XXX
   Depends on: CRYPTO [=y] && CRYPTO_HW [=y] && (X86 [=y] || COMPILE_TEST [=n]) && PCI [=y]
   Location:
     -> Cryptographic API (CRYPTO [=y])
       -> Hardware crypto devices (CRYPTO_HW [=y])
         -> Support for Intel(R) QuickAssist Technology QAT_6XXX (CRYPTO_DEV_QAT_6XXX [=n])
 Selects: CRYPTO_DEV_QAT [=m]

 Commit: 17fd7514ae68 (crypto: qat - add qat_6xxx driver)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>
Signed-off-by: Vladis Dronov <vdronov@redhat.com>