[redhat] New configs in drivers/char
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: TCG_SVSM [=n]
 Type  : tristate
 Defined at drivers/char/tpm/Kconfig:237
   Prompt: SNP SVSM vTPM interface
   Depends on: TCG_TPM [=y] && AMD_MEM_ENCRYPT [=y]
   Location:
     -> Device Drivers
       -> Character devices
         -> TPM Hardware Support (TCG_TPM [=y])
           -> SNP SVSM vTPM interface (TCG_SVSM [=n])

 Commit: 93b7c6b3ce91 (tpm: Add SNP SVSM vTPM driver)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>