[redhat] New configs in drivers/mfd
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: MFD_MAX77759 [=n]
 Type  : tristate
 Defined at drivers/mfd/Kconfig:946
   Prompt: Maxim Integrated MAX77759 PMIC
   Depends on: HAS_IOMEM [=y] && I2C [=y] && OF [=y]
   Location:
     -> Device Drivers
       -> Multifunction device drivers
         -> Maxim Integrated MAX77759 PMIC (MFD_MAX77759 [=n])
 Selects: IRQ_DOMAIN [=y] && MFD_CORE [=m] && REGMAP_I2C [=m] && REGMAP_IRQ [=n]

 Commit: bb71e40db1b7 (mfd: max77759: Add Maxim MAX77759 core driver)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>