[redhat] New configs in drivers/pwm
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: PWM_MC33XS2410 [=n]
 Type  : tristate
 Defined at drivers/pwm/Kconfig:426
   Prompt: MC33XS2410 PWM support
   Depends on: PWM [=y] && OF [=y] && SPI [=y]
   Location:
     -> Device Drivers
       -> Pulse-Width Modulation (PWM) Support (PWM [=y])
         -> MC33XS2410 PWM support (PWM_MC33XS2410 [=n])

 Commit: 2006016ec6b3 (pwm: add support for NXPs high-side switch MC33XS2410)

---

 Symbol: PWM_RENESAS_RCAR [=n]
 Type  : tristate
 Defined at drivers/pwm/Kconfig:537
   Prompt: Renesas R-Car PWM support
   Depends on: PWM [=y] && (ARCH_RENESAS [=y] || COMPILE_TEST [=n]) && HAS_IOMEM [=y]
   Location:
     -> Device Drivers
       -> Pulse-Width Modulation (PWM) Support (PWM [=y])
         -> Renesas R-Car PWM support (PWM_RENESAS_RCAR [=n])

 Commit: 25ac4834cae9 (pwm: Tidyup PWM menu for Renesas)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>