redhat/configs: enable IWLMLD for rhel
New WiFi7 intel cards are going to be handled with the new iwlmld driver
in order to support the new firmwares.
It will be used for the rebases to linux upstream kernel v6.15.
Unify configuration with fedora (it was previously enabled).
Do not enable it for automotive like other wireless intel drivers.

Signed-off-by: Jose Ignacio Tornos Martinez <jtornosm@redhat.com>