[redhat] New configs in crypto/krb5
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: CRYPTO_KRB5_SELFTESTS [=n]
 Type  : bool
 Defined at crypto/krb5/Kconfig:21
   Prompt: Kerberos 5 crypto selftests
   Depends on: CRYPTO [=y] && CRYPTO_KRB5 [=m]
   Location:
     -> Cryptographic API (CRYPTO [=y])
       -> Kerberos 5 crypto (CRYPTO_KRB5 [=m])
         -> Kerberos 5 crypto selftests (CRYPTO_KRB5_SELFTESTS [=n])

 Commit: fc0cf10c04f4 (crypto/krb5: Implement crypto self-testing)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>