[redhat] New configs in drivers/net/ethernet/stmicro
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: DWMAC_RENESAS_GBETH [=m]
 Type  : tristate
 Defined at drivers/net/ethernet/stmicro/stmmac/Kconfig:135
   Prompt: Renesas RZ/V2H(P) GBETH support
   Depends on: NETDEVICES [=y] && ETHERNET [=y] && NET_VENDOR_STMICRO [=y] && STMMAC_ETH [=m] && STMMAC_PLATFORM [=m] && OF [=y] && (ARCH_RENESAS [=y] || COMPILE_TEST [=n])
   Location:
     -> Device Drivers
       -> Network device support (NETDEVICES [=y])
         -> Ethernet driver support (ETHERNET [=y])
           -> STMicroelectronics devices (NET_VENDOR_STMICRO [=y])
             -> STMicroelectronics Multi-Gigabit Ethernet driver (STMMAC_ETH [=m])
               -> STMMAC Platform bus support (STMMAC_PLATFORM [=m])
                 -> Renesas RZ/V2H(P) GBETH support (DWMAC_RENESAS_GBETH [=m])

 Commit: 461f6529e594 (net: stmmac: Add DWMAC glue layer for Renesas GBETH)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>