[redhat] New configs in drivers/net/Kconfig
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: OVPN [=n]
 Type  : tristate
 Defined at drivers/net/Kconfig:118
   Prompt: OpenVPN data channel offload
   Depends on: NETDEVICES [=y] && NET_CORE [=y] && NET [=y] && INET [=y] && (IPV6 [=y] || !IPV6 [=y])
   Location:
     -> Device Drivers
       -> Network device support (NETDEVICES [=y])
         -> Network core driver support (NET_CORE [=y])
           -> OpenVPN data channel offload (OVPN [=n])
 Selects: DST_CACHE [=y] && NET_UDP_TUNNEL [=m] && CRYPTO [=y] && CRYPTO_AES [=y] && CRYPTO_GCM [=y] && CRYPTO_CHACHA20POLY1305 [=m] && STREAM_PARSER [=y]

 Commit: 9f23d943eb6b (net: introduce OpenVPN Data Channel Offload (ovpn))

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>