[redhat] New configs in drivers/net/phy
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: AS21XXX_PHY [=n]
 Type  : tristate
 Defined at drivers/net/phy/Kconfig:79
   Prompt: Aeonsemi AS21xxx PHYs
   Depends on: NETDEVICES [=y] && PHYLIB [=y]
   Location:
     -> Device Drivers
       -> Network device support (NETDEVICES [=y])
         -> PHY Device support and infrastructure (PHYLIB [=y])
           -> Aeonsemi AS21xxx PHYs (AS21XXX_PHY [=n])

 Commit: 830877d89edc (net: phy: Add support for Aeonsemi AS21xxx PHYs)

---

 Symbol: MAXLINEAR_86110_PHY [=n]
 Type  : tristate
 Defined at drivers/net/phy/Kconfig:278
   Prompt: MaxLinear MXL86110 PHY support
   Depends on: NETDEVICES [=y] && PHYLIB [=y]
   Location:
     -> Device Drivers
       -> Network device support (NETDEVICES [=y])
         -> PHY Device support and infrastructure (PHYLIB [=y])
           -> MaxLinear MXL86110 PHY support (MAXLINEAR_86110_PHY [=n])

 Commit: b2908a989c59 (net: phy: add driver for MaxLinear MxL86110 PHY)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>