[redhat] New configs in lib/Kconfig.debug
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: RANDSTRUCT_KUNIT_TEST [=m]
 Type  : tristate
 Defined at lib/Kconfig.debug:2866
   Prompt: Test randstruct structure layout randomization at runtime
   Depends on: RUNTIME_TESTING_MENU [=y] && KUNIT [=m]
   Visible if: RUNTIME_TESTING_MENU [=y] && KUNIT [=m] && !KUNIT_ALL_TESTS [=m]
   Location:
     -> Kernel hacking
       -> Kernel Testing and Coverage
         -> Runtime Testing (RUNTIME_TESTING_MENU [=y])
           -> Test randstruct structure layout randomization at runtime (RANDSTRUCT_KUNIT_TEST [=m])

 Commit: b370f7eacdcf (lib/tests: Add randstruct KUnit test)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>