[redhat] New configs in net/rxrpc
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: RXGK [=n]
 Type  : bool
 Defined at net/rxrpc/Kconfig:70
   Prompt: RxRPC GSSAPI security
   Depends on: NET [=y] && AF_RXRPC [=m]
   Location:
     -> Networking support (NET [=y])
       -> RxRPC session sockets (AF_RXRPC [=m])
         -> RxRPC GSSAPI security (RXGK [=n])
 Selects: CRYPTO_KRB5 [=m] && CRYPTO_MANAGER [=y] && CRYPTO_KRB5ENC [=m] && CRYPTO_AUTHENC [=y] && CRYPTO_SKCIPHER [=y] && CRYPTO_HASH_INFO [=y] && CRYPTO_HMAC [=y] && CRYPTO_CMAC [=y] && CRYPTO_SHA1 [=y] && CRYPTO_SHA256 [=y] && CRYPTO_SHA512 [=y] && CRYPTO_CBC [=y] && CRYPTO_CTS [=y] && CRYPTO_AES [=y] && CRYPTO_CAMELLIA [=m]

 Commit: c86f9b963dc6 (rxrpc: rxgk: Provide infrastructure and key derivation)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>