[redhat] New configs in security/integrity
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: IMA_KEXEC_EXTRA_MEMORY_KB [=0]
 Type  : integer
 Range : [0 40]
 Defined at security/integrity/ima/Kconfig:324
   Prompt: Extra memory for IMA measurements added during kexec soft reboot
   Depends on: INTEGRITY [=y] && IMA [=y] && IMA_KEXEC [=y]
   Location:
     -> Security options
       -> Integrity subsystem (INTEGRITY [=y])
         -> Integrity Measurement Architecture(IMA) (IMA [=y])
           -> Extra memory for IMA measurements added during kexec soft reboot (IMA_KEXEC_EXTRA_MEMORY_KB [=0])

 Commit: 0ad93987c9c1 (ima: make the kexec extra memory configurable)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>