[redhat] New configs in arch/arm64
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: AMPERE_ERRATUM_AC04_CPU_23 [=y]
 Type  : bool
 Defined at arch/arm64/Kconfig:467
   Prompt: AmpereOne: AC04_CPU_23:  Failure to synchronize writes to HCR_EL2 may corrupt address translations.
   Location:
     -> Kernel Features
       -> ARM errata workarounds via the alternatives framework
         -> AmpereOne: AC04_CPU_23:  Failure to synchronize writes to HCR_EL2 may corrupt address translations. (AMPERE_ERRATUM_AC04_CPU_23 [=y])

 Commit: fed55f49fad1 (arm64: errata: Work around AmpereOne's erratum AC04_CPU_23)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>