[redhat] New configs in samples/Kconfig
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: SAMPLE_TSM_MR [=n]
 Type  : tristate
 Defined at samples/Kconfig:187
   Prompt: TSM measurement sample
   Depends on: SAMPLES [=y]
   Location:
     -> Kernel hacking
       -> Sample kernel code (SAMPLES [=y])
         -> TSM measurement sample (SAMPLE_TSM_MR [=n])
 Selects: TSM_MEASUREMENTS [=y] && VIRT_DRIVERS [=y]

 Commit: f6953f1f9ec4 (tsm-mr: Add tsm-mr sample code)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>