[redhat] New configs in drivers/misc
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: TI_FPC202 [=n]
 Type  : tristate
 Defined at drivers/misc/Kconfig:117
   Prompt: TI FPC202 Dual Port Controller
   Depends on: I2C [=y]
   Location:
     -> Device Drivers
       -> Misc devices
         -> TI FPC202 Dual Port Controller (TI_FPC202 [=n])
 Selects: GPIOLIB [=y] && I2C_ATR [=n]

 Commit: 1e5c9b1efa1c (misc: add FPC202 dual port controller driver)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>