[redhat] New configs in drivers/reset
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: RESET_RZV2H_USB2PHY [=n]
 Type  : tristate
 Defined at drivers/reset/Kconfig:228
   Prompt: Renesas RZ/V2H(P) (and similar SoCs) USB2PHY Reset driver
   Depends on: RESET_CONTROLLER [=y] && (ARCH_RENESAS [=y] || COMPILE_TEST [=n])
   Location:
     -> Device Drivers
       -> Reset Controller Support (RESET_CONTROLLER [=y])
         -> Renesas RZ/V2H(P) (and similar SoCs) USB2PHY Reset driver (RESET_RZV2H_USB2PHY [=n])

 Commit: e3911d7f865b (reset: Add USB2PHY port reset driver for Renesas RZ/V2H(P))

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>