[redhat] New configs in fs/squashfs
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: SQUASHFS_COMP_CACHE_FULL [=n]
 Type  : bool
 Defined at fs/squashfs/Kconfig:152
   Prompt: Enable full caching of compressed blocks
   Depends on: MISC_FILESYSTEMS [=y] && SQUASHFS [=m]
   Location:
     -> File systems
       -> Miscellaneous filesystems (MISC_FILESYSTEMS [=y])
         -> SquashFS 4.0 - Squashed file system support (SQUASHFS [=m])
           -> Enable full caching of compressed blocks (SQUASHFS_COMP_CACHE_FULL [=n])

 Commit: 2e227ff5e272 (squashfs: add optional full compressed block caching)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>