[redhat] New configs in fs/sysfs
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: SYSFS [=y]
 Type  : bool
 Defined at fs/sysfs/Kconfig:2
   Prompt: sysfs file system support
   Visible if: EXPERT [=y]
   Location:
     -> File systems
       -> Pseudo filesystems
         -> sysfs file system support (SYSFS [=y])
 Selects: KERNFS [=y]

 Commit: 5f3a211a8b02 (fs/Kconfig: move sysfs out)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>