[redhat] New configs in drivers/hwmon
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: SENSORS_LT3074 [=n]
 Type  : tristate
 Defined at drivers/hwmon/pmbus/Kconfig:221
   Prompt: Analog Devices LT3074
   Depends on: HWMON [=y] && PMBUS [=m]
   Location:
     -> Device Drivers
       -> Hardware Monitoring support (HWMON [=y])
         -> PMBus support (PMBUS [=m])
           -> Analog Devices LT3074 (SENSORS_LT3074 [=n])

 Commit: c66c5bda7f24 (hwmon: (pmbus/lt3074) add support for lt3074)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>