[redhat] New configs in drivers/leds
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: LEDS_KUNIT_TEST [=m]
 Type  : tristate
 Defined at drivers/leds/Kconfig:58
   Prompt: KUnit tests for LEDs
   Depends on: NEW_LEDS [=y] && KUNIT [=m] && LEDS_CLASS [=y]
   Location:
     -> Device Drivers
       -> LED Support (NEW_LEDS [=y])
         -> LED Class Support (LEDS_CLASS [=y])
           -> KUnit tests for LEDs (LEDS_KUNIT_TEST [=m])

 Commit: 5039a33fed88 (leds: Provide skeleton KUnit testing for the LEDs framework)

---

 Symbol: LEDS_TPS6131X [=n]
 Type  : tristate
 Defined at drivers/leds/flash/Kconfig:135
   Prompt: LED support for TI TPS6131x flash LED driver
   Depends on: NEW_LEDS [=y] && LEDS_CLASS_FLASH [=m] && I2C [=y] && OF [=y] && GPIOLIB [=y]
   Location:
     -> Device Drivers
       -> LED Support (NEW_LEDS [=y])
         -> LED support for TI TPS6131x flash LED driver (LEDS_TPS6131X [=n])
 Selects: REGMAP_I2C [=y]

 Commit: b338a2ae9b31 (leds: tps6131x: Add support for Texas Instruments TPS6131X flash LED driver)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>