[redhat] New configs in drivers/mfd
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: MFD_SEC_I2C [=n]
 Type  : tristate
 Defined at drivers/mfd/Kconfig:1336
   Prompt: Samsung Electronics S2MPA/S2MPS1X/S2MPU/S5M series PMICs
   Depends on: HAS_IOMEM [=y] && I2C [=y]=y [=y] && OF [=y]
   Location:
     -> Device Drivers
       -> Multifunction device drivers
         -> Samsung Electronics S2MPA/S2MPS1X/S2MPU/S5M series PMICs (MFD_SEC_I2C [=n])
 Selects: MFD_SEC_CORE [=n] && REGMAP_I2C [=m]

 Commit: 680ef57915db (mfd: sec: Split into core and transport (i2c) drivers)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>