[redhat] New configs in drivers/pci
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: PCI_PWRCTRL_SLOT [=n]
 Type  : tristate
 Defined at drivers/pci/pwrctrl/Kconfig:14
   Prompt: PCI Power Control driver for PCI slots
   Depends on: PCI [=y]
   Location:
     -> Device Drivers
       -> PCI support (PCI [=y])
         -> PCI Power Control driver for PCI slots (PCI_PWRCTRL_SLOT [=n])
 Selects: PCI_PWRCTRL [=n]

 Commit: 13bbf6a5f065 (PCI/pwrctrl: Rename pwrctrl Kconfig symbols and slot module)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>