[redhat] New configs in drivers/dma
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: ARM_DMA350 [=n]
 Type  : tristate
 Defined at drivers/dma/Kconfig:96
   Prompt: Arm DMA-350 support
   Depends on: DMADEVICES [=y] && (ARM || ARM64 [=y] || COMPILE_TEST [=n])
   Location:
     -> Device Drivers
       -> DMA Engine support (DMADEVICES [=y])
         -> Arm DMA-350 support (ARM_DMA350 [=n])
 Selects: DMA_ENGINE [=y] && DMA_VIRTUAL_CHANNELS [=m]

 Commit: 5d099706449d (dmaengine: Add Arm DMA-350 driver)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>