[redhat] New configs in drivers/hwtracing
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: CORESIGHT_KUNIT_TESTS [=m]
 Type  : tristate
 Defined at drivers/hwtracing/coresight/Kconfig:263
   Prompt: Enable Coresight unit tests
   Depends on: CORESIGHT [=m] && KUNIT [=m]
   Location:
     -> Kernel hacking
       -> arm64 Debugging
         -> CoreSight Tracing Support (CORESIGHT [=m])
           -> Enable Coresight unit tests (CORESIGHT_KUNIT_TESTS [=m])

 Commit: b104a941a94e (coresight: Add a KUnit test for coresight_find_default_sink())

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>