Simplify include Makefile.rhelver
The current problem with where the include Makefile.rhelver is, is that
it is used before any of the standard definitions are defined.  This
makes it challenging to use stock methods and instead use kludgy ifneq
definitions.

There is no real reason to have the include this high in the Makefile.
It was just a legacy place that didn't change much thus minimizing merge
conflicts.

Let's move the 'include' later in the file _after_ the definitions and
in a place that is surrounded by code that haven't changed in 20 years.

This allows us to use expected definitions that hopefully make things
more future proof.

This change will be later rolled into ark-infra.

Signed-off-by: Don Zickus <dzickus@redhat.com>