Remove include Makefile.rhelver from ark-infra
The reason 'include Makefile.rhelver' is there is to define the RHEL_
variables in version.h.  Those RHEL_ definitions are there to help
third-party modules if-def their driver code around RHEL versions.

This isn't useful in ARK's os-build as the version is static but
upstream continuously changes.  But is kept for inclusion for the next
RHEL snapshots.  For the ark-infra branch this makes even less sense as
the point is to compile upstream code not in Linus's tree yet, not third
party modules.

Considering the RHEL_ definitions were not in the Makefile since the
dawn of ark-infra and no one complained, removing 'include
Makefile.rhelver' should be a noopt and allows us to future proof
changes and merge conflicts.

Therefore removing.

Signed-off-by: Don Zickus <dzickus@redhat.com>