[redhat] New configs in crypto/Kconfig
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 ~~~
 Symbol: CRYPTO_SELFTESTS_FULL [=n]
 Type  : bool
 Defined at crypto/Kconfig:196
   Prompt: Enable the full set of cryptographic self-tests
   Depends on: CRYPTO [=y] && CRYPTO_SELFTESTS [=y]
   Location:
     -> Cryptographic API (CRYPTO [=y])
       -> Crypto core or helper
         -> Enable cryptographic self-tests (CRYPTO_SELFTESTS [=y])
           -> Enable the full set of cryptographic self-tests (CRYPTO_SELFTESTS_FULL [=n])

 ~~~
 Commit: ac90aad0e9bf (crypto: testmgr - reinstate kconfig control over full self-tests)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>