redhat/configs: automotive: enable extra system cert
Reserve an area in the certificate ring to add another one without
recompiling. Arbitrarily choose 4k, but odds are this can be lowered.

This is intended to be used in the automotive images where the modules
are signed when the image is composed instead of when the kernel and
modules are built.

Signed-off-by: Eric Chanudet <echanude@redhat.com>