Revert "redhat/configs: automotive: Turn off ACPI Processor package for aarch64"
This reverts commit ea77accd2927cbb3d25b77e1b793ed5a90cbd765.

While allowing the cpu to enter idle states can/will introduce (non-deterministic)
latency and this is undesired for the automotive use case. The fact that
we also need to disable ACPI Processor package unconditionally can
introduce some problems on those platforms that are ACPI compliant. For
that reason, we're going to pass `cpuidle.off=1` in the kernel cmdline
to avoid the cpu to enter idle states for the platforms that are
supported in automotive.

Signed-off-by: Enric Balletbo i Serra <eballetb@redhat.com>