redhat: ark-merge-driver: initial commit
This introduces a script that can be either used as a git merge driver or a
git mergetool to resolve merge conflicts using AI.

The script supports Anthropic's claude model and Google's gemini model
and will choose based on which key is provided via an environment
variable (ANTHROPIC_API_KEY or GOOGLE_GEMINI_API_KEY).

The approach used here is kernel-ark specific and it simply ask the
model to reapply our Red Hat changes to the upstream file and therefore
it cannot be used to resolve complex backport merge conflicts yet.

Because the script does not use a conflict resolution strategy as
successful as git's ort strategy, it is currently better to use this as
a mergetool rather than a merge driver to resolve those conflicts that
"git merge" could not.

Signed-off-by: Scott Weaver <scweaver@redhat.com>