kernel.spec: introduce with_automotive_build
The with_automotiveonly flag was originally intended to allow someone to
build the automotive variant of the kernel independent of the stock
kernel packages and all of the other variant packages.

When we build the kernel-automotive package (with AUTOMOTIVE_BUILD
enabled) we aren't actually building the variant but instead a separate
standalone package. To achieve this and maintain the distinction with
the variant, a new build option called with_automotive_build has been
added.

Signed-off-by: Scott Weaver <scweaver@redhat.com>