kernel.spec: always provide kernel-devel-uname-r
Currently the akmods package requires kernel-devel-uname-r and therefore
we need to always provide this in case a downstream distro using either
kernel-automotive or kernel-redhat (gemini) were to include akmods in
their compose.

When %{name} is kernel, the duplicate Provides does not currently cause
an RPM error and duplicate Provides are not reported by the final rpm
package. Therefore, there is no need to guard against including
duplicate Provides.

Signed-off-by: Scott Weaver <scweaver@redhat.com>