redhat/configs: Consolidate the CONFIG_RANDOMIZE_IDENTITY_BASE switch
CONFIG_RANDOMIZE_IDENTITY_BASE is disabled in both, Fedora and RHEL, so
we can consolidate this switch in the common folder. This also makes
sure that the switch now resides in a s390x sub-folder where it belongs
to (in the rhel config, it was previously mis-placed in the generic main
folder instead).

Signed-off-by: Thomas Huth <thuth@redhat.com>