[redhat] New configs in drivers/devfreq
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 ~~~
 Symbol: ARM_HISI_UNCORE_DEVFREQ [=n]
 Type  : tristate
 Defined at drivers/devfreq/Kconfig:93
   Prompt: HiSilicon uncore DEVFREQ Driver
   Depends on: PM_DEVFREQ [=y] && ACPI [=y] && ACPI_PPTT [=y] && PCC [=y]
   Location:
     -> Device Drivers
       -> Generic Dynamic Voltage and Frequency Scaling (DVFS) support (PM_DEVFREQ [=y])
         -> HiSilicon uncore DEVFREQ Driver (ARM_HISI_UNCORE_DEVFREQ [=n])
 Selects: DEVFREQ_GOV_PERFORMANCE [=n] && DEVFREQ_GOV_USERSPACE [=m]

 ~~~
 Commit: 7da2fdaaa1e6 (PM / devfreq: Add HiSilicon uncore frequency scaling driver)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>