[redhat] New configs in drivers/regulator
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 ~~~
 Symbol: REGULATOR_RASPBERRYPI_TOUCHSCREEN_V2 [=n]
 Type  : tristate
 Defined at drivers/regulator/Kconfig:1156
   Prompt: Raspberry Pi 7-inch touchscreen panel V2 regulator
   Depends on: REGULATOR [=y] && GPIOLIB [=y] && I2C [=y] && OF [=y]
   Location:
     -> Device Drivers
       -> Voltage and Current Regulator Support (REGULATOR [=y])
         -> Raspberry Pi 7-inch touchscreen panel V2 regulator (REGULATOR_RASPBERRYPI_TOUCHSCREEN_V2 [=n])
 Selects: GPIO_REGMAP [=n] && REGMAP_I2C [=y]

 ~~~
 Commit: d49305862fdc (regulator: rpi-panel-v2: Add regulator for 7" Raspberry Pi 720x1280)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>