[redhat] New configs in drivers/spi
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 ~~~
 Symbol: SPI_RZV2H_RSPI [=n]
 Type  : tristate
 Defined at drivers/spi/Kconfig:935
   Prompt: Renesas RZ/V2H RSPI controller
   Depends on: SPI [=y] && SPI_MASTER [=y] && (ARCH_RENESAS [=y] || COMPILE_TEST [=n])
   Location:
     -> Device Drivers
       -> SPI support (SPI [=y])
         -> Renesas RZ/V2H RSPI controller (SPI_RZV2H_RSPI [=n])

 ~~~
 Commit: 8b61c8919dff (spi: Add driver for the RZ/V2H(P) RSPI IP)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>