[redhat] New configs in arch/riscv
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 ~~~
 Symbol: ARCH_ANDES [=n]
 Type  : bool
 Defined at arch/riscv/Kconfig.socs:3
   Prompt: Andes SoCs
   Depends on: MMU [=y] && !XIP_KERNEL [=n]
   Location:
     -> SoC selection
       -> Andes SoCs (ARCH_ANDES [=n])
 Selects: ERRATA_ANDES [=n]

 ~~~
 Commit: 00dba19aa005 (riscv: add Andes SoC family Kconfig support)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>