[redhat] New configs in drivers/bus
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 ~~~
 Symbol: IMX_AIPSTZ [=n]
 Type  : tristate
 Defined at drivers/bus/Kconfig:90
   Prompt: Support for IMX Secure AHB to IP Slave bus (AIPSTZ) bridge
   Depends on: ARCH_MXC [=y]
   Location:
     -> Device Drivers
       -> Bus devices
         -> Support for IMX Secure AHB to IP Slave bus (AIPSTZ) bridge (IMX_AIPSTZ [=n])

 ~~~
 Commit: 796cba2dd4d9 (bus: add driver for IMX AIPSTZ bridge)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>