[redhat] New configs in drivers/firmware
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 ~~~
 Symbol: TEGRA_BPMP [=n]
 Type  : bool
 Defined at drivers/firmware/tegra/Kconfig:14
   Prompt: Tegra BPMP driver
   Depends on: ARCH_TEGRA [=y] && TEGRA_HSP_MBOX [=y] && !CPU_BIG_ENDIAN [=n]
   Location:
     -> Device Drivers
       -> Firmware Drivers
         -> Tegra firmware driver
           -> Tegra BPMP driver (TEGRA_BPMP [=n])
 Selects: TEGRA_IVC [=n]

 ~~~
 Commit: 983de5f97169 (firmware: tegra: Add BPMP support)

---

 ~~~
 Symbol: TI_SCI_PROTOCOL [=y]
 Type  : tristate
 Defined at drivers/firmware/Kconfig:224
   Prompt: TI System Control Interface (TISCI) Message Protocol
   Depends on: TI_MESSAGE_MANAGER [=y]
   Location:
     -> Device Drivers
       -> Firmware Drivers
         -> TI System Control Interface (TISCI) Message Protocol (TI_SCI_PROTOCOL [=y])

 ~~~
 Commit: aa276781a64a (firmware: Add basic support for TI System Control Interface (TI-SCI) protocol)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>

v2:
- CONFIG_TEGRA_BPMP changes dropped; handled in !4085.

Signed-off-by: Patrick Talbert <ptalbert@redhat.com>