[redhat] New configs in drivers/perf
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 ~~~
 Symbol: ARM64_BRBE [=y]
 Type  : bool
 Defined at drivers/perf/Kconfig:226
   Prompt: Enable support for branch stack sampling using FEAT_BRBE
   Depends on: PERF_EVENTS [=y] && ARM_PMUV3 [=y] && ARM64 [=y]
   Location:
     -> Device Drivers
       -> Performance monitor support
         -> Enable support for branch stack sampling using FEAT_BRBE (ARM64_BRBE [=y])

 ~~~
 Commit: 58074a0fce66 (perf: arm_pmuv3: Add support for the Branch Record Buffer Extension (BRBE))

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>