[redhat] New configs in drivers/platform
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 ~~~
 Symbol: AMD_HFI [=n]
 Type  : bool
 Defined at drivers/platform/x86/amd/hfi/Kconfig:6
   Prompt: AMD Hetero Core Hardware Feedback Driver
   Depends on: X86_PLATFORM_DEVICES [=y] && ACPI [=y] && CPU_SUP_AMD [=y] && SCHED_MC_PRIO [=y]
   Location:
     -> Device Drivers
       -> X86 Platform Specific Device Drivers (X86_PLATFORM_DEVICES [=y])
         -> AMD Hetero Core Hardware Feedback Driver (AMD_HFI [=n])

 ~~~
 Commit: 5d902ee5609a (platform/x86: hfi: Introduce AMD Hardware Feedback Interface Driver)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>

v2:
- CONFIG_AMD_HFI enabled for rhel.

Signed-off-by: Patrick Talbert <ptalbert@redhat.com>