[redhat] New configs in init/Kconfig
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 ~~~
 Symbol: SCHED_PROXY_EXEC [=n]
 Type  : bool
 Defined at init/Kconfig:881
   Prompt: Proxy Execution
   Depends on: !PREEMPT_RT [=n] && !SCHED_CLASS_EXT [=n] && EXPERT [=y]
   Location:
     -> General setup
       -> Scheduler features
         -> Proxy Execution (SCHED_PROXY_EXEC [=n])

 ~~~
 Commit: 25c411fce735 (sched: Add CONFIG_SCHED_PROXY_EXEC & boot argument to enable/disable)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>