[redhat] New configs in kernel/time
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 ~~~
 Symbol: POSIX_AUX_CLOCKS [=n]
 Type  : bool
 Defined at kernel/time/Kconfig:213
   Prompt: Enable auxiliary POSIX clocks
   Depends on: POSIX_TIMERS [=y]
   Location:
     -> General setup
       -> Timers subsystem
         -> Enable auxiliary POSIX clocks (POSIX_AUX_CLOCKS [=n])

 ~~~
 Commit: 9094c72c3d81 (time: Introduce auxiliary POSIX clocks)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>