[redhat] New configs in lib/kunit
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 ~~~
 Symbol: KUNIT_DEFAULT_TIMEOUT [=300]
 Type  : integer
 Defined at lib/kunit/Kconfig:96
   Prompt: Default value of the timeout module parameter
   Depends on: KUNIT [=m]
   Location:
     -> Kernel hacking
       -> Kernel Testing and Coverage
         -> KUnit - Enable support for unit tests (KUNIT [=m])
           -> Default value of the timeout module parameter (KUNIT_DEFAULT_TIMEOUT [=300])

 ~~~
 Commit: 5ac244b9cc8f (kunit: Make default kunit_test timeout configurable via both a module parameter and a Kconfig option)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>