Remove CONFIG_TEST_MISC_MINOR as deps are no longer met
Upstream commit 74d8361be3441 changed TEST_MISC_MINOR to a bool and made
it depend on KUNIT=y.  While we have KUNIT=m the deps for
TEST_MISC_MINOR are no longer satisfied. Delete the configs for it
rather than forcing it off, so that we are notified if deps are met
again in the future.

Signed-off-by: Justin M. Forbes <jforbes@fedoraproject.org>