[redhat] New configs in drivers/dma-buf
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 ~~~
 Symbol: DMABUF_HEAPS_CMA_LEGACY [=y]
 Type  : bool
 Defined at drivers/dma-buf/heaps/Kconfig:16
   Prompt: Legacy DMA-BUF CMA Heap
   Depends on: DMABUF_HEAPS_CMA [=y]
   Location:
     -> Device Drivers
       -> DMABUF options
         -> DMA-BUF CMA Heap (DMABUF_HEAPS_CMA [=y])
           -> Legacy DMA-BUF CMA Heap (DMABUF_HEAPS_CMA_LEGACY [=y])

 ~~~
 Commit: 854acbe75ff4 (dma-buf: heaps: Give default CMA heap a fixed name)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>