[redhat] New configs in drivers/dpll
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 ~~~
 Symbol: ZL3073X [=n]
 Type  : tristate
 Defined at drivers/dpll/zl3073x/Kconfig:3
   Prompt: Microchip Azurite DPLL/PTP/SyncE devices
   Depends on: NET [=y]
   Location:
     -> Device Drivers
       -> DPLL device support
         -> Microchip Azurite DPLL/PTP/SyncE devices (ZL3073X [=n])
 Selects: DPLL [=y] && NET_DEVLINK [=y] && REGMAP [=y]

 ~~~
 Commit: 2df8e64e01c1 (dpll: Add basic Microchip ZL3073x support)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>

Removed pending-fedora file.

Signed-off-by: Patrick Talbert <ptalbert@redhat.com>