[redhat] New configs in drivers/mtd
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 ~~~
 Symbol: MTD_INTEL_DG [=n]
 Type  : tristate
 Defined at drivers/mtd/devices/Kconfig:186
   Prompt: Intel Discrete Graphics non-volatile memory driver
   Depends on: HAS_IOMEM [=y] && AUXILIARY_BUS [=y] && MTD [=m]
   Location:
     -> Device Drivers
       -> Memory Technology Device (MTD) support (MTD [=m])
         -> Self-contained MTD device drivers
           -> Intel Discrete Graphics non-volatile memory driver (MTD_INTEL_DG [=n])

 ~~~
 Commit: ceb5ab3cb646 (mtd: add driver for intel graphics non-volatile memory device)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>