[redhat] New configs in drivers/net/ethernet/broadcom
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 ~~~
 Symbol: BNGE [=n]
 Type  : tristate
 Defined at drivers/net/ethernet/broadcom/Kconfig:256
   Prompt: Broadcom Ethernet device support
   Depends on: NETDEVICES [=y] && ETHERNET [=y] && NET_VENDOR_BROADCOM [=y] && PCI [=y]
   Location:
     -> Device Drivers
       -> Network device support (NETDEVICES [=y])
         -> Ethernet driver support (ETHERNET [=y])
           -> Broadcom devices (NET_VENDOR_BROADCOM [=y])
             -> Broadcom Ethernet device support (BNGE [=n])
 Selects: NET_DEVLINK [=y]

 ~~~
 Commit: 74715c4ab0fa (bng_en: Add PCI interface)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>