[redhat] New configs in net/sched
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 ~~~
 Symbol: NET_SCH_DUALPI2 [=n]
 Type  : tristate
 Defined at net/sched/Kconfig:418
   Prompt: Dual Queue PI Square (DUALPI2) scheduler
   Depends on: NET [=y] && NET_SCHED [=y]
   Location:
     -> Networking support (NET [=y])
       -> Networking options
         -> QoS and/or fair queueing (NET_SCHED [=y])
           -> Dual Queue PI Square (DUALPI2) scheduler (NET_SCH_DUALPI2 [=n])

 ~~~
 Commit: 8f9516daedd6 (sched: Add enqueue/dequeue of dualpi2 qdisc)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>