Move CONFIG_SCHED_PROXY_EXEC to the zfcpdump directory
Putting this in generic is a bit misleading. This config only impacts
the s390 zfcpdump kernel because SCHED_PROXY_EXEC depends on
!SCHED_CLASS_EXT.  We have SCHED_CLASS_EXT=y everywhere except for the
zfcpdump kernels, where it is only turned off because SCHED_CLASS_EXT
depends on BPF_SYSCALL which is turned off for zfcpdump kernels.

In all other configurations, SCHED_PROXY_EXEC is disabled because the
deps are not met.  We want this in the correct directory so that we can
make an appropriate decision later if deps are met.

Signed-off-by: Justin M. Forbes <jforbes@fedoraproject.org>