[redhat] New configs in drivers/crypto
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 ~~~
 Symbol: CRYPTO_PHMAC_S390 [=n]
 Type  : tristate
 Defined at drivers/crypto/Kconfig:191
   Prompt: PHMAC cipher algorithms
   Depends on: CRYPTO [=y] && CRYPTO_HW [=y] && S390 [=y] && PKEY [=m]
   Location:
     -> Cryptographic API (CRYPTO [=y])
       -> Hardware crypto devices (CRYPTO_HW [=y])
         -> Kernel API for protected key handling (PKEY [=m])
           -> PHMAC cipher algorithms (CRYPTO_PHMAC_S390 [=n])
 Selects: CRYPTO_HASH [=y] && CRYPTO_ENGINE [=m]

 ~~~
 Commit: cbbc675506cc (crypto: s390 - New s390 specific protected key hash phmac)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>