[redhat] New configs in drivers/media
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 ~~~
 Symbol: VIDEO_CAMERA_LENS [=y]
 Type  : bool
 Defined at drivers/media/i2c/Kconfig:769
   Prompt: Lens drivers
   Depends on: MEDIA_SUPPORT [=m] && VIDEO_DEV [=m] && MEDIA_CAMERA_SUPPORT [=y] && I2C [=y]
   Location:
     -> Device Drivers
       -> Multimedia support (MEDIA_SUPPORT [=m])
         -> Media ancillary drivers
           -> Lens drivers (VIDEO_CAMERA_LENS [=y])
 Selects: MEDIA_CONTROLLER [=y] && V4L2_FWNODE [=m] && VIDEO_V4L2_SUBDEV_API [=y]

 ~~~
 Commit: 871a99ff4d72 (media: i2c: Automatically select common options for lens drivers)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>