[redhat] New configs in mm/Kconfig
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 ~~~
 Symbol: DEVICE_PRIVATE [=n]
 Type  : bool
 Defined at mm/Kconfig:1159
   Prompt: Unaddressable device memory (GPU memory, ...)
   Depends on: ZONE_DEVICE [=y]
   Location:
     -> Memory Management options
       -> Unaddressable device memory (GPU memory, ...) (DEVICE_PRIVATE [=n])
 Selects: GET_FREE_REGION [=y]

 ~~~
 Commit: 5042db43cc26 (mm/ZONE_DEVICE: new type of ZONE_DEVICE for unaddressable memory)

---

 ~~~
 Symbol: PAGE_BLOCK_MAX_ORDER [=10]
 Type  : integer
 Range : [1 10]
 Defined at mm/Kconfig:1015
   Prompt: Page Block Order Upper Limit
   Location:
     -> Memory Management options
       -> Page Block Order Upper Limit (PAGE_BLOCK_MAX_ORDER [=10])

 ~~~
 Commit: 3800d5525097 (mm: rename CONFIG_PAGE_BLOCK_ORDER to CONFIG_PAGE_BLOCK_MAX_ORDER)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>