[redhat] New configs in drivers/pci
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 ~~~
 Symbol: PCI_ENDPOINT_MSI_DOORBELL [=n]
 Type  : bool
 Defined at drivers/pci/endpoint/Kconfig:31
   Prompt: PCI Endpoint MSI Doorbell Support
   Depends on: PCI [=y] && PCI_ENDPOINT [=y] && GENERIC_MSI_IRQ [=y]
   Location:
     -> Device Drivers
       -> PCI support (PCI [=y])
         -> PCI Endpoint
           -> PCI Endpoint Support (PCI_ENDPOINT [=y])
             -> PCI Endpoint MSI Doorbell Support (PCI_ENDPOINT_MSI_DOORBELL [=n])

 ~~~
 Commit: 1c3b002c6bf6 (PCI: endpoint: Add RC-to-EP doorbell support using platform MSI controller)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>