[redhat] New configs in drivers/vhost
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 ~~~
 Symbol: VHOST_ENABLE_FORK_OWNER_CONTROL [=y]
 Type  : bool
 Defined at drivers/vhost/Kconfig:98
   Prompt: Enable VHOST_ENABLE_FORK_OWNER_CONTROL
   Depends on: VHOST_MENU [=y]
   Location:
     -> Device Drivers
       -> VHOST drivers (VHOST_MENU [=y])
         -> Enable VHOST_ENABLE_FORK_OWNER_CONTROL (VHOST_ENABLE_FORK_OWNER_CONTROL [=y])

 ~~~
 Commit: 7d9896e9f6d0 (vhost: Reintroduce kthread API and add mode selection)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>