[redhat] New configs in kernel/trace
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 ~~~
 Symbol: EPROBE_EVENTS [=y]
 Type  : bool
 Defined at kernel/trace/Kconfig:796
   Prompt: Enable event-based dynamic events
   Depends on: FTRACE [=y] && TRACING [=y] && HAVE_REGS_AND_STACK_ACCESS_API [=y]
   Location:
     -> Kernel hacking
       -> Tracers (FTRACE [=y])
         -> Enable event-based dynamic events (EPROBE_EVENTS [=y])
 Selects: PROBE_EVENTS [=y] && DYNAMIC_EVENTS [=y]

 ~~~
 Commit: 0dd1274a053f (tracing: Have eprobes have their own config option)

---

 ~~~
 Symbol: TRACEFS_AUTOMOUNT_DEPRECATED [=y]
 Type  : bool
 Defined at kernel/trace/Kconfig:203
   Prompt: Automount tracefs on debugfs [DEPRECATED]
   Depends on: FTRACE [=y] && TRACING [=y]
   Location:
     -> Kernel hacking
       -> Tracers (FTRACE [=y])
         -> Automount tracefs on debugfs [DEPRECATED] (TRACEFS_AUTOMOUNT_DEPRECATED [=y])

 ~~~
 Commit: 9ba817fb7c6a (tracing: Deprecate auto-mounting tracefs in debugfs)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>