[redhat] New configs in sound/hda
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 ~~~
 Symbol: SND_HDA_CODEC_HDMI_GENERIC [=m]
 Type  : tristate
 Defined at sound/hda/codecs/hdmi/Kconfig:8
   Prompt: Generic HDMI/DisplayPort HD-audio codec support
   Depends on: SOUND [=m] && SND [=m] && SND_HDA [=m] && SND_HDA_CODEC_HDMI [=m]
   Location:
     -> Device Drivers
       -> Sound card support (SOUND [=m])
         -> Advanced Linux Sound Architecture (SND [=m])
           -> HD-Audio
             -> HD-audio HDMI codec support (SND_HDA_CODEC_HDMI [=m])
               -> Generic HDMI/DisplayPort HD-audio codec support (SND_HDA_CODEC_HDMI_GENERIC [=m])
 Selects: SND_DYNAMIC_MINORS [=y] && SND_PCM_ELD [=y]
 Selected by [n]:
   - SND_HDA_CODEC_HDMI_INTEL [=n] && SOUND [=m] && SND [=m] && SND_HDA [=m] && SND_HDA_CODEC_HDMI [=m]
   - SND_HDA_CODEC_HDMI_ATI [=n] && SOUND [=m] && SND [=m] && SND_HDA [=m] && SND_HDA_CODEC_HDMI [=m]
   - SND_HDA_CODEC_HDMI_NVIDIA [=n] && SOUND [=m] && SND [=m] && SND_HDA [=m] && SND_HDA_CODEC_HDMI [=m]
   - SND_HDA_CODEC_HDMI_TEGRA [=n] && SOUND [=m] && SND [=m] && SND_HDA [=m] && SND_HDA_CODEC_HDMI [=m]

 ~~~
 Commit: 81231ad173d8 (ALSA: hda/hdmi: Enable drivers as default)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>