[redhat] New configs in drivers/i3c
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 ~~~
 Symbol: RENESAS_I3C [=n]
 Type  : tristate
 Defined at drivers/i3c/master/Kconfig:68
   Prompt: Renesas I3C controller driver
   Depends on: I3C [=m] && HAS_IOMEM [=y] && (ARCH_RENESAS [=y] || COMPILE_TEST [=n])
   Location:
     -> Device Drivers
       -> I3C support (I3C [=m])
         -> Renesas I3C controller driver (RENESAS_I3C [=n])

 ~~~
 Commit: d028219a9f14 (i3c: master: Add basic driver for the Renesas I3C controller)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>