Turn off LIBBPF_DYNAMIC for perf builds
Libbpf has been split out to a separate package for a while now, but
perf moves quickly and the general purpose libbpf does not always keep
up with the version in the kernel tree.
More importantly, system libbpf does have other consumers, so if a
rebase were to bump the soname, that could be a problem. Let's solve
both issues by letting perf in the kernel tree use the libbpf in the
kernel tree.

Signed-off-by: Justin M. Forbes <jforbes@fedoraproject.org>