redhat/configs: Move RHEL/Fedora lockdown configs to common
In 3290ab0bcd33 ("redhat/configs: Enable early lockdown for Arm"), the
configuration was placed in the common/ hierarchy, but both Fedora and
RHEL config hierarchies contained overrides. Since the configuration on
both is the same, this moves everything over to common/. This has the
side-effect of actually enabling early lockdown on Arm.

Signed-off-by: Jeremy Cline <jeremycline@linux.microsoft.com>