redhat/Makefile: auto select -z-test-pesign target for z-stream
For RHEL, we want to use test-pesign target that matches current
stream. With the exception of when stream moves to z-stream.

We have to always update target (by a patch) when stream enters z-stream,
and it turns out we have been forgetting a lot.

Add a rule that automatically switches to use of z-stream target when
(ZSTREAM=YES) AND (GA tag exists). The reason for the GA tag check is
that we typically enable ZSTREAM already at ITM26 to use zstream nvr
numbering early to avoid nvr collisions with next minor release.

Signed-off-by: Jan Stancek <jstancek@redhat.com>