Fix packaging for libcpupower python binding debuginfo
Because we did not add _raw_pylibcpupower.so.debug to the debuginfo
list, it ended up the top level kernel-debuginfo-common package.  This
creates file conflicts because the kernel package keeps multiple copies
on most systems.
This just moves _raw_pylibcpupower.so.debug to the
kernel-tools-debuginfo package where it belongs.

Fixes rhbz 2393622

Signed-off-by: Justin M. Forbes <jforbes@fedoraproject.org>