[redhat] New configs in arch/x86
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 ~~~
 Symbol: MITIGATION_VMSCAPE [=y]
 Type  : bool
 Defined at arch/x86/Kconfig:2705
   Prompt: Mitigate VMSCAPE
   Depends on: CPU_MITIGATIONS [=y] && KVM [=m]
   Location:
     -> Mitigations for CPU vulnerabilities (CPU_MITIGATIONS [=y])
       -> Mitigate VMSCAPE (MITIGATION_VMSCAPE [=y])

 ~~~
 Commit: 556c1ad666ad (x86/vmscape: Enable the mitigation)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>