[redhat] New configs in arch/arm64
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 ~~~
 Symbol: ARM64_GCS [=y]
 Type  : bool
 Defined at arch/arm64/Kconfig:2211
   Prompt: Enable support for Guarded Control Stack (GCS)
   Location:
     -> Kernel Features
       -> ARMv9.4 architectural features
         -> Enable support for Guarded Control Stack (GCS) (ARM64_GCS [=y])
 Selects: ARCH_HAS_USER_SHADOW_STACK [=y] && ARCH_USES_HIGH_VMA_FLAGS [=y]

 ~~~
 Commit: 5d8b172e7005 (arm64: Add Kconfig for Guarded Control Stack (GCS))

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>