[redhat] New configs in arch/Kconfig
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 ~~~
 Symbol: CFI [=n]
 Type  : bool
 Defined at arch/Kconfig:882
   Prompt: Use Kernel Control Flow Integrity (kCFI)
   Depends on: ARCH_SUPPORTS_CFI [=y]
   Location:
     -> General architecture-dependent options
       -> Use Kernel Control Flow Integrity (kCFI) (CFI [=n])

 ~~~
 Commit: 23ef9d439769 (kcfi: Rename CONFIG_CFI_CLANG to CONFIG_CFI)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>