[redhat] New configs in drivers/perf
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 ~~~
 Symbol: FUJITSU_UNCORE_PMU [=n]
 Type  : tristate
 Defined at drivers/perf/Kconfig:181
   Prompt: Fujitsu Uncore PMU
   Depends on: PERF_EVENTS [=y] && (ARM64 [=y] && ACPI [=y] || COMPILE_TEST [=n] && 64BIT [=y])
   Location:
     -> Device Drivers
       -> Performance monitor support
         -> Fujitsu Uncore PMU (FUJITSU_UNCORE_PMU [=n])

 ~~~
 Commit: bad11557eed2 (perf: Fujitsu: Add the Uncore PMU driver)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>