[redhat] New configs in drivers/edac
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 ~~~
 Symbol: EDAC_CORTEX_A72 [=n]
 Type  : tristate
 Defined at drivers/edac/Kconfig:579
   Prompt: ARM Cortex A72
   Depends on: EDAC [=y] && ARM64 [=y]
   Location:
     -> Device Drivers
       -> EDAC (Error Detection And Correction) reporting (EDAC [=y])
         -> ARM Cortex A72 (EDAC_CORTEX_A72 [=n])

 ~~~
 Commit: fb13ae067ad7 (EDAC: Add EDAC driver for ARM Cortex A72 cores)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>