[redhat] New configs in kernel/Kconfig.preempt
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 ~~~
 Symbol: PREEMPT_RT_NEEDS_BH_LOCK [=n]
 Type  : bool
 Defined at kernel/Kconfig.preempt:106
   Prompt: Enforce softirq synchronisation on PREEMPT_RT
   Depends on: PREEMPT_RT [=y]
   Location:
     -> General setup
       -> Fully Preemptible Kernel (Real-Time) (PREEMPT_RT [=y])
         -> Enforce softirq synchronisation on PREEMPT_RT (PREEMPT_RT_NEEDS_BH_LOCK [=n])

 ~~~
 Commit: 3253cb49cbad (softirq: Allow to drop the softirq-BKL lock on PREEMPT_RT)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>