[redhat] New configs in drivers/i3c
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 ~~~
 Symbol: ADI_I3C_MASTER [=n]
 Type  : tristate
 Defined at drivers/i3c/master/Kconfig:2
   Prompt: Analog Devices I3C master driver
   Depends on: I3C [=m] && HAS_IOMEM [=y]
   Location:
     -> Device Drivers
       -> I3C support (I3C [=m])
         -> Analog Devices I3C master driver (ADI_I3C_MASTER [=n])

 ~~~
 Commit: a79ac2cdc91d (i3c: master: Add driver for Analog Devices I3C Controller IP)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>