[redhat] New configs in drivers/input
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 ~~~
 Symbol: INPUT_TPS6594_PWRBUTTON [=n]
 Type  : tristate
 Defined at drivers/input/misc/Kconfig:519
   Prompt: TPS6594 Power button driver
   Depends on: INPUT [=y] && INPUT_MISC [=y] && MFD_TPS6594 [=m]
   Location:
     -> Device Drivers
       -> Input device support
         -> Generic input layer (needed for keyboard, mouse, ...) (INPUT [=y])
           -> Miscellaneous devices (INPUT_MISC [=y])
             -> TPS6594 Power button driver (INPUT_TPS6594_PWRBUTTON [=n])

 ~~~
 Commit: 170031ff27dd (input: tps6594-pwrbutton: Add power button functionality)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>