[redhat] New configs in drivers/pwm
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 ~~~
 Symbol: PWM_PROVIDE_GPIO [=n]
 Type  : bool
 Defined at drivers/pwm/Kconfig:41
   Prompt: Provide a GPIO chip for each PWM chip
   Depends on: PWM [=y] && GPIOLIB [=y]
   Location:
     -> Device Drivers
       -> Pulse-Width Modulation (PWM) Support (PWM [=y])
         -> Provide a GPIO chip for each PWM chip (PWM_PROVIDE_GPIO [=n])

 ~~~
 Commit: e7c9b66b1069 (pwm: Provide a gpio device for waveform drivers)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>