[redhat] New configs in drivers/soc
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 ~~~
 Symbol: ARCH_R8A78000 [=y]
 Type  : bool
 Defined at drivers/soc/renesas/Kconfig:355
   Prompt: ARM64 Platform support for R8A78000 (R-Car X5H)
   Depends on: SOC_RENESAS [=y] && ARM64 [=y]
   Location:
     -> Device Drivers
       -> SOC (System On Chip) specific Drivers
         -> Renesas SoC driver support (SOC_RENESAS [=y])
           -> ARM64 Platform support for R8A78000 (R-Car X5H) (ARCH_R8A78000 [=y])
 Selects: ARCH_RCAR_GEN5 [=y]

 ~~~
 Commit: 5284d0b09d1b (soc: renesas: Identify R-Car X5H)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>