[redhat] New configs in drivers/spi
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 ~~~
 Symbol: SPI_VIRTIO [=n]
 Type  : tristate
 Defined at drivers/spi/Kconfig:1238
   Prompt: Virtio SPI Controller
   Depends on: SPI [=y] && SPI_MASTER [=y] && VIRTIO [=y]
   Location:
     -> Device Drivers
       -> SPI support (SPI [=y])
         -> Virtio SPI Controller (SPI_VIRTIO [=n])

 ~~~
 Commit: f98cabe3f6cf (SPI: Add virtio SPI driver)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>