[redhat] New configs in drivers/tee
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 ~~~
 Symbol: QCOMTEE [=n]
 Type  : tristate
 Defined at drivers/tee/qcomtee/Kconfig:3
   Prompt: Qualcomm TEE Support
   Depends on: TEE [=m] && !CPU_BIG_ENDIAN
   Location:
     -> Device Drivers
       -> Trusted Execution Environment support (TEE [=m])
         -> Qualcomm TEE Support (QCOMTEE [=n])
 Selects: QCOM_SCM [=n] && QCOM_TZMEM_MODE_SHMBRIDGE [=n]

 ~~~
 Commit: d6e290837e50 (tee: add Qualcomm TEE driver)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>