[redhat] New configs in drivers/dibs
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 ~~~
 Symbol: DIBS [=n]
 Type  : tristate
 Defined at drivers/dibs/Kconfig:2
   Prompt: DIBS support
   Depends on: NET [=y]
   Location:
     -> Networking support (NET [=y])
       -> Networking options
         -> DIBS support (DIBS [=n])

 ~~~
 Commit: 35758b0032c0 (dibs: Create drivers/dibs)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>