[redhat] New configs in drivers/net/ethernet/qualcomm
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 ~~~
 Symbol: QCOM_PPE [=n]
 Type  : tristate
 Defined at drivers/net/ethernet/qualcomm/Kconfig:63
   Prompt: Qualcomm Technologies, Inc. PPE Ethernet support
   Depends on: NETDEVICES [=y] && ETHERNET [=y] && NET_VENDOR_QUALCOMM [=y] && COMMON_CLK [=y] && HAS_IOMEM [=y] && OF [=y] && (ARCH_QCOM [=y] || COMPILE_TEST [=n])
   Location:
     -> Device Drivers
       -> Network device support (NETDEVICES [=y])
         -> Ethernet driver support (ETHERNET [=y])
           -> Qualcomm devices (NET_VENDOR_QUALCOMM [=y])
             -> Qualcomm Technologies, Inc. PPE Ethernet support (QCOM_PPE [=n])
 Selects: REGMAP_MMIO [=y]

 ~~~
 Commit: 353a0f1d5b27 (net: ethernet: qualcomm: Add PPE driver for IPQ9574 SoC)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>