[redhat] New configs in drivers/net/pcs
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 ~~~
 Symbol: PCS_RZN1_MIIC [=m]
 Type  : tristate
 Defined at drivers/net/pcs/Kconfig:28
   Prompt: Renesas RZ/N1, RZ/N2H, RZ/T2H MII converter
   Depends on: NETDEVICES [=y] && OF [=y] && (ARCH_RENESAS [=y] || COMPILE_TEST [=n])
   Location:
     -> Device Drivers
       -> Network device support (NETDEVICES [=y])
         -> PCS device drivers
           -> Renesas RZ/N1, RZ/N2H, RZ/T2H MII converter (PCS_RZN1_MIIC [=m])
 Selected by [m]:
   - DWMAC_RENESAS_GBETH [=m] && NETDEVICES [=y] && ETHERNET [=y] && NET_VENDOR_STMICRO [=y] && STMMAC_ETH [=m] && STMMAC_PLATFORM [=m] && OF [=y] && (ARCH_RENESAS [=y] || COMPILE_TEST [=n])
 Selected by [n]:
   - NET_DSA_RZN1_A5PSW [=n] && NETDEVICES [=y] && NET_DSA [=n] && OF [=y] && (ARCH_RZN1 [=n] || COMPILE_TEST [=n])
   - DWMAC_RZN1 [=n] && NETDEVICES [=y] && ETHERNET [=y] && NET_VENDOR_STMICRO [=y] && STMMAC_ETH [=m] && STMMAC_PLATFORM [=m] && OF [=y] && (ARCH_RZN1 [=n] || COMPILE_TEST [=n])

 ~~~
 Commit: 7dc54d3b8d91 (net: pcs: add Renesas MII converter driver)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>