[redhat] New configs in drivers/ptp
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 ~~~
 Symbol: PTP_NETC_V4_TIMER [=n]
 Type  : tristate
 Defined at drivers/ptp/Kconfig:256
   Prompt: NXP NETC V4 Timer PTP Driver
   Depends on: PTP_1588_CLOCK [=y] && PCI_MSI [=y]
   Location:
     -> Device Drivers
       -> PTP clock support
         -> NXP NETC V4 Timer PTP Driver (PTP_NETC_V4_TIMER [=n])

 ~~~
 Commit: 87a201d59963 (ptp: netc: add NETC V4 Timer PTP driver support)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>