[redhat] New configs in kernel/Kconfig.kexec
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 ~~~
 Symbol: CRASH_DUMP_KUNIT_TEST [=m]
 Type  : tristate
 Defined at kernel/Kconfig.kexec:151
   Prompt: Unit Tests for kernel crash dumps
   Depends on: CRASH_DUMP [=y] && KUNIT [=m]
   Visible if: CRASH_DUMP [=y] && KUNIT [=m] && !KUNIT_ALL_TESTS [=m]
   Location:
     -> General setup
       -> Kexec and crash features
         -> Unit Tests for kernel crash dumps (CRASH_DUMP_KUNIT_TEST [=m])

 ~~~
 Commit: 913e65a2fe1a (crash: add KUnit tests for crash_exclude_mem_range)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>