[redhat] New configs in lib/Kconfig.debug
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 ~~~
 Symbol: FIND_BIT_BENCHMARK_RUST [=n]
 Type  : tristate
 Defined at lib/Kconfig.debug:2618
   Prompt: Test find_bit functions in Rust
   Depends on: RUNTIME_TESTING_MENU [=y] && RUST [=y]
   Location:
     -> Kernel hacking
       -> Kernel Testing and Coverage
         -> Runtime Testing (RUNTIME_TESTING_MENU [=y])
           -> Test find_bit functions in Rust (FIND_BIT_BENCHMARK_RUST [=n])

 ~~~
 Commit: 38cc91db2e87 (rust: add find_bit_benchmark_rust module.)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>