[redhat] New configs in net/psp
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 ~~~
 Symbol: INET_PSP [=n]
 Type  : bool
 Defined at net/psp/Kconfig:5
   Prompt: PSP Security Protocol support
   Depends on: NET [=y] && INET [=y]
   Location:
     -> Networking support (NET [=y])
       -> Networking options
         -> PSP Security Protocol support (INET_PSP [=n])
 Selects: SKB_DECRYPTED [=y] && SOCK_VALIDATE_XMIT [=y]

 ~~~
 Commit: 00c94ca2b99e (psp: base PSP device support)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>