[redhat] New configs in net/sctp
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 ~~~
 Symbol: SCTP_DEFAULT_COOKIE_HMAC_SHA256 [=y]
 Type  : bool
 Defined at net/sctp/Kconfig:63
   Prompt: HMAC-SHA256
   Depends on: NET [=y] && IP_SCTP [=m]
   Location:
     -> Networking support (NET [=y])
       -> Networking options
         -> The SCTP Protocol (IP_SCTP [=m])
           -> Default SCTP cookie authentication method (<choice> [=n])
             -> HMAC-SHA256 (SCTP_DEFAULT_COOKIE_HMAC_SHA256 [=y])

 ~~~
 Commit: 2f3dd6ec901f (sctp: Convert cookie authentication to use HMAC-SHA256)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>