[redhat] New configs in security/Kconfig.hardening
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 ~~~
 Symbol: RUST_BITMAP_HARDENED [=n]
 Type  : bool
 Defined at security/Kconfig.hardening:258
   Prompt: Check integrity of bitmap Rust API
   Depends on: RUST [=y]
   Location:
     -> Security options
       -> Kernel hardening options
         -> Hardening of kernel data structures
           -> Check integrity of bitmap Rust API (RUST_BITMAP_HARDENED [=n])

 ~~~
 Commit: 11eca92a2cae (rust: add bitmap API.)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>
Signed-off-by: Vladis Dronov <vdronov@redhat.com>