[redhat] New configs in sound/usb
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 ~~~
 Symbol: SND_USB_US144MKII [=m]
 Type  : tristate
 Defined at sound/usb/Kconfig:120
   Prompt: Tascam US-144MKII USB driver
   Depends on: SOUND [=m] && SND [=m] && SND_USB [=y] && USB [=y] && (X86 [=y] || COMPILE_TEST [=n])
   Location:
     -> Device Drivers
       -> Sound card support (SOUND [=m])
         -> Advanced Linux Sound Architecture (SND [=m])
           -> USB sound devices (SND_USB [=y])
             -> Tascam US-144MKII USB driver (SND_USB_US144MKII [=n])
 Selects: SND_RAWMIDI [=m] && SND_PCM [=m]

 ~~~
 Commit: 0ec417aa8546 (ALSA: usb-audio: Add infrastructure for TASCAM US-144MKII)

Enable for RHEL by default except automotive.

---

Signed-off-by: Jaroslav Kysela <jkysela@redhat.com>
Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>