[redhat] New configs in drivers/md
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 ~~~
 Symbol: DM_PCACHE [=n]
 Type  : tristate
 Defined at drivers/md/dm-pcache/Kconfig:1
   Prompt: Persistent cache for Block Device (Experimental)
   Depends on: MD [=y] && BLK_DEV_DM [=m] && DEV_DAX [=m]
   Location:
     -> Device Drivers
       -> Multiple devices driver support (RAID and LVM) (MD [=y])
         -> Device mapper support (BLK_DEV_DM [=m])
           -> Persistent cache for Block Device (Experimental) (DM_PCACHE [=n])

 ~~~
 Commit: 1d57628ff95b (dm-pcache: add persistent cache target in device-mapper)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>