[redhat] New configs in net/sunrpc
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 ~~~
 Symbol: SUNRPC_DEBUG_TRACE [=n]
 Type  : bool
 Defined at net/sunrpc/Kconfig:104
   Prompt: RPC: Send dfprintk() output to the trace buffer
   Depends on: NETWORK_FILESYSTEMS [=y] && SUNRPC_DEBUG [=y] && TRACING [=y]
   Location:
     -> File systems
       -> Network File Systems (NETWORK_FILESYSTEMS [=y])
         -> RPC: Enable dprintk debugging (SUNRPC_DEBUG [=y])
           -> RPC: Send dfprintk() output to the trace buffer (SUNRPC_DEBUG_TRACE [=n])

 ~~~
 Commit: ec7d8e68ef0e (sunrpc: add a Kconfig option to redirect dfprintk() output to trace buffer)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>