[redhat] New configs in drivers/clk
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 ~~~
 Symbol: COMMON_CLK_RPMI [=y]
 Type  : tristate
 Defined at drivers/clk/Kconfig:504
   Prompt: Clock driver based on RISC-V RPMI
   Depends on: COMMON_CLK [=y] && (RISCV [=y] || COMPILE_TEST [=n]) && MAILBOX [=y]
   Location:
     -> Device Drivers
       -> Common Clock Framework (COMMON_CLK [=y])
         -> Clock driver based on RISC-V RPMI (COMMON_CLK_RPMI [=y])

 ~~~
 Commit: 5ba9f520f41a (clk: Add clock driver for the RISC-V RPMI clock service group)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>

v2:
- CONFIG_COMMON_CLK_RPMI: moved under rhel/generic/riscv/riscv64/.

Signed-off-by: Patrick Talbert <ptalbert@redhat.com>