[redhat] New configs in drivers/extcon
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 ~~~
 Symbol: EXTCON_MAX14526 [=n]
 Type  : tristate
 Defined at drivers/extcon/Kconfig:137
   Prompt: Maxim MAX14526 EXTCON Support
   Depends on: EXTCON [=y] && I2C [=y]
   Location:
     -> Device Drivers
       -> External Connector Class (extcon) support (EXTCON [=y])
         -> Maxim MAX14526 EXTCON Support (EXTCON_MAX14526 [=n])
 Selects: IRQ_DOMAIN [=y] && REGMAP_I2C [=y]

 ~~~
 Commit: 145af3ddd1cd (extcon: Add basic support for Maxim MAX14526 MUIC)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>