[redhat] New configs in drivers/mailbox
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 ~~~
 Symbol: RISCV_SBI_MPXY_MBOX [=y]
 Type  : tristate
 Defined at drivers/mailbox/Kconfig:372
   Prompt: RISC-V SBI Message Proxy (MPXY) Mailbox
   Depends on: MAILBOX [=y] && RISCV_SBI [=y]
   Location:
     -> Device Drivers
       -> Mailbox Hardware Support (MAILBOX [=y])
         -> RISC-V SBI Message Proxy (MPXY) Mailbox (RISCV_SBI_MPXY_MBOX [=y])

 ~~~
 Commit: bf3022a4eb11 (mailbox: Add RISC-V SBI message proxy (MPXY) based mailbox driver)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>