[redhat] New configs in drivers/misc
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 ~~~
 Symbol: EEPROM_M24LR [=n]
 Type  : tristate
 Defined at drivers/misc/eeprom/Kconfig:123
   Prompt: STMicroelectronics M24LR RFID/NFC EEPROM support
   Depends on: I2C [=y] && SYSFS [=y]
   Location:
     -> Device Drivers
       -> Misc devices
         -> EEPROM support
           -> STMicroelectronics M24LR RFID/NFC EEPROM support (EEPROM_M24LR [=n])
 Selects: REGMAP_I2C [=m] && NVMEM [=y] && NVMEM_SYSFS [=y]

 ~~~
 Commit: cd5c5e0231c5 (eeprom: add driver for ST M24LR series RFID/NFC EEPROM chips)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>