[redhat] New configs in drivers/platform
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 ~~~
 Symbol: EC_LENOVO_THINKPAD_T14S [=n]
 Type  : tristate
 Defined at drivers/platform/arm64/Kconfig:73
   Prompt: Lenovo Thinkpad T14s Embedded Controller driver
   Depends on: ARM64_PLATFORM_DEVICES [=y] && (ARCH_QCOM [=y] || COMPILE_TEST [=n]) && I2C [=y] && INPUT [=y]
   Location:
     -> Device Drivers
       -> ARM64 Platform-Specific Device Drivers (ARM64_PLATFORM_DEVICES [=y])
         -> Lenovo Thinkpad T14s Embedded Controller driver (EC_LENOVO_THINKPAD_T14S [=n])
 Selects: INPUT_SPARSEKMAP [=m] && LEDS_CLASS [=y] && NEW_LEDS [=y] && SND_CTL_LED [=m]

 ~~~
 Commit: 60b7ab6ce030 (platform: arm64: thinkpad-t14s-ec: new driver)

---

 ~~~
 Symbol: REDMI_WMI [=n]
 Type  : tristate
 Defined at drivers/platform/x86/Kconfig:121
   Prompt: Redmibook WMI key driver
   Depends on: X86_PLATFORM_DEVICES [=y] && ACPI_WMI [=m] && INPUT [=y]
   Location:
     -> Device Drivers
       -> X86 Platform Specific Device Drivers (X86_PLATFORM_DEVICES [=y])
         -> Redmibook WMI key driver (REDMI_WMI [=n])
 Selects: INPUT_SPARSEKMAP [=m]

 ~~~
 Commit: 98aadf8e494d (platform/x86: Add WMI driver for Redmibook keyboard)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>