[redhat] New configs in drivers/tty
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 ~~~
 Symbol: SERIAL_QCOM_GENI_UART_PORTS [=8]
 Type  : integer
 Defined at drivers/tty/serial/Kconfig:931
   Prompt: Maximum number of GENI UART ports
   Depends on: TTY [=y] && HAS_IOMEM [=y] && SERIAL_QCOM_GENI [=y]
   Location:
     -> Device Drivers
       -> Character devices
         -> Enable TTY (TTY [=y])
           -> Serial drivers
             -> QCOM on-chip GENI based serial port support (SERIAL_QCOM_GENI [=y])
               -> Maximum number of GENI UART ports (SERIAL_QCOM_GENI_UART_PORTS [=8])

 ~~~
 Commit: 9391ab1ed9b3 (serial: qcom-geni: Make UART port count configurable via Kconfig)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>