[redhat] New configs in drivers/usb
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 ~~~
 Symbol: USB_DWC3_GENERIC_PLAT [=m]
 Type  : tristate
 Defined at drivers/usb/dwc3/Kconfig:192
   Prompt: DWC3 Generic Platform Driver
   Depends on: USB_SUPPORT [=y] && USB_DWC3 [=m] && OF [=y] && COMMON_CLK [=y]
   Location:
     -> Device Drivers
       -> USB support (USB_SUPPORT [=y])
         -> DesignWare USB3 DRD Core Support (USB_DWC3 [=m])
           -> DWC3 Generic Platform Driver (USB_DWC3_GENERIC_PLAT [=m])

 ~~~
 Commit: e0b6dc00c701 (usb: dwc3: add generic driver to support flattened)

---

 ~~~
 Symbol: USB_USBIO [=n]
 Type  : tristate
 Defined at drivers/usb/misc/Kconfig:182
   Prompt: Intel USBIO Bridge support
   Depends on: USB_SUPPORT [=y] && USB [=y] && ACPI [=y]
   Location:
     -> Device Drivers
       -> USB support (USB_SUPPORT [=y])
         -> Intel USBIO Bridge support (USB_USBIO [=n])
 Selects: AUXILIARY_BUS [=y]

 ~~~
 Commit: 121a0f839dbb (usb: misc: Add Intel USBIO bridge driver)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>