[redhat] New configs in drivers/base
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 ~~~
 Symbol: PM_RUNTIME_KUNIT_TEST [=m]
 Type  : tristate
 Defined at drivers/base/Kconfig:170
   Prompt: KUnit Tests for runtime PM
   Depends on: KUNIT [=m] && PM [=y]
   Visible if: KUNIT [=m] && PM [=y] && !KUNIT_ALL_TESTS [=m]
   Location:
     -> Device Drivers
       -> Generic Driver Options
         -> KUnit Tests for runtime PM (PM_RUNTIME_KUNIT_TEST [=m])

 ~~~
 Commit: 7f7acd193ba8 (PM: runtime: Add basic kunit tests for API contracts)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>