redhat: introduce RELEASE_LOCALVERSION variable
Sometimes we run into situation where people forget to set proper
localversion during official builds and end up with .test in NVR.

Introduce a new variable to Makefile.variables, which will be persisted
in source-git as well (contrast to localversion) and take priority when
used for dist-rtg targets. The scenarios where we need custom localversion
for official build should be very rare, meaning that the maintainer doesn't
need to touch this variable and default (empty localversion) will be used
instead of whatever may be left in 'localversion' (or DISTLOCALVERSION).

Signed-off-by: Jan Stancek <jstancek@redhat.com>