Configs: Mark SCHED_MC as enabled for powerpc
Upstream consolidated some of the arch/topology scheduler configs
with 7bd291abe2da ("sched: Unify the SCHED_{SMT,CLUSTER,MC} Kconfig")
which has the side effect of blocking the ability to disable them
selectively if the arch actually supports the feature. In RHEL we had
CONFIG_SCHED_MC off on powerpc but that is no longer possible.
Remove the new pending-rhel file (=y) and the old (is not set) file.

Signed-off-by: Phil Auld <pauld@redhat.com>