[redhat] New configs in lib/Kconfig.kmsan
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 ~~~
 Symbol: KMSAN [=n]
 Type  : bool
 Defined at lib/Kconfig.kmsan:8
   Prompt: KMSAN: detector of uninitialized values use
   Depends on: HAVE_ARCH_KMSAN [=y] && HAVE_KMSAN_COMPILER [=y] && DEBUG_KERNEL [=y] && !KASAN [=n] && !KCSAN [=n] && !PREEMPT_RT [=n]
   Location:
     -> Kernel hacking
       -> Memory Debugging
         -> KMSAN: detector of uninitialized values use (KMSAN [=n])
 Selects: STACKDEPOT [=y] && STACKDEPOT_ALWAYS_INIT [=n]

 ~~~
 Commit: f80be4571b19 (kmsan: add KMSAN runtime core)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>