redhat/configs: Set CONFIG_SCHED_PROXY_EXEC=n
The new CONFIG_SCHED_PROXY_EXEC kernel option is for enabling a
new experimental proxy execution feature to address the problem of
priority inversion, especially within the context of real-time or
latency-sensitive workloads like those found in Android systems..

As this feature is still experimental and not fully completed and
tested, disable it for now. This also solves the "make dist-configs"
failure problem.

Signed-off-by: Waiman Long <longman@redhat.com>