[redhat] New configs in kernel/Kconfig.kexec
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 ~~~
 Symbol: KEXEC_HANDOVER_DEBUG [=n]
 Type  : bool
 Defined at kernel/Kconfig.kexec:112
   Prompt: Enable Kexec Handover debug checks
   Depends on: KEXEC_HANDOVER [=y]
   Location:
     -> General setup
       -> Kexec and crash features
         -> kexec handover (KEXEC_HANDOVER [=y])
           -> Enable Kexec Handover debug checks (KEXEC_HANDOVER_DEBUG [=n])

 ~~~
 Commit: e38f65d317df (kho: warn and fail on metadata or preserved memory in scratch area)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>