Rename with_up to with_stock.
The term 'up' used to refer to the specific kernel for a uni-processor.
Today, this term "up" remains in the spec file but it no longer retains
any of its original meaning. Currently 'up' refers to the stock kernel
build that produces the kernel, kernel-core, kernel-modules, etc. packages.
Rename the with_up variable to with_stock to avoid confusion and improve
spec file readability.

Keep the possibility to use with_up for backwards compatibility.

Signed-off-by: Alexandra Hájková <ahajkova@redhat.com>