[redhat] New configs in drivers/edac
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 ~~~
 Symbol: EDAC_IMH [=n]
 Type  : tristate
 Defined at drivers/edac/Kconfig:286
   Prompt: Intel Integrated Memory/IO Hub MC
   Depends on: EDAC [=y] && X86_64 [=y] && X86_MCE_INTEL [=y] && ACPI [=y] && (ACPI_NFIT [=m] || !ACPI_NFIT [=m])
   Location:
     -> Device Drivers
       -> EDAC (Error Detection And Correction) reporting (EDAC [=y])
         -> Intel Integrated Memory/IO Hub MC (EDAC_IMH [=n])
 Selects: DMI [=y] && ACPI_ADXL [=y]

 ~~~
 Commit: 9fc67b11703f (EDAC/{skx_common,imh}: Add EDAC driver for Intel Diamond Rapids servers)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>

v2:
- CONFIG_EDAC_IMH: enable for rhel as a module.

Signed-off-by: Patrick Talbert <ptalbert@redhat.com>