[redhat] New configs in drivers/thermal
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 ~~~
 Symbol: IMX91_THERMAL [=n]
 Type  : tristate
 Defined at drivers/thermal/Kconfig:299
   Prompt: Temperature sensor driver for NXP i.MX91 SoC
   Depends on: THERMAL [=y] && (ARCH_MXC [=y] || COMPILE_TEST [=n]) && OF [=y]
   Location:
     -> Device Drivers
       -> Thermal drivers (THERMAL [=y])
         -> Temperature sensor driver for NXP i.MX91 SoC (IMX91_THERMAL [=n])

 ~~~
 Commit: c411d8bf0699 (thermal/drivers/imx91: Add support for i.MX91 thermal monitoring unit)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>