[redhat] New configs in kernel/power
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 ~~~
 Symbol: PM_QOS_CPU_SYSTEM_WAKEUP [=n]
 Type  : bool
 Defined at kernel/power/Kconfig:205
   Prompt: User space interface for CPU system wakeup QoS
   Depends on: CPU_IDLE [=y]
   Location:
     -> Power management and ACPI options
       -> User space interface for CPU system wakeup QoS (PM_QOS_CPU_SYSTEM_WAKEUP [=n])

 ~~~
 Commit: a4e6512a79d8 (PM: QoS: Introduce a CPU system wakeup QoS limit)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>