[redhat] New configs in drivers/media
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 ~~~
 Symbol: VIDEO_IMX111 [=n]
 Type  : tristate
 Defined at drivers/media/i2c/Kconfig:140
   Prompt: Sony IMX111 sensor support
   Depends on: MEDIA_SUPPORT [=m] && VIDEO_DEV [=m] && VIDEO_CAMERA_SENSOR [=y]
   Location:
     -> Device Drivers
       -> Multimedia support (MEDIA_SUPPORT [=m])
         -> Media ancillary drivers
           -> Camera sensor devices (VIDEO_CAMERA_SENSOR [=y])
             -> Sony IMX111 sensor support (VIDEO_IMX111 [=n])
 Selects: V4L2_CCI_I2C [=n]

 ~~~
 Commit: d363bdfa0ec6 (media: i2c: add Sony IMX111 CMOS camera sensor driver)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>