[redhat] New configs in drivers/mfd
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 ~~~
 Symbol: MFD_PF1550 [=n]
 Type  : tristate
 Defined at drivers/mfd/Kconfig:609
   Prompt: NXP PF1550 PMIC Support
   Depends on: HAS_IOMEM [=y] && I2C [=y]=y [=y] && OF [=y]
   Location:
     -> Device Drivers
       -> Multifunction device drivers
         -> NXP PF1550 PMIC Support (MFD_PF1550 [=n])
 Selects: MFD_CORE [=m] && REGMAP_I2C [=m] && REGMAP_IRQ [=n]

 ~~~
 Commit: ebaec90ec0b5 (mfd: pf1550: Add core driver for the PF1550 PMIC)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>