[redhat] New configs in drivers/power
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 ~~~
 Symbol: CHARGER_RT9756 [=n]
 Type  : tristate
 Defined at drivers/power/supply/Kconfig:945
   Prompt: Richtek RT9756 smart cap divider charger driver
   Depends on: POWER_SUPPLY [=y] && I2C [=y]
   Location:
     -> Device Drivers
       -> Power supply class support (POWER_SUPPLY [=y])
         -> Richtek RT9756 smart cap divider charger driver (CHARGER_RT9756 [=n])
 Selects: REGMAP_I2C [=m] && LINEAR_RANGES [=y]

 ~~~
 Commit: b6f0796d4398 (power: supply: rt9756: Add Richtek RT9756 smart cap divider charger)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>