[redhat] New configs in drivers/thermal
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 ~~~
 Symbol: DEVFREQ_THERMAL [=n]
 Type  : bool
 Defined at drivers/thermal/Kconfig:207
   Prompt: Generic device cooling support
   Depends on: THERMAL [=y] && PM_DEVFREQ [=y] && PM_OPP [=y]
   Location:
     -> Device Drivers
       -> Thermal drivers (THERMAL [=y])
         -> Generic device cooling support (DEVFREQ_THERMAL [=n])

 ~~~
 Commit: a76caf55e5b3 (thermal: Add devfreq cooling)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>