[redhat] New configs in drivers/video
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 ~~~
 Symbol: BACKLIGHT_AW99706 [=n]
 Type  : tristate
 Defined at drivers/video/backlight/Kconfig:159
   Prompt: Backlight Driver for Awinic AW99706
   Depends on: HAS_IOMEM [=y] && BACKLIGHT_CLASS_DEVICE [=y] && I2C [=y]
   Location:
     -> Device Drivers
       -> Graphics support
         -> Backlight & LCD device support
           -> Lowlevel Backlight controls (BACKLIGHT_CLASS_DEVICE [=y])
             -> Backlight Driver for Awinic AW99706 (BACKLIGHT_AW99706 [=n])
 Selects: REGMAP_I2C [=m]

 ~~~
 Commit: 147b38a5ad06 (backlight: aw99706: Add support for Awinic AW99706 backlight)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>