[redhat] New configs in sound/hda
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 ~~~
 Symbol: SND_HDA_SCODEC_CS35L56_CAL_DEBUGFS [=n]
 Type  : bool
 Defined at sound/hda/codecs/side-codecs/Kconfig:94
   Prompt: CS35L56 create debugfs for factory calibration
   Depends on: SOUND [=m] && SND [=m] && SND_HDA [=m] && SND_HDA_SCODEC_CS35L56 [=m] && DEBUG_FS [=y]
   Location:
     -> Device Drivers
       -> Sound card support (SOUND [=m])
         -> Advanced Linux Sound Architecture (SND [=m])
           -> HD-Audio
             -> CS35L56 driver options
               -> CS35L56 create debugfs for factory calibration (SND_HDA_SCODEC_CS35L56_CAL_DEBUGFS [=n])
 Selects: SND_SOC_CS35L56_CAL_DEBUGFS_COMMON [=n]

 ~~~
 Commit: 46a3df50b0ca (ALSA: hda/cs35l56: Create debugfs files for factory calibration)
 Enable this option for debug kernels.

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>
Signed-off-by: Jaroslav Kysela <jkysela@redhat.com>