[redhat] New configs in arch/riscv
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 ~~~
 Symbol: ARCH_ANLOGIC [=n]
 Type  : bool
 Defined at arch/riscv/Kconfig.socs:10
   Prompt: Anlogic SoCs
   Location:
     -> SoC selection
       -> Anlogic SoCs (ARCH_ANLOGIC [=n])

 ~~~
 Commit: 9c96219602b1 (riscv: Add Anlogic SoC famly Kconfig support)

---

 ~~~
 Symbol: ARCH_TENSTORRENT [=n]
 Type  : bool
 Defined at arch/riscv/Kconfig.socs:71
   Prompt: Tenstorrent SoCs
   Location:
     -> SoC selection
       -> Tenstorrent SoCs (ARCH_TENSTORRENT [=n])

 ~~~
 Commit: 16eb4e02a631 (riscv: Kconfig.socs: Add ARCH_TENSTORRENT for Tenstorrent SoCs)

---

 ~~~
 Symbol: RISCV_ISA_SVRSW60T59B [=y]
 Type  : bool
 Defined at arch/riscv/Kconfig:854
   Prompt: Svrsw60t59b extension support for using PTE bits 60 and 59
   Depends on: MMU [=y] && 64BIT [=y] && RISCV_ALTERNATIVE [=y]
   Location:
     -> Platform type
       -> Svrsw60t59b extension support for using PTE bits 60 and 59 (RISCV_ISA_SVRSW60T59B [=y])

 ~~~
 Commit: 59f6acb4be02 (riscv: add RISC-V Svrsw60t59b extension support)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>