[redhat] New configs in arch/arm64
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 ~~~
 Symbol: ARCH_BST [=n]
 Type  : bool
 Defined at arch/arm64/Kconfig.platforms:122
   Prompt: Black Sesame Technologies SoC Family
   Location:
     -> Platform selection
       -> Black Sesame Technologies SoC Family (ARCH_BST [=n])

 ~~~
 Commit: 154121941643 (arm64: Kconfig: add ARCH_BST for Black Sesame Technologies SoCs)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>