[redhat] New configs in drivers/cache
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 ~~~
 Symbol: CACHEMAINT_FOR_DMA [=y]
 Type  : bool
 Defined at drivers/cache/Kconfig:3
   Prompt: Cache management for noncoherent DMA
   Depends on: RISCV [=y]
   Location:
     -> Device Drivers
       -> Cache management for noncoherent DMA (CACHEMAINT_FOR_DMA [=y])

 ~~~
 Commit: 4d1608d0ab33 (cache: Make top level Kconfig menu a boolean dependent on RISCV)

---

 ~~~
 Symbol: CACHEMAINT_FOR_HOTPLUG [=n]
 Type  : bool
 Defined at drivers/cache/Kconfig:36
   Prompt: Cache management for memory hot plug like operations
   Depends on: GENERIC_CPU_CACHE_MAINTENANCE [=y]
   Location:
     -> Device Drivers
       -> Cache management for memory hot plug like operations (CACHEMAINT_FOR_HOTPLUG [=n])

 ~~~
 Commit: 2ec3b54a6ff0 (cache: Support cache maintenance for HiSilicon SoC Hydra Home Agent)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>