[redhat] New configs in drivers/clk
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 ~~~
 Symbol: SM_VIDEOCC_8750 [=n]
 Type  : tristate
 Defined at drivers/clk/qcom/Kconfig:1534
   Prompt: SM8750 Video Clock Controller
   Depends on: COMMON_CLK [=y] && COMMON_CLK_QCOM [=y] && (ARM64 [=y] || COMPILE_TEST [=n])
   Location:
     -> Device Drivers
       -> Common Clock Framework (COMMON_CLK [=y])
         -> Support for Qualcomm's clock controllers (COMMON_CLK_QCOM [=y])
           -> SM8750 Video Clock Controller (SM_VIDEOCC_8750 [=n])
 Selects: SM_GCC_8750 [=n] && QCOM_GDSC [=y]

 ~~~
 Commit: a160860529b5 (clk: qcom: videocc-sm8750: Add video clock controller driver for SM8750)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>