[redhat] New configs in drivers/platform
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 ~~~
 Symbol: BCM_VIDEOCORE [=m]
 Type  : tristate
 Defined at drivers/platform/raspberrypi/Kconfig:3
   Prompt: Broadcom VideoCore support
   Depends on: OF [=y] && (RASPBERRYPI_FIRMWARE [=m] || COMPILE_TEST [=n] && !RASPBERRYPI_FIRMWARE [=m])
   Location:
     -> Device Drivers
       -> Broadcom VideoCore support (BCM_VIDEOCORE [=m])

 ~~~
 Commit: aa125180c673 (platform/raspberrypi: Destage VCHIQ interface)

---

 ~~~
 Symbol: VCHIQ_CDEV [=y]
 Type  : bool
 Defined at drivers/platform/raspberrypi/Kconfig:34
   Prompt: VCHIQ Character Driver
   Depends on: BCM_VIDEOCORE [=m] && BCM2835_VCHIQ [=m]
   Location:
     -> Device Drivers
       -> Broadcom VideoCore support (BCM_VIDEOCORE [=m])
         -> BCM2835 VCHIQ (BCM2835_VCHIQ [=m])
           -> VCHIQ Character Driver (VCHIQ_CDEV [=y])
 Implied by [m]:
   - BCM2835_VCHIQ [=m] && BCM_VIDEOCORE [=m] && HAS_DMA [=y]

 ~~~
 Commit: aa125180c673 (platform/raspberrypi: Destage VCHIQ interface)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>