[redhat] New configs in drivers/pci
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 ~~~
 Symbol: PCI_TSM [=n]
 Type  : bool
 Defined at drivers/pci/Kconfig:128
   Prompt: PCI TSM: Device security protocol support
   Depends on: PCI [=y]
   Location:
     -> Device Drivers
       -> PCI support (PCI [=y])
         -> PCI TSM: Device security protocol support (PCI_TSM [=n])
 Selects: PCI_IDE [=n] && PCI_DOE [=y] && TSM [=n]
 Selected by [n]:
   - CRYPTO_DEV_SP_PSP [=n] && CRYPTO [=y] && CRYPTO_HW [=y] && CRYPTO_DEV_CCP [=y] && CRYPTO_DEV_CCP_DD [=m] && X86_64 [=y] && AMD_IOMMU [=y] && PCI [=y]

 ~~~
 Commit: 3225f52cde56 (PCI/TSM: Establish Secure Sessions and Link Encryption)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>