[redhat] New configs in drivers/uio
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 ~~~
 Symbol: UIO_PCI_GENERIC_SVA [=n]
 Type  : tristate
 Defined at drivers/uio/Kconfig:168
   Prompt: Generic driver for PCI Express that supports sva
   Depends on: UIO [=m] && PCI [=y] && IOMMU_SVA [=y]
   Location:
     -> Device Drivers
       -> Userspace I/O drivers (UIO [=m])
         -> Generic driver for PCI Express that supports sva (UIO_PCI_GENERIC_SVA [=n])

 ~~~
 Commit: 3397c3cd859a (uio: Add SVA support for PCI devices via uio_pci_generic_sva.c)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>