[redhat] New configs in drivers/watchdog
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 ~~~
 Symbol: RENESAS_WWDT [=n]
 Type  : tristate
 Defined at drivers/watchdog/Kconfig:972
   Prompt: Renesas Window WWDT Watchdog
   Depends on: WATCHDOG [=y] && (ARCH_RENESAS [=y] || COMPILE_TEST [=n])
   Location:
     -> Device Drivers
       -> Watchdog Timer Support (WATCHDOG [=y])
         -> Renesas Window WWDT Watchdog (RENESAS_WWDT [=n])
 Selects: WATCHDOG_CORE [=m]

 ~~~
 Commit: babe81b06158 (watchdog: renesas_wwdt: add driver)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>