[redhat] New configs in init/Kconfig
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 ~~~
 Symbol: CMDLINE_LOG_WRAP_IDEAL_LEN [=1021]
 Type  : integer
 Range : [0 1021]
 Defined at init/Kconfig:1522
   Prompt: Length to try to wrap the cmdline when logged at boot
   Location:
     -> General setup
       -> Length to try to wrap the cmdline when logged at boot (CMDLINE_LOG_WRAP_IDEAL_LEN [=1021])

 ~~~
 Commit: 032a730268a3 (init/main.c: wrap long kernel cmdline when printing to logs)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>