[redhat] New configs in kernel/trace
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 ~~~
 Symbol: FUNCTION_SELF_TRACING [=n]
 Type  : bool
 Defined at kernel/trace/Kconfig:345
   Prompt: Function trace tracing code
   Depends on: FTRACE [=y] && FUNCTION_TRACER [=y]
   Location:
     -> Kernel hacking
       -> Tracers (FTRACE [=y])
         -> Function trace tracing code (FUNCTION_SELF_TRACING [=n])

 ~~~
 Commit: f93a7d0caccd (ftrace: Allow tracing of some of the tracing code)

---

 ~~~
 Symbol: TRACE_SYSCALL_BUF_SIZE_DEFAULT [=63]
 Type  : integer
 Range : [0 165]
 Defined at kernel/trace/Kconfig:604
   Prompt: System call user read max size
   Depends on: FTRACE [=y] && FTRACE_SYSCALLS [=y]
   Location:
     -> Kernel hacking
       -> Tracers (FTRACE [=y])
         -> Trace syscalls (FTRACE_SYSCALLS [=y])
           -> System call user read max size (TRACE_SYSCALL_BUF_SIZE_DEFAULT [=63])

 ~~~
 Commit: 299ea67e6a2b (tracing: Add a config and syscall_user_buf_size file to limit amount written)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>