[redhat] New configs in lib/Kconfig.debug
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 ~~~
 Symbol: BASE64_KUNIT [=m]
 Type  : tristate
 Defined at lib/Kconfig.debug:2820
   Prompt: KUnit test for base64 decoding and encoding
   Depends on: RUNTIME_TESTING_MENU [=y] && KUNIT [=m]
   Visible if: RUNTIME_TESTING_MENU [=y] && KUNIT [=m] && !KUNIT_ALL_TESTS [=m]
   Location:
     -> Kernel hacking
       -> Kernel Testing and Coverage
         -> Runtime Testing (RUNTIME_TESTING_MENU [=y])
           -> KUnit test for base64 decoding and encoding (BASE64_KUNIT [=m])

 ~~~
 Commit: 8b365c4f5be9 (lib: add KUnit tests for base64 encoding/decoding)

---

 ~~~
 Symbol: BOOTPARAM_HUNG_TASK_PANIC [=0]
 Type  : integer
 Defined at lib/Kconfig.debug:1261
   Prompt: Number of hung tasks to trigger kernel panic
   Depends on: DETECT_HUNG_TASK [=y]
   Location:
     -> Kernel hacking
       -> Debug Oops, Lockups and Hangs
         -> Detect Hung Tasks (DETECT_HUNG_TASK [=y])
           -> Number of hung tasks to trigger kernel panic (BOOTPARAM_HUNG_TASK_PANIC [=0])

 ~~~
 Commit: 0610c8a8a258 (order memory debugging Kconfig options)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>