[redhat] New configs in drivers/dma
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 ~~~
 Symbol: RZ_DMAC [=n]
 Type  : tristate
 Defined at drivers/dma/sh/Kconfig:51
   Prompt: Renesas RZ DMA Controller
   Depends on: DMADEVICES [=y] && (ARCH_RENESAS [=y] || COMPILE_TEST [=n])
   Location:
     -> Device Drivers
       -> DMA Engine support (DMADEVICES [=y])
         -> Renesas RZ DMA Controller (RZ_DMAC [=n])
 Selects: RENESAS_DMA [=y] && DMA_VIRTUAL_CHANNELS [=m]

 ~~~
 Commit: 5000d37042a6 (dmaengine: sh: Add DMAC driver for RZ/G2L SoC)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>