[redhat] New configs in drivers/phy
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 ~~~
 Symbol: PHY_RZ_G3E_USB3 [=n]
 Type  : tristate
 Defined at drivers/phy/renesas/Kconfig:44
   Prompt: Renesas RZ/G3E USB 3.0 PHY driver
   Depends on: ARCH_RENESAS [=y] || COMPILE_TEST [=n]
   Location:
     -> Device Drivers
       -> PHY Subsystem
         -> Renesas RZ/G3E USB 3.0 PHY driver (PHY_RZ_G3E_USB3 [=n])
 Selects: GENERIC_PHY [=y]

 ~~~
 Commit: ee5f1a3f90a4 (phy: renesas: Add Renesas RZ/G3E USB3.0 PHY driver)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>