[redhat] New configs in drivers/rtc
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 ~~~
 Symbol: RTC_DRV_NVIDIA_VRS10 [=n]
 Type  : tristate
 Defined at drivers/rtc/Kconfig:419
   Prompt: NVIDIA VRS10 RTC device
   Depends on: RTC_CLASS [=y] && I2C [=y]
   Location:
     -> Device Drivers
       -> Real Time Clock (RTC_CLASS [=y])
         -> NVIDIA VRS10 RTC device (RTC_DRV_NVIDIA_VRS10 [=n])

 ~~~
 Commit: 9d6d6b06933c (rtc: nvvrs: add NVIDIA VRS RTC device driver)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>

v2:
- CONFIG_RTC_DRV_NVIDIA_VRS10 enable for aarch64.

Signed-off-by: Patrick Talbert <ptalbert@redhat.com>