configs: riscv: Enable common hardware drivers for CentOS
These configs are all set to the same value as in aarch64 and should be
useful for a wide variety of current and future hardware. This doesn't
include any drivers that are specific to a single SoC vendor or board.

Every one of these configs exactly match an existing config for rhel
arm64, except one: I2C_MUX_PINCTRL is disabled because no
currently-known RISC-V systems use it.

Signed-off-by: Jennifer Berringer <jberring@redhat.com>