configs: riscv: Enable StarFive drivers for CentOS
Enable drivers for StarFive VisionFive 2 (JH7110). This SoC uses four
SiFive U74 cores. I have also tested this on Milk-V Mars, another board
which uses the same SoC. Other boards with the same SoC and merged
devicetrees (Star64, FLM13V01, Mars CM) are likely to work after this
change too but I haven't tested them.

This system does not support RVA23. Its ISA string is equivalent to
"rva20_zba_zbb". It is likely that CentOS Stream 11 will move to RVA23
as an ISA baseline for RISC-V, meaning it won't be able to run on this
system, but it can still be useful to run the ELN kernel on a system
running CentOS/RHEL 10 userspace and to ensure the rhel build flavor of
kernel-ark can run on a nonzero number of RISC-V hardware systems in the
short term.

Signed-off-by: Jennifer Berringer <jberring@redhat.com>