configs: riscv: Enable ARCH_ESWIN for CentOS
Enable configs for ESWIN EIC7700 and the SiFive HiFive Premier P550.
As of 6.19 this does not yet include clock, pcie, or sata drivers.

This is the same platform that was used for the RHEL 10 RISC-V Developer
Preview. This system does not support RVA23. Its ISA extension set is
equivalent to RVA20 plus Zba, Zbb, Sscofpmf, and H draft version 0.6.

Signed-off-by: Jennifer Berringer <jberring@redhat.com>