[redhat] New configs in drivers/net/can
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 ~~~
 Symbol: CAN_DEV [=y]
 Type  : bool
 Defined at drivers/net/can/Kconfig:3
   Prompt: CAN Device Drivers
   Depends on: NETDEVICES [=y] && CAN [=m]
   Location:
     -> Device Drivers
       -> Network device support (NETDEVICES [=y])
         -> CAN Device Drivers (CAN_DEV [=y])

 ~~~
 Commit: 6a5286442fb6 (can: Kconfig: turn menu "CAN Device Drivers" into a menuconfig using CAN_DEV)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>