tools/verification: fix typos in automata docstring and comments
Fix two typos in the Automata class documentation that have been
present since the initial implementation. The class docstring
incorrectly stated "part it" instead of "parses it" when
describing how the class processes DOT files. Additionally, a
comment describing transition labels contained the misspelling
"lables" instead of "labels".

Fix a typo in the comment describing the insertion of the initial
state into the states list: "bein og" should be "beginning of".

Signed-off-by: Wander Lairson Costa <wander@redhat.com>